Cache method that accommodates environments, S4 methods, Rasters

Cache method that accommodates environments, S4 methods, Rasters

Cache(FUN, ..., notOlderThan = NULL, objects = NULL, outputObjects = NULL,
  algo = "xxhash64", cacheRepo = NULL, compareRasterFileLength = 1e+06,
  userTags = c(), digestPathContent = TRUE, omitArgs = NULL,
  classOptions = list(), debugCache = character(), sideEffect = FALSE,
  makeCopy = FALSE, quick = FALSE)

# S4 method for ANY
Cache(FUN, ..., notOlderThan = NULL, objects = NULL,
  outputObjects = NULL, algo = "xxhash64", cacheRepo = NULL,
  compareRasterFileLength = 1e+06, userTags = c(),
  digestPathContent = TRUE, omitArgs = NULL, classOptions = list(),
  debugCache = character(), sideEffect = FALSE, makeCopy = FALSE,
  quick = FALSE)

Arguments

FUN
Either a function or an unevaluated function call (e.g., using quote.

...
Arguments of FUN function .

notOlderThan
load an artifact from the database only if it was created after notOlderThan.

objects
Character vector of objects to be digested. This is only applicable if there is a list, environment or simList with named objects within it. Only this/these objects will be considered for caching, i.e., only use a subset of the list, environment or simList objects.

outputObjects
Optional character vector indicating which objects to return. This is only relevant for simList objects

algo
The algorithms to be used; currently available choices are md5, which is also the default, sha1, crc32, sha256, sha512, xxhash32, xxhash64 and murmur32.

cacheRepo
A repository used for storing cached objects. This is optional if Cache is used inside a SpaDES module.

compareRasterFileLength
Numeric. Optional. When there are Rasters, that have file-backed storage, this is passed to the length arg in digest when determining if the Raster file is already in the database. Note: uses digest for file-backed Raster. Default 1e6. Passed to .prepareFileBackedRaster.

userTags
A character vector with Tags. These Tags will be added to the repository along with the artifact.

digestPathContent
Logical. Should arguments that are of class Path (see examples below) have their name digested (FALSE), or their file contents (TRUE; default).

omitArgs
Optional character string of arguments in the FUN to omit from the digest.

classOptions
Optional list. This will pass into .robustDigest for specific classes. Should be options that the .robustDigest knows what to do with.

debugCache
Character or Logical. Either "complete" or "quick" (uses partial matching, so "c" or "q" work). TRUE is equivalent to "complete". If "complete", then the returned object from the Cache function will have two attributes, debugCache1 and debugCache2, which are the entire list(...) and that same object, but after all .robustDigest calls, at the moment that it is digested using fastdigest, respectively. This attr(mySimOut, "debugCache2") can then be compared to a subsequent call and individual items within the object attr(mySimOut, "debugCache1") can be compared. If "quick", then it will return the same two objects directly, without evalutating the FUN(...).

sideEffect
Logical or path. Deteremines where the function will look for new files following function completion. See Details. NOTE: this argument is experimental and may change in future releases.

makeCopy
Logical. If sideEffect = TRUE, and makeCopy = TRUE, a copy of the downloaded files will be made and stored in the cacheRepo to speed up subsequent file recovery in the case where the original copy of the downloaded files are corrupted or missing. Currently only works when set to TRUE during the first run of Cache. Default is FALSE. NOTE: this argument is experimental and may change in future releases.

quick
Logical. If sideEffect = TRUE, setting this to TRUE, will hash the file's metadata (i.e., filename and file size) instead of hashing the contents of the file(s). If set to FALSE (default), the contents of the file(s) are hashed. NOTE: this argument is experimental and may change in future releases.

Value

As with cache, returns the value of the function call or the cached version (i.e., the result from a previous call to this same cached function with identical arguments).

Details

Caching R objects using cache has four important limitations:

the archivist package detects different environments as different;

it also does not detect S4 methods correctly due to method inheritance;

it does not detect objects that have file-base storage of information (specifically RasterLayer-class objects);

the default hashing algorithm is relatively slow.

This version of the Cache function accommodates those four special, though quite common, cases by:

converting any environments into list equivalents;

identifying the dispatched S4 method (including those made through inheritance) before hashing so the correct method is being cached;

by hashing the linked file, rather than the Raster object. Currently, only file-backed Raster* objects are digested (e.g., not ff objects, or any other R object where the data are on disk instead of in RAM);

using fastdigest internally when the object is in RAM, which can be up to ten times faster than digest. Note that file-backed objects are still hashed using digest.

If Cache is called within a SpaDES module, then the cached entry will automatically get 3 extra userTags: eventTime, eventType, and moduleName. These can then be used in clearCache to selectively remove cached objects by eventTime, eventType or moduleName. Cache will add a tag to the artifact in the database called accessed, which will assign the time that it was accessed, either read or write. That way, artifacts can be shown (using showCache) or removed (using clearCache) selectively, based on their access dates, rather than only by their creation dates. See example in clearCache. Cache (uppercase C) is used here so that it is not confused with, and does not mask, the archivist::cache function.

Note

As indicated above, several objects require pre-treatment before caching will work as expected. The function .robustDigest accommodates this. It is an S4 generic, meaning that developers can produce their own methods for different classes of objects. Currently, there are methods for several types of classes. See .robustDigest.

See .robustDigest for other specifics for other classes.

Filepaths

If a function has a path argument, there is some ambiguity about what should be done. Possibilities include:

hash the string as is (this will be very system specific, meaning a Cache call will not work if copied between systems or directories);

hash the basename(path);

hash the contents of the file.

If paths are passed in as is (i.e,. character string), the result will not be predictable. Instead, one should use the wrapper function asPath(path), which sets the class of the string to a Path, and one should decide whether one wants to digest the content of the file (using digestPathContent = TRUE), or just the filename ((digestPathContent = FALSE)). See examples.

Stochasticity

In general, it is expected that caching will only be used when stochasticity is not relevant, or if a user has achieved sufficient stochasticity (e.g., via sufficient number of calls to experiment) such that no new explorations of stochastic outcomes are required. It will also be very useful in a reproducible workflow.

<code>sideEffect</code>

If sideEffect is not FALSE, then metadata about any files that added to sideEffect will be added as an attribute to the cached copy. Subsequent calls to this function will assess for the presence of the new files in the sideEffect location. If the files are identical (quick = FALSE) or their file size is identical (quick = TRUE), then the cached copy of the function will be returned (and no files changed). If there are missing or incorrect files, then the function will re-run. This will accommodate the situation where the function call is identical, but somehow the side effect files were modified. If sideEffect is logical, then the function will check the cacheRepo; if it is a path, then it will check the path. The function will assess whether the files to be downloaded are found locally prior to download. If it fails the local test, then it will try to recover from a local copy if (makeCopy had been set to TRUE the first time the function was run. Currently, local recovery will only work ifmakeCOpy was set to TRUE the first time Cache was run). Default is FALSE.

See also

cache, .robustDigest

Examples

library(raster)

#> Loading required package: sp

#> 
#> Attaching package: ‘raster’

#> The following object is masked from ‘package:reproducible’:
#> 
#>     init

tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir), silent = TRUE) # just to make sure it is clear

# Basic use
ranNumsA <- Cache(rnorm, 10, 16, cacheRepo = tmpDir)

# All same
ranNumsB <- Cache(rnorm, 10, 16, cacheRepo = tmpDir) # recovers cached copy

#>   loading cached result from previous rnorm call.

ranNumsC <- rnorm(10, 16) %>% Cache(cacheRepo = tmpDir) # recovers cached copy

#>   loading cached result from previous 'rnorm' pipe sequence call.

ranNumsD <- Cache(quote(rnorm(n = 10, 16)), cacheRepo = tmpDir) # recovers cached copy

#>   loading cached result from previous rnorm call.

# Any minor change makes it different
ranNumsE <- rnorm(10, 6) %>% Cache(cacheRepo = tmpDir) # different

## Example 1: basic cache use with tags
ranNumsA <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(runif, 4, cacheRepo = tmpDir, userTags = "objectName:b")

showCache(tmpDir, userTags = c("objectName"))

#>                             artifact         tagKey
#>  1: 2d58e2618e8ef93a0a337693d7229d10         format
#>  2: 2d58e2618e8ef93a0a337693d7229d10           name
#>  3: 2d58e2618e8ef93a0a337693d7229d10          class
#>  4: 2d58e2618e8ef93a0a337693d7229d10           date
#>  5: 2d58e2618e8ef93a0a337693d7229d10     objectName
#>  6: 2d58e2618e8ef93a0a337693d7229d10       function
#>  7: 2d58e2618e8ef93a0a337693d7229d10    object.size
#>  8: 2d58e2618e8ef93a0a337693d7229d10       accessed
#>  9: 2d58e2618e8ef93a0a337693d7229d10 otherFunctions
#> 10: 2d58e2618e8ef93a0a337693d7229d10 otherFunctions
#> 11: 2d58e2618e8ef93a0a337693d7229d10 otherFunctions
#> 12: 2d58e2618e8ef93a0a337693d7229d10        cacheId
#> 13: 98d7c74270fee5196e9701c02a08081a         format
#> 14: 98d7c74270fee5196e9701c02a08081a           name
#> 15: 98d7c74270fee5196e9701c02a08081a          class
#> 16: 98d7c74270fee5196e9701c02a08081a           date
#> 17: 98d7c74270fee5196e9701c02a08081a     objectName
#> 18: 98d7c74270fee5196e9701c02a08081a       function
#> 19: 98d7c74270fee5196e9701c02a08081a    object.size
#> 20: 98d7c74270fee5196e9701c02a08081a       accessed
#> 21: 98d7c74270fee5196e9701c02a08081a otherFunctions
#> 22: 98d7c74270fee5196e9701c02a08081a otherFunctions
#> 23: 98d7c74270fee5196e9701c02a08081a otherFunctions
#> 24: 98d7c74270fee5196e9701c02a08081a        cacheId
#>                             artifact         tagKey
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:47:38
#>  2:                               Cache 2018-01-25 08:47:38
#>  3:                             numeric 2018-01-25 08:47:38
#>  4:                 2018-01-25 08:47:38 2018-01-25 08:47:38
#>  5:                                   b 2018-01-25 08:47:38
#>  6:                               runif 2018-01-25 08:47:38
#>  7:                                 528 2018-01-25 08:47:38
#>  8:                 2018-01-25 08:47:38 2018-01-25 08:47:38
#>  9:      otherFunctions:build_reference 2018-01-25 08:47:38
#> 10: otherFunctions:data_reference_topic 2018-01-25 08:47:38
#> 11:                  otherFunctions:try 2018-01-25 08:47:38
#> 12:    d89a70080574eedee9d56a2a09d7ca01 2018-01-25 08:47:38
#> 13:                                 rda 2018-01-25 08:47:38
#> 14:                               Cache 2018-01-25 08:47:38
#> 15:                             numeric 2018-01-25 08:47:38
#> 16:                 2018-01-25 08:47:38 2018-01-25 08:47:38
#> 17:                                   a 2018-01-25 08:47:38
#> 18:                               rnorm 2018-01-25 08:47:38
#> 19:                                 528 2018-01-25 08:47:38
#> 20:                 2018-01-25 08:47:38 2018-01-25 08:47:38
#> 21:      otherFunctions:build_reference 2018-01-25 08:47:38
#> 22: otherFunctions:data_reference_topic 2018-01-25 08:47:38
#> 23:                  otherFunctions:try 2018-01-25 08:47:38
#> 24:    f05172cef0faa451e18d7ddc500dbd0e 2018-01-25 08:47:38
#>                                tagValue         createdDate

showCache(tmpDir, userTags = c("^a$")) # regular expression ... "a" exactly

#>                             artifact         tagKey
#>  1: 98d7c74270fee5196e9701c02a08081a         format
#>  2: 98d7c74270fee5196e9701c02a08081a           name
#>  3: 98d7c74270fee5196e9701c02a08081a          class
#>  4: 98d7c74270fee5196e9701c02a08081a           date
#>  5: 98d7c74270fee5196e9701c02a08081a     objectName
#>  6: 98d7c74270fee5196e9701c02a08081a       function
#>  7: 98d7c74270fee5196e9701c02a08081a    object.size
#>  8: 98d7c74270fee5196e9701c02a08081a       accessed
#>  9: 98d7c74270fee5196e9701c02a08081a otherFunctions
#> 10: 98d7c74270fee5196e9701c02a08081a otherFunctions
#> 11: 98d7c74270fee5196e9701c02a08081a otherFunctions
#> 12: 98d7c74270fee5196e9701c02a08081a        cacheId
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:47:38
#>  2:                               Cache 2018-01-25 08:47:38
#>  3:                             numeric 2018-01-25 08:47:38
#>  4:                 2018-01-25 08:47:38 2018-01-25 08:47:38
#>  5:                                   a 2018-01-25 08:47:38
#>  6:                               rnorm 2018-01-25 08:47:38
#>  7:                                 528 2018-01-25 08:47:38
#>  8:                 2018-01-25 08:47:38 2018-01-25 08:47:38
#>  9:      otherFunctions:build_reference 2018-01-25 08:47:38
#> 10: otherFunctions:data_reference_topic 2018-01-25 08:47:38
#> 11:                  otherFunctions:try 2018-01-25 08:47:38
#> 12:    f05172cef0faa451e18d7ddc500dbd0e 2018-01-25 08:47:38

showCache(tmpDir, userTags = c("runif")) # show only cached objects made during runif call

#>                             artifact         tagKey
#>  1: 2d58e2618e8ef93a0a337693d7229d10         format
#>  2: 2d58e2618e8ef93a0a337693d7229d10           name
#>  3: 2d58e2618e8ef93a0a337693d7229d10          class
#>  4: 2d58e2618e8ef93a0a337693d7229d10           date
#>  5: 2d58e2618e8ef93a0a337693d7229d10     objectName
#>  6: 2d58e2618e8ef93a0a337693d7229d10       function
#>  7: 2d58e2618e8ef93a0a337693d7229d10    object.size
#>  8: 2d58e2618e8ef93a0a337693d7229d10       accessed
#>  9: 2d58e2618e8ef93a0a337693d7229d10 otherFunctions
#> 10: 2d58e2618e8ef93a0a337693d7229d10 otherFunctions
#> 11: 2d58e2618e8ef93a0a337693d7229d10 otherFunctions
#> 12: 2d58e2618e8ef93a0a337693d7229d10        cacheId
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:47:38
#>  2:                               Cache 2018-01-25 08:47:38
#>  3:                             numeric 2018-01-25 08:47:38
#>  4:                 2018-01-25 08:47:38 2018-01-25 08:47:38
#>  5:                                   b 2018-01-25 08:47:38
#>  6:                               runif 2018-01-25 08:47:38
#>  7:                                 528 2018-01-25 08:47:38
#>  8:                 2018-01-25 08:47:38 2018-01-25 08:47:38
#>  9:      otherFunctions:build_reference 2018-01-25 08:47:38
#> 10: otherFunctions:data_reference_topic 2018-01-25 08:47:38
#> 11:                  otherFunctions:try 2018-01-25 08:47:38
#> 12:    d89a70080574eedee9d56a2a09d7ca01 2018-01-25 08:47:38

clearCache(tmpDir, userTags = c("runif")) # remove only cached objects made during runif call
showCache(tmpDir) # only those made during rnorm call

#>                             artifact         tagKey
#>  1: 8f32b763226ecc49a4aed377031cda7b         format
#>  2: 8f32b763226ecc49a4aed377031cda7b           name
#>  3: 8f32b763226ecc49a4aed377031cda7b          class
#>  4: 8f32b763226ecc49a4aed377031cda7b           date
#>  5: 8f32b763226ecc49a4aed377031cda7b       function
#>  6: 8f32b763226ecc49a4aed377031cda7b    object.size
#>  7: 8f32b763226ecc49a4aed377031cda7b       accessed
#>  8: 8f32b763226ecc49a4aed377031cda7b otherFunctions
#>  9: 8f32b763226ecc49a4aed377031cda7b otherFunctions
#> 10: 8f32b763226ecc49a4aed377031cda7b otherFunctions
#> 11: 8f32b763226ecc49a4aed377031cda7b        cacheId
#> 12: 98d7c74270fee5196e9701c02a08081a         format
#> 13: 98d7c74270fee5196e9701c02a08081a           name
#> 14: 98d7c74270fee5196e9701c02a08081a          class
#> 15: 98d7c74270fee5196e9701c02a08081a           date
#> 16: 98d7c74270fee5196e9701c02a08081a     objectName
#> 17: 98d7c74270fee5196e9701c02a08081a       function
#> 18: 98d7c74270fee5196e9701c02a08081a    object.size
#> 19: 98d7c74270fee5196e9701c02a08081a       accessed
#> 20: 98d7c74270fee5196e9701c02a08081a otherFunctions
#> 21: 98d7c74270fee5196e9701c02a08081a otherFunctions
#> 22: 98d7c74270fee5196e9701c02a08081a otherFunctions
#> 23: 98d7c74270fee5196e9701c02a08081a        cacheId
#> 24: eb8e9864f48565413eceddf69996a7a3         format
#> 25: eb8e9864f48565413eceddf69996a7a3           name
#> 26: eb8e9864f48565413eceddf69996a7a3          class
#> 27: eb8e9864f48565413eceddf69996a7a3           date
#> 28: eb8e9864f48565413eceddf69996a7a3       function
#> 29: eb8e9864f48565413eceddf69996a7a3    object.size
#> 30: eb8e9864f48565413eceddf69996a7a3       accessed
#> 31: eb8e9864f48565413eceddf69996a7a3 otherFunctions
#> 32: eb8e9864f48565413eceddf69996a7a3 otherFunctions
#> 33: eb8e9864f48565413eceddf69996a7a3 otherFunctions
#> 34: eb8e9864f48565413eceddf69996a7a3        cacheId
#> 35: eb8e9864f48565413eceddf69996a7a3       accessed
#> 36: eb8e9864f48565413eceddf69996a7a3       accessed
#> 37: eb8e9864f48565413eceddf69996a7a3       accessed
#>                             artifact         tagKey
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:47:38
#>  2:                               Cache 2018-01-25 08:47:38
#>  3:                             numeric 2018-01-25 08:47:38
#>  4:                 2018-01-25 08:47:38 2018-01-25 08:47:38
#>  5:                 rnorm pipe sequence 2018-01-25 08:47:38
#>  6:                                 624 2018-01-25 08:47:38
#>  7:                 2018-01-25 08:47:38 2018-01-25 08:47:38
#>  8:      otherFunctions:build_reference 2018-01-25 08:47:38
#>  9: otherFunctions:data_reference_topic 2018-01-25 08:47:38
#> 10:                  otherFunctions:try 2018-01-25 08:47:38
#> 11:    25729d525b445f9a94475253cd095f48 2018-01-25 08:47:38
#> 12:                                 rda 2018-01-25 08:47:38
#> 13:                               Cache 2018-01-25 08:47:38
#> 14:                             numeric 2018-01-25 08:47:38
#> 15:                 2018-01-25 08:47:38 2018-01-25 08:47:38
#> 16:                                   a 2018-01-25 08:47:38
#> 17:                               rnorm 2018-01-25 08:47:38
#> 18:                                 528 2018-01-25 08:47:38
#> 19:                 2018-01-25 08:47:38 2018-01-25 08:47:38
#> 20:      otherFunctions:build_reference 2018-01-25 08:47:38
#> 21: otherFunctions:data_reference_topic 2018-01-25 08:47:38
#> 22:                  otherFunctions:try 2018-01-25 08:47:38
#> 23:    f05172cef0faa451e18d7ddc500dbd0e 2018-01-25 08:47:38
#> 24:                                 rda 2018-01-25 08:47:37
#> 25:                               Cache 2018-01-25 08:47:37
#> 26:                             numeric 2018-01-25 08:47:37
#> 27:                 2018-01-25 08:47:37 2018-01-25 08:47:37
#> 28:                               rnorm 2018-01-25 08:47:37
#> 29:                                 624 2018-01-25 08:47:37
#> 30:                 2018-01-25 08:47:37 2018-01-25 08:47:37
#> 31:      otherFunctions:build_reference 2018-01-25 08:47:37
#> 32: otherFunctions:data_reference_topic 2018-01-25 08:47:37
#> 33:                  otherFunctions:try 2018-01-25 08:47:37
#> 34:    c9b9e3d462fdcd976aac998bbbc60c7a 2018-01-25 08:47:37
#> 35:                 2018-01-25 08:47:38 2018-01-25 08:47:38
#> 36:                 2018-01-25 08:47:38 2018-01-25 08:47:38
#> 37:                 2018-01-25 08:47:38 2018-01-25 08:47:38
#>                                tagValue         createdDate

clearCache(tmpDir)

## Example 2: using the "accessed" tag
ranNumsA <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(runif, 4, cacheRepo = tmpDir, userTags = "objectName:b")

# access it again, from Cache
ranNumsA <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:a")

#>   loading cached result from previous rnorm call.

wholeCache <- showCache(tmpDir)

# keep only items accessed "recently" (i.e., only objectName:a)
onlyRecentlyAccessed <- showCache(tmpDir, userTags = max(wholeCache[tagKey == "accessed"]$tagValue))

# inverse join with 2 data.tables ... using: a[!b]
# i.e., return all of wholeCache that was not recently accessed
toRemove <- unique(wholeCache[!onlyRecentlyAccessed], by = "artifact")$artifact
clearCache(tmpDir, toRemove) # remove ones not recently accessed
showCache(tmpDir) # still has more recently accessed

#> Empty data.table (0 rows) of 3 cols: md5hash,name,createdDate

clearCache(tmpDir)

## Example 3: using keepCache
ranNumsA <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(runif, 4, cacheRepo = tmpDir, userTags = "objectName:b")

# keep only those cached items from the last 24 hours
oneDay <- 60 * 60 * 24
keepCache(tmpDir, after = Sys.time() - oneDay)

#>                             artifact         tagKey
#>  1: 032db0f6aea3114b467dbbbfa6d8c0bf         format
#>  2: 032db0f6aea3114b467dbbbfa6d8c0bf           name
#>  3: 032db0f6aea3114b467dbbbfa6d8c0bf          class
#>  4: 032db0f6aea3114b467dbbbfa6d8c0bf           date
#>  5: 032db0f6aea3114b467dbbbfa6d8c0bf     objectName
#>  6: 032db0f6aea3114b467dbbbfa6d8c0bf       function
#>  7: 032db0f6aea3114b467dbbbfa6d8c0bf    object.size
#>  8: 032db0f6aea3114b467dbbbfa6d8c0bf       accessed
#>  9: 032db0f6aea3114b467dbbbfa6d8c0bf otherFunctions
#> 10: 032db0f6aea3114b467dbbbfa6d8c0bf otherFunctions
#> 11: 032db0f6aea3114b467dbbbfa6d8c0bf otherFunctions
#> 12: 032db0f6aea3114b467dbbbfa6d8c0bf        cacheId
#> 13: 5a59a3ff316978f926ed7ca36431fd16         format
#> 14: 5a59a3ff316978f926ed7ca36431fd16           name
#> 15: 5a59a3ff316978f926ed7ca36431fd16          class
#> 16: 5a59a3ff316978f926ed7ca36431fd16           date
#> 17: 5a59a3ff316978f926ed7ca36431fd16     objectName
#> 18: 5a59a3ff316978f926ed7ca36431fd16       function
#> 19: 5a59a3ff316978f926ed7ca36431fd16    object.size
#> 20: 5a59a3ff316978f926ed7ca36431fd16       accessed
#> 21: 5a59a3ff316978f926ed7ca36431fd16 otherFunctions
#> 22: 5a59a3ff316978f926ed7ca36431fd16 otherFunctions
#> 23: 5a59a3ff316978f926ed7ca36431fd16 otherFunctions
#> 24: 5a59a3ff316978f926ed7ca36431fd16        cacheId
#>                             artifact         tagKey
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:47:39
#>  2:                               Cache 2018-01-25 08:47:39
#>  3:                             numeric 2018-01-25 08:47:39
#>  4:                 2018-01-25 08:47:39 2018-01-25 08:47:39
#>  5:                                   a 2018-01-25 08:47:39
#>  6:                               rnorm 2018-01-25 08:47:39
#>  7:                                 528 2018-01-25 08:47:39
#>  8:                 2018-01-25 08:47:39 2018-01-25 08:47:39
#>  9:      otherFunctions:build_reference 2018-01-25 08:47:39
#> 10: otherFunctions:data_reference_topic 2018-01-25 08:47:39
#> 11:                  otherFunctions:try 2018-01-25 08:47:39
#> 12:    f05172cef0faa451e18d7ddc500dbd0e 2018-01-25 08:47:39
#> 13:                                 rda 2018-01-25 08:47:39
#> 14:                               Cache 2018-01-25 08:47:39
#> 15:                             numeric 2018-01-25 08:47:39
#> 16:                 2018-01-25 08:47:39 2018-01-25 08:47:39
#> 17:                                   b 2018-01-25 08:47:39
#> 18:                               runif 2018-01-25 08:47:39
#> 19:                                 528 2018-01-25 08:47:39
#> 20:                 2018-01-25 08:47:39 2018-01-25 08:47:39
#> 21:      otherFunctions:build_reference 2018-01-25 08:47:39
#> 22: otherFunctions:data_reference_topic 2018-01-25 08:47:39
#> 23:                  otherFunctions:try 2018-01-25 08:47:39
#> 24:    d89a70080574eedee9d56a2a09d7ca01 2018-01-25 08:47:39
#>                                tagValue         createdDate

# Keep all Cache items created with an rnorm() call
keepCache(tmpDir, userTags = "rnorm")

#>                             artifact         tagKey
#>  1: 032db0f6aea3114b467dbbbfa6d8c0bf         format
#>  2: 032db0f6aea3114b467dbbbfa6d8c0bf           name
#>  3: 032db0f6aea3114b467dbbbfa6d8c0bf          class
#>  4: 032db0f6aea3114b467dbbbfa6d8c0bf           date
#>  5: 032db0f6aea3114b467dbbbfa6d8c0bf     objectName
#>  6: 032db0f6aea3114b467dbbbfa6d8c0bf       function
#>  7: 032db0f6aea3114b467dbbbfa6d8c0bf    object.size
#>  8: 032db0f6aea3114b467dbbbfa6d8c0bf       accessed
#>  9: 032db0f6aea3114b467dbbbfa6d8c0bf otherFunctions
#> 10: 032db0f6aea3114b467dbbbfa6d8c0bf otherFunctions
#> 11: 032db0f6aea3114b467dbbbfa6d8c0bf otherFunctions
#> 12: 032db0f6aea3114b467dbbbfa6d8c0bf        cacheId
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:47:39
#>  2:                               Cache 2018-01-25 08:47:39
#>  3:                             numeric 2018-01-25 08:47:39
#>  4:                 2018-01-25 08:47:39 2018-01-25 08:47:39
#>  5:                                   a 2018-01-25 08:47:39
#>  6:                               rnorm 2018-01-25 08:47:39
#>  7:                                 528 2018-01-25 08:47:39
#>  8:                 2018-01-25 08:47:39 2018-01-25 08:47:39
#>  9:      otherFunctions:build_reference 2018-01-25 08:47:39
#> 10: otherFunctions:data_reference_topic 2018-01-25 08:47:39
#> 11:                  otherFunctions:try 2018-01-25 08:47:39
#> 12:    f05172cef0faa451e18d7ddc500dbd0e 2018-01-25 08:47:39

# Remove all Cache items that happened within a rnorm() call
clearCache(tmpDir, userTags = "rnorm")

showCache(tmpDir) ## empty

#> Empty data.table (0 rows) of 3 cols: md5hash,name,createdDate

## Example 4: searching for multiple objects in the cache

# default userTags is "and" matching; for "or" matching use |
ranNumsA <- Cache(runif, 4, cacheRepo = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:b")

# show all objects (runif and rnorm in this case)
showCache(tmpDir)

#>                             artifact         tagKey
#>  1: 9632d7d9cdb4bcb11c7f62021d342a90         format
#>  2: 9632d7d9cdb4bcb11c7f62021d342a90           name
#>  3: 9632d7d9cdb4bcb11c7f62021d342a90          class
#>  4: 9632d7d9cdb4bcb11c7f62021d342a90           date
#>  5: 9632d7d9cdb4bcb11c7f62021d342a90     objectName
#>  6: 9632d7d9cdb4bcb11c7f62021d342a90       function
#>  7: 9632d7d9cdb4bcb11c7f62021d342a90    object.size
#>  8: 9632d7d9cdb4bcb11c7f62021d342a90       accessed
#>  9: 9632d7d9cdb4bcb11c7f62021d342a90 otherFunctions
#> 10: 9632d7d9cdb4bcb11c7f62021d342a90 otherFunctions
#> 11: 9632d7d9cdb4bcb11c7f62021d342a90 otherFunctions
#> 12: 9632d7d9cdb4bcb11c7f62021d342a90        cacheId
#> 13: de8432ac8f988025e51f60e237bd4e2c         format
#> 14: de8432ac8f988025e51f60e237bd4e2c           name
#> 15: de8432ac8f988025e51f60e237bd4e2c          class
#> 16: de8432ac8f988025e51f60e237bd4e2c           date
#> 17: de8432ac8f988025e51f60e237bd4e2c     objectName
#> 18: de8432ac8f988025e51f60e237bd4e2c       function
#> 19: de8432ac8f988025e51f60e237bd4e2c    object.size
#> 20: de8432ac8f988025e51f60e237bd4e2c       accessed
#> 21: de8432ac8f988025e51f60e237bd4e2c otherFunctions
#> 22: de8432ac8f988025e51f60e237bd4e2c otherFunctions
#> 23: de8432ac8f988025e51f60e237bd4e2c otherFunctions
#> 24: de8432ac8f988025e51f60e237bd4e2c        cacheId
#>                             artifact         tagKey
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:47:40
#>  2:                               Cache 2018-01-25 08:47:40
#>  3:                             numeric 2018-01-25 08:47:40
#>  4:                 2018-01-25 08:47:40 2018-01-25 08:47:40
#>  5:                                   b 2018-01-25 08:47:40
#>  6:                               rnorm 2018-01-25 08:47:40
#>  7:                                 528 2018-01-25 08:47:40
#>  8:                 2018-01-25 08:47:40 2018-01-25 08:47:40
#>  9:      otherFunctions:build_reference 2018-01-25 08:47:40
#> 10: otherFunctions:data_reference_topic 2018-01-25 08:47:40
#> 11:                  otherFunctions:try 2018-01-25 08:47:40
#> 12:    f05172cef0faa451e18d7ddc500dbd0e 2018-01-25 08:47:40
#> 13:                                 rda 2018-01-25 08:47:39
#> 14:                               Cache 2018-01-25 08:47:39
#> 15:                             numeric 2018-01-25 08:47:39
#> 16:                 2018-01-25 08:47:39 2018-01-25 08:47:39
#> 17:                                   a 2018-01-25 08:47:39
#> 18:                               runif 2018-01-25 08:47:39
#> 19:                                 528 2018-01-25 08:47:39
#> 20:                 2018-01-25 08:47:39 2018-01-25 08:47:39
#> 21:      otherFunctions:build_reference 2018-01-25 08:47:39
#> 22: otherFunctions:data_reference_topic 2018-01-25 08:47:40
#> 23:                  otherFunctions:try 2018-01-25 08:47:40
#> 24:    d89a70080574eedee9d56a2a09d7ca01 2018-01-25 08:47:40
#>                                tagValue         createdDate

# show objects that are both runif and rnorm
# (i.e., none in this case, because objecs are either or, not both)
showCache(tmpDir, userTags = c("runif", "rnorm")) ## empty

#> Empty data.table (0 rows) of 4 cols: artifact,tagKey,tagValue,createdDate

# show objects that are either runif or rnorm ("or" search)
showCache(tmpDir, userTags = "runif|rnorm")

#>                             artifact         tagKey
#>  1: 9632d7d9cdb4bcb11c7f62021d342a90         format
#>  2: 9632d7d9cdb4bcb11c7f62021d342a90           name
#>  3: 9632d7d9cdb4bcb11c7f62021d342a90          class
#>  4: 9632d7d9cdb4bcb11c7f62021d342a90           date
#>  5: 9632d7d9cdb4bcb11c7f62021d342a90     objectName
#>  6: 9632d7d9cdb4bcb11c7f62021d342a90       function
#>  7: 9632d7d9cdb4bcb11c7f62021d342a90    object.size
#>  8: 9632d7d9cdb4bcb11c7f62021d342a90       accessed
#>  9: 9632d7d9cdb4bcb11c7f62021d342a90 otherFunctions
#> 10: 9632d7d9cdb4bcb11c7f62021d342a90 otherFunctions
#> 11: 9632d7d9cdb4bcb11c7f62021d342a90 otherFunctions
#> 12: 9632d7d9cdb4bcb11c7f62021d342a90        cacheId
#> 13: de8432ac8f988025e51f60e237bd4e2c         format
#> 14: de8432ac8f988025e51f60e237bd4e2c           name
#> 15: de8432ac8f988025e51f60e237bd4e2c          class
#> 16: de8432ac8f988025e51f60e237bd4e2c           date
#> 17: de8432ac8f988025e51f60e237bd4e2c     objectName
#> 18: de8432ac8f988025e51f60e237bd4e2c       function
#> 19: de8432ac8f988025e51f60e237bd4e2c    object.size
#> 20: de8432ac8f988025e51f60e237bd4e2c       accessed
#> 21: de8432ac8f988025e51f60e237bd4e2c otherFunctions
#> 22: de8432ac8f988025e51f60e237bd4e2c otherFunctions
#> 23: de8432ac8f988025e51f60e237bd4e2c otherFunctions
#> 24: de8432ac8f988025e51f60e237bd4e2c        cacheId
#>                             artifact         tagKey
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:47:40
#>  2:                               Cache 2018-01-25 08:47:40
#>  3:                             numeric 2018-01-25 08:47:40
#>  4:                 2018-01-25 08:47:40 2018-01-25 08:47:40
#>  5:                                   b 2018-01-25 08:47:40
#>  6:                               rnorm 2018-01-25 08:47:40
#>  7:                                 528 2018-01-25 08:47:40
#>  8:                 2018-01-25 08:47:40 2018-01-25 08:47:40
#>  9:      otherFunctions:build_reference 2018-01-25 08:47:40
#> 10: otherFunctions:data_reference_topic 2018-01-25 08:47:40
#> 11:                  otherFunctions:try 2018-01-25 08:47:40
#> 12:    f05172cef0faa451e18d7ddc500dbd0e 2018-01-25 08:47:40
#> 13:                                 rda 2018-01-25 08:47:39
#> 14:                               Cache 2018-01-25 08:47:39
#> 15:                             numeric 2018-01-25 08:47:39
#> 16:                 2018-01-25 08:47:39 2018-01-25 08:47:39
#> 17:                                   a 2018-01-25 08:47:39
#> 18:                               runif 2018-01-25 08:47:39
#> 19:                                 528 2018-01-25 08:47:39
#> 20:                 2018-01-25 08:47:39 2018-01-25 08:47:39
#> 21:      otherFunctions:build_reference 2018-01-25 08:47:39
#> 22: otherFunctions:data_reference_topic 2018-01-25 08:47:40
#> 23:                  otherFunctions:try 2018-01-25 08:47:40
#> 24:    d89a70080574eedee9d56a2a09d7ca01 2018-01-25 08:47:40
#>                                tagValue         createdDate

# keep only objects that are either runif or rnorm ("or" search)
keepCache(tmpDir, userTags = "runif|rnorm")

#>                             artifact         tagKey
#>  1: 9632d7d9cdb4bcb11c7f62021d342a90         format
#>  2: 9632d7d9cdb4bcb11c7f62021d342a90           name
#>  3: 9632d7d9cdb4bcb11c7f62021d342a90          class
#>  4: 9632d7d9cdb4bcb11c7f62021d342a90           date
#>  5: 9632d7d9cdb4bcb11c7f62021d342a90     objectName
#>  6: 9632d7d9cdb4bcb11c7f62021d342a90       function
#>  7: 9632d7d9cdb4bcb11c7f62021d342a90    object.size
#>  8: 9632d7d9cdb4bcb11c7f62021d342a90       accessed
#>  9: 9632d7d9cdb4bcb11c7f62021d342a90 otherFunctions
#> 10: 9632d7d9cdb4bcb11c7f62021d342a90 otherFunctions
#> 11: 9632d7d9cdb4bcb11c7f62021d342a90 otherFunctions
#> 12: 9632d7d9cdb4bcb11c7f62021d342a90        cacheId
#> 13: de8432ac8f988025e51f60e237bd4e2c         format
#> 14: de8432ac8f988025e51f60e237bd4e2c           name
#> 15: de8432ac8f988025e51f60e237bd4e2c          class
#> 16: de8432ac8f988025e51f60e237bd4e2c           date
#> 17: de8432ac8f988025e51f60e237bd4e2c     objectName
#> 18: de8432ac8f988025e51f60e237bd4e2c       function
#> 19: de8432ac8f988025e51f60e237bd4e2c    object.size
#> 20: de8432ac8f988025e51f60e237bd4e2c       accessed
#> 21: de8432ac8f988025e51f60e237bd4e2c otherFunctions
#> 22: de8432ac8f988025e51f60e237bd4e2c otherFunctions
#> 23: de8432ac8f988025e51f60e237bd4e2c otherFunctions
#> 24: de8432ac8f988025e51f60e237bd4e2c        cacheId
#>                             artifact         tagKey
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:47:40
#>  2:                               Cache 2018-01-25 08:47:40
#>  3:                             numeric 2018-01-25 08:47:40
#>  4:                 2018-01-25 08:47:40 2018-01-25 08:47:40
#>  5:                                   b 2018-01-25 08:47:40
#>  6:                               rnorm 2018-01-25 08:47:40
#>  7:                                 528 2018-01-25 08:47:40
#>  8:                 2018-01-25 08:47:40 2018-01-25 08:47:40
#>  9:      otherFunctions:build_reference 2018-01-25 08:47:40
#> 10: otherFunctions:data_reference_topic 2018-01-25 08:47:40
#> 11:                  otherFunctions:try 2018-01-25 08:47:40
#> 12:    f05172cef0faa451e18d7ddc500dbd0e 2018-01-25 08:47:40
#> 13:                                 rda 2018-01-25 08:47:39
#> 14:                               Cache 2018-01-25 08:47:39
#> 15:                             numeric 2018-01-25 08:47:39
#> 16:                 2018-01-25 08:47:39 2018-01-25 08:47:39
#> 17:                                   a 2018-01-25 08:47:39
#> 18:                               runif 2018-01-25 08:47:39
#> 19:                                 528 2018-01-25 08:47:39
#> 20:                 2018-01-25 08:47:39 2018-01-25 08:47:39
#> 21:      otherFunctions:build_reference 2018-01-25 08:47:39
#> 22: otherFunctions:data_reference_topic 2018-01-25 08:47:40
#> 23:                  otherFunctions:try 2018-01-25 08:47:40
#> 24:    d89a70080574eedee9d56a2a09d7ca01 2018-01-25 08:47:40
#>                                tagValue         createdDate

clearCache(tmpDir)

## Example 5: using caching to speed up rerunning expensive computations
ras <- raster(extent(0, 10, 0, 10), res = 1,
              vals = sample(1:5, replace = TRUE, size = 1e2),
              crs = "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +ellps=WGS84")

# A slow operation, like GIS operation
notCached <- suppressWarnings(
  # project raster generates warnings when run non-interactively
  projectRaster(ras, crs = crs(ras), res = 5, cacheRepo = tmpDir)
)

cached <- suppressWarnings(
  # project raster generates warnings when run non-interactively
  Cache(projectRaster, ras, crs = crs(ras), res = 5, cacheRepo = tmpDir)
)

# second time is much faster
reRun <- suppressWarnings(
  # project raster generates warnings when run non-interactively
  Cache(projectRaster, ras, crs = crs(ras), res = 5, cacheRepo = tmpDir)
)

#>   loading cached result from previous projectRaster call.

# recovered cached version is same as non-cached version
all.equal(notCached, reRun) ## TRUE

#> [1] TRUE

## Example 6: working with file paths

# if passing a character string, it will take 2 complete passes to before
#  a cached copy is used when it is a save event (read or load is different)
obj <- 1:10
fname <- tempfile(fileext = ".RData")
Cache(saveRDS, obj, file = fname, cacheRepo = tmpDir)

#> NULL

Cache(saveRDS, obj, file = fname, cacheRepo = tmpDir)

#> NULL

Cache(saveRDS, obj, file = fname, cacheRepo = tmpDir) # cached copy is loaded

#>   loading cached result from previous saveRDS call.

#> [1] "NULL"
#> attr(,"tags")
#> [1] "cacheId:256e9b67000fa93742bffc0f38015cea"
#> attr(,"call")
#> [1] ""

# however, using asPath(), cached version retrieved after being run once
fname1 <- tempfile(fileext = ".RData")
Cache(saveRDS, obj, file = asPath(fname1), cacheRepo = tmpDir)

#> NULL

Cache(saveRDS, obj, file = asPath(fname1), cacheRepo = tmpDir) # cached copy is loaded

#> NULL

clearCache(tmpDir)

## cleanup
unlink(c("filename.rda", "filename1.rda"))
unlink(dirname(tmpDir), recursive = TRUE)