A set of helpers for Cache

These are internal only.

A few helpers to get specific things from the cache repository

getFunctionName(FUN, originalDots, ..., overrideCall, isPipe)

.digestRasterLayer(object, length, algo, quick)

getUserTags(cacheRepo, shownCache, cacheId, concatenated = TRUE)

getCacheId(cacheRepo, shownCache, artifact)

getArtifact(cacheRepo, shownCache, cacheId)

Arguments

FUN
A function

...
passing the ... from outer function, which will include potential arguments to the FUN

overrideCall
A character string indicating a different (not "Cache") function name to search for. Mostly so that this works with deprecated "cache".

isPipe
Logical. If the call to getFunctionName is coming from a pipe, there is more information available. Specifically, ._lhs which is already a call.

length
Numeric. If the element passed to Cache is a Path class object (from e.g., asPath(filename)) or it is a Raster with file-backing, then this will be passed to digest::digest, essentially limiting the number of bytes to digest (for speed). This will only be used if quick = FALSE. Default is getOption("reproducible.length"), which is set to Inf.

algo
The algorithms to be used; currently available choices are md5, which is also the default, sha1, crc32, sha256, sha512, xxhash32, xxhash64, murmur32, spookyhash and blake3.

quick
Logical or character. If TRUE, no disk-based information will be assessed, i.e., only memory content. See Details section about quick in Cache.

cacheRepo
A repository used for storing cached objects. This is optional if Cache is used inside a SpaDES module.

shownCache
Primary way of supplying cacheRepo; the data.table obj resulting from showCache, i.e., it will override cacheRepo. If this and cacheRepo are missing, then it will default to getOption('reproducible.cachePath')

cacheId
A character vector of cacheId values to use in the cache

concatenated
Logical. If TRUE, the returned userTags will be concatenated tagKey:tagValue.

artifact
Character vector of artifact values in the artifact column of showCache

Value

getCacheId returns the cacheId values for 1 or more artifacts in the cache.

getArtifact returns the artifact value for 1 or more entries in the cache, by cacheId.

Note

If the function cannot figure out a clean function name, it returns "internal"

Author

Eliot McIntire