Check for presence of checkFolderID (for Cache(useCloud))

Will check for presence of a cloudFolderID and make a new one if one not present on Google Drive, with a warning.

checkAndMakeCloudFolderID(
  cloudFolderID = getOption("reproducible.cloudFolderID", NULL),
  cachePath = NULL,
  create = FALSE,
  overwrite = FALSE,
  verbose = getOption("reproducible.verbose", 1),
  team_drive = NULL
)

Arguments

cloudFolderID

The google folder ID where cloud caching will occur.

cachePath

A repository used for storing cached objects. This is optional if Cache is used inside a SpaDES module.

create

Logical. If TRUE, then the cloudFolderID will be created. This should be used with caution as there are no checks for overwriting. See googledrive::drive_mkdir. Default FALSE.

overwrite

Logical. Passed to googledrive::drive_mkdir.

verbose

Numeric, -1 silent (where possible), 0 being very quiet, 1 showing more messaging, 2 being more messaging, etc. Default is 1. Above 3 will output much more information about the internals of Caching, which may help diagnose Caching challenges. Can set globally with an option, e.g., options('reproducible.verbose' = 0) to reduce to minimal

team_drive

Logical indicating whether to check team drives.

Value

Returns the character string of the cloud folder ID created or reported