Download file from generic source url

Download file from generic source url

dlGeneric(url, destinationPath, verbose = getOption("reproducible.verbose", 1))

Arguments

url

The url (link) to the file.

destinationPath

Character string of a directory in which to download and save the file that comes from url and is also where the function will look for archive or targetFile. NOTE (still experimental): To prevent repeated downloads in different locations, the user can also set options("reproducible.inputPaths") to one or more local file paths to search for the file before attempting to download. Default for that option is NULL meaning do not search locally.

verbose

Numeric, -1 silent (where possible), 0 being very quiet, 1 showing more messaging, 2 being more messaging, etc. Default is 1. Above 3 will output much more information about the internals of Caching, which may help diagnose Caching challenges. Can set globally with an option, e.g., options('reproducible.verbose' = 0) to reduce to minimal

needChecksums

Logical indicating whether to generate checksums. ## TODO: add overwrite arg to the function?

Author

Eliot McIntire and Alex Chubaty