Remove quote and determine if call uses ...

Minor cleaning up of the FUN and ... to be used subsequently. This does only very minor things as it is run even if useCache = FALSE, i.e., even if the Cache is skipped.

matchCall2(definition, call, envir, envir2 = parent.frame(), FUN)

Arguments

definition

a function, by default the function from which match.call is called. See details.

call

an unevaluated call to the function specified by definition, as generated by call.

envir

an environment, from which the ... in call are retrieved, if any.

envir2

Environment. The environment where matchCall2 was called.

FUN

Either a function (e.g., rnorm), a function call (e.g., rnorm(1)), or an unevaluated function call (e.g., using quote()).

Value

A named list with call (the original call, without quote), FUNorig, the original value passed by user to FUN, and usesDots which is a logical indicating whether the ... are used.