Merge two cache repositories together

mergeCache(
  cacheTo,
  cacheFrom,
  drvTo = getOption("reproducible.drv", RSQLite::SQLite()),
  drvFrom = getOption("reproducible.drv", RSQLite::SQLite()),
  connTo = NULL,
  connFrom = NULL
)

# S4 method for ANY
mergeCache(
  cacheTo,
  cacheFrom,
  drvTo = getOption("reproducible.drv", RSQLite::SQLite()),
  drvFrom = getOption("reproducible.drv", RSQLite::SQLite()),
  connTo = NULL,
  connFrom = NULL
)

Arguments

cacheTo
The cache repository (character string of the file path) that will become larger, i.e., merge into this

cacheFrom
The cache repository (character string of the file path) from which all objects will be taken and copied from

drvTo
The database driver for the cacheTo.

drvFrom
The database driver for the cacheFrom

connTo
The connection for the cacheTo. If not provided, then a new one will be made from drvTo and cacheTo

connFrom
The database for the cacheFrom. If not provided, then a new one will be made from drvFrom and cacheFrom

Value

The character string of the path of cacheTo, i.e., not the objects themselves.

Details

All the cacheFrom artifacts will be put into cacheTo repository. All userTags will be copied verbatim, including accessed, with 1 exception: date will be the current Sys.time() at the time of merging. The createdDate column will be similarly the current time of merging.