Memoised version of package_dependencies

This have a 6 minute memory time window.

package_dependenciesMem(packages = NULL, db = NULL,
  which = c("Depends", "Imports", "LinkingTo"), recursive = FALSE,
  reverse = FALSE, verbose = getOption("verbose"))

Arguments

packages
a character vector of package names.

db
character matrix as from available.packages() (with the default NULL the results of this call) or data frame variants thereof. Alternatively, a package database like the one available from https://cran.r-project.org/web/packages/packages.rds.

which
a character vector listing the types of dependencies, a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances"). Character string "all" is shorthand for that vector, character string "most" for the same vector without "Enhances".

recursive
logical: should (reverse) dependencies of (reverse) dependencies (and so on) be included?

reverse
logical: if FALSE (default), regular dependencies are calculated, otherwise reverse dependencies.

verbose
logical indicating if output should monitor the package search cycles.