Take a snapshot of all the packages and version numbers

This can be used later by installVersions to install or re-install the correct versions.

pkgSnapshot(packageVersionFile, libPath, standAlone = FALSE)

Arguments

packageVersionFile
A filename to save the packages and their currently installed version numbers. Defaults to ".packageVersions.txt".

libPath
The path to the local library where packages are installed. Defaults to the .libPaths()[1]

standAlone
Logical. If TRUE, all packages will be installed and loaded strictly from the libPaths only. If FALSE, all .libPaths will be used to find the correct versions. This can be create dramatically faster installs if the user has a substantial number of the packages already in their personal library. In the case of TRUE, there will be a hidden file place in the libPath directory that lists all the packages that were needed during the Require call. Default FALSE to minimize package installing.

Details

A file is written with the package names and versions of all packages within libPath. This can later be passed to Require.

Examples

pkgSnapFile <- tempfile()
pkgSnapshot(pkgSnapFile, .libPaths()[1])

#> There is no /tmp/RtmpTxnw1O/reproducible/cache/.Require/_tmp_Rtmpl5Zmit_temp_libpath55264dac90de._packageVersionsAuto.txt and standAlone is FALSE. This snapshot will not be accurate because it will include all packages in /tmp/Rtmpl5Zmit/temp_libpath55264dac90de, /tmp/Rtmpl5Zmit/temp_libpath55267c2b62d3, /home/achubaty/R/x86_64-pc-linux-gnu-library/3.6, /usr/local/lib/R/site-library, /usr/lib/R/site-library, /usr/lib/R/library

#>              instPkgs    instVers
#> 1        reproducible       1.1.0
#> 2        reproducible       1.1.0
#> 3        reproducible       1.1.0
#> 4        reproducible       1.1.0
#> 5        reproducible       1.1.0
#> 6        reproducible       1.1.0
#> 7               abind       1.4-5
#> 8             acepack       1.4.1
#> 9                 amc  0.1.6.9000
#> 10          animation         2.6
#> 11                ape         5.3
#> 12                aqp        1.19
#> 13          archivist       2.3.4
#> 14                arm      1.11-1
#> 15            askpass         1.1
#> 16         assertthat       0.2.1
#> 17          available       1.0.4
#> 18          backports       1.1.6
#> 19          base64enc       0.1-4
#> 20         bayestestR       0.6.0
#> 21           beanplot         1.2
#> 22                 BH    1.72.0-3
#> 23       BiocGenerics      0.32.0
#> 24        BiocManager     1.30.10
#> 25        BiocVersion      3.10.1
#> 26                bit    1.1-15.2
#> 27              bit64       0.9-7
#> 28             bitops       1.0-6
#> 29               blme       1.0-4
#> 30               blob       1.2.1
#> 31               brew       1.0-6
#> 32              broom       0.5.6
#> 33              callr       3.4.3
#> 34                car       3.0-7
#> 35            carData       3.0-3
#> 36         cellranger       1.1.0
#> 37          checkmate       2.0.0
#> 38          CircStats       0.2-6
#> 39           classInt       0.4-3
#> 40                cli       2.0.2
#> 41              clipr       0.7.0
#> 42         clisymbols       1.2.0
#> 43               coda      0.19-3
#> 44        CodeDepends       0.6.5
#> 45         colorRamps         2.3
#> 46         colorspace       1.4-1
#> 47         commonmark         1.7
#> 48             config         0.3
#> 49               covr       3.5.0
#> 50            cowplot       1.0.0
#> 51          crancache  0.0.0.9001
#> 52           cranlike       1.0.2
#> 53             crayon  1.3.4.9000
#> 54          crosstalk     1.1.0.1
#> 55               crul       0.9.0
#> 56               curl         4.3
#> 57          cyclocomp       1.1.0
#> 58         data.table      1.12.8
#> 59                DBI       1.1.0
#> 60            debugme       1.1.0
#> 61             deldir      0.1-25
#> 62            DEoptim       2.2-5
#> 63               desc       1.2.0
#> 64           devtools       2.3.0
#> 65         DiagrammeR     1.0.6.1
#> 66      DiagrammeRsvg         0.1
#> 67             digest      0.6.25
#> 68              dismo       1.1-4
#> 69         doParallel      1.0.15
#> 70          dotCall64       1.0-0
#> 71         downloader         0.4
#> 72              dplyr       0.8.5
#> 73                 DT        0.13
#> 74              e1071       1.7-3
#> 75            effsize       0.8.0
#> 76           ellipsis       0.3.0
#> 77           EnvStats       2.3.1
#> 78           evaluate      0.14.1
#> 79      exactextractr       0.3.0
#> 80               expm     0.999-4
#> 81              fansi       0.4.1
#> 82             farver       2.0.3
#> 83         fastdigest       0.6-3
#> 84          fasterize       1.0.2
#> 85            fastmap       1.0.1
#> 86          fastmatch       1.1-1
#> 87            fastshp       0.1-2
#> 88            fBasics   3042.89.1
#> 89                 ff    2.2-14.2
#> 90             ffbase      0.12.8
#> 91             fields        10.3
#> 92     fireSenseUtils  0.0.0.9025
#> 93              flock         0.7
#> 94                FNN       1.1.3
#> 95            forcats       0.5.0
#> 96            foreach       1.5.0
#> 97            Formula       1.2-3
#> 98          fpCompare       0.2.3
#> 99                 fs       1.4.1
#> 100            future      1.17.0
#> 101      future.apply       1.5.0
#> 102      future.callr       0.5.0
#> 103            gamlss       5.1-6
#> 104       gamlss.data       5.1-4
#> 105       gamlss.dist       5.1-6
#> 106            gargle       0.5.0
#> 107         gdalUtils     2.0.3.2
#> 108             gdata      2.18.0
#> 109          generics       0.0.2
#> 110           geojson       0.3.2
#> 111         geojsonio       0.9.2
#> 112       geojsonlint       0.4.0
#> 113           ggimage       0.2.8
#> 114           ggplot2       3.3.0
#> 115         ggplotify       0.0.5
#> 116          ggthemes       4.2.0
#> 117                gh       1.1.0
#> 118             git2r      0.27.1
#> 119              glmm       1.3.0
#> 120           globals      0.12.5
#> 121              glue       1.4.1
#> 122            gmailr       1.0.0
#> 123           gmodels      2.18.1
#> 124           goftest       1.2-2
#> 125       googleAuthR       1.2.1
#> 126       googledrive       1.0.1
#> 127          googleID    0.0.9001
#> 128        grainscape       0.4.3
#> 129             graph      1.64.0
#> 130          gridBase       0.4-7
#> 131         gridExtra         2.3
#> 132      gridGraphics       0.5-0
#> 133               gss       2.2-0
#> 134             gstat       2.0-5
#> 135            gtable       0.3.0
#> 136            gtools       3.8.2
#> 137            gurobi       9.0-1
#> 138             haven       2.2.0
#> 139            hexbin      1.28.1
#> 140        hexSticker       0.4.6
#> 141         highlight       0.5.0
#> 142             highr       0.8.1
#> 143             Hmisc       4.4-0
#> 144               hms       0.5.3
#> 145         htmlTable      1.13.3
#> 146         htmltools       0.4.0
#> 147       htmlwidgets       1.5.1
#> 148          httpcode       0.3.0
#> 149            httpuv       1.5.2
#> 150              httr       1.4.1
#> 151          hunspell         3.0
#> 152            igraph       1.2.5
#> 153        influenceR       0.1.0
#> 154               ini       0.3.1
#> 155           insight       0.8.3
#> 156         intervals      0.15.2
#> 157           isoband       0.2.1
#> 158          ISOcodes  2020.03.16
#> 159         iterators      1.0.12
#> 160         itertools       0.1-3
#> 161       janeaustenr       0.1.5
#> 162           janitor       2.0.1
#> 163              jpeg     0.1-8.1
#> 164               jqr       1.1.0
#> 165          jsonlite       1.6.1
#> 166      jsonvalidate       1.1.0
#> 167        kableExtra       1.1.0
#> 168             knitr        1.28
#> 169            koRpus      0.11-5
#> 170    koRpus.lang.en       0.1-3
#> 171          kSamples       1.2-9
#> 172          labeling         0.3
#> 173             LandR       0.0.5
#> 174          LandR.CS  0.0.2.0002
#> 175  landscapemetrics       1.4.3
#> 176    landscapetools       0.5.0
#> 177      LandWebUtils       0.1.0
#> 178             later       1.0.0
#> 179      latticeExtra      0.6-29
#> 180          lazyeval       0.2.2
#> 181           leaflet       2.0.3
#> 182    leaflet.extras       1.0.0
#> 183 leaflet.providers       1.9.0
#> 184        LearnBayes      2.15.1
#> 185         lifecycle       0.2.0
#> 186             lintr  2.0.1.9000
#> 187           listenv       0.8.0
#> 188              lme4      1.1-23
#> 189           logging    0.10-108
#> 190        lpsymphony      1.14.0
#> 191         lubridate       1.7.8
#> 192            lwgeom       0.2-3
#> 193            magick         2.3
#> 194          magrittr         1.5
#> 195               map       0.0.1
#> 196           mapdata       2.3.0
#> 197              maps       3.3.0
#> 198          maptools       0.9-9
#> 199          markdown       1.1.1
#> 200      MatrixModels       0.4-1
#> 201       matrixStats      0.56.0
#> 202           memoise       1.1.0
#> 203          merTools       0.5.0
#> 204    microbenchmark       1.4-7
#> 205              mime       0.9.1
#> 206            miniUI     0.1.1.1
#> 207             minqa       1.2.4
#> 208          multcomp      1.4-13
#> 209             MuMIn     1.43.17
#> 210           munsell       0.5.0
#> 211           mvtnorm       1.1-0
#> 212            nloptr     1.2.2.1
#> 213           nortest       1.0-4
#> 214           openssl       1.4.1
#> 215          openxlsx       4.1.5
#> 216           packrat       0.5.0
#> 217         parsedate       1.2.0
#> 218         patchwork       1.0.0
#> 219          pbkrtest     0.4-8.6
#> 220            pemisc  0.0.2.9000
#> 221       performance       0.4.6
#> 222           permute       0.9-5
#> 223            pillar       1.4.4
#> 224            pixmap      0.4-11
#> 225          pkgbuild       1.0.8
#> 226         pkgconfig       2.0.3
#> 227           pkgdown       1.5.1
#> 228           pkgload       1.0.2
#> 229             plogr       0.2.0
#> 230           plotKML       0.6-1
#> 231           plotrix       3.7-8
#> 232              plyr       1.8.6
#> 233               png       0.1-8
#> 234         polspline      1.1.18
#> 235          polyclip      1.10-0
#> 236            praise       1.0.0
#> 237        prettycode       1.1.0
#> 238       prettyunits       1.1.1
#> 239        prioritizr       4.1.5
#> 240          processx       3.4.2
#> 241           profvis       0.3.6
#> 242          progress       1.2.2
#> 243          promises       1.1.0
#> 244             proto       1.0.0
#> 245         protolite         2.1
#> 246              pryr       0.1.4
#> 247                ps       1.3.3
#> 248         PtProcess      3.3-13
#> 249             purrr       0.3.4
#> 250                qs      0.21.2
#> 251          quantreg        5.55
#> 252         quickPlot  0.1.6.9001
#> 253           R.cache      0.14.0
#> 254         R.devices      2.16.1
#> 255       R.methodsS3       1.8.0
#> 256              R.oo      1.23.0
#> 257           R.utils       2.9.2
#> 258                R6       2.4.1
#> 259      RandomFields       3.3.8
#> 260 RandomFieldsUtils       0.5.3
#> 261              RANN       2.6.1
#> 262     RApiSerialize       0.1.0
#> 263          rappdirs       0.3.1
#> 264            raster       3.1-5
#> 265         rasterVis        0.47
#> 266            Rclean       1.1.8
#> 267         rcmdcheck  1.3.3.9000
#> 268      RColorBrewer       1.1-2
#> 269              Rcpp     1.0.4.6
#> 270     RcppArmadillo 0.9.870.2.0
#> 271         RcppEigen   0.3.3.7.0
#> 272             RCurl    1.98-1.2
#> 273              rdoc       0.1.0
#> 274             readr       1.3.1
#> 275            readxl       1.3.1
#> 276           rematch       1.0.1
#> 277          rematch2       2.1.2
#> 278           remotes       2.1.1
#> 279      reproducible       1.1.0
#> 280      reproducible       1.1.0
#> 281      reproducible       1.1.0
#> 282           reshape       0.8.8
#> 283          reshape2       1.4.4
#> 284       revdepcheck  1.0.0.9001
#> 285               rex       1.2.0
#> 286             rgdal       1.4-8
#> 287             rgeos       0.5-3
#> 288         Rgraphviz      2.30.0
#> 289               rio      0.5.16
#> 290             rJava      0.9-12
#> 291             rlang       0.4.6
#> 292        rmapshaper       0.4.4
#> 293         rmarkdown         2.1
#> 294               rms       5.1-4
#> 295          roxygen2       7.1.0
#> 296         RPostgres       1.2.0
#> 297         rprojroot       1.3-2
#> 298             RSAGA       1.3.0
#> 299            Rserve     1.7-3.1
#> 300           RSQLite       2.2.0
#> 301        rstudioapi        0.11
#> 302         Rsymphony      0.1-28
#> 303               rts      1.0-49
#> 304           rvcheck       0.1.8
#> 305         rversions       2.0.1
#> 306             rvest       0.3.5
#> 307          sandwich       2.5-1
#> 308            scales       1.1.1
#> 309               sdm      1.0-89
#> 310           selectr       0.4-2
#> 311       sessioninfo       1.1.1
#> 312                sf       0.9-3
#> 313        shapefiles         0.7
#> 314             shiny     1.4.0.2
#> 315           shinyBS        0.61
#> 316   shinycssloaders         0.3
#> 317    shinydashboard       0.7.1
#> 318           shinyjs         1.1
#> 319      shinyWidgets       0.5.1
#> 320          showtext         0.8
#> 321        showtextdb         2.0
#> 322            slackr       1.4.2
#> 323              slam      0.1-47
#> 324         snakecase      0.11.0
#> 325              snow       0.4-3
#> 326         SnowballC       0.7.0
#> 327       sourcetools       0.1.7
#> 328                sp       1.4-1
#> 329         spacetime       1.2-3
#> 330            SpaDES       2.0.4
#> 331     SpaDES.addins       0.1.2
#> 332       SpaDES.core  1.0.0.9004
#> 333      SpaDES.shiny       0.2.0
#> 334      SpaDES.tools  0.3.4.9001
#> 335              spam       2.5-1
#> 336           SparseM        1.78
#> 337        spatialEco       1.3-1
#> 338       SpatialPack       0.3-8
#> 339          spatstat      1.63-3
#> 340     spatstat.data       1.4-3
#> 341    spatstat.utils      1.17-0
#> 342            spData       0.3.5
#> 343             spdep       1.1-3
#> 344          spelling         2.1
#> 345        stabledist       0.7-1
#> 346             stars       0.4-1
#> 347           statmod      1.4.34
#> 348         stopwords         2.0
#> 349        stringdist     0.9.5.5
#> 350           stringi       1.4.6
#> 351           stringr       1.4.0
#> 352            styler  1.3.2.9000
#> 353         SuppDists     1.1-9.5
#> 354             sylly       0.1-5
#> 355          sylly.en       0.1-3
#> 356               sys         3.3
#> 357          sysfonts       0.8.1
#> 358            tensor         1.5
#> 359          testthat       2.3.2
#> 360           TH.data      1.0-10
#> 361            tibble       3.0.1
#> 362            tictoc         1.0
#> 363             tidyr       1.0.3
#> 364        tidyselect       1.1.0
#> 365          tidytext       0.2.4
#> 366             tiler       0.2.4
#> 367          timeDate    3043.102
#> 368        timeSeries    3062.100
#> 369          TimeWarp      1.0.15
#> 370           tinytex        0.22
#> 371        tokenizers       0.2.1
#> 372         triebeard       0.3.0
#> 373         truncnorm       1.0-8
#> 374             trust       0.1-8
#> 375             udapi       0.1.3
#> 376             units       0.6-6
#> 377         unixtools       0.1-1
#> 378          urltools       1.7.3
#> 379        usefulFuns       0.1.0
#> 380           usethis       1.6.1
#> 381              utf8       1.1.4
#> 382              uuid       0.1-4
#> 383                V8       3.0.2
#> 384             vctrs       0.3.0
#> 385             vegan       2.5-6
#> 386             velox       0.2.0
#> 387          versions         0.3
#> 388              VGAM       1.1-3
#> 389           viridis       0.5.1
#> 390       viridisLite       0.3.0
#> 391        visNetwork       2.0.9
#> 392           webshot       0.5.2
#> 393           whisker         0.4
#> 394            whoami       1.3.0
#> 395             withr       2.2.0
#> 396    wordcountaddin  0.3.0.9000
#> 397              xfun        0.13
#> 398              xlsx       0.6.3
#> 399          xlsxjars       0.6.1
#> 400               XML    3.99-0.3
#> 401              xml2       1.3.2
#> 402      xmlparsedata       1.0.3
#> 403             xopen       1.0.0
#> 404            xtable       1.8-4
#> 405               xts      0.12-0
#> 406          yaImpute      1.0-32
#> 407              yaml       2.2.1
#> 408             yesno       0.1.1
#> 409               zip       2.0.4
#> 410               zoo       1.8-8
#> 411           littler       0.3.9
#> 412              base       3.6.3
#> 413              boot      1.3-25
#> 414             class      7.3-17
#> 415           cluster       2.1.0
#> 416         codetools      0.2-16
#> 417          compiler       3.6.3
#> 418          datasets       3.6.3
#> 419           foreign      0.8-76
#> 420          graphics       3.6.3
#> 421         grDevices       3.6.3
#> 422              grid       3.6.3
#> 423        KernSmooth     2.23-17
#> 424           lattice     0.20-41
#> 425              MASS    7.3-51.6
#> 426            Matrix      1.2-18
#> 427           methods       3.6.3
#> 428              mgcv      1.8-31
#> 429              nlme     3.1-147
#> 430              nnet      7.3-14
#> 431          parallel       3.6.3
#> 432             rpart      4.1-15
#> 433           spatial      7.3-12
#> 434           splines       3.6.3
#> 435             stats       3.6.3
#> 436            stats4       3.6.3
#> 437          survival      3.1-12
#> 438             tcltk       3.6.3
#> 439             tools       3.6.3
#> 440      translations       3.6.3
#> 441             utils       3.6.3

data.table::fread(pkgSnapFile)

#>          instPkgs instVers
#>   1: reproducible    1.1.0
#>   2: reproducible    1.1.0
#>   3: reproducible    1.1.0
#>   4: reproducible    1.1.0
#>   5: reproducible    1.1.0
#>  ---                      
#> 437:     survival   3.1-12
#> 438:        tcltk    3.6.3
#> 439:        tools    3.6.3
#> 440: translations    3.6.3
#> 441:        utils    3.6.3