Make any modifications to object recovered from cacheRepo

This is a generic definition that can be extended according to class.

.prepareOutput(object, cacheRepo, ...)

# S4 method for ANY
.prepareOutput(object, cacheRepo, ...)

Arguments

object

Any R object

cacheRepo

A repository used for storing cached objects. This is optional if Cache is used inside a SpaDES module.

...

Arguments passed to FUN

Value

The object, modified

Author

Eliot McIntire

Examples

a <- 1
.prepareOutput(a) # does nothing
#> [1] 1

b <- "NULL"
.prepareOutput(b) # converts to NULL
#> NULL

library(raster)
r <- raster(extent(0,10,0,10), vals = 1:100)

# write to disk manually -- will be in tempdir()
r <- writeRaster(r, file = tempfile())

# copy it to the cache repository
r <- .prepareOutput(r, tempdir())