Alternative to readLines that is faster

This alternative is from https://gist.github.com/hadley/6353939

readLinesRcpp(path)

Arguments

path
Path to text file to read.

Value

Similar to readLines. It may not return identical results.

Examples

readLinesRcpp(system.file(package = "reproducible", "DESCRIPTION"))

#>  [1] "Package: reproducible"                                                                                               
#>  [2] "Type: Package"                                                                                                       
#>  [3] "Title: A Set of Tools that Enhance Reproducibility Beyond Package Management"                                        
#>  [4] "Description: Built on top of 'git2r' and 'archivist', this package aims at"                                          
#>  [5] "    making high-level, robust, machine and OS independent tools for making"                                          
#>  [6] "    deeply reproducible and reusable content in R. This includes light weight"                                       
#>  [7] "    package management (similar to 'packrat' and 'checkpoint', but more flexible,"                                   
#>  [8] "    lightweight and simpler than both), plus it includes"                                                            
#>  [9] "    tools for caching, and accessing GitHub repositories."                                                           
#> [10] "URL: "                                                                                                               
#> [11] "    https://predictiveecology.github.io/reproducible/,"                                                              
#> [12] "    https://github.com/PredictiveEcology/reproducible"                                                               
#> [13] "Date: 2018-01-24"                                                                                                    
#> [14] "Version: 0.1.4"                                                                                                      
#> [15] "Authors@R: c("                                                                                                       
#> [16] "    person(\"Eliot J B\", \"McIntire\", email = \"eliot.mcintire@canada.ca\","                                       
#> [17] "    role = c(\"aut\", \"cre\")),"                                                                                    
#> [18] "    person(\"Alex M\", \"Chubaty\", email = \"alexander.chubaty@canada.ca\","                                        
#> [19] "    role = c(\"aut\")),"                                                                                             
#> [20] "    person(\"Her Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada\","
#> [21] "    role = \"cph\")"                                                                                                 
#> [22] "    )"                                                                                                               
#> [23] "Depends:"                                                                                                            
#> [24] "    R (>= 3.3.0)"                                                                                                    
#> [25] "Imports:"                                                                                                            
#> [26] "    archivist (>= 2.1.2),"                                                                                           
#> [27] "    crayon,"                                                                                                         
#> [28] "    data.table (>= 1.10.4),"                                                                                         
#> [29] "    devtools,"                                                                                                       
#> [30] "    digest,"                                                                                                         
#> [31] "    fastdigest,"                                                                                                     
#> [32] "    git2r (>= 0.18),"                                                                                                
#> [33] "    magrittr,"                                                                                                       
#> [34] "    memoise,"                                                                                                        
#> [35] "    methods,"                                                                                                        
#> [36] "    raster,"                                                                                                         
#> [37] "    Rcpp (>= 0.12.13),"                                                                                              
#> [38] "    RCurl,"                                                                                                          
#> [39] "    sp,"                                                                                                             
#> [40] "    versions"                                                                                                        
#> [41] "Suggests:"                                                                                                           
#> [42] "    hunspell,"                                                                                                       
#> [43] "    knitr,"                                                                                                          
#> [44] "    rgdal,"                                                                                                          
#> [45] "    rmarkdown,"                                                                                                      
#> [46] "    testthat"                                                                                                        
#> [47] "LinkingTo: Rcpp"                                                                                                     
#> [48] "License: GPL-3"                                                                                                      
#> [49] "BugReports: https://github.com/PredictiveEcology/reproducible/issues"                                                
#> [50] "ByteCompile: yes"                                                                                                    
#> [51] "RoxygenNote: 6.0.1"                                                                                                  
#> [52] "Collate: "                                                                                                           
#> [53] "    'RcppExports.R'"                                                                                                 
#> [54] "    'cache-helpers.R'"                                                                                               
#> [55] "    'cache-tools.R'"                                                                                                 
#> [56] "    'robustDigest.R'"                                                                                                
#> [57] "    'cache.R'"                                                                                                       
#> [58] "    'consistentPaths.R'"                                                                                             
#> [59] "    'git.R'"                                                                                                         
#> [60] "    'helpers.R'"                                                                                                     
#> [61] "    'packages.R'"                                                                                                    
#> [62] "    'pipe.R'"                                                                                                        
#> [63] "    'reproducible-package.R'"                                                                                        
#> [64] "    'search.R'"                                                                                                      
#> [65] "R'"                                                                                                                  
#> [66] "    'pipe.R'"                                                                                                        
#> [67] "    'reproducible-package.R'"                                                                                        
#> [68] "    'search.R'"