Create reproducible digests of objects in R

Not all aspects of R objects are captured by current hashing tools in R (e.g. digest::digest, fastdigest::fastdigest, knitr caching, archivist::cache). This is mostly because many objects have "transient" (e.g., functions have environments), or "disk-backed" features. This function allows for these accommodations to be made and uses fastdigest internally. Since the goal of using reproducibility is to have tools that are not session specific, this function attempts to strip all session specific information so that the fastdigest works between sessions and operating systems. It is tested under many conditions and object types, there are bound to be others that don't work correctly.

.robustDigest(object, objects, compareRasterFileLength = 1e+06,
  algo = "xxhash64", digestPathContent = FALSE, classOptions = list())

# S4 method for ANY
.robustDigest(object, compareRasterFileLength, algo,
  digestPathContent, classOptions)

# S4 method for cluster
.robustDigest(object, compareRasterFileLength, algo,
  digestPathContent, classOptions)

# S4 method for `function`
.robustDigest(object, compareRasterFileLength, algo,
  digestPathContent, classOptions)

# S4 method for expression
.robustDigest(object, compareRasterFileLength, algo,
  digestPathContent, classOptions)

# S4 method for character
.robustDigest(object, compareRasterFileLength, algo,
  digestPathContent, classOptions)

# S4 method for Path
.robustDigest(object, compareRasterFileLength, algo,
  digestPathContent, classOptions)

# S4 method for environment
.robustDigest(object, compareRasterFileLength, algo,
  digestPathContent, classOptions)

# S4 method for list
.robustDigest(object, compareRasterFileLength, algo,
  digestPathContent, classOptions)

# S4 method for Raster
.robustDigest(object, compareRasterFileLength, algo,
  digestPathContent, classOptions)

# S4 method for Spatial
.robustDigest(object, compareRasterFileLength, algo,
  digestPathContent, classOptions)

Arguments

object
an object to digest.

objects
Optional character vector indicating which objects are to be considered while making digestible. This is only relevant if the object being passed is an environment or list or the like.

compareRasterFileLength
Numeric. Optional. When there are Rasters, that have file-backed storage, this is passed to the length arg in digest when determining if the Raster file is already in the database. Note: uses digest for file-backed Raster. Default 1e6. Passed to .prepareFileBackedRaster.

algo
The algorithms to be used; currently available choices are md5, which is also the default, sha1, crc32, sha256, sha512, xxhash32, xxhash64 and murmur32.

digestPathContent
Logical. Should arguments that are of class Path (see examples below) have their name digested (FALSE), or their file contents (TRUE; default).

classOptions
Optional list. This will pass into .robustDigest for specific classes. Should be options that the .robustDigest knows what to do with.

Value

A hash i.e., digest of the object passed in.

Classes

Raster* objects have the potential for disk-backed storage. If the object in the R session is cached using archivist::cache, only the header component will be assessed for caching. Thus, objects like this require more work. Also, because Raster* can have a built-in representation for having their data content located on disk, this format will be maintained if the raster already is file-backed, i.e., to create .tif or .grd backed rasters, use writeRaster first, then Cache. The .tif or .grd will be copied to the "raster" subdirectory of the cacheRepo. Their RAM representation (as an R object) will still be in the usual gallery/ directory. For inMemory raster objects, they will remain as binary .RData files.

Functions (which are contained within environments) are converted to a text representation via a call to format(FUN).

Objects contained within a list or environment are recursively hashed using fastdigest, while removing all references to environments.

Character strings are first assessed with dir.exists and file.exists to check for paths. If they are found to be paths, then the path is hashed with only its filename via basename(filename). If it is actually a path, we suggest using asPath(thePath)

See also

cache. fastdigest.

Examples

a <- 2
tmpfile1 <- tempfile()
tmpfile2 <- tempfile()
save(a, file = tmpfile1)
save(a, file = tmpfile2)

# treats as character string, so 2 filenames are different
fastdigest::fastdigest(tmpfile1)

#> [1] "7616b3f08f75b8e46b6e6e5099259606"

fastdigest::fastdigest(tmpfile2)

#> [1] "558ca583c69b03f52a3e3a8d875fbbd2"

# tests to see whether character string is representing a file
.robustDigest(tmpfile1)

#> [1] "46016fc41d2dbcf4"

.robustDigest(tmpfile2) # same

#> [1] "46016fc41d2dbcf4"

# if you tell it that it is a path, then you can decide if you want it to be
#  treated as a character string or as a file path
.robustDigest(asPath(tmpfile1))

#> [1] "8345cbf07267dee5f8f511ec20c8a062"

.robustDigest(asPath(tmpfile2)) # different because you have a choice

#> [1] "2ffbd2f505cc7de0b64ce2c9895d2c21"

.robustDigest(asPath(tmpfile1), digestPathContent = TRUE)

#> [[1]]
#> [1] "46016fc41d2dbcf4"
#> 

.robustDigest(asPath(tmpfile2), digestPathContent = TRUE) # same

#> [[1]]
#> [1] "46016fc41d2dbcf4"
#> 

# Rasters are interesting because it is not know a priori if it
#   it has a file name associated with it.
library(raster)
r <- raster(extent(0,10,0,10), vals = 1:100)

# write to disk
r1 <- writeRaster(r, file = tmpfile1)
r2 <- writeRaster(r, file = tmpfile2)

digest::digest(r1)

#> [1] "ad819d19cbb8edf7d40b659108a7a00a"

digest::digest(r2) # different

#> [1] "79e6cb8ed6625234805ba9d7de899215"

fastdigest::fastdigest(r1)

#> [1] "f9b841727bfd42ae83d53fdfd1273e71"

fastdigest::fastdigest(r2) # different

#> [1] "4971a320b754b8efb35077eddf602ef4"

.robustDigest(r1)

#> [1] "79317ebbbb2d0e6b05fa1e968b9f5880"

.robustDigest(r2) # same... data are the same in the file

#> [1] "79317ebbbb2d0e6b05fa1e968b9f5880"

# note, this is not true for comparing memory and file-backed rasters
.robustDigest(r)

#> [1] "36cb66a1b7fedc105c3b25649645ae91"

.robustDigest(r1) # different

#> [1] "79317ebbbb2d0e6b05fa1e968b9f5880"