Write module inputs on disk

Can be used to write prepared inputs on disk.

writeOutputs(
  x,
  filename2,
  overwrite = getOption("reproducible.overwrite", NULL),
  ...
)

# S3 method for Raster
writeOutputs(
  x,
  filename2 = NULL,
  overwrite = getOption("reproducible.overwrite", FALSE),
  verbose = getOption("reproducible.verbose", 1),
  ...
)

# S3 method for Spatial
writeOutputs(
  x,
  filename2 = NULL,
  overwrite = getOption("reproducible.overwrite", TRUE),
  ...
)

# S3 method for sf
writeOutputs(
  x,
  filename2 = NULL,
  overwrite = getOption("reproducible.overwrite", FALSE),
  verbose = getOption("reproducible.verbose", 1),
  ...
)

# S3 method for quosure
writeOutputs(x, filename2, ...)

# S3 method for default
writeOutputs(x, filename2, ...)

Arguments

x
The object save to disk i.e., write outputs

filename2
File name passed to writeRaster, or shapefile or st_write (dsn argument).

overwrite
Logical. Should file being written overwrite an existing file if it exists.

...
Passed into shapefile or writeRaster or st_write

verbose
Numeric, -1 silent (where possible), 0 being very quiet, 1 showing more messaging, 2 being more messaging, etc. Default is 1. Above 3 will output much more information about the internals of Caching, which may help diagnose Caching challenges. Can set globally with an option, e.g., options('reproducible.verbose' = 0) to reduce to minimal

Author

Eliot McIntire and Jean Marchal

Examples

# Add a study area to Crop and Mask to
# Create a "study area"
library(sp)
library(raster)
ow <- setwd(tempdir())

# make a SpatialPolygon
coords1 <- structure(c(-123.98, -117.1, -80.2, -100, -123.98, 60.9, 67.73, 65.58, 51.79, 60.9),
                     .Dim = c(5L, 2L))
Sr1 <- Polygon(coords1)
Srs1 <- Polygons(list(Sr1), "s1")
shpEcozone <- SpatialPolygons(list(Srs1), 1L)
crs(shpEcozone) <- "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"

# make a "study area" that is subset of larger dataset
coords <- structure(c(-118.98, -116.1, -99.2, -106, -118.98, 59.9, 65.73, 63.58, 54.79, 59.9),
                    .Dim = c(5L, 2L))
Sr1 <- Polygon(coords)
Srs1 <- Polygons(list(Sr1), "s1")
StudyArea <- SpatialPolygons(list(Srs1), 1L)
crs(StudyArea) <- "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"

##########
shpEcozonePostProcessed <- postProcess(shpEcozone, studyArea = StudyArea)

#> useCache is FALSE; skipping Cache on function cropInputs)

#>     cropping ...

#> although coordinates are longitude/latitude, st_intersection assumes that they are planar

#> No cacheRepo supplied and getOption('reproducible.cachePath') is inside a temporary directory;
#>   this will not persist across R sessions.

#>   ...(Object to retrieve (c5139ef6645d9ae3.rds))

#>      loaded cached result from previous projectInputs call, 

#> maskInputs with sf class objects is still experimental

#>     intersecting ...

#> Checking for errors in SimpleFeature

#> Found errors in SimpleFeature. Attempting to correct.

#> useCache is FALSE; skipping Cache on function FUN)

#> dist is assumed to be in decimal degrees (arc_degrees).

#> Warning: st_buffer does not correctly buffer longitude/latitude data

#>   Some or all of the errors fixed.

#> although coordinates are longitude/latitude, st_intersection assumes that they are planar

#>   Skipping writeOutputs; filename2 is NULL

# Try manually, individual pieces
shpEcozoneReprojected <- projectInputs(shpEcozone, StudyArea)

#>     reprojecting ...

shpEcozoneCropped <- cropInputs(shpEcozone, StudyArea)

#>     cropping ...

#> although coordinates are longitude/latitude, st_intersection assumes that they are planar

shpEcozoneClean <- fixErrors(shpEcozone)

#> Checking for errors in SpatialPolygon

#>   Found no errors.

shpEcozoneMasked <- maskInputs(shpEcozone, StudyArea)

#> maskInputs with sf class objects is still experimental

#>     intersecting ...

#> Checking for errors in SimpleFeature

#> Found errors in SimpleFeature. Attempting to correct.

#> useCache is FALSE; skipping Cache on function FUN)

#> dist is assumed to be in decimal degrees (arc_degrees).

#> Warning: st_buffer does not correctly buffer longitude/latitude data

#>   Some or all of the errors fixed.

#> although coordinates are longitude/latitude, st_intersection assumes that they are planar

setwd(ow)