Clear erroneous archivist artifacts

Stub artifacts can result from several causes. The most common being erroneous removal of a file in the SQLite database. This can be caused sometimes if an archive object is being saved multiple times by multiple threads. This function will clear entries in the SQLite database which have no actual file with data.

clearStubArtifacts(repoDir = NULL)

# S4 method for ANY
clearStubArtifacts(repoDir = NULL)

Arguments

repoDir
A character denoting an existing directory of the repository for which metadata will be returned. If NULL (default), it will use the repoDir specified in archivist::setLocalRepo.

Value

Invoked for its side effect on the repoDir.

Examples

tmpDir <- file.path(tempdir(), "reproducible_examples", "clearStubArtifacts")

lapply(c(runif, rnorm), function(f) {
  reproducible::Cache(f, 10, cacheRepo = tmpDir)
})

#> [[1]]
#>  [1] 0.859625102 0.757621128 0.926860752 0.081588733 0.049136837 0.345142181
#>  [7] 0.002873588 0.678100822 0.604252266 0.119109327
#> attr(,"tags")
#> [1] "cacheId:c544d867e89c93fb20f0fccb922cc028"
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"call")
#> [1] ""
#> 
#> [[2]]
#>  [1]  1.02457159 -0.65918555 -0.38630438 -0.24249895  0.22059324 -1.48109003
#>  [7]  0.72469120 -2.37751470 -0.06540921 -0.46970481
#> attr(,"tags")
#> [1] "cacheId:03dcbe8e18b571c397bbc2e72230d216"
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"call")
#> [1] ""
#> 

# clear out any stub artifacts
showCache(tmpDir)

#> Cache size: 

#>   Total (including Rasters): 560 bytes

#>   Selected objects (not including Rasters): 560 bytes

#>                             artifact         tagKey
#>  1: 3553a073a928c7c1737d6e5475d8b0f1         format
#>  2: 3553a073a928c7c1737d6e5475d8b0f1           name
#>  3: 3553a073a928c7c1737d6e5475d8b0f1          class
#>  4: 3553a073a928c7c1737d6e5475d8b0f1           date
#>  5: 3553a073a928c7c1737d6e5475d8b0f1        cacheId
#>  6: 3553a073a928c7c1737d6e5475d8b0f1       function
#>  7: 3553a073a928c7c1737d6e5475d8b0f1    object.size
#>  8: 3553a073a928c7c1737d6e5475d8b0f1       accessed
#>  9: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 10: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 11: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 12: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 13: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 14: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 15: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 16: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 17: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 18: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 19: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 20: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 21: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 22: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 23: 3553a073a928c7c1737d6e5475d8b0f1      preDigest
#> 24: 3553a073a928c7c1737d6e5475d8b0f1      preDigest
#> 25: 95a41085c0a8dbfc008cc40afec6c50f         format
#> 26: 95a41085c0a8dbfc008cc40afec6c50f           name
#> 27: 95a41085c0a8dbfc008cc40afec6c50f          class
#> 28: 95a41085c0a8dbfc008cc40afec6c50f           date
#> 29: 95a41085c0a8dbfc008cc40afec6c50f        cacheId
#> 30: 95a41085c0a8dbfc008cc40afec6c50f       function
#> 31: 95a41085c0a8dbfc008cc40afec6c50f    object.size
#> 32: 95a41085c0a8dbfc008cc40afec6c50f       accessed
#> 33: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 34: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 35: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 36: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 37: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 38: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 39: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 40: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 41: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 42: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 43: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 44: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 45: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 46: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 47: 95a41085c0a8dbfc008cc40afec6c50f      preDigest
#> 48: 95a41085c0a8dbfc008cc40afec6c50f      preDigest
#>                             artifact         tagKey
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-11-07 10:12:43
#>  2:      3553a073a928c7c1737d6e5475d8b0f1 2018-11-07 10:12:43
#>  3:                               numeric 2018-11-07 10:12:43
#>  4:                   2018-11-07 10:12:43 2018-11-07 10:12:43
#>  5:      c544d867e89c93fb20f0fccb922cc028 2018-11-07 10:12:43
#>  6:                                     f 2018-11-07 10:12:43
#>  7:                                  1120 2018-11-07 10:12:43
#>  8:                   2018-11-07 10:12:43 2018-11-07 10:12:43
#>  9:                   withCallingHandlers 2018-11-07 10:12:43
#> 10:                               saveRDS 2018-11-07 10:12:43
#> 11:                               do.call 2018-11-07 10:12:43
#> 12:                      build_site_local 2018-11-07 10:12:43
#> 13:                       build_reference 2018-11-07 10:12:43
#> 14:                  data_reference_topic 2018-11-07 10:12:43
#> 15:                               as_data 2018-11-07 10:12:43
#> 16:                  as_data.tag_examples 2018-11-07 10:12:43
#> 17:                             timing_fn 2018-11-07 10:12:43
#> 18:                                handle 2018-11-07 10:12:43
#> 19:                                   try 2018-11-07 10:12:43
#> 20:                           withVisible 2018-11-07 10:12:43
#> 21:                                lapply 2018-11-07 10:12:43
#> 22:                                   FUN 2018-11-07 10:12:43
#> 23:    n:c192d58bc0a9ddf2b5f916d929bccc08 2018-11-07 10:12:43
#> 24: .FUN:179092fbad3ec261e58bc89bc8a703c3 2018-11-07 10:12:43
#> 25:                                   rda 2018-11-07 10:12:43
#> 26:      95a41085c0a8dbfc008cc40afec6c50f 2018-11-07 10:12:43
#> 27:                               numeric 2018-11-07 10:12:43
#> 28:                   2018-11-07 10:12:43 2018-11-07 10:12:44
#> 29:      03dcbe8e18b571c397bbc2e72230d216 2018-11-07 10:12:44
#> 30:                                     f 2018-11-07 10:12:44
#> 31:                                  1120 2018-11-07 10:12:44
#> 32:                   2018-11-07 10:12:43 2018-11-07 10:12:44
#> 33:                   withCallingHandlers 2018-11-07 10:12:44
#> 34:                               saveRDS 2018-11-07 10:12:44
#> 35:                               do.call 2018-11-07 10:12:44
#> 36:                      build_site_local 2018-11-07 10:12:44
#> 37:                       build_reference 2018-11-07 10:12:44
#> 38:                  data_reference_topic 2018-11-07 10:12:44
#> 39:                               as_data 2018-11-07 10:12:44
#> 40:                  as_data.tag_examples 2018-11-07 10:12:44
#> 41:                             timing_fn 2018-11-07 10:12:44
#> 42:                                handle 2018-11-07 10:12:44
#> 43:                                   try 2018-11-07 10:12:44
#> 44:                           withVisible 2018-11-07 10:12:44
#> 45:                                lapply 2018-11-07 10:12:44
#> 46:                                   FUN 2018-11-07 10:12:44
#> 47:    n:c192d58bc0a9ddf2b5f916d929bccc08 2018-11-07 10:12:44
#> 48: .FUN:a979d0700a6ef8c7c18736096ff1d522 2018-11-07 10:12:44
#>                                  tagValue         createdDate

file2Remove <- dir(file.path(tmpDir, "gallery"), full.name = TRUE)[1]
file.remove(file2Remove)

#> [1] TRUE

showCache(tmpDir) # repository directory still thinks files are there

#> Cache size: 

#>   Total (including Rasters): 560 bytes

#>   Selected objects (not including Rasters): 560 bytes

#>                             artifact         tagKey
#>  1: 3553a073a928c7c1737d6e5475d8b0f1         format
#>  2: 3553a073a928c7c1737d6e5475d8b0f1           name
#>  3: 3553a073a928c7c1737d6e5475d8b0f1          class
#>  4: 3553a073a928c7c1737d6e5475d8b0f1           date
#>  5: 3553a073a928c7c1737d6e5475d8b0f1        cacheId
#>  6: 3553a073a928c7c1737d6e5475d8b0f1       function
#>  7: 3553a073a928c7c1737d6e5475d8b0f1    object.size
#>  8: 3553a073a928c7c1737d6e5475d8b0f1       accessed
#>  9: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 10: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 11: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 12: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 13: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 14: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 15: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 16: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 17: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 18: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 19: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 20: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 21: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 22: 3553a073a928c7c1737d6e5475d8b0f1 otherFunctions
#> 23: 3553a073a928c7c1737d6e5475d8b0f1      preDigest
#> 24: 3553a073a928c7c1737d6e5475d8b0f1      preDigest
#> 25: 95a41085c0a8dbfc008cc40afec6c50f         format
#> 26: 95a41085c0a8dbfc008cc40afec6c50f           name
#> 27: 95a41085c0a8dbfc008cc40afec6c50f          class
#> 28: 95a41085c0a8dbfc008cc40afec6c50f           date
#> 29: 95a41085c0a8dbfc008cc40afec6c50f        cacheId
#> 30: 95a41085c0a8dbfc008cc40afec6c50f       function
#> 31: 95a41085c0a8dbfc008cc40afec6c50f    object.size
#> 32: 95a41085c0a8dbfc008cc40afec6c50f       accessed
#> 33: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 34: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 35: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 36: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 37: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 38: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 39: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 40: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 41: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 42: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 43: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 44: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 45: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 46: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 47: 95a41085c0a8dbfc008cc40afec6c50f      preDigest
#> 48: 95a41085c0a8dbfc008cc40afec6c50f      preDigest
#>                             artifact         tagKey
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-11-07 10:12:43
#>  2:      3553a073a928c7c1737d6e5475d8b0f1 2018-11-07 10:12:43
#>  3:                               numeric 2018-11-07 10:12:43
#>  4:                   2018-11-07 10:12:43 2018-11-07 10:12:43
#>  5:      c544d867e89c93fb20f0fccb922cc028 2018-11-07 10:12:43
#>  6:                                     f 2018-11-07 10:12:43
#>  7:                                  1120 2018-11-07 10:12:43
#>  8:                   2018-11-07 10:12:43 2018-11-07 10:12:43
#>  9:                   withCallingHandlers 2018-11-07 10:12:43
#> 10:                               saveRDS 2018-11-07 10:12:43
#> 11:                               do.call 2018-11-07 10:12:43
#> 12:                      build_site_local 2018-11-07 10:12:43
#> 13:                       build_reference 2018-11-07 10:12:43
#> 14:                  data_reference_topic 2018-11-07 10:12:43
#> 15:                               as_data 2018-11-07 10:12:43
#> 16:                  as_data.tag_examples 2018-11-07 10:12:43
#> 17:                             timing_fn 2018-11-07 10:12:43
#> 18:                                handle 2018-11-07 10:12:43
#> 19:                                   try 2018-11-07 10:12:43
#> 20:                           withVisible 2018-11-07 10:12:43
#> 21:                                lapply 2018-11-07 10:12:43
#> 22:                                   FUN 2018-11-07 10:12:43
#> 23:    n:c192d58bc0a9ddf2b5f916d929bccc08 2018-11-07 10:12:43
#> 24: .FUN:179092fbad3ec261e58bc89bc8a703c3 2018-11-07 10:12:43
#> 25:                                   rda 2018-11-07 10:12:43
#> 26:      95a41085c0a8dbfc008cc40afec6c50f 2018-11-07 10:12:43
#> 27:                               numeric 2018-11-07 10:12:43
#> 28:                   2018-11-07 10:12:43 2018-11-07 10:12:44
#> 29:      03dcbe8e18b571c397bbc2e72230d216 2018-11-07 10:12:44
#> 30:                                     f 2018-11-07 10:12:44
#> 31:                                  1120 2018-11-07 10:12:44
#> 32:                   2018-11-07 10:12:43 2018-11-07 10:12:44
#> 33:                   withCallingHandlers 2018-11-07 10:12:44
#> 34:                               saveRDS 2018-11-07 10:12:44
#> 35:                               do.call 2018-11-07 10:12:44
#> 36:                      build_site_local 2018-11-07 10:12:44
#> 37:                       build_reference 2018-11-07 10:12:44
#> 38:                  data_reference_topic 2018-11-07 10:12:44
#> 39:                               as_data 2018-11-07 10:12:44
#> 40:                  as_data.tag_examples 2018-11-07 10:12:44
#> 41:                             timing_fn 2018-11-07 10:12:44
#> 42:                                handle 2018-11-07 10:12:44
#> 43:                                   try 2018-11-07 10:12:44
#> 44:                           withVisible 2018-11-07 10:12:44
#> 45:                                lapply 2018-11-07 10:12:44
#> 46:                                   FUN 2018-11-07 10:12:44
#> 47:    n:c192d58bc0a9ddf2b5f916d929bccc08 2018-11-07 10:12:44
#> 48: .FUN:a979d0700a6ef8c7c18736096ff1d522 2018-11-07 10:12:44
#>                                  tagValue         createdDate

# run clearStubArtifacts
suppressWarnings(clearStubArtifacts(tmpDir))
showCache(tmpDir) # stubs are removed

#> Cache size: 

#>   Total (including Rasters): 280 bytes

#>   Selected objects (not including Rasters): 280 bytes

#>                             artifact         tagKey
#>  1: 95a41085c0a8dbfc008cc40afec6c50f         format
#>  2: 95a41085c0a8dbfc008cc40afec6c50f           name
#>  3: 95a41085c0a8dbfc008cc40afec6c50f          class
#>  4: 95a41085c0a8dbfc008cc40afec6c50f           date
#>  5: 95a41085c0a8dbfc008cc40afec6c50f        cacheId
#>  6: 95a41085c0a8dbfc008cc40afec6c50f       function
#>  7: 95a41085c0a8dbfc008cc40afec6c50f    object.size
#>  8: 95a41085c0a8dbfc008cc40afec6c50f       accessed
#>  9: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 10: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 11: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 12: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 13: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 14: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 15: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 16: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 17: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 18: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 19: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 20: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 21: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 22: 95a41085c0a8dbfc008cc40afec6c50f otherFunctions
#> 23: 95a41085c0a8dbfc008cc40afec6c50f      preDigest
#> 24: 95a41085c0a8dbfc008cc40afec6c50f      preDigest
#>                             artifact         tagKey
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-11-07 10:12:43
#>  2:      95a41085c0a8dbfc008cc40afec6c50f 2018-11-07 10:12:43
#>  3:                               numeric 2018-11-07 10:12:43
#>  4:                   2018-11-07 10:12:43 2018-11-07 10:12:44
#>  5:      03dcbe8e18b571c397bbc2e72230d216 2018-11-07 10:12:44
#>  6:                                     f 2018-11-07 10:12:44
#>  7:                                  1120 2018-11-07 10:12:44
#>  8:                   2018-11-07 10:12:43 2018-11-07 10:12:44
#>  9:                   withCallingHandlers 2018-11-07 10:12:44
#> 10:                               saveRDS 2018-11-07 10:12:44
#> 11:                               do.call 2018-11-07 10:12:44
#> 12:                      build_site_local 2018-11-07 10:12:44
#> 13:                       build_reference 2018-11-07 10:12:44
#> 14:                  data_reference_topic 2018-11-07 10:12:44
#> 15:                               as_data 2018-11-07 10:12:44
#> 16:                  as_data.tag_examples 2018-11-07 10:12:44
#> 17:                             timing_fn 2018-11-07 10:12:44
#> 18:                                handle 2018-11-07 10:12:44
#> 19:                                   try 2018-11-07 10:12:44
#> 20:                           withVisible 2018-11-07 10:12:44
#> 21:                                lapply 2018-11-07 10:12:44
#> 22:                                   FUN 2018-11-07 10:12:44
#> 23:    n:c192d58bc0a9ddf2b5f916d929bccc08 2018-11-07 10:12:44
#> 24: .FUN:a979d0700a6ef8c7c18736096ff1d522 2018-11-07 10:12:44
#>                                  tagValue         createdDate

# cleanup
clearCache(tmpDir)
unlink(tmpDir, recursive = TRUE)