Clear erroneous archivist artifacts

Stub artifacts can result from several causes. The most common being erroneous removal of a file in the sqlite database. This can be caused sometimes if an archive object is being saved multiple times by multiple threads. This function will clear entries in the sqlite database which have no actual file with data.

clearStubArtifacts(repoDir = NULL)

# S4 method for ANY
clearStubArtifacts(repoDir = NULL)

Arguments

repoDir
A character denoting an existing directory of the repository for which metadata will be returned. If NULL (default), it will use the repoDir specified in archivist::setLocalRepo.

Value

Invoked for its side effect on the repoDir.

Examples

tmpDir <- file.path(tempdir(), "reproducible_examples", "clearStubArtifacts")

lapply(c(runif, rnorm), function(f) {
  reproducible::Cache(f, 10, cacheRepo = tmpDir)
})

#> [[1]]
#>  [1] 0.94467527 0.66079779 0.62911404 0.06178627 0.20597457 0.17655675
#>  [7] 0.68702285 0.38410372 0.76984142 0.49769924
#> attr(,"tags")
#> [1] "cacheId:f024b63398f829eb00833e3e0d1cc753"
#> attr(,"newCache")
#> [1] TRUE
#> attr(,"call")
#> [1] ""
#> 
#> [[2]]
#>  [1]  0.57578135 -0.30538839  1.51178117  0.38984324 -0.62124058 -2.21469989
#>  [7]  1.12493092 -0.04493361 -0.01619026  0.94383621
#> attr(,"tags")
#> [1] "cacheId:b7eae7c361dfb6d161192321de6bc893"
#> attr(,"newCache")
#> [1] TRUE
#> attr(,"call")
#> [1] ""
#> 

# clear out any stub artifacts
showCache(tmpDir)

#> Cache size: 

#>   Total (including Rasters): 12.4 Kb

#>   Selected objects (not including Rasters): 426 bytes

#>                             artifact         tagKey
#>  1: 2f832b611ee22ec8a43069b5c431fe70         format
#>  2: 2f832b611ee22ec8a43069b5c431fe70           name
#>  3: 2f832b611ee22ec8a43069b5c431fe70          class
#>  4: 2f832b611ee22ec8a43069b5c431fe70           date
#>  5: 2f832b611ee22ec8a43069b5c431fe70        cacheId
#>  6: 2f832b611ee22ec8a43069b5c431fe70       function
#>  7: 2f832b611ee22ec8a43069b5c431fe70    object.size
#>  8: 2f832b611ee22ec8a43069b5c431fe70       accessed
#>  9: 2f832b611ee22ec8a43069b5c431fe70 otherFunctions
#> 10: 2f832b611ee22ec8a43069b5c431fe70 otherFunctions
#> 11: 2f832b611ee22ec8a43069b5c431fe70 otherFunctions
#> 12: 2f832b611ee22ec8a43069b5c431fe70      preDigest
#> 13: 2f832b611ee22ec8a43069b5c431fe70      preDigest
#> 14: d2780cccdef206458bba01d19d3ff208         format
#> 15: d2780cccdef206458bba01d19d3ff208           name
#> 16: d2780cccdef206458bba01d19d3ff208          class
#> 17: d2780cccdef206458bba01d19d3ff208           date
#> 18: d2780cccdef206458bba01d19d3ff208        cacheId
#> 19: d2780cccdef206458bba01d19d3ff208       function
#> 20: d2780cccdef206458bba01d19d3ff208    object.size
#> 21: d2780cccdef206458bba01d19d3ff208       accessed
#> 22: d2780cccdef206458bba01d19d3ff208 otherFunctions
#> 23: d2780cccdef206458bba01d19d3ff208 otherFunctions
#> 24: d2780cccdef206458bba01d19d3ff208 otherFunctions
#> 25: d2780cccdef206458bba01d19d3ff208      preDigest
#> 26: d2780cccdef206458bba01d19d3ff208      preDigest
#>                             artifact         tagKey
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-05-17 13:15:18
#>  2:                                 Cache 2018-05-17 13:15:18
#>  3:                               numeric 2018-05-17 13:15:18
#>  4:                   2018-05-17 13:15:18 2018-05-17 13:15:18
#>  5:      f024b63398f829eb00833e3e0d1cc753 2018-05-17 13:15:18
#>  6:                                     f 2018-05-17 13:15:18
#>  7:                                   784 2018-05-17 13:15:18
#>  8:                   2018-05-17 13:15:18 2018-05-17 13:15:18
#>  9:                       build_reference 2018-05-17 13:15:18
#> 10:                  data_reference_topic 2018-05-17 13:15:18
#> 11:                                   try 2018-05-17 13:15:18
#> 12:    n:52ad08b1270ae9be9c0a12805d408433 2018-05-17 13:15:18
#> 13: .FUN:d2631d24c3b38b89c7bdd4ab7faaaac3 2018-05-17 13:15:18
#> 14:                                   rda 2018-05-17 13:15:18
#> 15:                                 Cache 2018-05-17 13:15:18
#> 16:                               numeric 2018-05-17 13:15:18
#> 17:                   2018-05-17 13:15:18 2018-05-17 13:15:18
#> 18:      b7eae7c361dfb6d161192321de6bc893 2018-05-17 13:15:18
#> 19:                                     f 2018-05-17 13:15:18
#> 20:                                   784 2018-05-17 13:15:18
#> 21:                   2018-05-17 13:15:18 2018-05-17 13:15:18
#> 22:                       build_reference 2018-05-17 13:15:18
#> 23:                  data_reference_topic 2018-05-17 13:15:18
#> 24:                                   try 2018-05-17 13:15:18
#> 25:    n:52ad08b1270ae9be9c0a12805d408433 2018-05-17 13:15:18
#> 26: .FUN:7e9a928f110f80b3612e71883a6ec1f4 2018-05-17 13:15:18
#>                                  tagValue         createdDate

file2Remove <- dir(file.path(tmpDir, "gallery"), full.name = TRUE)[1]
file.remove(file2Remove)

#> [1] TRUE

showCache(tmpDir) # repository directory still thinks files are there

#> Cache size: 

#>   Total (including Rasters): 12.2 Kb

#>   Selected objects (not including Rasters): 224 bytes

#>                             artifact         tagKey
#>  1: 2f832b611ee22ec8a43069b5c431fe70         format
#>  2: 2f832b611ee22ec8a43069b5c431fe70           name
#>  3: 2f832b611ee22ec8a43069b5c431fe70          class
#>  4: 2f832b611ee22ec8a43069b5c431fe70           date
#>  5: 2f832b611ee22ec8a43069b5c431fe70        cacheId
#>  6: 2f832b611ee22ec8a43069b5c431fe70       function
#>  7: 2f832b611ee22ec8a43069b5c431fe70    object.size
#>  8: 2f832b611ee22ec8a43069b5c431fe70       accessed
#>  9: 2f832b611ee22ec8a43069b5c431fe70 otherFunctions
#> 10: 2f832b611ee22ec8a43069b5c431fe70 otherFunctions
#> 11: 2f832b611ee22ec8a43069b5c431fe70 otherFunctions
#> 12: 2f832b611ee22ec8a43069b5c431fe70      preDigest
#> 13: 2f832b611ee22ec8a43069b5c431fe70      preDigest
#> 14: d2780cccdef206458bba01d19d3ff208         format
#> 15: d2780cccdef206458bba01d19d3ff208           name
#> 16: d2780cccdef206458bba01d19d3ff208          class
#> 17: d2780cccdef206458bba01d19d3ff208           date
#> 18: d2780cccdef206458bba01d19d3ff208        cacheId
#> 19: d2780cccdef206458bba01d19d3ff208       function
#> 20: d2780cccdef206458bba01d19d3ff208    object.size
#> 21: d2780cccdef206458bba01d19d3ff208       accessed
#> 22: d2780cccdef206458bba01d19d3ff208 otherFunctions
#> 23: d2780cccdef206458bba01d19d3ff208 otherFunctions
#> 24: d2780cccdef206458bba01d19d3ff208 otherFunctions
#> 25: d2780cccdef206458bba01d19d3ff208      preDigest
#> 26: d2780cccdef206458bba01d19d3ff208      preDigest
#>                             artifact         tagKey
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-05-17 13:15:18
#>  2:                                 Cache 2018-05-17 13:15:18
#>  3:                               numeric 2018-05-17 13:15:18
#>  4:                   2018-05-17 13:15:18 2018-05-17 13:15:18
#>  5:      f024b63398f829eb00833e3e0d1cc753 2018-05-17 13:15:18
#>  6:                                     f 2018-05-17 13:15:18
#>  7:                                   784 2018-05-17 13:15:18
#>  8:                   2018-05-17 13:15:18 2018-05-17 13:15:18
#>  9:                       build_reference 2018-05-17 13:15:18
#> 10:                  data_reference_topic 2018-05-17 13:15:18
#> 11:                                   try 2018-05-17 13:15:18
#> 12:    n:52ad08b1270ae9be9c0a12805d408433 2018-05-17 13:15:18
#> 13: .FUN:d2631d24c3b38b89c7bdd4ab7faaaac3 2018-05-17 13:15:18
#> 14:                                   rda 2018-05-17 13:15:18
#> 15:                                 Cache 2018-05-17 13:15:18
#> 16:                               numeric 2018-05-17 13:15:18
#> 17:                   2018-05-17 13:15:18 2018-05-17 13:15:18
#> 18:      b7eae7c361dfb6d161192321de6bc893 2018-05-17 13:15:18
#> 19:                                     f 2018-05-17 13:15:18
#> 20:                                   784 2018-05-17 13:15:18
#> 21:                   2018-05-17 13:15:18 2018-05-17 13:15:18
#> 22:                       build_reference 2018-05-17 13:15:18
#> 23:                  data_reference_topic 2018-05-17 13:15:18
#> 24:                                   try 2018-05-17 13:15:18
#> 25:    n:52ad08b1270ae9be9c0a12805d408433 2018-05-17 13:15:18
#> 26: .FUN:7e9a928f110f80b3612e71883a6ec1f4 2018-05-17 13:15:18
#>                                  tagValue         createdDate

# run clearStubArtifacts
suppressWarnings(clearStubArtifacts(tmpDir))
showCache(tmpDir) # stubs are removed

#> Cache size: 

#>   Total (including Rasters): 12.2 Kb

#>   Selected objects (not including Rasters): 224 bytes

#>                             artifact         tagKey
#>  1: d2780cccdef206458bba01d19d3ff208         format
#>  2: d2780cccdef206458bba01d19d3ff208           name
#>  3: d2780cccdef206458bba01d19d3ff208          class
#>  4: d2780cccdef206458bba01d19d3ff208           date
#>  5: d2780cccdef206458bba01d19d3ff208        cacheId
#>  6: d2780cccdef206458bba01d19d3ff208       function
#>  7: d2780cccdef206458bba01d19d3ff208    object.size
#>  8: d2780cccdef206458bba01d19d3ff208       accessed
#>  9: d2780cccdef206458bba01d19d3ff208 otherFunctions
#> 10: d2780cccdef206458bba01d19d3ff208 otherFunctions
#> 11: d2780cccdef206458bba01d19d3ff208 otherFunctions
#> 12: d2780cccdef206458bba01d19d3ff208      preDigest
#> 13: d2780cccdef206458bba01d19d3ff208      preDigest
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-05-17 13:15:18
#>  2:                                 Cache 2018-05-17 13:15:18
#>  3:                               numeric 2018-05-17 13:15:18
#>  4:                   2018-05-17 13:15:18 2018-05-17 13:15:18
#>  5:      b7eae7c361dfb6d161192321de6bc893 2018-05-17 13:15:18
#>  6:                                     f 2018-05-17 13:15:18
#>  7:                                   784 2018-05-17 13:15:18
#>  8:                   2018-05-17 13:15:18 2018-05-17 13:15:18
#>  9:                       build_reference 2018-05-17 13:15:18
#> 10:                  data_reference_topic 2018-05-17 13:15:18
#> 11:                                   try 2018-05-17 13:15:18
#> 12:    n:52ad08b1270ae9be9c0a12805d408433 2018-05-17 13:15:18
#> 13: .FUN:7e9a928f110f80b3612e71883a6ec1f4 2018-05-17 13:15:18

# cleanup
clearCache(tmpDir)
unlink(tmpDir, recursive = TRUE)