Clear erroneous archivist artifacts

Stub artifacts can result from several causes. The most common being erroneous removal of a file in the sqlite database. This can be caused sometimes if an archive object is being saved multiple times by multiple threads. This function will clear entries in the sqlite database which have no actual file with data.

clearStubArtifacts(repoDir = NULL)

# S4 method for ANY
clearStubArtifacts(repoDir = NULL)

Arguments

repoDir
A character denoting an existing directory of the repository for which metadata will be returned. If NULL (default), it will use the repoDir specified in archivist::setLocalRepo.

Value

Invoked for its side effect on the repoDir.

Examples

tmpDir <- file.path(tempdir(), "reproducible_examples", "clearStubArtifacts")

lapply(c(runif, rnorm), function(f) {
  reproducible::Cache(f, 10, cacheRepo = tmpDir)
})

#> [[1]]
#>  [1] 0.6865692 0.1636232 0.9528248 0.3218546 0.3615341 0.8877234 0.8280144
#>  [8] 0.1006565 0.9060516 0.7727304
#> attr(,"tags")
#> [1] "cacheId:c69b2b95bcd43bf312ca6d89049726f4"
#> attr(,"call")
#> [1] ""
#> 
#> [[2]]
#>  [1] -0.2966400 -0.3872136 -0.7854327 -1.0567369 -0.7955414 -1.7562754
#>  [7] -0.6905379 -0.5585420 -0.5366633  0.2271271
#> attr(,"tags")
#> [1] "cacheId:7428e865f914063acaf23ecc7c817315"
#> attr(,"call")
#> [1] ""
#> 

# clear out any stub artifacts
showCache(tmpDir)

#>                             artifact         tagKey
#>  1: 09a4f143d744044c0c8e2e60c6d5fbbb         format
#>  2: 09a4f143d744044c0c8e2e60c6d5fbbb           name
#>  3: 09a4f143d744044c0c8e2e60c6d5fbbb          class
#>  4: 09a4f143d744044c0c8e2e60c6d5fbbb           date
#>  5: 09a4f143d744044c0c8e2e60c6d5fbbb       function
#>  6: 09a4f143d744044c0c8e2e60c6d5fbbb    object.size
#>  7: 09a4f143d744044c0c8e2e60c6d5fbbb       accessed
#>  8: 09a4f143d744044c0c8e2e60c6d5fbbb otherFunctions
#>  9: 09a4f143d744044c0c8e2e60c6d5fbbb otherFunctions
#> 10: 09a4f143d744044c0c8e2e60c6d5fbbb otherFunctions
#> 11: 09a4f143d744044c0c8e2e60c6d5fbbb        cacheId
#> 12: b372b382fe0902130b2848f82546618d         format
#> 13: b372b382fe0902130b2848f82546618d           name
#> 14: b372b382fe0902130b2848f82546618d          class
#> 15: b372b382fe0902130b2848f82546618d           date
#> 16: b372b382fe0902130b2848f82546618d       function
#> 17: b372b382fe0902130b2848f82546618d    object.size
#> 18: b372b382fe0902130b2848f82546618d       accessed
#> 19: b372b382fe0902130b2848f82546618d otherFunctions
#> 20: b372b382fe0902130b2848f82546618d otherFunctions
#> 21: b372b382fe0902130b2848f82546618d otherFunctions
#> 22: b372b382fe0902130b2848f82546618d        cacheId
#>                             artifact         tagKey
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:47:51
#>  2:                               Cache 2018-01-25 08:47:51
#>  3:                             numeric 2018-01-25 08:47:51
#>  4:                 2018-01-25 08:47:51 2018-01-25 08:47:51
#>  5:                                   f 2018-01-25 08:47:51
#>  6:                                 624 2018-01-25 08:47:51
#>  7:                 2018-01-25 08:47:51 2018-01-25 08:47:51
#>  8:      otherFunctions:build_reference 2018-01-25 08:47:51
#>  9: otherFunctions:data_reference_topic 2018-01-25 08:47:51
#> 10:                  otherFunctions:try 2018-01-25 08:47:51
#> 11:    c69b2b95bcd43bf312ca6d89049726f4 2018-01-25 08:47:51
#> 12:                                 rda 2018-01-25 08:47:51
#> 13:                               Cache 2018-01-25 08:47:51
#> 14:                             numeric 2018-01-25 08:47:51
#> 15:                 2018-01-25 08:47:51 2018-01-25 08:47:51
#> 16:                                   f 2018-01-25 08:47:51
#> 17:                                 624 2018-01-25 08:47:51
#> 18:                 2018-01-25 08:47:51 2018-01-25 08:47:51
#> 19:      otherFunctions:build_reference 2018-01-25 08:47:51
#> 20: otherFunctions:data_reference_topic 2018-01-25 08:47:51
#> 21:                  otherFunctions:try 2018-01-25 08:47:51
#> 22:    7428e865f914063acaf23ecc7c817315 2018-01-25 08:47:51
#>                                tagValue         createdDate

file2Remove <- dir(file.path(tmpDir, "gallery"), full.name = TRUE)[1]
file.remove(file2Remove)

#> [1] TRUE

showCache(tmpDir) # repository directory still thinks files are there

#>                             artifact         tagKey
#>  1: 09a4f143d744044c0c8e2e60c6d5fbbb         format
#>  2: 09a4f143d744044c0c8e2e60c6d5fbbb           name
#>  3: 09a4f143d744044c0c8e2e60c6d5fbbb          class
#>  4: 09a4f143d744044c0c8e2e60c6d5fbbb           date
#>  5: 09a4f143d744044c0c8e2e60c6d5fbbb       function
#>  6: 09a4f143d744044c0c8e2e60c6d5fbbb    object.size
#>  7: 09a4f143d744044c0c8e2e60c6d5fbbb       accessed
#>  8: 09a4f143d744044c0c8e2e60c6d5fbbb otherFunctions
#>  9: 09a4f143d744044c0c8e2e60c6d5fbbb otherFunctions
#> 10: 09a4f143d744044c0c8e2e60c6d5fbbb otherFunctions
#> 11: 09a4f143d744044c0c8e2e60c6d5fbbb        cacheId
#> 12: b372b382fe0902130b2848f82546618d         format
#> 13: b372b382fe0902130b2848f82546618d           name
#> 14: b372b382fe0902130b2848f82546618d          class
#> 15: b372b382fe0902130b2848f82546618d           date
#> 16: b372b382fe0902130b2848f82546618d       function
#> 17: b372b382fe0902130b2848f82546618d    object.size
#> 18: b372b382fe0902130b2848f82546618d       accessed
#> 19: b372b382fe0902130b2848f82546618d otherFunctions
#> 20: b372b382fe0902130b2848f82546618d otherFunctions
#> 21: b372b382fe0902130b2848f82546618d otherFunctions
#> 22: b372b382fe0902130b2848f82546618d        cacheId
#>                             artifact         tagKey
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:47:51
#>  2:                               Cache 2018-01-25 08:47:51
#>  3:                             numeric 2018-01-25 08:47:51
#>  4:                 2018-01-25 08:47:51 2018-01-25 08:47:51
#>  5:                                   f 2018-01-25 08:47:51
#>  6:                                 624 2018-01-25 08:47:51
#>  7:                 2018-01-25 08:47:51 2018-01-25 08:47:51
#>  8:      otherFunctions:build_reference 2018-01-25 08:47:51
#>  9: otherFunctions:data_reference_topic 2018-01-25 08:47:51
#> 10:                  otherFunctions:try 2018-01-25 08:47:51
#> 11:    c69b2b95bcd43bf312ca6d89049726f4 2018-01-25 08:47:51
#> 12:                                 rda 2018-01-25 08:47:51
#> 13:                               Cache 2018-01-25 08:47:51
#> 14:                             numeric 2018-01-25 08:47:51
#> 15:                 2018-01-25 08:47:51 2018-01-25 08:47:51
#> 16:                                   f 2018-01-25 08:47:51
#> 17:                                 624 2018-01-25 08:47:51
#> 18:                 2018-01-25 08:47:51 2018-01-25 08:47:51
#> 19:      otherFunctions:build_reference 2018-01-25 08:47:51
#> 20: otherFunctions:data_reference_topic 2018-01-25 08:47:51
#> 21:                  otherFunctions:try 2018-01-25 08:47:51
#> 22:    7428e865f914063acaf23ecc7c817315 2018-01-25 08:47:51
#>                                tagValue         createdDate

# run clearStubArtifacts
suppressWarnings(clearStubArtifacts(tmpDir))
showCache(tmpDir) # stubs are removed

#>                             artifact         tagKey
#>  1: b372b382fe0902130b2848f82546618d         format
#>  2: b372b382fe0902130b2848f82546618d           name
#>  3: b372b382fe0902130b2848f82546618d          class
#>  4: b372b382fe0902130b2848f82546618d           date
#>  5: b372b382fe0902130b2848f82546618d       function
#>  6: b372b382fe0902130b2848f82546618d    object.size
#>  7: b372b382fe0902130b2848f82546618d       accessed
#>  8: b372b382fe0902130b2848f82546618d otherFunctions
#>  9: b372b382fe0902130b2848f82546618d otherFunctions
#> 10: b372b382fe0902130b2848f82546618d otherFunctions
#> 11: b372b382fe0902130b2848f82546618d        cacheId
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:47:51
#>  2:                               Cache 2018-01-25 08:47:51
#>  3:                             numeric 2018-01-25 08:47:51
#>  4:                 2018-01-25 08:47:51 2018-01-25 08:47:51
#>  5:                                   f 2018-01-25 08:47:51
#>  6:                                 624 2018-01-25 08:47:51
#>  7:                 2018-01-25 08:47:51 2018-01-25 08:47:51
#>  8:      otherFunctions:build_reference 2018-01-25 08:47:51
#>  9: otherFunctions:data_reference_topic 2018-01-25 08:47:51
#> 10:                  otherFunctions:try 2018-01-25 08:47:51
#> 11:    7428e865f914063acaf23ecc7c817315 2018-01-25 08:47:51

# cleanup
clearCache(tmpDir)
unlink(tmpDir, recursive = TRUE)