Clear erroneous archivist artifacts

Stub artifacts can result from several causes. The most common being erroneous removal of a file in the sqlite database. This can be caused sometimes if an archive object is being saved multiple times by multiple threads. This function will clear entries in the sqlite database which have no actual file with data.

clearStubArtifacts(repoDir = NULL)

# S4 method for ANY
clearStubArtifacts(repoDir = NULL)

Arguments

repoDir
A character denoting an existing directory of the repository for which metadata will be returned. If NULL (default), it will use the repoDir specified in archivist::setLocalRepo.

Value

Invoked for its side effect on the repoDir.

Examples

tmpDir <- file.path(tempdir(), "reproducible_examples", "clearStubArtifacts")

lapply(c(runif, rnorm), function(f) {
  reproducible::Cache(f, 10, cacheRepo = tmpDir)
})

#> [[1]]
#>  [1] 0.46118646 0.31524175 0.17467589 0.53157354 0.49363702 0.77930863
#>  [7] 0.20417834 0.71339728 0.06521611 0.35420680
#> attr(,"tags")
#> [1] "cacheId:f024b63398f829eb00833e3e0d1cc753"
#> attr(,"newCache")
#> [1] TRUE
#> attr(,"call")
#> [1] ""
#> 
#> [[2]]
#>  [1]  0.9353632  0.1764886  0.2436855  1.6235489  0.1120381 -0.1339970
#>  [7] -1.9100875 -0.2792372 -0.3134460  1.0673079
#> attr(,"tags")
#> [1] "cacheId:b7eae7c361dfb6d161192321de6bc893"
#> attr(,"newCache")
#> [1] TRUE
#> attr(,"call")
#> [1] ""
#> 

# clear out any stub artifacts
showCache(tmpDir)

#> Cache size: 

#>   Total (including Rasters): 12.4 Kb

#>   Selected objects (not including Rasters): 431 bytes

#>                             artifact         tagKey
#>  1: 5423dd716caabc0841ada0514c274670         format
#>  2: 5423dd716caabc0841ada0514c274670           name
#>  3: 5423dd716caabc0841ada0514c274670          class
#>  4: 5423dd716caabc0841ada0514c274670           date
#>  5: 5423dd716caabc0841ada0514c274670        cacheId
#>  6: 5423dd716caabc0841ada0514c274670       function
#>  7: 5423dd716caabc0841ada0514c274670    object.size
#>  8: 5423dd716caabc0841ada0514c274670       accessed
#>  9: 5423dd716caabc0841ada0514c274670 otherFunctions
#> 10: 5423dd716caabc0841ada0514c274670 otherFunctions
#> 11: 5423dd716caabc0841ada0514c274670 otherFunctions
#> 12: 5423dd716caabc0841ada0514c274670 otherFunctions
#> 13: 5423dd716caabc0841ada0514c274670 otherFunctions
#> 14: 5423dd716caabc0841ada0514c274670 otherFunctions
#> 15: 5423dd716caabc0841ada0514c274670 otherFunctions
#> 16: 5423dd716caabc0841ada0514c274670      preDigest
#> 17: 5423dd716caabc0841ada0514c274670      preDigest
#> 18: c2e2fe1370e9d419a960bbccbe58fee6         format
#> 19: c2e2fe1370e9d419a960bbccbe58fee6           name
#> 20: c2e2fe1370e9d419a960bbccbe58fee6          class
#> 21: c2e2fe1370e9d419a960bbccbe58fee6           date
#> 22: c2e2fe1370e9d419a960bbccbe58fee6        cacheId
#> 23: c2e2fe1370e9d419a960bbccbe58fee6       function
#> 24: c2e2fe1370e9d419a960bbccbe58fee6    object.size
#> 25: c2e2fe1370e9d419a960bbccbe58fee6       accessed
#> 26: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 27: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 28: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 29: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 30: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 31: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 32: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 33: c2e2fe1370e9d419a960bbccbe58fee6      preDigest
#> 34: c2e2fe1370e9d419a960bbccbe58fee6      preDigest
#>                             artifact         tagKey
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-06-14 10:55:00
#>  2:                                 Cache 2018-06-14 10:55:00
#>  3:                               numeric 2018-06-14 10:55:00
#>  4:                   2018-06-14 10:55:00 2018-06-14 10:55:00
#>  5:      b7eae7c361dfb6d161192321de6bc893 2018-06-14 10:55:00
#>  6:                                     f 2018-06-14 10:55:00
#>  7:                                   784 2018-06-14 10:55:00
#>  8:                   2018-06-14 10:55:00 2018-06-14 10:55:00
#>  9:                   withCallingHandlers 2018-06-14 10:55:00
#> 10:                               saveRDS 2018-06-14 10:55:00
#> 11:                               do.call 2018-06-14 10:55:00
#> 12:                      build_site_local 2018-06-14 10:55:00
#> 13:                       build_reference 2018-06-14 10:55:00
#> 14:                  data_reference_topic 2018-06-14 10:55:00
#> 15:                                   try 2018-06-14 10:55:00
#> 16:    n:52ad08b1270ae9be9c0a12805d408433 2018-06-14 10:55:00
#> 17: .FUN:7e9a928f110f80b3612e71883a6ec1f4 2018-06-14 10:55:00
#> 18:                                   rda 2018-06-14 10:55:00
#> 19:                                 Cache 2018-06-14 10:55:00
#> 20:                               numeric 2018-06-14 10:55:00
#> 21:                   2018-06-14 10:55:00 2018-06-14 10:55:00
#> 22:      f024b63398f829eb00833e3e0d1cc753 2018-06-14 10:55:00
#> 23:                                     f 2018-06-14 10:55:00
#> 24:                                   784 2018-06-14 10:55:00
#> 25:                   2018-06-14 10:55:00 2018-06-14 10:55:00
#> 26:                   withCallingHandlers 2018-06-14 10:55:00
#> 27:                               saveRDS 2018-06-14 10:55:00
#> 28:                               do.call 2018-06-14 10:55:00
#> 29:                      build_site_local 2018-06-14 10:55:00
#> 30:                       build_reference 2018-06-14 10:55:00
#> 31:                  data_reference_topic 2018-06-14 10:55:00
#> 32:                                   try 2018-06-14 10:55:00
#> 33:    n:52ad08b1270ae9be9c0a12805d408433 2018-06-14 10:55:00
#> 34: .FUN:d2631d24c3b38b89c7bdd4ab7faaaac3 2018-06-14 10:55:00
#>                                  tagValue         createdDate

file2Remove <- dir(file.path(tmpDir, "gallery"), full.name = TRUE)[1]
file.remove(file2Remove)

#> [1] TRUE

showCache(tmpDir) # repository directory still thinks files are there

#> Cache size: 

#>   Total (including Rasters): 12.2 Kb

#>   Selected objects (not including Rasters): 207 bytes

#>                             artifact         tagKey
#>  1: 5423dd716caabc0841ada0514c274670         format
#>  2: 5423dd716caabc0841ada0514c274670           name
#>  3: 5423dd716caabc0841ada0514c274670          class
#>  4: 5423dd716caabc0841ada0514c274670           date
#>  5: 5423dd716caabc0841ada0514c274670        cacheId
#>  6: 5423dd716caabc0841ada0514c274670       function
#>  7: 5423dd716caabc0841ada0514c274670    object.size
#>  8: 5423dd716caabc0841ada0514c274670       accessed
#>  9: 5423dd716caabc0841ada0514c274670 otherFunctions
#> 10: 5423dd716caabc0841ada0514c274670 otherFunctions
#> 11: 5423dd716caabc0841ada0514c274670 otherFunctions
#> 12: 5423dd716caabc0841ada0514c274670 otherFunctions
#> 13: 5423dd716caabc0841ada0514c274670 otherFunctions
#> 14: 5423dd716caabc0841ada0514c274670 otherFunctions
#> 15: 5423dd716caabc0841ada0514c274670 otherFunctions
#> 16: 5423dd716caabc0841ada0514c274670      preDigest
#> 17: 5423dd716caabc0841ada0514c274670      preDigest
#> 18: c2e2fe1370e9d419a960bbccbe58fee6         format
#> 19: c2e2fe1370e9d419a960bbccbe58fee6           name
#> 20: c2e2fe1370e9d419a960bbccbe58fee6          class
#> 21: c2e2fe1370e9d419a960bbccbe58fee6           date
#> 22: c2e2fe1370e9d419a960bbccbe58fee6        cacheId
#> 23: c2e2fe1370e9d419a960bbccbe58fee6       function
#> 24: c2e2fe1370e9d419a960bbccbe58fee6    object.size
#> 25: c2e2fe1370e9d419a960bbccbe58fee6       accessed
#> 26: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 27: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 28: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 29: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 30: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 31: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 32: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 33: c2e2fe1370e9d419a960bbccbe58fee6      preDigest
#> 34: c2e2fe1370e9d419a960bbccbe58fee6      preDigest
#>                             artifact         tagKey
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-06-14 10:55:00
#>  2:                                 Cache 2018-06-14 10:55:00
#>  3:                               numeric 2018-06-14 10:55:00
#>  4:                   2018-06-14 10:55:00 2018-06-14 10:55:00
#>  5:      b7eae7c361dfb6d161192321de6bc893 2018-06-14 10:55:00
#>  6:                                     f 2018-06-14 10:55:00
#>  7:                                   784 2018-06-14 10:55:00
#>  8:                   2018-06-14 10:55:00 2018-06-14 10:55:00
#>  9:                   withCallingHandlers 2018-06-14 10:55:00
#> 10:                               saveRDS 2018-06-14 10:55:00
#> 11:                               do.call 2018-06-14 10:55:00
#> 12:                      build_site_local 2018-06-14 10:55:00
#> 13:                       build_reference 2018-06-14 10:55:00
#> 14:                  data_reference_topic 2018-06-14 10:55:00
#> 15:                                   try 2018-06-14 10:55:00
#> 16:    n:52ad08b1270ae9be9c0a12805d408433 2018-06-14 10:55:00
#> 17: .FUN:7e9a928f110f80b3612e71883a6ec1f4 2018-06-14 10:55:00
#> 18:                                   rda 2018-06-14 10:55:00
#> 19:                                 Cache 2018-06-14 10:55:00
#> 20:                               numeric 2018-06-14 10:55:00
#> 21:                   2018-06-14 10:55:00 2018-06-14 10:55:00
#> 22:      f024b63398f829eb00833e3e0d1cc753 2018-06-14 10:55:00
#> 23:                                     f 2018-06-14 10:55:00
#> 24:                                   784 2018-06-14 10:55:00
#> 25:                   2018-06-14 10:55:00 2018-06-14 10:55:00
#> 26:                   withCallingHandlers 2018-06-14 10:55:00
#> 27:                               saveRDS 2018-06-14 10:55:00
#> 28:                               do.call 2018-06-14 10:55:00
#> 29:                      build_site_local 2018-06-14 10:55:00
#> 30:                       build_reference 2018-06-14 10:55:00
#> 31:                  data_reference_topic 2018-06-14 10:55:00
#> 32:                                   try 2018-06-14 10:55:00
#> 33:    n:52ad08b1270ae9be9c0a12805d408433 2018-06-14 10:55:00
#> 34: .FUN:d2631d24c3b38b89c7bdd4ab7faaaac3 2018-06-14 10:55:00
#>                                  tagValue         createdDate

# run clearStubArtifacts
suppressWarnings(clearStubArtifacts(tmpDir))
showCache(tmpDir) # stubs are removed

#> Cache size: 

#>   Total (including Rasters): 12.2 Kb

#>   Selected objects (not including Rasters): 207 bytes

#>                             artifact         tagKey
#>  1: c2e2fe1370e9d419a960bbccbe58fee6         format
#>  2: c2e2fe1370e9d419a960bbccbe58fee6           name
#>  3: c2e2fe1370e9d419a960bbccbe58fee6          class
#>  4: c2e2fe1370e9d419a960bbccbe58fee6           date
#>  5: c2e2fe1370e9d419a960bbccbe58fee6        cacheId
#>  6: c2e2fe1370e9d419a960bbccbe58fee6       function
#>  7: c2e2fe1370e9d419a960bbccbe58fee6    object.size
#>  8: c2e2fe1370e9d419a960bbccbe58fee6       accessed
#>  9: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 10: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 11: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 12: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 13: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 14: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 15: c2e2fe1370e9d419a960bbccbe58fee6 otherFunctions
#> 16: c2e2fe1370e9d419a960bbccbe58fee6      preDigest
#> 17: c2e2fe1370e9d419a960bbccbe58fee6      preDigest
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-06-14 10:55:00
#>  2:                                 Cache 2018-06-14 10:55:00
#>  3:                               numeric 2018-06-14 10:55:00
#>  4:                   2018-06-14 10:55:00 2018-06-14 10:55:00
#>  5:      f024b63398f829eb00833e3e0d1cc753 2018-06-14 10:55:00
#>  6:                                     f 2018-06-14 10:55:00
#>  7:                                   784 2018-06-14 10:55:00
#>  8:                   2018-06-14 10:55:00 2018-06-14 10:55:00
#>  9:                   withCallingHandlers 2018-06-14 10:55:00
#> 10:                               saveRDS 2018-06-14 10:55:00
#> 11:                               do.call 2018-06-14 10:55:00
#> 12:                      build_site_local 2018-06-14 10:55:00
#> 13:                       build_reference 2018-06-14 10:55:00
#> 14:                  data_reference_topic 2018-06-14 10:55:00
#> 15:                                   try 2018-06-14 10:55:00
#> 16:    n:52ad08b1270ae9be9c0a12805d408433 2018-06-14 10:55:00
#> 17: .FUN:d2631d24c3b38b89c7bdd4ab7faaaac3 2018-06-14 10:55:00

# cleanup
clearCache(tmpDir)
unlink(tmpDir, recursive = TRUE)