Clear erroneous archivist artifacts

Stub artifacts can result from several causes. The most common being erroneous removal of a file in the SQLite database. This can be caused sometimes if an archive object is being saved multiple times by multiple threads. This function will clear entries in the SQLite database which have no actual file with data.

clearStubArtifacts(repoDir = NULL)

# S4 method for ANY
clearStubArtifacts(repoDir = NULL)

Arguments

repoDir
A character denoting an existing directory of the repository for which metadata will be returned. If NULL (default), it will use the repoDir specified in archivist::setLocalRepo.

Value

Invoked for its side effect on the repoDir.

Examples

tmpDir <- file.path(tempdir(), "reproducible_examples", "clearStubArtifacts")

lapply(c(runif, rnorm), function(f) {
  reproducible::Cache(f, 10, cacheRepo = tmpDir)
})

#> [[1]]
#>  [1] 0.1375330 0.8478029 0.4852220 0.4702427 0.7910454 0.3129833 0.3087645
#>  [8] 0.3451553 0.8284433 0.2739955
#> attr(,"tags")
#> [1] "cacheId:f024b63398f829eb00833e3e0d1cc753"
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"call")
#> [1] ""
#> 
#> [[2]]
#>  [1] -0.1742460 -1.1062360 -0.9459850  0.2890896  0.8769131 -1.1489039
#>  [7] -1.1376128 -1.4372467 -0.4941435  0.8408018
#> attr(,"tags")
#> [1] "cacheId:b7eae7c361dfb6d161192321de6bc893"
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"call")
#> [1] ""
#> 

# clear out any stub artifacts
showCache(tmpDir)

#> Cache size: 

#>   Total (including Rasters): 524 bytes

#>   Selected objects (not including Rasters): 524 bytes

#>                             artifact         tagKey
#>  1: 26b7aa9fd99afcfd0f44f4b32952b7ce         format
#>  2: 26b7aa9fd99afcfd0f44f4b32952b7ce           name
#>  3: 26b7aa9fd99afcfd0f44f4b32952b7ce          class
#>  4: 26b7aa9fd99afcfd0f44f4b32952b7ce           date
#>  5: 26b7aa9fd99afcfd0f44f4b32952b7ce        cacheId
#>  6: 26b7aa9fd99afcfd0f44f4b32952b7ce       function
#>  7: 26b7aa9fd99afcfd0f44f4b32952b7ce    object.size
#>  8: 26b7aa9fd99afcfd0f44f4b32952b7ce       accessed
#>  9: 26b7aa9fd99afcfd0f44f4b32952b7ce otherFunctions
#> 10: 26b7aa9fd99afcfd0f44f4b32952b7ce otherFunctions
#> 11: 26b7aa9fd99afcfd0f44f4b32952b7ce otherFunctions
#> 12: 26b7aa9fd99afcfd0f44f4b32952b7ce otherFunctions
#> 13: 26b7aa9fd99afcfd0f44f4b32952b7ce otherFunctions
#> 14: 26b7aa9fd99afcfd0f44f4b32952b7ce otherFunctions
#> 15: 26b7aa9fd99afcfd0f44f4b32952b7ce otherFunctions
#> 16: 26b7aa9fd99afcfd0f44f4b32952b7ce      preDigest
#> 17: 26b7aa9fd99afcfd0f44f4b32952b7ce      preDigest
#> 18: c033af4ea722076351fbb49a2f3ae5da         format
#> 19: c033af4ea722076351fbb49a2f3ae5da           name
#> 20: c033af4ea722076351fbb49a2f3ae5da          class
#> 21: c033af4ea722076351fbb49a2f3ae5da           date
#> 22: c033af4ea722076351fbb49a2f3ae5da        cacheId
#> 23: c033af4ea722076351fbb49a2f3ae5da       function
#> 24: c033af4ea722076351fbb49a2f3ae5da    object.size
#> 25: c033af4ea722076351fbb49a2f3ae5da       accessed
#> 26: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 27: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 28: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 29: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 30: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 31: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 32: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 33: c033af4ea722076351fbb49a2f3ae5da      preDigest
#> 34: c033af4ea722076351fbb49a2f3ae5da      preDigest
#>                             artifact         tagKey
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-07-23 12:02:09
#>  2:      26b7aa9fd99afcfd0f44f4b32952b7ce 2018-07-23 12:02:09
#>  3:                               numeric 2018-07-23 12:02:09
#>  4:                   2018-07-23 12:02:09 2018-07-23 12:02:09
#>  5:      f024b63398f829eb00833e3e0d1cc753 2018-07-23 12:02:09
#>  6:                                     f 2018-07-23 12:02:09
#>  7:                                  1048 2018-07-23 12:02:09
#>  8:                   2018-07-23 12:02:09 2018-07-23 12:02:09
#>  9:                   withCallingHandlers 2018-07-23 12:02:09
#> 10:                               saveRDS 2018-07-23 12:02:09
#> 11:                               do.call 2018-07-23 12:02:09
#> 12:                      build_site_local 2018-07-23 12:02:09
#> 13:                       build_reference 2018-07-23 12:02:09
#> 14:                  data_reference_topic 2018-07-23 12:02:09
#> 15:                                   try 2018-07-23 12:02:09
#> 16:    n:52ad08b1270ae9be9c0a12805d408433 2018-07-23 12:02:09
#> 17: .FUN:d2631d24c3b38b89c7bdd4ab7faaaac3 2018-07-23 12:02:09
#> 18:                                   rda 2018-07-23 12:02:09
#> 19:      c033af4ea722076351fbb49a2f3ae5da 2018-07-23 12:02:09
#> 20:                               numeric 2018-07-23 12:02:09
#> 21:                   2018-07-23 12:02:09 2018-07-23 12:02:09
#> 22:      b7eae7c361dfb6d161192321de6bc893 2018-07-23 12:02:09
#> 23:                                     f 2018-07-23 12:02:09
#> 24:                                  1048 2018-07-23 12:02:09
#> 25:                   2018-07-23 12:02:09 2018-07-23 12:02:09
#> 26:                   withCallingHandlers 2018-07-23 12:02:09
#> 27:                               saveRDS 2018-07-23 12:02:09
#> 28:                               do.call 2018-07-23 12:02:09
#> 29:                      build_site_local 2018-07-23 12:02:09
#> 30:                       build_reference 2018-07-23 12:02:09
#> 31:                  data_reference_topic 2018-07-23 12:02:09
#> 32:                                   try 2018-07-23 12:02:09
#> 33:    n:52ad08b1270ae9be9c0a12805d408433 2018-07-23 12:02:09
#> 34: .FUN:7e9a928f110f80b3612e71883a6ec1f4 2018-07-23 12:02:09
#>                                  tagValue         createdDate

file2Remove <- dir(file.path(tmpDir, "gallery"), full.name = TRUE)[1]
file.remove(file2Remove)

#> [1] TRUE

showCache(tmpDir) # repository directory still thinks files are there

#> Cache size: 

#>   Total (including Rasters): 524 bytes

#>   Selected objects (not including Rasters): 524 bytes

#>                             artifact         tagKey
#>  1: 26b7aa9fd99afcfd0f44f4b32952b7ce         format
#>  2: 26b7aa9fd99afcfd0f44f4b32952b7ce           name
#>  3: 26b7aa9fd99afcfd0f44f4b32952b7ce          class
#>  4: 26b7aa9fd99afcfd0f44f4b32952b7ce           date
#>  5: 26b7aa9fd99afcfd0f44f4b32952b7ce        cacheId
#>  6: 26b7aa9fd99afcfd0f44f4b32952b7ce       function
#>  7: 26b7aa9fd99afcfd0f44f4b32952b7ce    object.size
#>  8: 26b7aa9fd99afcfd0f44f4b32952b7ce       accessed
#>  9: 26b7aa9fd99afcfd0f44f4b32952b7ce otherFunctions
#> 10: 26b7aa9fd99afcfd0f44f4b32952b7ce otherFunctions
#> 11: 26b7aa9fd99afcfd0f44f4b32952b7ce otherFunctions
#> 12: 26b7aa9fd99afcfd0f44f4b32952b7ce otherFunctions
#> 13: 26b7aa9fd99afcfd0f44f4b32952b7ce otherFunctions
#> 14: 26b7aa9fd99afcfd0f44f4b32952b7ce otherFunctions
#> 15: 26b7aa9fd99afcfd0f44f4b32952b7ce otherFunctions
#> 16: 26b7aa9fd99afcfd0f44f4b32952b7ce      preDigest
#> 17: 26b7aa9fd99afcfd0f44f4b32952b7ce      preDigest
#> 18: c033af4ea722076351fbb49a2f3ae5da         format
#> 19: c033af4ea722076351fbb49a2f3ae5da           name
#> 20: c033af4ea722076351fbb49a2f3ae5da          class
#> 21: c033af4ea722076351fbb49a2f3ae5da           date
#> 22: c033af4ea722076351fbb49a2f3ae5da        cacheId
#> 23: c033af4ea722076351fbb49a2f3ae5da       function
#> 24: c033af4ea722076351fbb49a2f3ae5da    object.size
#> 25: c033af4ea722076351fbb49a2f3ae5da       accessed
#> 26: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 27: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 28: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 29: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 30: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 31: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 32: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 33: c033af4ea722076351fbb49a2f3ae5da      preDigest
#> 34: c033af4ea722076351fbb49a2f3ae5da      preDigest
#>                             artifact         tagKey
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-07-23 12:02:09
#>  2:      26b7aa9fd99afcfd0f44f4b32952b7ce 2018-07-23 12:02:09
#>  3:                               numeric 2018-07-23 12:02:09
#>  4:                   2018-07-23 12:02:09 2018-07-23 12:02:09
#>  5:      f024b63398f829eb00833e3e0d1cc753 2018-07-23 12:02:09
#>  6:                                     f 2018-07-23 12:02:09
#>  7:                                  1048 2018-07-23 12:02:09
#>  8:                   2018-07-23 12:02:09 2018-07-23 12:02:09
#>  9:                   withCallingHandlers 2018-07-23 12:02:09
#> 10:                               saveRDS 2018-07-23 12:02:09
#> 11:                               do.call 2018-07-23 12:02:09
#> 12:                      build_site_local 2018-07-23 12:02:09
#> 13:                       build_reference 2018-07-23 12:02:09
#> 14:                  data_reference_topic 2018-07-23 12:02:09
#> 15:                                   try 2018-07-23 12:02:09
#> 16:    n:52ad08b1270ae9be9c0a12805d408433 2018-07-23 12:02:09
#> 17: .FUN:d2631d24c3b38b89c7bdd4ab7faaaac3 2018-07-23 12:02:09
#> 18:                                   rda 2018-07-23 12:02:09
#> 19:      c033af4ea722076351fbb49a2f3ae5da 2018-07-23 12:02:09
#> 20:                               numeric 2018-07-23 12:02:09
#> 21:                   2018-07-23 12:02:09 2018-07-23 12:02:09
#> 22:      b7eae7c361dfb6d161192321de6bc893 2018-07-23 12:02:09
#> 23:                                     f 2018-07-23 12:02:09
#> 24:                                  1048 2018-07-23 12:02:09
#> 25:                   2018-07-23 12:02:09 2018-07-23 12:02:09
#> 26:                   withCallingHandlers 2018-07-23 12:02:09
#> 27:                               saveRDS 2018-07-23 12:02:09
#> 28:                               do.call 2018-07-23 12:02:09
#> 29:                      build_site_local 2018-07-23 12:02:09
#> 30:                       build_reference 2018-07-23 12:02:09
#> 31:                  data_reference_topic 2018-07-23 12:02:09
#> 32:                                   try 2018-07-23 12:02:09
#> 33:    n:52ad08b1270ae9be9c0a12805d408433 2018-07-23 12:02:09
#> 34: .FUN:7e9a928f110f80b3612e71883a6ec1f4 2018-07-23 12:02:09
#>                                  tagValue         createdDate

# run clearStubArtifacts
suppressWarnings(clearStubArtifacts(tmpDir))
showCache(tmpDir) # stubs are removed

#> Cache size: 

#>   Total (including Rasters): 262 bytes

#>   Selected objects (not including Rasters): 262 bytes

#>                             artifact         tagKey
#>  1: c033af4ea722076351fbb49a2f3ae5da         format
#>  2: c033af4ea722076351fbb49a2f3ae5da           name
#>  3: c033af4ea722076351fbb49a2f3ae5da          class
#>  4: c033af4ea722076351fbb49a2f3ae5da           date
#>  5: c033af4ea722076351fbb49a2f3ae5da        cacheId
#>  6: c033af4ea722076351fbb49a2f3ae5da       function
#>  7: c033af4ea722076351fbb49a2f3ae5da    object.size
#>  8: c033af4ea722076351fbb49a2f3ae5da       accessed
#>  9: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 10: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 11: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 12: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 13: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 14: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 15: c033af4ea722076351fbb49a2f3ae5da otherFunctions
#> 16: c033af4ea722076351fbb49a2f3ae5da      preDigest
#> 17: c033af4ea722076351fbb49a2f3ae5da      preDigest
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-07-23 12:02:09
#>  2:      c033af4ea722076351fbb49a2f3ae5da 2018-07-23 12:02:09
#>  3:                               numeric 2018-07-23 12:02:09
#>  4:                   2018-07-23 12:02:09 2018-07-23 12:02:09
#>  5:      b7eae7c361dfb6d161192321de6bc893 2018-07-23 12:02:09
#>  6:                                     f 2018-07-23 12:02:09
#>  7:                                  1048 2018-07-23 12:02:09
#>  8:                   2018-07-23 12:02:09 2018-07-23 12:02:09
#>  9:                   withCallingHandlers 2018-07-23 12:02:09
#> 10:                               saveRDS 2018-07-23 12:02:09
#> 11:                               do.call 2018-07-23 12:02:09
#> 12:                      build_site_local 2018-07-23 12:02:09
#> 13:                       build_reference 2018-07-23 12:02:09
#> 14:                  data_reference_topic 2018-07-23 12:02:09
#> 15:                                   try 2018-07-23 12:02:09
#> 16:    n:52ad08b1270ae9be9c0a12805d408433 2018-07-23 12:02:09
#> 17: .FUN:7e9a928f110f80b3612e71883a6ec1f4 2018-07-23 12:02:09

# cleanup
clearCache(tmpDir)

#> Cache size: 

#>   Total (including Rasters): 262 bytes

#>   Selected objects (not including Rasters): 262 bytes

unlink(tmpDir, recursive = TRUE)