Clear erroneous archivist artifacts

clearStubArtifacts(repoDir = NULL)

# S4 method for ANY
clearStubArtifacts(repoDir = NULL)

Arguments

repoDir

A character denoting an existing directory of the repository for which metadata will be returned. If NULL (default), it will use the repoDir specified in archivist::setLocalRepo.

Value

Invoked for its side effect on the repoDir.

Details

Stub artifacts can result from several causes. The most common being erroneous removal of a file in the SQLite database. This can be caused sometimes if an archive object is being saved multiple times by multiple threads. This function will clear entries in the SQLite database which have no actual file with data.

Author

Eliot McIntire

Examples

tmpDir <- file.path(tempdir(), "reproducible_examples", "clearStubArtifacts")

lapply(c(runif, rnorm), function(f) {
  reproducible::Cache(f, 10, cacheRepo = tmpDir)
})
#> [[1]]
#>  [1] 0.6221827 0.7119298 0.3585995 0.1272161 0.2255348 0.2910905 0.1378617
#>  [8] 0.7698906 0.5060579 0.4986368
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:88dabbb6fbb67941"
#> attr(,"call")
#> [1] ""
#> 
#> [[2]]
#>  [1] -1.969754450 -0.927111409  1.197422463 -0.769678593 -0.002123372
#>  [6]  0.004642366 -0.436383681 -1.292466880 -0.818865742 -0.870473932
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:87eee6212ba1ae0a"
#> attr(,"call")
#> [1] ""
#> 

# clear out any stub artifacts
showCache(tmpDir)
#> Cache size: 
#>   Total (including Rasters): 552 bytes
#>   Selected objects (not including Rasters): 552 bytes
#>              cacheId              tagKey              tagValue
#>  1: 88dabbb6fbb67941            function                     f
#>  2: 88dabbb6fbb67941               class               numeric
#>  3: 88dabbb6fbb67941         object.size                  1104
#>  4: 88dabbb6fbb67941            accessed   2022-08-16 16:04:17
#>  5: 88dabbb6fbb67941             inCloud                 FALSE
#>  6: 88dabbb6fbb67941          resultHash                      
#>  7: 88dabbb6fbb67941   elapsedTimeDigest     0.0006113052 secs
#>  8: 88dabbb6fbb67941 elapsedTimeFirstRun     3.361702e-05 secs
#>  9: 88dabbb6fbb67941      otherFunctions            build_site
#> 10: 88dabbb6fbb67941      otherFunctions      build_site_local
#> 11: 88dabbb6fbb67941      otherFunctions       build_reference
#> 12: 88dabbb6fbb67941      otherFunctions  data_reference_topic
#> 13: 88dabbb6fbb67941      otherFunctions          run_examples
#> 14: 88dabbb6fbb67941      otherFunctions    highlight_examples
#> 15: 88dabbb6fbb67941      otherFunctions             timing_fn
#> 16: 88dabbb6fbb67941      otherFunctions                handle
#> 17: 88dabbb6fbb67941      otherFunctions                   try
#> 18: 88dabbb6fbb67941      otherFunctions                lapply
#> 19: 88dabbb6fbb67941           preDigest    n:c5775c3b366fb719
#> 20: 88dabbb6fbb67941           preDigest .FUN:881ec847b7161f3c
#> 21: 88dabbb6fbb67941           file.size                   208
#> 22: 87eee6212ba1ae0a            function                     f
#> 23: 87eee6212ba1ae0a               class               numeric
#> 24: 87eee6212ba1ae0a         object.size                  1104
#> 25: 87eee6212ba1ae0a            accessed   2022-08-16 16:04:17
#> 26: 87eee6212ba1ae0a             inCloud                 FALSE
#> 27: 87eee6212ba1ae0a          resultHash                      
#> 28: 87eee6212ba1ae0a   elapsedTimeDigest      0.000371933 secs
#> 29: 87eee6212ba1ae0a elapsedTimeFirstRun     3.218651e-05 secs
#> 30: 87eee6212ba1ae0a      otherFunctions            build_site
#> 31: 87eee6212ba1ae0a      otherFunctions      build_site_local
#> 32: 87eee6212ba1ae0a      otherFunctions       build_reference
#> 33: 87eee6212ba1ae0a      otherFunctions  data_reference_topic
#> 34: 87eee6212ba1ae0a      otherFunctions          run_examples
#> 35: 87eee6212ba1ae0a      otherFunctions    highlight_examples
#> 36: 87eee6212ba1ae0a      otherFunctions             timing_fn
#> 37: 87eee6212ba1ae0a      otherFunctions                handle
#> 38: 87eee6212ba1ae0a      otherFunctions                   try
#> 39: 87eee6212ba1ae0a      otherFunctions                lapply
#> 40: 87eee6212ba1ae0a           preDigest    n:c5775c3b366fb719
#> 41: 87eee6212ba1ae0a           preDigest .FUN:4f604aa46882b368
#> 42: 87eee6212ba1ae0a           file.size                   226
#>              cacheId              tagKey              tagValue
#>             createdDate
#>  1: 2022-08-16 16:04:17
#>  2: 2022-08-16 16:04:17
#>  3: 2022-08-16 16:04:17
#>  4: 2022-08-16 16:04:17
#>  5: 2022-08-16 16:04:17
#>  6: 2022-08-16 16:04:17
#>  7: 2022-08-16 16:04:17
#>  8: 2022-08-16 16:04:17
#>  9: 2022-08-16 16:04:17
#> 10: 2022-08-16 16:04:17
#> 11: 2022-08-16 16:04:17
#> 12: 2022-08-16 16:04:17
#> 13: 2022-08-16 16:04:17
#> 14: 2022-08-16 16:04:17
#> 15: 2022-08-16 16:04:17
#> 16: 2022-08-16 16:04:17
#> 17: 2022-08-16 16:04:17
#> 18: 2022-08-16 16:04:17
#> 19: 2022-08-16 16:04:17
#> 20: 2022-08-16 16:04:17
#> 21: 2022-08-16 16:04:17
#> 22: 2022-08-16 16:04:17
#> 23: 2022-08-16 16:04:17
#> 24: 2022-08-16 16:04:17
#> 25: 2022-08-16 16:04:17
#> 26: 2022-08-16 16:04:17
#> 27: 2022-08-16 16:04:17
#> 28: 2022-08-16 16:04:17
#> 29: 2022-08-16 16:04:17
#> 30: 2022-08-16 16:04:17
#> 31: 2022-08-16 16:04:17
#> 32: 2022-08-16 16:04:17
#> 33: 2022-08-16 16:04:17
#> 34: 2022-08-16 16:04:17
#> 35: 2022-08-16 16:04:17
#> 36: 2022-08-16 16:04:17
#> 37: 2022-08-16 16:04:17
#> 38: 2022-08-16 16:04:17
#> 39: 2022-08-16 16:04:17
#> 40: 2022-08-16 16:04:17
#> 41: 2022-08-16 16:04:17
#> 42: 2022-08-16 16:04:17
#>             createdDate

file2Remove <- dir(CacheStorageDir(tmpDir), full.name = TRUE)[1]
file.remove(file2Remove)
#> [1] TRUE
showCache(tmpDir) # repository directory still thinks files are there
#> Cache size: 
#>   Total (including Rasters): 552 bytes
#>   Selected objects (not including Rasters): 552 bytes
#>              cacheId              tagKey              tagValue
#>  1: 88dabbb6fbb67941            function                     f
#>  2: 88dabbb6fbb67941               class               numeric
#>  3: 88dabbb6fbb67941         object.size                  1104
#>  4: 88dabbb6fbb67941            accessed   2022-08-16 16:04:17
#>  5: 88dabbb6fbb67941             inCloud                 FALSE
#>  6: 88dabbb6fbb67941          resultHash                      
#>  7: 88dabbb6fbb67941   elapsedTimeDigest     0.0006113052 secs
#>  8: 88dabbb6fbb67941 elapsedTimeFirstRun     3.361702e-05 secs
#>  9: 88dabbb6fbb67941      otherFunctions            build_site
#> 10: 88dabbb6fbb67941      otherFunctions      build_site_local
#> 11: 88dabbb6fbb67941      otherFunctions       build_reference
#> 12: 88dabbb6fbb67941      otherFunctions  data_reference_topic
#> 13: 88dabbb6fbb67941      otherFunctions          run_examples
#> 14: 88dabbb6fbb67941      otherFunctions    highlight_examples
#> 15: 88dabbb6fbb67941      otherFunctions             timing_fn
#> 16: 88dabbb6fbb67941      otherFunctions                handle
#> 17: 88dabbb6fbb67941      otherFunctions                   try
#> 18: 88dabbb6fbb67941      otherFunctions                lapply
#> 19: 88dabbb6fbb67941           preDigest    n:c5775c3b366fb719
#> 20: 88dabbb6fbb67941           preDigest .FUN:881ec847b7161f3c
#> 21: 88dabbb6fbb67941           file.size                   208
#> 22: 87eee6212ba1ae0a            function                     f
#> 23: 87eee6212ba1ae0a               class               numeric
#> 24: 87eee6212ba1ae0a         object.size                  1104
#> 25: 87eee6212ba1ae0a            accessed   2022-08-16 16:04:17
#> 26: 87eee6212ba1ae0a             inCloud                 FALSE
#> 27: 87eee6212ba1ae0a          resultHash                      
#> 28: 87eee6212ba1ae0a   elapsedTimeDigest      0.000371933 secs
#> 29: 87eee6212ba1ae0a elapsedTimeFirstRun     3.218651e-05 secs
#> 30: 87eee6212ba1ae0a      otherFunctions            build_site
#> 31: 87eee6212ba1ae0a      otherFunctions      build_site_local
#> 32: 87eee6212ba1ae0a      otherFunctions       build_reference
#> 33: 87eee6212ba1ae0a      otherFunctions  data_reference_topic
#> 34: 87eee6212ba1ae0a      otherFunctions          run_examples
#> 35: 87eee6212ba1ae0a      otherFunctions    highlight_examples
#> 36: 87eee6212ba1ae0a      otherFunctions             timing_fn
#> 37: 87eee6212ba1ae0a      otherFunctions                handle
#> 38: 87eee6212ba1ae0a      otherFunctions                   try
#> 39: 87eee6212ba1ae0a      otherFunctions                lapply
#> 40: 87eee6212ba1ae0a           preDigest    n:c5775c3b366fb719
#> 41: 87eee6212ba1ae0a           preDigest .FUN:4f604aa46882b368
#> 42: 87eee6212ba1ae0a           file.size                   226
#>              cacheId              tagKey              tagValue
#>             createdDate
#>  1: 2022-08-16 16:04:17
#>  2: 2022-08-16 16:04:17
#>  3: 2022-08-16 16:04:17
#>  4: 2022-08-16 16:04:17
#>  5: 2022-08-16 16:04:17
#>  6: 2022-08-16 16:04:17
#>  7: 2022-08-16 16:04:17
#>  8: 2022-08-16 16:04:17
#>  9: 2022-08-16 16:04:17
#> 10: 2022-08-16 16:04:17
#> 11: 2022-08-16 16:04:17
#> 12: 2022-08-16 16:04:17
#> 13: 2022-08-16 16:04:17
#> 14: 2022-08-16 16:04:17
#> 15: 2022-08-16 16:04:17
#> 16: 2022-08-16 16:04:17
#> 17: 2022-08-16 16:04:17
#> 18: 2022-08-16 16:04:17
#> 19: 2022-08-16 16:04:17
#> 20: 2022-08-16 16:04:17
#> 21: 2022-08-16 16:04:17
#> 22: 2022-08-16 16:04:17
#> 23: 2022-08-16 16:04:17
#> 24: 2022-08-16 16:04:17
#> 25: 2022-08-16 16:04:17
#> 26: 2022-08-16 16:04:17
#> 27: 2022-08-16 16:04:17
#> 28: 2022-08-16 16:04:17
#> 29: 2022-08-16 16:04:17
#> 30: 2022-08-16 16:04:17
#> 31: 2022-08-16 16:04:17
#> 32: 2022-08-16 16:04:17
#> 33: 2022-08-16 16:04:17
#> 34: 2022-08-16 16:04:17
#> 35: 2022-08-16 16:04:17
#> 36: 2022-08-16 16:04:17
#> 37: 2022-08-16 16:04:17
#> 38: 2022-08-16 16:04:17
#> 39: 2022-08-16 16:04:17
#> 40: 2022-08-16 16:04:17
#> 41: 2022-08-16 16:04:17
#> 42: 2022-08-16 16:04:17
#>             createdDate

# run clearStubArtifacts
suppressWarnings(clearStubArtifacts(tmpDir))
showCache(tmpDir) # stubs are removed
#> Cache size: 
#>   Total (including Rasters): 552 bytes
#>   Selected objects (not including Rasters): 552 bytes
#>              cacheId              tagKey              tagValue
#>  1: 88dabbb6fbb67941            function                     f
#>  2: 88dabbb6fbb67941               class               numeric
#>  3: 88dabbb6fbb67941         object.size                  1104
#>  4: 88dabbb6fbb67941            accessed   2022-08-16 16:04:17
#>  5: 88dabbb6fbb67941             inCloud                 FALSE
#>  6: 88dabbb6fbb67941          resultHash                      
#>  7: 88dabbb6fbb67941   elapsedTimeDigest     0.0006113052 secs
#>  8: 88dabbb6fbb67941 elapsedTimeFirstRun     3.361702e-05 secs
#>  9: 88dabbb6fbb67941      otherFunctions            build_site
#> 10: 88dabbb6fbb67941      otherFunctions      build_site_local
#> 11: 88dabbb6fbb67941      otherFunctions       build_reference
#> 12: 88dabbb6fbb67941      otherFunctions  data_reference_topic
#> 13: 88dabbb6fbb67941      otherFunctions          run_examples
#> 14: 88dabbb6fbb67941      otherFunctions    highlight_examples
#> 15: 88dabbb6fbb67941      otherFunctions             timing_fn
#> 16: 88dabbb6fbb67941      otherFunctions                handle
#> 17: 88dabbb6fbb67941      otherFunctions                   try
#> 18: 88dabbb6fbb67941      otherFunctions                lapply
#> 19: 88dabbb6fbb67941           preDigest    n:c5775c3b366fb719
#> 20: 88dabbb6fbb67941           preDigest .FUN:881ec847b7161f3c
#> 21: 88dabbb6fbb67941           file.size                   208
#> 22: 87eee6212ba1ae0a            function                     f
#> 23: 87eee6212ba1ae0a               class               numeric
#> 24: 87eee6212ba1ae0a         object.size                  1104
#> 25: 87eee6212ba1ae0a            accessed   2022-08-16 16:04:17
#> 26: 87eee6212ba1ae0a             inCloud                 FALSE
#> 27: 87eee6212ba1ae0a          resultHash                      
#> 28: 87eee6212ba1ae0a   elapsedTimeDigest      0.000371933 secs
#> 29: 87eee6212ba1ae0a elapsedTimeFirstRun     3.218651e-05 secs
#> 30: 87eee6212ba1ae0a      otherFunctions            build_site
#> 31: 87eee6212ba1ae0a      otherFunctions      build_site_local
#> 32: 87eee6212ba1ae0a      otherFunctions       build_reference
#> 33: 87eee6212ba1ae0a      otherFunctions  data_reference_topic
#> 34: 87eee6212ba1ae0a      otherFunctions          run_examples
#> 35: 87eee6212ba1ae0a      otherFunctions    highlight_examples
#> 36: 87eee6212ba1ae0a      otherFunctions             timing_fn
#> 37: 87eee6212ba1ae0a      otherFunctions                handle
#> 38: 87eee6212ba1ae0a      otherFunctions                   try
#> 39: 87eee6212ba1ae0a      otherFunctions                lapply
#> 40: 87eee6212ba1ae0a           preDigest    n:c5775c3b366fb719
#> 41: 87eee6212ba1ae0a           preDigest .FUN:4f604aa46882b368
#> 42: 87eee6212ba1ae0a           file.size                   226
#>              cacheId              tagKey              tagValue
#>             createdDate
#>  1: 2022-08-16 16:04:17
#>  2: 2022-08-16 16:04:17
#>  3: 2022-08-16 16:04:17
#>  4: 2022-08-16 16:04:17
#>  5: 2022-08-16 16:04:17
#>  6: 2022-08-16 16:04:17
#>  7: 2022-08-16 16:04:17
#>  8: 2022-08-16 16:04:17
#>  9: 2022-08-16 16:04:17
#> 10: 2022-08-16 16:04:17
#> 11: 2022-08-16 16:04:17
#> 12: 2022-08-16 16:04:17
#> 13: 2022-08-16 16:04:17
#> 14: 2022-08-16 16:04:17
#> 15: 2022-08-16 16:04:17
#> 16: 2022-08-16 16:04:17
#> 17: 2022-08-16 16:04:17
#> 18: 2022-08-16 16:04:17
#> 19: 2022-08-16 16:04:17
#> 20: 2022-08-16 16:04:17
#> 21: 2022-08-16 16:04:17
#> 22: 2022-08-16 16:04:17
#> 23: 2022-08-16 16:04:17
#> 24: 2022-08-16 16:04:17
#> 25: 2022-08-16 16:04:17
#> 26: 2022-08-16 16:04:17
#> 27: 2022-08-16 16:04:17
#> 28: 2022-08-16 16:04:17
#> 29: 2022-08-16 16:04:17
#> 30: 2022-08-16 16:04:17
#> 31: 2022-08-16 16:04:17
#> 32: 2022-08-16 16:04:17
#> 33: 2022-08-16 16:04:17
#> 34: 2022-08-16 16:04:17
#> 35: 2022-08-16 16:04:17
#> 36: 2022-08-16 16:04:17
#> 37: 2022-08-16 16:04:17
#> 38: 2022-08-16 16:04:17
#> 39: 2022-08-16 16:04:17
#> 40: 2022-08-16 16:04:17
#> 41: 2022-08-16 16:04:17
#> 42: 2022-08-16 16:04:17
#>             createdDate

# cleanup
clearCache(tmpDir, ask = FALSE)
unlink(tmpDir, recursive = TRUE)