Convert all ways of calling a function into canonical form, including defaults

e.g., stats::rnorm(1) –> rnorm(n = 1, mean = 0, sd = 1)

convertCallToCommonFormat(call, usesDots, isSquiggly, .callingEnv)

Arguments

call

The full captured call as it was passed by user.

usesDots

Logical. Whether the original Cache call used ...

isSquiggly

Logical. Whether there are curly braces e.g., as in a pipe sequence.

.callingEnv

Environment. The environment from which Cache was called.