Merge two cache repositories together

mergeCache(
  cacheTo,
  cacheFrom,
  drvTo = getDrv(getOption("reproducible.drv", NULL)),
  drvFrom = getDrv(getOption("reproducible.drv", NULL)),
  connTo = NULL,
  connFrom = NULL,
  verbose = getOption("reproducible.verbose")
)

# S4 method for ANY
mergeCache(
  cacheTo,
  cacheFrom,
  drvTo = getDrv(getOption("reproducible.drv", NULL)),
  drvFrom = getDrv(getOption("reproducible.drv", NULL)),
  connTo = NULL,
  connFrom = NULL,
  verbose = getOption("reproducible.verbose")
)

Arguments

cacheTo

The cache repository (character string of the file path) that will become larger, i.e., merge into this

cacheFrom

The cache repository (character string of the file path) from which all objects will be taken and copied from

drvTo

The database driver for the cacheTo.

drvFrom

The database driver for the cacheFrom

connTo

The connection for the cacheTo. If not provided, then a new one will be made from drvTo and cacheTo

connFrom

The database for the cacheFrom. If not provided, then a new one will be made from drvFrom and cacheFrom

verbose

Numeric, -1 silent (where possible), 0 being very quiet, 1 showing more messaging, 2 being more messaging, etc. Default is 1. Above 3 will output much more information about the internals of Caching, which may help diagnose Caching challenges. Can set globally with an option, e.g., options('reproducible.verbose' = 0) to reduce to minimal

Value

The character string of the path of cacheTo, i.e., not the objects themselves.

Details

All the cacheFrom artifacts will be put into cacheTo repository. All userTags will be copied verbatim, including accessed, with 1 exception: date will be the current Sys.time() at the time of merging. The createdDate column will be similarly the current time of merging.