Recursive object.size

This has methods for various types of things that may not correctly report their object size using object.size. Also, for lists and environments, it will return the object size separately for each element. These are estimates only, and could be inaccurate. Alternative, similar functions include object.size and pryr::object_size. See Details for the special case of functions and their enclosing environments.

objSize(x, quick, enclosingEnvs, .prevEnvirs, ...)

# S3 method for default
objSize(
  x,
  quick = getOption("reproducible.quick", FALSE),
  enclosingEnvs = TRUE,
  .prevEnvirs = list(),
  ...
)

# S3 method for list
objSize(
  x,
  quick = getOption("reproducible.quick", FALSE),
  enclosingEnvs = TRUE,
  .prevEnvirs = list(),
  ...
)

# S3 method for environment
objSize(
  x,
  quick = getOption("reproducible.quick", FALSE),
  enclosingEnvs = TRUE,
  .prevEnvirs = list(),
  ...
)

# S3 method for Path
objSize(
  x,
  quick = getOption("reproducible.quick", FALSE),
  enclosingEnvs = TRUE,
  .prevEnvirs = list(),
  ...
)

# S3 method for `function`
objSize(
  x,
  quick = getOption("reproducible.quick", FALSE),
  enclosingEnvs = TRUE,
  .prevEnvirs = list(),
  ...
)

objSizeSession(sumLevel = Inf, enclosingEnvs = TRUE, .prevEnvirs = list())

Arguments

x
An object

quick
Logical. Only some methods use this. e.g., Path class objects. In which case, file.size will be used instead of object.size.

enclosingEnvs
Logical indicating whether to include enclosing environments. Default TRUE.

.prevEnvirs
For internal account keeping to identify and prevent duplicate counting

...
Additional arguments (currently unused)

sumLevel
Numeric, indicating at which depth in the list of objects should the object sizes be summed (summarized). Default is Inf, meaning no sums. Currently, the only option other than Inf is 1: objSizeSession(1), which gives the size of each package.

Details

For functions, a user can include the enclosing environment as described https://www.r-bloggers.com/using-closures-as-objects-in-r/ and http://adv-r.had.co.nz/memory.html. It is not entirely clear which estimate is better. However, if the enclosing environment is the .GlobalEnv, it will not be included even though enclosingEnvs = TRUE.

objSizeSession will give the size of the whole session, including loaded packages. Because of the difficulties in calculating the object size of base and methods packages and Autoloads, these are omitted.

Examples

library(utils)

foo <- new.env()
foo$b <- 1:10
foo$d <- 1:10

objSize(foo) # all the elements in the environment

#> $`foo$b`
#> 96 bytes
#> 
#> $`foo$d`
#> 96 bytes
#> 
#> $foo
#> 56 bytes
#> 

object.size(foo) # different - only measuring the environment as an object

#> 56 bytes

object.size(prepInputs) # only the function, without its enclosing environment

#> 125136 bytes

objSize(prepInputs)     # the function, plus its enclosing environment

#> $`%<%`
#> 47288 bytes
#> 
#> $`%C%`
#> 548360 bytes
#> 
#> $appendChecksumsTable
#> 270136 bytes
#> 
#> $argsToRemove
#> 1968 bytes
#> 
#> $asPath
#> 6944 bytes
#> 
#> $asPath.character
#> 20184 bytes
#> 
#> $asPath.null
#> 6496 bytes
#> 
#> $assessDataType
#> 6800 bytes
#> 
#> $assessDataType.default
#> 9968 bytes
#> 
#> $assessDataType.Raster
#> 109840 bytes
#> 
#> $assessDataType.RasterStack
#> 58944 bytes
#> 
#> $assessDataTypeGDAL
#> 9520 bytes
#> 
#> $assessGoogle
#> 198120 bytes
#> 
#> $attemptGDAL
#> 35088 bytes
#> 
#> $basename2
#> 24656 bytes
#> 
#> $bigRastersTmpFile
#> 5288 bytes
#> 
#> $bigRastersTmpFolder
#> 7344 bytes
#> 
#> $bufferWarningSuppress
#> 59264 bytes
#> 
#> $Cache
#> 27720 bytes
#> 
#> $CacheDBFile
#> 97544 bytes
#> 
#> $CacheDBTableName
#> 119416 bytes
#> 
#> $CacheDigest
#> 272008 bytes
#> 
#> $CacheIsACache
#> 65104 bytes
#> 
#> $CacheStorageDir
#> 29000 bytes
#> 
#> $CacheStoredFile
#> 100432 bytes
#> 
#> $captureWarningsToAttr
#> 106112 bytes
#> 
#> $cc
#> 38672 bytes
#> 
#> $checkAndMakeCloudFolderID
#> 76904 bytes
#> 
#> $checkColors
#> 23888 bytes
#> 
#> $checkFutures
#> 36272 bytes
#> 
#> $checkGDALVersion
#> 53224 bytes
#> 
#> $checkoutVersion
#> 8040 bytes
#> 
#> $Checksums
#> 12872 bytes
#> 
#> $chooseCRANmirror2
#> 23000 bytes
#> 
#> $clearCache
#> 17392 bytes
#> 
#> $clearStubArtifacts
#> 9728 bytes
#> 
#> $cloudCache
#> 10440 bytes
#> 
#> $cloudCheckOld
#> 10664 bytes
#> 
#> $cloudDownload
#> 32968 bytes
#> 
#> $cloudDownloadRasterBackend
#> 69496 bytes
#> 
#> $cloudFolderFromCacheRepo
#> 14200 bytes
#> 
#> $cloudSyncCacheOld
#> 12280 bytes
#> 
#> $cloudUpload
#> 54440 bytes
#> 
#> $cloudUploadFromCache
#> 31848 bytes
#> 
#> $cloudUploadRasterBackends
#> 24512 bytes
#> 
#> $cloudWriteOld
#> 11472 bytes
#> 
#> $compareNA
#> 20800 bytes
#> 
#> $convertPaths
#> 61888 bytes
#> 
#> $convertRasterPaths
#> 141952 bytes
#> 
#> $Copy
#> 10648 bytes
#> 
#> $copyFile
#> 33328 bytes
#> 
#> $copySingleFile
#> 636400 bytes
#> 
#> $createCache
#> 35344 bytes
#> 
#> $cropInputs
#> 7072 bytes
#> 
#> $cropInputs.default
#> 6608 bytes
#> 
#> $cropInputs.sf
#> 551496 bytes
#> 
#> $cropInputs.spatialClasses
#> 183000 bytes
#> 
#> $cropReprojMaskWGDAL
#> 175448 bytes
#> 
#> $datatypeVals
#> 8080 bytes
#> 
#> $dbConnectAll
#> 15288 bytes
#> 
#> $dealWithCores
#> 33968 bytes
#> 
#> $dealWithRasters
#> 67968 bytes
#> 
#> $dealWithRastersOnRecovery
#> 382240 bytes
#> 
#> $determineFilename
#> 488176 bytes
#> 
#> $determineNestedTags
#> 249576 bytes
#> 
#> $devModeFn1
#> 73712 bytes
#> 
#> $differentRasters
#> 53960 bytes
#> 
#> $dlGeneric
#> 92664 bytes
#> 
#> $dlGoogle
#> 414768 bytes
#> 
#> $downloadFile
#> 247672 bytes
#> 
#> $downloadRemote
#> 165192 bytes
#> 
#> $driveLs
#> 30384 bytes
#> 
#> $dtp
#> 864 bytes
#> 
#> $dtps
#> 1232 bytes
#> 
#> $extractFromArchive
#> 158576 bytes
#> 
#> $fastMask
#> 145744 bytes
#> 
#> $file.move
#> 76440 bytes
#> 
#> $fileExt
#> 25048 bytes
#> 
#> $Filenames
#> 10296 bytes
#> 
#> $filePathSansExt
#> 10008 bytes
#> 
#> $findGDAL
#> 141416 bytes
#> 
#> $fixErrors
#> 7600 bytes
#> 
#> $fixErrors.default
#> 7136 bytes
#> 
#> $fixErrors.Raster
#> 63976 bytes
#> 
#> $fixErrors.sf
#> 176856 bytes
#> 
#> $fixErrors.SpatialPolygons
#> 214568 bytes
#> 
#> $getArtifact
#> 84704 bytes
#> 
#> $getCacheId
#> 86856 bytes
#> 
#> $getCacheRepos
#> 75776 bytes
#> 
#> $getCRANrepos
#> 94720 bytes
#> 
#> $getFunctionName
#> 1310448 bytes
#> 
#> $getGDALVersion
#> 108072 bytes
#> 
#> $getUserTags
#> 124248 bytes
#> 
#> $internetExists
#> 129688 bytes
#> 
#> $isAbsolutePath
#> 82408 bytes
#> 
#> $isDirectory
#> 48928 bytes
#> 
#> $isFile
#> 45248 bytes
#> 
#> $isInteractive
#> 2816 bytes
#> 
#> $isLongLat
#> 15000 bytes
#> 
#> $isOrHasRaster
#> 203088 bytes
#> 
#> $isTRUEorForce
#> 12072 bytes
#> 
#> $isWindows
#> 5664 bytes
#> 
#> $keepCache
#> 15696 bytes
#> 
#> $knownArchiveExtensions
#> 432 bytes
#> 
#> $knownInternalArchiveExtensions
#> 304 bytes
#> 
#> $knownSystemArchiveExtensions
#> 176 bytes
#> 
#> $linkOrCopy
#> 300424 bytes
#> 
#> $loadFile
#> 17960 bytes
#> 
#> $loadFromCache
#> 232576 bytes
#> 
#> $makeMemoisable
#> 6568 bytes
#> 
#> $makeMemoisable.default
#> 6104 bytes
#> 
#> $maskInputs
#> 6904 bytes
#> 
#> $maskInputs.Raster
#> 31120 bytes
#> 
#> $maskInputs.sf
#> 236024 bytes
#> 
#> $maskInputs.Spatial
#> 11376 bytes
#> 
#> $maskWithRasterNAs
#> 44888 bytes
#> 
#> $MaxVals
#> 1232 bytes
#> 
#> $MaxValsFlts
#> 464 bytes
#> 
#> $mergeCache
#> 14912 bytes
#> 
#> $messageDF
#> 227952 bytes
#> 
#> $MinVals
#> 1232 bytes
#> 
#> $MinValsFlts
#> 464 bytes
#> 
#> $missingFiles
#> 67968 bytes
#> 
#> $missingUnrarMess
#> 232 bytes
#> 
#> $moveAttributes
#> 74328 bytes
#> 
#> $movedCache
#> 52464 bytes
#> 
#> $nextNumericName
#> 205104 bytes
#> 
#> $objSize
#> 7448 bytes
#> 
#> $objSize.default
#> 9608 bytes
#> 
#> $objSize.environment
#> 60744 bytes
#> 
#> $objSize.function
#> 116680 bytes
#> 
#> $objSize.list
#> 325296 bytes
#> 
#> $objSize.Path
#> 28016 bytes
#> 
#> $objSizeSession
#> 377264 bytes
#> 
#> $postProcess
#> 6736 bytes
#> 
#> $postProcess.default
#> 6272 bytes
#> 
#> $postProcess.list
#> 22112 bytes
#> 
#> $postProcess.quosure
#> 6368 bytes
#> 
#> $postProcess.sf
#> 90016 bytes
#> 
#> $postProcess.spatialClasses
#> 12280 bytes
#> 
#> $postProcessAllSpatial
#> 178888 bytes
#> 
#> $postProcessChecks
#> 113400 bytes
#> 
#> $prepInputs
#> 125136 bytes
#> 
#> $preProcess
#> 494584 bytes
#> 
#> $preProcessParams
#> 510816 bytes
#> 
#> $proj6Warn
#> 136 bytes
#> 
#> $projectInputs
#> 6904 bytes
#> 
#> $projectInputs.default
#> 6440 bytes
#> 
#> $projectInputs.Raster
#> 265112 bytes
#> 
#> $projectInputs.sf
#> 139832 bytes
#> 
#> $projectInputs.Spatial
#> 133528 bytes
#> 
#> $projNotWKT2warn
#> 136 bytes
#> 
#> $RCurlMess
#> 232 bytes
#> 
#> $rebuildColors
#> 112552 bytes
#> 
#> $reproducibleOptions
#> 20768 bytes
#> 
#> $requireNamespaceMsg
#> 37632 bytes
#> 
#> $retry
#> 146632 bytes
#> 
#> $rmFromCache
#> 90664 bytes
#> 
#> $rmFromCloudFolder
#> 42728 bytes
#> 
#> $rndstr
#> 62304 bytes
#> 
#> $roundTo6Dec
#> 40552 bytes
#> 
#> $roundToRes
#> 54736 bytes
#> 
#> $saveFileInCacheFolder
#> 176184 bytes
#> 
#> $saveToCache
#> 99008 bytes
#> 
#> $searchFull
#> 84664 bytes
#> 
#> $searchFullEx
#> 6152 bytes
#> 
#> $setMinMaxIfNeeded
#> 170400 bytes
#> 
#> $showCache
#> 14680 bytes
#> 
#> $studyAreaName
#> 10232 bytes
#> 
#> $suppressWarningsSpecific
#> 125848 bytes
#> 
#> $unmakeMemoisable
#> 6600 bytes
#> 
#> $unmakeMemoisable.default
#> 6104 bytes
#> 
#> $updateFilenameSlots
#> 153160 bytes
#> 
#> $useDBI
#> 24144 bytes
#> 
#> $useETM
#> 39696 bytes
#> 
#> $verboseDF1
#> 76232 bytes
#> 
#> $verboseDF2
#> 70480 bytes
#> 
#> $verboseDF3
#> 70576 bytes
#> 
#> $verboseMessage1
#> 17648 bytes
#> 
#> $verboseMessage2
#> 12128 bytes
#> 
#> $verboseMessage3
#> 23296 bytes
#> 
#> $verboseTime
#> 16200 bytes
#> 
#> $writeChecksumsTable
#> 38280 bytes
#> 
#> $writeFuture
#> 26408 bytes
#> 
#> $writeOutputs
#> 7376 bytes
#> 
#> $writeOutputs.default
#> 10608 bytes
#> 
#> $writeOutputs.quosure
#> 14400 bytes
#> 
#> $writeOutputs.Raster
#> 1225200 bytes
#> 
#> $writeOutputs.sf
#> 146768 bytes
#> 
#> $writeOutputs.Spatial
#> 21872 bytes
#> 

# Size of all packages; includes their imported functions
if (FALSE) {
  bar <- objSizeSession(1)
  print(bar, units = "auto")
}

os1 <- object.size(as.environment("package:reproducible"))
os2 <- objSize(as.environment("package:reproducible"))
(os1) # very small -- just the environment container

#> 568 bytes

sum(unlist(os2)) # around 13 MB, with all functions, objects

#> [1] 28702056

                 # and imported functions