Recursive object.size

This has methods for various types of things that may not correctly report their object size using object.size. Also, for lists and environments, it will return the object size separately for each element. These are estimates only, and could be inaccurate. Alternative, similar functions include object.size and pryr::object_size. See Details for the special case of functions and their enclosing environments.

objSize(x, quick, enclosingEnvs, .prevEnvirs, ...)

# S3 method for default
<<<<<<< HEAD
objSize(x, quick = getOption("reproducible.quick",
  FALSE), enclosingEnvs = TRUE, .prevEnvirs = list(), ...)

# S3 method for list
objSize(x, quick = getOption("reproducible.quick", FALSE),
  enclosingEnvs = TRUE, .prevEnvirs = list(), ...)

# S3 method for environment
objSize(x, quick = getOption("reproducible.quick",
  FALSE), enclosingEnvs = TRUE, .prevEnvirs = list(), ...)

# S3 method for Path
objSize(x, quick = getOption("reproducible.quick", FALSE),
  enclosingEnvs = TRUE, .prevEnvirs = list(), ...)

# S3 method for function
objSize(x, quick = getOption("reproducible.quick",
  FALSE), enclosingEnvs = TRUE, .prevEnvirs = list(), ...)

objSizeSession(sumLevel = Inf, enclosingEnvs = TRUE,
  .prevEnvirs = list())
======= objSize( x, quick = getOption("reproducible.quick", FALSE), enclosingEnvs = TRUE, .prevEnvirs = list(), ... ) # S3 method for list objSize( x, quick = getOption("reproducible.quick", FALSE), enclosingEnvs = TRUE, .prevEnvirs = list(), ... ) # S3 method for environment objSize( x, quick = getOption("reproducible.quick", FALSE), enclosingEnvs = TRUE, .prevEnvirs = list(), ... ) # S3 method for Path objSize( x, quick = getOption("reproducible.quick", FALSE), enclosingEnvs = TRUE, .prevEnvirs = list(), ... ) # S3 method for `function` objSize( x, quick = getOption("reproducible.quick", FALSE), enclosingEnvs = TRUE, .prevEnvirs = list(), ... ) objSizeSession(sumLevel = Inf, enclosingEnvs = TRUE, .prevEnvirs = list()) >>>>>>> development
Arguments

x
An object

quick
Logical. Only some methods use this. e.g., Path class objects. In which case, file.size will be used instead of object.size.

enclosingEnvs
Logical indicating whether to include enclosing environments. Default TRUE.

.prevEnvirs
For internal account keeping to identify and prevent duplicate counting

<<<<<<< HEAD ======= 	...
Additional arguments (currently unused)

>>>>>>> development 	sumLevel
Numeric, indicating at which depth in the list of objects should the object sizes be summed (summarized). Default is Inf, meaning no sums. Currently, the only option other than Inf is 1: objSizeSession(1), which gives the size of each package.

Details

For functions, a user can include the enclosing environment as described https://www.r-bloggers.com/using-closures-as-objects-in-r/ and http://adv-r.had.co.nz/memory.html. It is not entirely clear which estimate is better. However, if the enclosing environment is the .GlobalEnv, it will not be included even though enclosingEnvs = TRUE.

objSizeSession will give the size of the whole session, including loaded packages. Because of the difficulties in calculating the object size of base and methods packages and Autoloads, these are omitted.

Examples

library(utils)

foo <- new.env()
foo$b <- 1:10
foo$d <- 1:10

<<<<<<< HEAD
objSize(foo) # all the elements in the environment

#> [1] "<environment: 0x5635668649e8>"

#> $`foo$b`
=======
objSize(foo) # all the elements in the environment

#> [1] "<environment: 0x5607ed75cd90>"

#> $`foo$b`
>>>>>>> development
#> 96 bytes
#> 
#> $`foo$d`
#> 96 bytes
#> 
#> $foo
#> 56 bytes
#> 

object.size(foo) # different - only measuring the environment as an object

#> 56 bytes

object.size(prepInputs) # only the function, without its enclosing environment

#> 341416 bytes

objSize(prepInputs)     # the function, plus its enclosing environment

#> $`%<%`
#> 24120 bytes
#> 
#> $`%C%`
#> 292504 bytes
#> 
#> $appendChecksumsTable
#> 134552 bytes
#> 
#> $argsToRemove
#> 1896 bytes
#> 
#> $asPath
#> 1928 bytes
#> 
#> $asPath.character
#> 10824 bytes
#> 
#> $asPath.null
#> 1480 bytes
#> 
#> $assessDataType
#> 1784 bytes
#> 
#> $assessDataType.default
#> 3808 bytes
#> 
#> $assessDataType.Raster
#> 175000 bytes
#> 
#> $assessDataType.RasterStack
#> 21832 bytes
#> 
#> $assessDataTypeGDAL
#> 109368 bytes
#> 
#> $assessGoogle
<<<<<<< HEAD
#> 70672 bytes
=======
#> 73320 bytes
>>>>>>> development
#> 
#> $available.packagesMem
#> 3792 bytes
#> 
#> $basename2
#> 4768 bytes
#> 
#> $Cache
#> 33504 bytes
#> 
#> $CacheDigest
#> 112568 bytes
#> 
#> $cc
#> 53296 bytes
#> 
#> $checkAndMakeCloudFolderID
<<<<<<< HEAD
#> 21504 bytes
=======
#> 21104 bytes
>>>>>>> development
#> 
#> $checkFutures
#> 54168 bytes
#> 
#> $checkGDALVersion
#> 12808 bytes
#> 
#> $checkoutVersion
#> 179968 bytes
#> 
#> $checkPath
#> 5256 bytes
#> 
#> $Checksums
#> 7840 bytes
#> 
<<<<<<< HEAD
=======
#> $chooseCRANmirror2
#> 1120 bytes
#> 
>>>>>>> development
#> $clearCache
#> 8792 bytes
#> 
#> $clearStubArtifacts
#> 4696 bytes
#> 
#> $cloudCache
<<<<<<< HEAD
#> 1136 bytes
#> 
#> $cloudCheckOld
#> 44784 bytes
=======
#> 3776 bytes
#> 
#> $cloudCheckOld
#> 4000 bytes
>>>>>>> development
#> 
#> $cloudDownload
#> 47272 bytes
#> 
<<<<<<< HEAD
#> $cloudDownloadChecksumsOld
#> 57136 bytes
#> 
#> $cloudDownloadOld
#> 59992 bytes
#> 
=======
>>>>>>> development
#> $cloudDownloadRasterBackend
#> 139056 bytes
#> 
#> $cloudSyncCacheOld
<<<<<<< HEAD
#> 757968 bytes
#> 
#> $cloudUpdateChecksumsOld
#> 9392 bytes
=======
#> 5616 bytes
>>>>>>> development
#> 
#> $cloudUpload
#> 47696 bytes
#> 
<<<<<<< HEAD
#> $cloudUploadFileAndChecksumsOld
#> 61664 bytes
#> 
=======
>>>>>>> development
#> $cloudUploadFromCache
#> 36080 bytes
#> 
#> $cloudUploadRasterBackends
#> 65544 bytes
#> 
#> $cloudWriteOld
<<<<<<< HEAD
#> 121776 bytes
=======
#> 4808 bytes
>>>>>>> development
#> 
#> $compareNA
#> 12544 bytes
#> 
#> $convertPaths
#> 28456 bytes
#> 
#> $convertRasterPaths
#> 49840 bytes
#> 
#> $Copy
<<<<<<< HEAD
#> 5728 bytes
=======
#> 5616 bytes
>>>>>>> development
#> 
#> $copyFile
#> 33328 bytes
#> 
#> $copySingleFile
<<<<<<< HEAD
#> 268992 bytes
=======
#> 276680 bytes
>>>>>>> development
#> 
#> $cropInputs
#> 2056 bytes
#> 
#> $cropInputs.default
#> 1592 bytes
#> 
#> $cropInputs.sf
<<<<<<< HEAD
#> 150008 bytes
#> 
#> $cropInputs.spatialObjects
#> 149256 bytes
#> 
#> $determineFilename
#> 125176 bytes
#> 
#> $determineNestedTags
#> 167272 bytes
=======
#> 154696 bytes
#> 
#> $cropInputs.spatialObjects
#> 212696 bytes
#> 
#> $determineFilename
#> 102848 bytes
#> 
#> $determineNestedTags
#> 104896 bytes
>>>>>>> development
#> 
#> $devModeFn1
#> 170568 bytes
#> 
#> $dlGeneric
#> 46688 bytes
#> 
#> $dlGoogle
<<<<<<< HEAD
#> 61560 bytes
=======
#> 208280 bytes
>>>>>>> development
#> 
#> $downloadFile
#> 677272 bytes
#> 
#> $downloadRemote
#> 342592 bytes
#> 
#> $extractFromArchive
<<<<<<< HEAD
#> 533432 bytes
#> 
#> $fastMask
#> 473288 bytes
=======
#> 538640 bytes
#> 
#> $fastMask
#> 532416 bytes
>>>>>>> development
#> 
#> $Filenames
#> 4744 bytes
#> 
#> $fixErrors
#> 2584 bytes
#> 
#> $fixErrors.default
#> 2120 bytes
#> 
#> $fixErrors.sf
<<<<<<< HEAD
#> 104352 bytes
=======
#> 109720 bytes
>>>>>>> development
#> 
#> $fixErrors.SpatialPolygons
#> 98616 bytes
#> 
#> $getArtifact
#> 27864 bytes
#> 
#> $getCacheId
#> 30208 bytes
#> 
#> $getCacheRepos
#> 23040 bytes
#> 
<<<<<<< HEAD
#> $getChecksumsFileIDOld
#> 52704 bytes
#> 
=======
>>>>>>> development
#> $getCRANrepos
#> 34016 bytes
#> 
#> $getFunctionName
#> 552952 bytes
#> 
#> $getGDALVersion
#> 43216 bytes
#> 
#> $getLocalTags
#> 40808 bytes
#> 
#> $getUserTags
#> 56040 bytes
#> 
#> $installedVersions
#> 110480 bytes
#> 
#> $installedVersionsQuick
#> 116976 bytes
#> 
#> $installVersions
<<<<<<< HEAD
#> 932544 bytes
=======
#> 926920 bytes
>>>>>>> development
#> 
#> $isInteractive
#> 840 bytes
#> 
#> $isOrHasRaster
<<<<<<< HEAD
#> 31056 bytes
=======
#> 42432 bytes
#> 
#> $isWindows
#> 3544 bytes
>>>>>>> development
#> 
#> $keepCache
#> 7864 bytes
#> 
<<<<<<< HEAD
#> $linkOrCopy
#> 126864 bytes
=======
#> $knownArchiveExtensions
#> 432 bytes
#> 
#> $knownInternalArchiveExtensions
#> 304 bytes
#> 
#> $knownSystemArchiveExtensions
#> 176 bytes
#> 
#> $linkOrCopy
#> 121392 bytes
>>>>>>> development
#> 
#> $makeMemoisable
#> 1552 bytes
#> 
#> $makeMemoisable.default
#> 1088 bytes
#> 
#> $maskInputs
#> 1888 bytes
#> 
#> $maskInputs.Raster
#> 28704 bytes
#> 
#> $maskInputs.sf
#> 107768 bytes
#> 
#> $maskInputs.Spatial
#> 107640 bytes
#> 
#> $mergeCache
#> 5264 bytes
#> 
#> $missingFiles
#> 38056 bytes
#> 
<<<<<<< HEAD
=======
#> $missingUnrarMess
#> 232 bytes
#> 
#> $moveAttributes
#> 28808 bytes
#> 
>>>>>>> development
#> $newLibPaths
#> 9816 bytes
#> 
#> $nextNumericName
#> 97200 bytes
#> 
#> $normPath
#> 4744 bytes
#> 
#> $objSize
#> 2432 bytes
#> 
#> $objSize.default
#> 3496 bytes
#> 
#> $objSize.environment
#> 37928 bytes
#> 
#> $objSize.function
#> 48056 bytes
#> 
#> $objSize.list
#> 127880 bytes
#> 
#> $objSize.Path
#> 6984 bytes
#> 
#> $objSizeSession
#> 128576 bytes
#> 
#> $package_dependenciesMem
#> 102240 bytes
#> 
#> $pickFirstVersion
#> 17000 bytes
#> 
#> $pkgDep
#> 1066856 bytes
#> 
#> $pkgDep2
#> 62152 bytes
#> 
#> $pkgSnapshot
#> 164016 bytes
#> 
#> $postProcess
#> 1720 bytes
#> 
#> $postProcess.default
#> 1256 bytes
#> 
#> $postProcess.list
#> 8872 bytes
#> 
#> $postProcess.quosure
#> 5944 bytes
#> 
#> $postProcess.sf
#> 38704 bytes
#> 
#> $postProcess.spatialObjects
#> 27496 bytes
#> 
#> $postProcessAllSpatial
<<<<<<< HEAD
#> 298096 bytes
=======
#> 327088 bytes
>>>>>>> development
#> 
#> $postProcessChecks
#> 49936 bytes
#> 
#> $prepInputs
#> 341416 bytes
#> 
#> $preProcess
<<<<<<< HEAD
#> 1048856 bytes
=======
#> 1024000 bytes
>>>>>>> development
#> 
#> $projectInputs
#> 1888 bytes
#> 
#> $projectInputs.default
#> 1424 bytes
#> 
#> $projectInputs.Raster
<<<<<<< HEAD
#> 707496 bytes
=======
#> 702992 bytes
>>>>>>> development
#> 
#> $projectInputs.sf
#> 91944 bytes
#> 
#> $projectInputs.Spatial
#> 43432 bytes
#> 
#> $reproducibleOptions
<<<<<<< HEAD
#> 15888 bytes
=======
#> 16400 bytes
>>>>>>> development
#> 
#> $Require
#> 354824 bytes
#> 
#> $retry
#> 40288 bytes
#> 
#> $rndstr
#> 29240 bytes
#> 
#> $searchFull
#> 37224 bytes
#> 
#> $searchFullEx
#> 1120 bytes
#> 
#> $showCache
#> 6848 bytes
#> 
#> $showLocalRepo2
#> 11400 bytes
#> 
#> $showLocalRepo3
#> 3272 bytes
#> 
#> $showLocalRepo3Mem
#> 88680 bytes
#> 
<<<<<<< HEAD
=======
#> $tempdir2
#> 5680 bytes
#> 
>>>>>>> development
#> $unmakeMemoisable
#> 1584 bytes
#> 
#> $unmakeMemoisable.default
#> 1088 bytes
#> 
<<<<<<< HEAD
=======
#> $useETM
#> 14544 bytes
#> 
>>>>>>> development
#> $verboseDF1
#> 39560 bytes
#> 
#> $verboseDF2
#> 36000 bytes
#> 
#> $verboseDF3
#> 36096 bytes
#> 
#> $verboseMessage1
#> 8792 bytes
#> 
#> $verboseMessage2
#> 5464 bytes
#> 
#> $verboseMessage3
#> 10160 bytes
#> 
#> $verboseTime
#> 4208 bytes
#> 
<<<<<<< HEAD
#> $warnonce
#> 34096 bytes
#> 
=======
>>>>>>> development
#> $writeChecksumsTable
#> 24960 bytes
#> 
#> $writeFuture
#> 64408 bytes
#> 
#> $writeOutputs
#> 2360 bytes
#> 
#> $writeOutputs.default
#> 4400 bytes
#> 
#> $writeOutputs.Raster
<<<<<<< HEAD
#> 95992 bytes
=======
#> 132056 bytes
>>>>>>> development
#> 
#> $writeOutputs.sf
#> 42088 bytes
#> 
#> $writeOutputs.Spatial
#> 61776 bytes
#> 

# Size of all packages; includes their imported functions
if (FALSE) {
  bar <- objSizeSession(1)
  print(bar, units = "auto")
}

os1 <- object.size(as.environment("package:reproducible"))
<<<<<<< HEAD
os2 <- objSize(as.environment("package:reproducible"))

#> [1] "<environment: package:reproducible>"

(os1) # very small -- just the environment container

#> 584 bytes

sum(unlist(os2)) # around 31 MB, with all functions, objects

#> [1] 14754864

                 # and imported functions
=======
os2 <- objSize(as.environment("package:reproducible"))

#> [1] "<environment: package:reproducible>"

(os1) # very small -- just the environment container

#> 584 bytes

sum(unlist(os2)) # around 31 MB, with all functions, objects

#> [1] 13881104

                 # and imported functions
>>>>>>> development