A cache-aware pipe (currently not working)

With updates to magrittr to version 2.0, this Cache pipe is now broken. We are working on an update.

This pipe can only be used at any point in a pipe chain, but must be preceded by Cache(...) (which allows other Cache() %C% ... remaining pipes arguments to be passed).

This will take the input arguments of the first function immediately following the Cache() and the pipe chain until the special %C%, evaluate them both against the cacheRepo argument in Cache. If they exist, then the entire pipe chain will be skipped, and only the previous final result will be given. If there is no previous cached copy of the initial function's arguments, then all chain elements will be evaluated. The final result will be cached for future use. Therefore, the entire chain must be identical. The required usage should be straight forward to insert into existing code that uses pipes (Cache() %C% ... remaining pipes).

Still experimental and may change. This form cannot pass any arguments to ]codeCache, such as cacheRepo, thus it is of limited utility. However, it is a clean alternative for simple cases.

lhs %C% rhs

lhs %<% rhs

Arguments

lhs

A name to assign to.

rhs

A function call

Examples

# THIS IS CURRENTLY BROKEN DUE TO UPGRADES TO INTERNALS OF magrittr %>%
if (FALSE)  # these can't be automatically run due to package conflicts with magrittr
library(magrittr) # standard pipe
tmpdir <- file.path(tempdir(), "testCache")
checkPath(tmpdir, create = TRUE)
#> [1] "/tmp/RtmpteydN6/testCache"
a <- rnorm(10, 16) %>%
     mean() %>%
     prod(., 6)
b <- Cache(cacheRepo = tmpdir) %C% # use of the %C% pipe!
     rnorm(10, 16) %>% # everything after here is NOT cached!
     mean() %>%
     prod(., 6)
#> Called from: Cache(cacheRepo = tmpdir) %C% rnorm(10, 16)
#> debug: args <- list(eval(lhs[[1]]), ._pipe = parse(text = paste(c(lhs, 
#>     rhss), collapse = " %>% ")), ._pipeFn = as.character(lhs[[1]]), 
#>     ._lhs = quote(lhs), ._rhss = quote(rhss), ._envir = parent)
#> debug: args <- append(args, lapply(cacheArgs, eval, envir = parent, 
#>     enclos = parent))
#> debug: result <- withVisible(do.call("Cache", args))
#> debug: cmpdPipe <- FALSE
#> debug: if (length(pipes) > 0) {
#>     if (identical(pipes[[1L]], quote(`%<>%`))) 
#>         cmpdPipe <- TRUE
#> }
#> debug: if (identical(pipes[[1L]], quote(`%<>%`))) cmpdPipe <- TRUE
#> debug: if (cmpdPipe) {
#>     ret <- eval(call("<-", lhs, result[["value"]]), parent, parent)
#> } else {
#>     if (result[["visible"]]) 
#>         ret <- result[["value"]]
#>     else ret <- invisible(result[["value"]])
#> }
#> debug: if (result[["visible"]]) ret <- result[["value"]] else ret <- invisible(result[["value"]])
#> debug: ret <- result[["value"]]
d <- Cache(cacheRepo = tmpdir) %C%
     rnorm(10, 16) %>%
     mean() %>%
     prod(., 6)
#> Called from: Cache(cacheRepo = tmpdir) %C% rnorm(10, 16)
#> debug: args <- list(eval(lhs[[1]]), ._pipe = parse(text = paste(c(lhs, 
#>     rhss), collapse = " %>% ")), ._pipeFn = as.character(lhs[[1]]), 
#>     ._lhs = quote(lhs), ._rhss = quote(rhss), ._envir = parent)
#> debug: args <- append(args, lapply(cacheArgs, eval, envir = parent, 
#>     enclos = parent))
#> debug: result <- withVisible(do.call("Cache", args))
#>   ...(Object to retrieve (c0ce50be720d40a8.rds))
#>      loaded cached result from previous 'rnorm, mean' pipe sequence call
#> debug: cmpdPipe <- FALSE
#> debug: if (length(pipes) > 0) {
#>     if (identical(pipes[[1L]], quote(`%<>%`))) 
#>         cmpdPipe <- TRUE
#> }
#> debug: if (identical(pipes[[1L]], quote(`%<>%`))) cmpdPipe <- TRUE
#> debug: if (cmpdPipe) {
#>     ret <- eval(call("<-", lhs, result[["value"]]), parent, parent)
#> } else {
#>     if (result[["visible"]]) 
#>         ret <- result[["value"]]
#>     else ret <- invisible(result[["value"]])
#> }
#> debug: if (result[["visible"]]) ret <- result[["value"]] else ret <- invisible(result[["value"]])
#> debug: ret <- result[["value"]]
e <- Cache(cacheRepo = tmpdir) %C%
     rnorm(10, 16) %>%
     mean() %>%
     prod(., 5) # changed
#> Called from: Cache(cacheRepo = tmpdir) %C% rnorm(10, 16)
#> debug: args <- list(eval(lhs[[1]]), ._pipe = parse(text = paste(c(lhs, 
#>     rhss), collapse = " %>% ")), ._pipeFn = as.character(lhs[[1]]), 
#>     ._lhs = quote(lhs), ._rhss = quote(rhss), ._envir = parent)
#> debug: args <- append(args, lapply(cacheArgs, eval, envir = parent, 
#>     enclos = parent))
#> debug: result <- withVisible(do.call("Cache", args))
#>   ...(Object to retrieve (c0ce50be720d40a8.rds))
#>      loaded cached result from previous 'rnorm, mean' pipe sequence call
#> debug: cmpdPipe <- FALSE
#> debug: if (length(pipes) > 0) {
#>     if (identical(pipes[[1L]], quote(`%<>%`))) 
#>         cmpdPipe <- TRUE
#> }
#> debug: if (identical(pipes[[1L]], quote(`%<>%`))) cmpdPipe <- TRUE
#> debug: if (cmpdPipe) {
#>     ret <- eval(call("<-", lhs, result[["value"]]), parent, parent)
#> } else {
#>     if (result[["visible"]]) 
#>         ret <- result[["value"]]
#>     else ret <- invisible(result[["value"]])
#> }
#> debug: if (result[["visible"]]) ret <- result[["value"]] else ret <- invisible(result[["value"]])
#> debug: ret <- result[["value"]]
all.equal(b,d) # TRUE
#> [1] TRUE
all.equal(a,d) # different because 'a' uses a unique rnorm, 'd' uses the Cached rnorm
#> [1] "Mean relative difference: 0.06114944"
               #   because the arguments to rnorm, i.e., 10 and 16, and
               #   the subsequent functions in the chain, are identical
all.equal(a,e) # different because the final function, prod, has a changed argument.
#> [1] "Mean relative difference: 0.1157088"

###########
# multiple random elements shows Cached sequence up to %C%
a1 <- Cache(cacheRepo = tmpdir) %>%
       seq(1, 10) %>%
       rnorm(2, mean = .) %>%
       mean() %C%                # Cache pipe here --
                                 # means this pipe is the last one that is Cached
       rnorm(3, mean = .) %>%
       mean(.) %>%
       rnorm(4, mean = .)  # Random 4 numbers, the mean is same each time
#> Called from: Cache(cacheRepo = tmpdir) %>% seq(1, 10) %>% rnorm(2, mean = .) %>% 
#>     mean() %C% rnorm(3, mean = .)
#> debug: args <- list(eval(lhs[[1]]), ._pipe = parse(text = paste(c(lhs, 
#>     rhss), collapse = " %>% ")), ._pipeFn = as.character(lhs[[1]]), 
#>     ._lhs = quote(lhs), ._rhss = quote(rhss), ._envir = parent)
#> debug: args <- append(args, lapply(cacheArgs, eval, envir = parent, 
#>     enclos = parent))
#> debug: result <- withVisible(do.call("Cache", args))
#> debug: cmpdPipe <- FALSE
#> debug: if (length(pipes) > 0) {
#>     if (identical(pipes[[1L]], quote(`%<>%`))) 
#>         cmpdPipe <- TRUE
#> }
#> debug: if (identical(pipes[[1L]], quote(`%<>%`))) cmpdPipe <- TRUE
#> debug: if (cmpdPipe) {
#>     ret <- eval(call("<-", lhs, result[["value"]]), parent, parent)
#> } else {
#>     if (result[["visible"]]) 
#>         ret <- result[["value"]]
#>     else ret <- invisible(result[["value"]])
#> }
#> debug: if (result[["visible"]]) ret <- result[["value"]] else ret <- invisible(result[["value"]])
#> debug: ret <- result[["value"]]
a2 <- Cache(cacheRepo = tmpdir) %>%
       seq(1, 10) %>%
       rnorm(2, mean = .) %>%
       mean() %C%                # Cache pipe here --
                                 # means this pipe is the last one that is Cached
       rnorm(3, mean = .) %>%
       mean(.) %>%
       rnorm(4, mean = .)  # Random 4 numbers, the mean is same each time
#> Called from: Cache(cacheRepo = tmpdir) %>% seq(1, 10) %>% rnorm(2, mean = .) %>% 
#>     mean() %C% rnorm(3, mean = .)
#> debug: args <- list(eval(lhs[[1]]), ._pipe = parse(text = paste(c(lhs, 
#>     rhss), collapse = " %>% ")), ._pipeFn = as.character(lhs[[1]]), 
#>     ._lhs = quote(lhs), ._rhss = quote(rhss), ._envir = parent)
#> debug: args <- append(args, lapply(cacheArgs, eval, envir = parent, 
#>     enclos = parent))
#> debug: result <- withVisible(do.call("Cache", args))
#>   ...(Object to retrieve (3ca383481492d0b0.rds))
#>      loaded cached result from previous 'seq, rnorm, mean, rnorm' pipe sequence call
#> debug: cmpdPipe <- FALSE
#> debug: if (length(pipes) > 0) {
#>     if (identical(pipes[[1L]], quote(`%<>%`))) 
#>         cmpdPipe <- TRUE
#> }
#> debug: if (identical(pipes[[1L]], quote(`%<>%`))) cmpdPipe <- TRUE
#> debug: if (cmpdPipe) {
#>     ret <- eval(call("<-", lhs, result[["value"]]), parent, parent)
#> } else {
#>     if (result[["visible"]]) 
#>         ret <- result[["value"]]
#>     else ret <- invisible(result[["value"]])
#> }
#> debug: if (result[["visible"]]) ret <- result[["value"]] else ret <- invisible(result[["value"]])
#> debug: ret <- result[["value"]]
sum(a1 - a2) # not 0 # i.e., numbers are different
#> [1] -0.7357814

# NOW DO WITH CACHE AT END
b1 <- Cache(cacheRepo = tmpdir) %>%
       seq(1, 10) %>%
       rnorm(2, mean = .) %>%
       mean() %>%
                                 # means this pipe is the last one that is Cached
       rnorm(3, mean = .) %>%
       mean(.) %C%               # Cache pipe here --
       rnorm(4, mean = .)        # These are samethe mean is same each time
#> Called from: Cache(cacheRepo = tmpdir) %>% seq(1, 10) %>% rnorm(2, mean = .) %>% 
#>     mean() %>% rnorm(3, mean = .) %>% mean(.) %C% rnorm(4, mean = .)
#> debug: args <- list(eval(lhs[[1]]), ._pipe = parse(text = paste(c(lhs, 
#>     rhss), collapse = " %>% ")), ._pipeFn = as.character(lhs[[1]]), 
#>     ._lhs = quote(lhs), ._rhss = quote(rhss), ._envir = parent)
#> debug: args <- append(args, lapply(cacheArgs, eval, envir = parent, 
#>     enclos = parent))
#> debug: result <- withVisible(do.call("Cache", args))
#> debug: cmpdPipe <- FALSE
#> debug: if (length(pipes) > 0) {
#>     if (identical(pipes[[1L]], quote(`%<>%`))) 
#>         cmpdPipe <- TRUE
#> }
#> debug: if (identical(pipes[[1L]], quote(`%<>%`))) cmpdPipe <- TRUE
#> debug: if (cmpdPipe) {
#>     ret <- eval(call("<-", lhs, result[["value"]]), parent, parent)
#> } else {
#>     if (result[["visible"]]) 
#>         ret <- result[["value"]]
#>     else ret <- invisible(result[["value"]])
#> }
#> debug: if (result[["visible"]]) ret <- result[["value"]] else ret <- invisible(result[["value"]])
#> debug: ret <- result[["value"]]
b2 <- Cache(cacheRepo = tmpdir) %>%
       seq(1, 10) %>%
       rnorm(2, mean = .) %>%
       mean() %>%
                                 # means this pipe is the last one that is Cached
       rnorm(3, mean = .) %>%
       mean(.) %C%               # Cache pipe here --
       rnorm(4, mean = .)        # These are samethe mean is same each time
#> Called from: Cache(cacheRepo = tmpdir) %>% seq(1, 10) %>% rnorm(2, mean = .) %>% 
#>     mean() %>% rnorm(3, mean = .) %>% mean(.) %C% rnorm(4, mean = .)
#> debug: args <- list(eval(lhs[[1]]), ._pipe = parse(text = paste(c(lhs, 
#>     rhss), collapse = " %>% ")), ._pipeFn = as.character(lhs[[1]]), 
#>     ._lhs = quote(lhs), ._rhss = quote(rhss), ._envir = parent)
#> debug: args <- append(args, lapply(cacheArgs, eval, envir = parent, 
#>     enclos = parent))
#> debug: result <- withVisible(do.call("Cache", args))
#>   ...(Object to retrieve (a048b89a8eeb8f74.rds))
#>      loaded cached result from previous 'seq, rnorm, mean, rnorm, mean, rnorm' pipe sequence call
#> debug: cmpdPipe <- FALSE
#> debug: if (length(pipes) > 0) {
#>     if (identical(pipes[[1L]], quote(`%<>%`))) 
#>         cmpdPipe <- TRUE
#> }
#> debug: if (identical(pipes[[1L]], quote(`%<>%`))) cmpdPipe <- TRUE
#> debug: if (cmpdPipe) {
#>     ret <- eval(call("<-", lhs, result[["value"]]), parent, parent)
#> } else {
#>     if (result[["visible"]]) 
#>         ret <- result[["value"]]
#>     else ret <- invisible(result[["value"]])
#> }
#> debug: if (result[["visible"]]) ret <- result[["value"]] else ret <- invisible(result[["value"]])
#> debug: ret <- result[["value"]]
sum(b1 - b2) # 0 # i.e., numbers are same
#> [1] 0

unlink(tmpdir, recursive = TRUE)

# Equivalent
a <- Cache(rnorm, 1)
#> No cacheRepo supplied and getOption('reproducible.cachePath') is inside a temporary directory;
#>   this will not persist across R sessions.
b %<% rnorm(1)
#> No cacheRepo supplied and getOption('reproducible.cachePath') is inside a temporary directory;
#>   this will not persist across R sessions.
#>   ...(Object to retrieve (7072c305d8c69df0.rds))
#>      loaded cached result from previous rnorm call