A cache-aware pipe that does not mask with %>%

STILL EXPERIMENTAL. THIS MAY NOT WORK AS ANTICIPATED.

lhs %C% rhs

Arguments

lhs
A value or the magrittr placeholder.

rhs
A function call using the magrittr semantics.

Details

This pipe can only be used at any point in a pipe chain, but must be preceded by Cache(...) (which allows other Cache arguments to be passed).

This will take the input arguments of the first function immediately following the Cache() and the pipe chain until the special %C%, evaluate them both against the cacheRepo argument in Cache. If they exist, then the entire pipe chain will be skipped, and only the previous final result will be given. If there is no previous cached copy of the initial function's arguments, then all chain elements will be evaluated. The final result will be cached for future use. Therefore, the entire chain must be identical. The required usage should be straight forward to insert into existing code that uses pipes (Cache() %C% ... remaining pipes).

Note

This is still experimental; use with care.

Examples

library(magrittr) # standard pipe
# dontrun{ # these can't be automatically run due to package conflicts with magrittr
tmpdir <- file.path(tempdir(), "testCache")
checkPath(tmpdir, create = TRUE)

#> [1] "/tmp/RtmpCbsdp6/testCache"

a <- rnorm(10, 16) %>%
     mean() %>%
     prod(., 6)
b <- Cache(cacheRepo = tmpdir) %C% # use of the %C% pipe!
     rnorm(10, 16) %>% # everything after here is NOT cached!
     mean() %>%
     prod(., 6)
d <- Cache(cacheRepo = tmpdir) %C%
     rnorm(10, 16) %>%
     mean() %>%
     prod(., 6)

#>   loading cached result from previous 'rnorm, mean' pipe sequence call, adding to memoised copy

e <- Cache(cacheRepo = tmpdir) %C%
     rnorm(10, 16) %>%
     mean() %>%
     prod(., 5) # changed

#>   loading memoised result from previous 'rnorm, mean' pipe sequence call.

all.equal(b,d) # TRUE

#> [1] TRUE

all.equal(a,d) # different because 'a' uses a unique rnorm, 'd' uses the Cached rnorm

#> [1] "Mean relative difference: 0.01152155"

               #   because the arguments to rnorm, i.e., 10 and 16, and
               #   the subsequent functions in the chain, are identical
all.equal(a,e) # different because the final function, prod, has a changed argument.

#> [1] "Mean relative difference: 0.176268"

###########
# multiple random elements shows Cached sequence up to %C%
a1 <- Cache(cacheRepo = tmpdir) %>%
       seq(1, 10) %>%
       rnorm(2, mean = .) %>%
       mean() %C%                # Cache pipe here --
                                 # means this pipe is the last one that is Cached
       rnorm(3, mean = .) %>%
       mean(.) %>%
       rnorm(4, mean = .)  # Random 4 numbers, the mean is same each time
a2 <- Cache(cacheRepo = tmpdir) %>%
       seq(1, 10) %>%
       rnorm(2, mean = .) %>%
       mean() %C%                # Cache pipe here --
                                 # means this pipe is the last one that is Cached
       rnorm(3, mean = .) %>%
       mean(.) %>%
       rnorm(4, mean = .)  # Random 4 numbers, the mean is same each time

#>   loading cached result from previous 'seq, rnorm, mean, rnorm' pipe sequence call, adding to memoised copy

sum(a1 - a2) # not 0 # i.e., numbers are different

#> [1] -3.281365

# NOW DO WITH CACHE AT END
b1 <- Cache(cacheRepo = tmpdir) %>%
       seq(1, 10) %>%
       rnorm(2, mean = .) %>%
       mean() %>%
                                 # means this pipe is the last one that is Cached
       rnorm(3, mean = .) %>%
       mean(.) %C%               # Cache pipe here --
       rnorm(4, mean = .)        # These are samethe mean is same each time
b2 <- Cache(cacheRepo = tmpdir) %>%
       seq(1, 10) %>%
       rnorm(2, mean = .) %>%
       mean() %>%
                                 # means this pipe is the last one that is Cached
       rnorm(3, mean = .) %>%
       mean(.) %C%               # Cache pipe here --
       rnorm(4, mean = .)        # These are samethe mean is same each time

#>   loading cached result from previous 'seq, rnorm, mean, rnorm, mean, rnorm' pipe sequence call, adding to memoised copy

sum(b1 - b2) # 0 # i.e., numbers are same

#> [1] 0

unlink(tmpdir, recursive = TRUE)
#}