Determine package dependencies, first looking at local filesystem

This is intended to replace package_dependencies or pkgDep in the miniCRAN package, but with modifications for speed. It will first check local package directories in libPath, and it if the function cannot find the packages there, then it will use package_dependencies.

pkgDep2 is a convenience wrapper of pkgDep that "goes one level in", i.e., the first order dependencies, and runs the pkgDep on those.

This is a wrapper around tools::dependsOnPkgs, but with the added option of sorted, which will sort them such that the packages at the top will have the least number of dependencies that are in pkgs. This is essentially a topological sort, but it is done heuristically. This can be used to e.g., detach or unloadNamespace packages in order so that they each of their dependencies are detached or unloaded first.

pkgDep(
  packages,
  libPath,
  recursive = TRUE,
  depends = TRUE,
  imports = TRUE,
  suggests = FALSE,
  linkingTo = TRUE,
  topoSort = FALSE,
  repos = getOption("repos"),
  refresh = FALSE,
  verbose = getOption("reproducible.verbose")
)

pkgDep2(
  packages,
  recursive = TRUE,
  depends = TRUE,
  imports = TRUE,
  suggests = FALSE,
  linkingTo = TRUE,
  repos = getOption("repos"),
  refresh = FALSE,
  verbose = getOption("reproducible.verbose"),
  sorted = TRUE
)

pkgDepTopoSort(
  pkgs,
  deps,
  reverse = FALSE,
  topoSort = TRUE,
  useAllInSearch = FALSE,
  returnFull = TRUE
)

Arguments

packages
a character vector of package names.

libPath
The library path where all packages should be installed, and looked for to load (i.e., call library)

recursive
Logical. Should dependencies of dependencies be searched, recursively. NOTE: Dependencies of suggests will not be recursive. Default TRUE.

depends
Logical. Include packages listed in "Depends". Default TRUE.

imports
Logical. Include packages listed in "Imports". Default TRUE.

suggests
Logical. Include packages listed in "Suggests". Default FALSE.

linkingTo
Logical. Include packages listed in "LinkingTo". Default TRUE.

topoSort
Logical. If TRUE, the default, then the returned list of packages will be in order with the least number of dependencies listed in pkgs at the top of the list.

repos
The remote repository (e.g., a CRAN mirror), passed to either install.packages, install_github or installVersions.

refresh
There is an internal type of caching. If the results are wrong, likely set refresh = TRUE.

verbose
logical indicating if output should monitor the package search cycles.

sorted
Logical. If TRUE, the default, the packages will be sorted in the returned list from most number of dependencies to least.

pkgs
A vector of package names to evaluate their reverse depends (i.e., the packages that use each of these packages)

deps
An optional named list of (reverse) dependencies. If not supplied, then tools::dependsOnPkgs(..., recursive = TRUE) will be used

reverse
logical: if FALSE (default), regular dependencies are calculated, otherwise reverse dependencies.

useAllInSearch
Logical. If TRUE, then all non-core R packages in search() will be appended to pkgs to allow those to also be identified

returnFull
Logical. If TRUE, then the full reverse dependencies will be returned; if FALSE, the default, only the reverse dependencies that are found within the pkgs (and search() if useAllInSearch = TRUE) will be returned.

Value

A possibly ordered, named (with packages as names) list where list elements are either full reverse depends.

Note

package_dependencies and pkgDep will differ under the following circumstances:

GitHub packages are not detected using tools::package_dependencies;

tools::package_dependencies does not detect the dependencies of base packages among themselves, e.g., methods depends on stats and graphics.

Examples

pkgDep("crayon")

#> $crayon
#> [1] "grDevices" "methods"   "utils"     "stats"     "graphics" 
#> 

pkgDep2("reproducible")

#> $quickPlot
#>  [1] "ggplot2"      "grid"         "gridBase"     "igraph"       "raster"      
#>  [6] "RColorBrewer" "rgdal"        "rgeos"        "sp"           "gtable"      
#> [11] "isoband"      "MASS"         "mgcv"         "scales"       "testthat"    
#> [16] "evaluate"     "pkgload"      "praise"       "methods"      "stats"       
#> [21] "graphics"     "grDevices"    "utils"        "desc"         "pkgbuild"    
#> [26] "rprojroot"    "rstudioapi"   "assertthat"   "R6"           "crayon"      
#> [31] "tools"        "callr"        "prettyunits"  "processx"     "ps"          
#> [36] "cli"          "withr"        "glue"         "fansi"        "backports"   
#> [41] "rlang"        "digest"       "ellipsis"     "magrittr"     "Rcpp"        
#> [46] "Matrix"       "splines"      "nlme"         "lattice"      "farver"      
#> [51] "labeling"     "munsell"      "viridisLite"  "colorspace"   "lifecycle"   
#> [56] "tibble"       "pillar"       "pkgconfig"    "vctrs"        "utf8"        
#> [61] "data.table"   "fpCompare"   
#> 
#> $usethis
#>  [1] "clipr"      "gh"         "git2r"      "rematch2"   "whisker"   
#>  [6] "yaml"       "utils"      "ini"        "cli"        "jsonlite"  
#> [11] "httr"       "assertthat" "tools"      "crayon"     "glue"      
#> [16] "methods"    "fansi"      "grDevices"  "stats"      "graphics"  
#> [21] "mime"       "openssl"    "R6"         "askpass"    "sys"       
#> [26] "curl"       "tibble"     "ellipsis"   "lifecycle"  "pillar"    
#> [31] "pkgconfig"  "vctrs"      "rlang"      "utf8"       "digest"    
#> [36] "magrittr"   "desc"       "fs"         "purrr"      "rprojroot" 
#> [41] "rstudioapi" "withr"      "backports" 
#> 
#> $googledrive
#>  [1] "curl"       "gargle"     "httr"       "jsonlite"   "magrittr"  
#>  [6] "purrr"      "rlang"      "tibble"     "uuid"       "fs"        
#> [11] "withr"      "methods"    "stats"      "graphics"   "grDevices" 
#> [16] "utils"      "glue"       "mime"       "openssl"    "R6"        
#> [21] "tools"      "askpass"    "sys"        "cli"        "ellipsis"  
#> [26] "fansi"      "lifecycle"  "pillar"     "pkgconfig"  "vctrs"     
#> [31] "assertthat" "crayon"     "utf8"       "digest"    
#> 
#> $testthat
#>  [1] "evaluate"    "pkgload"     "praise"      "methods"     "stats"      
#>  [6] "graphics"    "grDevices"   "utils"       "desc"        "pkgbuild"   
#> [11] "rprojroot"   "rstudioapi"  "assertthat"  "R6"          "crayon"     
#> [16] "tools"       "callr"       "prettyunits" "processx"    "ps"         
#> [21] "cli"         "withr"       "glue"        "fansi"       "backports"  
#> [26] "rlang"       "digest"      "ellipsis"    "magrittr"   
#> 
#> $RSQLite
#>  [1] "bit64"     "blob"      "BH"        "plogr"     "bit"       "utils"    
#>  [7] "methods"   "stats"     "graphics"  "grDevices" "rlang"     "vctrs"    
#> [13] "ellipsis"  "digest"    "glue"      "DBI"       "memoise"   "pkgconfig"
#> [19] "Rcpp"     
#> 
#> $sf
#>  [1] "classInt"   "units"      "e1071"      "class"      "KernSmooth"
#>  [6] "graphics"   "grDevices"  "stats"      "methods"    "utils"     
#> [11] "MASS"       "Rcpp"       "DBI"        "grid"       "magrittr"  
#> [16] "tools"     
#> 
#> $httr
#> [1] "mime"     "openssl"  "R6"       "tools"    "askpass"  "sys"      "curl"    
#> [8] "jsonlite"
#> 
#> $R.utils
#> [1] "R.methodsS3" "R.oo"        "utils"       "methods"     "stats"      
#> [6] "graphics"    "grDevices"   "tools"      
#> 
#> $lifecycle
#> [1] "glue"      "rlang"     "methods"   "stats"     "graphics"  "grDevices"
#> [7] "utils"    
#> 
#> $qs
#> [1] "Rcpp"          "RApiSerialize" "methods"       "utils"        
#> [5] "stats"         "graphics"      "grDevices"    
#> 
#> $raster
#> [1] "Rcpp"      "methods"   "sp"        "utils"     "stats"     "graphics" 
#> [7] "grDevices"
#> 
#> $sp
#> [1] "utils"     "stats"     "graphics"  "grDevices" "lattice"   "grid"     
#> [7] "methods"  
#> 
#> $remotes
#> [1] "methods"   "stats"     "tools"     "utils"     "graphics"  "grDevices"
#> 
#> $crayon
#> [1] "grDevices" "methods"   "utils"     "stats"     "graphics" 
#> 
#> $data.table
#> [1] "methods"   "stats"     "graphics"  "grDevices" "utils"    
#> 
#> $DBI
#> [1] "methods"   "stats"     "graphics"  "grDevices" "utils"    
#> 
#> $glue
#> [1] "methods"   "stats"     "graphics"  "grDevices" "utils"    
#> 
#> $methods
#> [1] "stats"     "graphics"  "grDevices" "utils"    
#> 
#> $memoise
#> [1] "digest" "utils" 
#> 
#> $backports
#> [1] "utils"
#> 
#> $digest
#> [1] "utils"
#> 
#> $fastdigest
#> character(0)
#> 
#> $fpCompare
#> character(0)
#> 
#> $magrittr
#> character(0)
#> 
#> $rlang
#> character(0)
#> 
#> $tools
#> character(0)
#> 
#> $utils
#> character(0)
#> 
#> $versions
#> character(0)
#>