Determine package dependencies, first looking at local filesystem

This is intended to replace package_dependencies or pkgDep in the miniCRAN package, but with modifications for speed. It will first check local package directories in libPath, and it if the function cannot find the packages there, then it will use package_dependencies.

pkgDep2 is a convenience wrapper of pkgDep that "goes one level in" i.e., the first order dependencies, and runs the pkgDep on those.

pkgDep(packages, libPath, recursive = TRUE, depends = TRUE,
  imports = TRUE, suggests = FALSE, linkingTo = TRUE,
  repos = getOption("repos"), refresh = FALSE,
  verbose = getOption("reproducible.verbose"))

pkgDep2(packages, recursive = TRUE, depends = TRUE, imports = TRUE,
  suggests = FALSE, linkingTo = TRUE, repos = getOption("repos"),
  refresh = FALSE, verbose = getOption("reproducible.verbose"),
  sorted = TRUE)

Arguments

packages
a character vector of package names.

libPath
The library path where all packages should be installed, and looked for to load (i.e., call library)

recursive
Logical. Should dependencies of dependencies be searched, recursively. NOTE: Dependencies of suggests will not be recursive. Default TRUE.

depends
Logical. Include packages listed in "Depends". Default TRUE.

imports
Logical. Include packages listed in "Imports". Default TRUE.

suggests
Logical. Include packages listed in "Suggests". Default FALSE.

linkingTo
Logical. Include packages listed in "LinkingTo". Default TRUE.

repos
The remote repository (e.g., a CRAN mirror), passed to either install.packages, install_github or installVersions.

refresh
There is an internal type of caching. If the results are wrong, likely set refresh = TRUE.

verbose
logical indicating if output should monitor the package search cycles.

sorted
Logical. If TRUE, the default, the packages will be sorted in the returned list from most number of dependencies to least.

Note

package_dependencies and pkgDep will differ under the following circumstances:

GitHub packages are not detected using tools::package_dependencies;

tools::package_dependencies does not detect the dependencies of base packages among themselves, e.g., methods depends on stats and graphics.

Examples

pkgDep("crayon")

#> $crayon
#> [1] "grDevices" "methods"   "utils"     "stats"     "graphics" 
#> 

pkgDep2("reproducible")

#> $quickPlot
#>  [1] "ggplot2"      "gridBase"     "igraph"       "RColorBrewer" "rgeos"       
#>  [6] "gtable"       "lazyeval"     "MASS"         "mgcv"         "reshape2"    
#> [11] "scales"       "viridisLite"  "grid"         "grDevices"    "utils"       
#> [16] "methods"      "graphics"     "stats"        "Matrix"       "splines"     
#> [21] "nlme"         "lattice"      "plyr"         "Rcpp"         "stringr"     
#> [26] "glue"         "stringi"      "tools"        "magrittr"     "labeling"    
#> [31] "munsell"      "colorspace"   "R6"           "digest"       "rlang"       
#> [36] "tibble"       "withr"        "cli"          "fansi"        "pillar"      
#> [41] "assertthat"   "crayon"       "utf8"         "vctrs"        "backports"   
#> [46] "ellipsis"     "zeallot"      "pkgconfig"    "sp"           "data.table"  
#> [51] "fpCompare"    "raster"       "rgdal"       
#> 
#> $archivist
#>  [1] "RCurl"       "digest"      "httr"        "DBI"         "lubridate"  
#>  [6] "RSQLite"     "magrittr"    "flock"       "bitops"      "methods"    
#> [11] "stats"       "graphics"    "grDevices"   "utils"       "curl"       
#> [16] "jsonlite"    "mime"        "openssl"     "R6"          "tools"      
#> [21] "askpass"     "sys"         "stringr"     "Rcpp"        "glue"       
#> [26] "stringi"     "bit64"       "blob"        "memoise"     "pkgconfig"  
#> [31] "BH"          "plogr"       "bit"         "prettyunits" "rlang"      
#> [36] "vctrs"       "assertthat"  "backports"   "ellipsis"    "zeallot"    
#> 
#> $googledrive
#>  [1] "gargle"     "uuid"       "fs"         "withr"      "methods"   
#>  [6] "Rcpp"       "stats"      "graphics"   "grDevices"  "utils"     
#> [11] "glue"       "httr"       "jsonlite"   "rlang"      "curl"      
#> [16] "mime"       "openssl"    "R6"         "tools"      "askpass"   
#> [21] "sys"        "magrittr"   "purrr"      "tibble"     "cli"       
#> [26] "fansi"      "pillar"     "assertthat" "crayon"     "utf8"      
#> [31] "vctrs"      "backports"  "ellipsis"   "zeallot"    "digest"    
#> [36] "pkgconfig" 
#> 
#> $dplyr
#>  [1] "tibble"     "tidyselect" "cli"        "fansi"      "pillar"    
#>  [6] "assertthat" "crayon"     "methods"    "utils"      "tools"     
#> [11] "grDevices"  "stats"      "graphics"   "utf8"       "rlang"     
#> [16] "vctrs"      "backports"  "ellipsis"   "zeallot"    "digest"    
#> [21] "glue"       "pkgconfig"  "purrr"      "magrittr"   "Rcpp"      
#> [26] "R6"         "BH"         "plogr"     
#> 
#> $gdalUtils
#>  [1] "foreach"     "R.utils"     "rgdal"       "codetools"   "iterators"  
#>  [6] "utils"       "R.methodsS3" "R.oo"        "methods"     "stats"      
#> [11] "graphics"    "grDevices"   "tools"       "sp"          "lattice"    
#> [16] "grid"        "raster"      "Rcpp"       
#> 
#> $sf
#>  [1] "classInt"   "units"      "e1071"      "class"      "KernSmooth"
#>  [6] "graphics"   "grDevices"  "stats"      "methods"    "utils"     
#> [11] "MASS"       "Rcpp"       "DBI"        "grid"       "magrittr"  
#> [16] "tools"     
#> 
#> $testthat
#>  [1] "evaluate"   "praise"     "methods"    "stats"      "graphics"  
#>  [6] "grDevices"  "utils"      "cli"        "crayon"     "digest"    
#> [11] "magrittr"   "R6"         "rlang"      "withr"      "assertthat"
#> [16] "tools"     
#> 
#> $httr
#>  [1] "curl"      "jsonlite"  "mime"      "openssl"   "R6"        "methods"  
#>  [7] "stats"     "graphics"  "grDevices" "utils"     "tools"     "askpass"  
#> [13] "sys"      
#> 
#> $fasterize
#>  [1] "raster"        "sp"            "RcppArmadillo" "Rcpp"         
#>  [5] "methods"       "utils"         "stats"         "graphics"     
#>  [9] "grDevices"     "lattice"       "grid"         
#> 
#> $R.utils
#> [1] "R.methodsS3" "R.oo"        "utils"       "methods"     "stats"      
#> [6] "graphics"    "grDevices"   "tools"      
#> 
#> $rgdal
#> [1] "grDevices" "graphics"  "stats"     "utils"     "methods"   "sp"       
#> [7] "lattice"   "grid"     
#> 
#> $rgeos
#> [1] "utils"     "stats"     "graphics"  "methods"   "sp"        "grDevices"
#> [7] "lattice"   "grid"     
#> 
#> $raster
#> [1] "Rcpp"      "methods"   "sp"        "utils"     "stats"     "graphics" 
#> [7] "grDevices"
#> 
#> $sp
#> [1] "lattice"   "grid"      "grDevices" "graphics"  "stats"     "utils"    
#> [7] "methods"  
#> 
#> $RCurl
#> [1] "bitops"    "methods"   "stats"     "graphics"  "grDevices" "utils"    
#> 
#> $remotes
#> [1] "methods"   "stats"     "tools"     "utils"     "graphics"  "grDevices"
#> 
#> $crayon
#> [1] "grDevices" "methods"   "utils"     "stats"     "graphics" 
#> 
#> $data.table
#> [1] "methods"   "stats"     "graphics"  "grDevices" "utils"    
#> 
#> $methods
#> [1] "stats"     "graphics"  "grDevices" "utils"    
#> 
#> $git2r
#> [1] "graphics"  "utils"     "grDevices"
#> 
#> $parallel
#> [1] "compiler" "tools"   
#> 
#> $backports
#> [1] "utils"
#> 
#> $memoise
#> [1] "digest"
#> 
#> $digest
#> character(0)
#> 
#> $fastdigest
#> character(0)
#> 
#> $fpCompare
#> character(0)
#> 
#> $magrittr
#> character(0)
#> 
#> $rlang
#> character(0)
#> 
#> $tools
#> character(0)
#> 
#> $utils
#> character(0)
#> 
#> $versions
#> character(0)
#>