Take a snapshot of all the packages and version numbers

This can be used later by installVersions to install or re-install the correct versions.

pkgSnapshot(packageVersionFile, libPath, standAlone = FALSE)

Arguments

packageVersionFile
A filename to save the packages and their currently installed version numbers. Defaults to ".packageVersions.txt".

libPath
The path to the local library where packages are installed. Defaults to the .libPaths()[1]

standAlone
Logical. If TRUE, all packages will be installed and loaded strictly from the libPaths only. If FALSE, all .libPaths will be used to find the correct versions. This can be create dramatically faster installs if the user has a substantial number of the packages already in their personal library. In the case of TRUE, there will be a hidden file place in the libPath directory that lists all the packages that were needed during the Require call. Default FALSE to minimize package installing.

Details

A file is written with the package names and versions of all packages within libPath. This can later be passed to Require.

Examples

pkgSnapFile <- tempfile()
pkgSnapshot(pkgSnapFile, .libPaths()[1])

#> There is no /tmp/RtmpXMlGMn/reproducible/cache/.Require/_tmp_RtmpBVXgYu_temp_libpath148974c3af400._packageVersionsAuto.txt and standAlone is FALSE. This snapshot will not be accurate because it will include all packages in /tmp/RtmpBVXgYu/temp_libpath148974c3af400, /home/achubaty/R/x86_64-pc-linux-gnu-library/3.6, /usr/local/lib/R/site-library, /usr/lib/R/site-library, /usr/lib/R/library

#>              instPkgs    instVers
#> 1        reproducible       1.1.1
#> 2        reproducible       1.1.1
#> 3               abind       1.4-5
#> 4             acepack       1.4.1
#> 5                 amc  0.1.6.9000
#> 6           animation         2.6
#> 7                 ape         5.3
#> 8                 aqp        1.19
#> 9           archivist       2.3.4
#> 10                arm      1.11-1
#> 11            askpass         1.1
#> 12         assertthat       0.2.1
#> 13          available       1.0.4
#> 14          backports       1.1.7
#> 15          base64enc       0.1-4
#> 16         bayestestR       0.6.0
#> 17           beanplot         1.2
#> 18                 BH    1.72.0-3
#> 19       BiocGenerics      0.32.0
#> 20        BiocManager     1.30.10
#> 21        BiocVersion      3.10.1
#> 22                bit    1.1-15.2
#> 23              bit64       0.9-7
#> 24             bitops       1.0-6
#> 25               blme       1.0-4
#> 26               blob       1.2.1
#> 27               brew       1.0-6
#> 28              broom       0.5.6
#> 29              callr       3.4.3
#> 30                car       3.0-7
#> 31            carData       3.0-3
#> 32         cellranger       1.1.0
#> 33          checkmate       2.0.0
#> 34          CircStats       0.2-6
#> 35           classInt       0.4-3
#> 36                cli       2.0.2
#> 37              clipr       0.7.0
#> 38         clisymbols       1.2.0
#> 39               coda      0.19-3
#> 40        CodeDepends       0.6.5
#> 41         colorRamps         2.3
#> 42         colorspace       1.4-1
#> 43         commonmark         1.7
#> 44             config         0.3
#> 45               covr       3.5.0
#> 46            cowplot       1.0.0
#> 47          crancache  0.0.0.9001
#> 48           cranlike       1.0.2
#> 49             crayon  1.3.4.9000
#> 50          crosstalk     1.1.0.1
#> 51               crul       0.9.0
#> 52               curl         4.3
#> 53          cyclocomp       1.1.0
#> 54         data.table      1.12.8
#> 55                DBI       1.1.0
#> 56            debugme       1.1.0
#> 57             deldir      0.1-25
#> 58            DEoptim       2.2-5
#> 59               desc       1.2.0
#> 60           devtools       2.3.0
#> 61         DiagrammeR     1.0.6.1
#> 62      DiagrammeRsvg         0.1
#> 63             digest      0.6.25
#> 64              dismo       1.1-4
#> 65         doParallel      1.0.15
#> 66          dotCall64       1.0-0
#> 67         downloader         0.4
#> 68              dplyr       0.8.5
#> 69                 DT        0.13
#> 70              e1071       1.7-3
#> 71            effsize       0.8.0
#> 72           ellipsis       0.3.1
#> 73           EnvStats       2.3.1
#> 74           evaluate      0.14.1
#> 75      exactextractr       0.3.0
#> 76               expm     0.999-4
#> 77              fansi       0.4.1
#> 78             farver       2.0.3
#> 79         fastdigest       0.6-3
#> 80          fasterize       1.0.2
#> 81            fastmap       1.0.1
#> 82          fastmatch       1.1-1
#> 83            fastshp       0.1-2
#> 84            fBasics   3042.89.1
#> 85                 ff    2.2-14.2
#> 86             ffbase      0.12.8
#> 87             fields        10.3
#> 88     fireSenseUtils  0.0.0.9025
#> 89              flock         0.7
#> 90                FNN       1.1.3
#> 91            forcats       0.5.0
#> 92            foreach       1.5.0
#> 93            Formula       1.2-3
#> 94          fpCompare       0.2.3
#> 95                 fs       1.4.1
#> 96             future      1.17.0
#> 97       future.apply       1.5.0
#> 98       future.callr       0.5.0
#> 99             gamlss       5.1-6
#> 100       gamlss.data       5.1-4
#> 101       gamlss.dist       5.1-6
#> 102            gargle       0.5.0
#> 103         gdalUtils     2.0.3.2
#> 104             gdata      2.18.0
#> 105          generics       0.0.2
#> 106           geojson       0.3.2
#> 107         geojsonio       0.9.2
#> 108       geojsonlint       0.4.0
#> 109           ggimage       0.2.8
#> 110           ggplot2       3.3.0
#> 111         ggplotify       0.0.5
#> 112          ggthemes       4.2.0
#> 113                gh       1.1.0
#> 114             git2r      0.27.1
#> 115              glmm       1.3.0
#> 116           globals      0.12.5
#> 117              glue       1.4.1
#> 118            gmailr       1.0.0
#> 119           gmodels      2.18.1
#> 120           goftest       1.2-2
#> 121       googleAuthR       1.2.1
#> 122       googledrive       1.0.1
#> 123          googleID    0.0.9001
#> 124        grainscape       0.4.3
#> 125             graph      1.64.0
#> 126          gridBase       0.4-7
#> 127         gridExtra         2.3
#> 128      gridGraphics       0.5-0
#> 129               gss       2.2-1
#> 130             gstat       2.0-6
#> 131            gtable       0.3.0
#> 132            gtools       3.8.2
#> 133            gurobi       9.0-1
#> 134             haven       2.2.0
#> 135            hexbin      1.28.1
#> 136        hexSticker       0.4.6
#> 137         highlight       0.5.0
#> 138             highr       0.8.1
#> 139             Hmisc       4.4-0
#> 140               hms       0.5.3
#> 141         htmlTable      1.13.3
#> 142         htmltools       0.4.0
#> 143       htmlwidgets       1.5.1
#> 144          httpcode       0.3.0
#> 145            httpuv       1.5.2
#> 146              httr       1.4.1
#> 147          hunspell         3.0
#> 148            igraph       1.2.5
#> 149        influenceR       0.1.0
#> 150               ini       0.3.1
#> 151           insight       0.8.4
#> 152         intervals      0.15.2
#> 153           isoband       0.2.1
#> 154          ISOcodes  2020.03.16
#> 155         iterators      1.0.12
#> 156         itertools       0.1-3
#> 157       janeaustenr       0.1.5
#> 158           janitor       2.0.1
#> 159              jpeg     0.1-8.1
#> 160               jqr       1.1.0
#> 161          jsonlite       1.6.1
#> 162      jsonvalidate       1.1.0
#> 163        kableExtra       1.1.0
#> 164             knitr        1.28
#> 165            koRpus      0.11-5
#> 166    koRpus.lang.en       0.1-3
#> 167          kSamples       1.2-9
#> 168          labeling         0.3
#> 169             LandR       0.0.5
#> 170          LandR.CS  0.0.2.0002
#> 171  landscapemetrics       1.4.4
#> 172    landscapetools       0.5.0
#> 173      LandWebUtils       0.1.0
#> 174             later       1.0.0
#> 175      latticeExtra      0.6-29
#> 176          lazyeval       0.2.2
#> 177           leaflet       2.0.3
#> 178    leaflet.extras       1.0.0
#> 179 leaflet.providers       1.9.0
#> 180        LearnBayes      2.15.1
#> 181         lifecycle       0.2.0
#> 182             lintr  2.0.1.9000
#> 183           listenv       0.8.0
#> 184              lme4      1.1-23
#> 185           logging    0.10-108
#> 186        lpsymphony      1.14.0
#> 187         lubridate       1.7.8
#> 188            lwgeom       0.2-3
#> 189            magick         2.3
#> 190          magrittr         1.5
#> 191               map       0.0.1
#> 192           mapdata       2.3.0
#> 193              maps       3.3.0
#> 194          maptools       1.0-1
#> 195          markdown       1.1.1
#> 196      MatrixModels       0.4-1
#> 197       matrixStats      0.56.0
#> 198           memoise       1.1.0
#> 199          merTools       0.5.0
#> 200    microbenchmark       1.4-7
#> 201              mime       0.9.1
#> 202            miniUI     0.1.1.1
#> 203             minqa       1.2.4
#> 204          multcomp      1.4-13
#> 205             MuMIn     1.43.17
#> 206           munsell       0.5.0
#> 207           mvtnorm       1.1-0
#> 208            nloptr     1.2.2.1
#> 209           nortest       1.0-4
#> 210           openssl       1.4.1
#> 211          openxlsx       4.1.5
#> 212           packrat       0.5.0
#> 213         parsedate       1.2.0
#> 214         patchwork       1.0.0
#> 215          pbkrtest     0.4-8.6
#> 216            pemisc  0.0.2.9000
#> 217       performance       0.4.6
#> 218           permute       0.9-5
#> 219            pillar       1.4.4
#> 220            pixmap      0.4-11
#> 221          pkgbuild       1.0.8
#> 222         pkgconfig       2.0.3
#> 223           pkgdown       1.5.1
#> 224           pkgload       1.0.2
#> 225             plogr       0.2.0
#> 226           plotKML       0.6-1
#> 227           plotrix       3.7-8
#> 228              plyr       1.8.6
#> 229               png       0.1-8
#> 230         polspline      1.1.19
#> 231          polyclip      1.10-0
#> 232            praise       1.0.0
#> 233        prettycode       1.1.0
#> 234       prettyunits       1.1.1
#> 235        prioritizr       5.0.1
#> 236          processx       3.4.2
#> 237           profvis       0.3.6
#> 238          progress       1.2.2
#> 239          promises       1.1.0
#> 240             proto       1.0.0
#> 241         protolite         2.1
#> 242              pryr       0.1.4
#> 243                ps       1.3.3
#> 244         PtProcess      3.3-13
#> 245             purrr       0.3.4
#> 246                qs      0.21.2
#> 247          quantreg        5.55
#> 248         quickPlot  0.1.6.9001
#> 249           R.cache      0.14.0
#> 250         R.devices      2.16.1
#> 251       R.methodsS3       1.8.0
#> 252              R.oo      1.23.0
#> 253           R.utils       2.9.2
#> 254                R6       2.4.1
#> 255      RandomFields       3.3.8
#> 256 RandomFieldsUtils       0.5.3
#> 257              RANN       2.6.1
#> 258     RApiSerialize       0.1.0
#> 259          rappdirs       0.3.1
#> 260            raster       3.1-5
#> 261         rasterVis        0.47
#> 262            Rclean       1.1.8
#> 263         rcmdcheck  1.3.3.9000
#> 264      RColorBrewer       1.1-2
#> 265              Rcpp     1.0.4.6
#> 266     RcppArmadillo 0.9.880.1.0
#> 267         RcppEigen   0.3.3.7.0
#> 268             RCurl    1.98-1.2
#> 269              rdoc       0.1.0
#> 270             readr       1.3.1
#> 271            readxl       1.3.1
#> 272           rematch       1.0.1
#> 273          rematch2       2.1.2
#> 274           remotes       2.1.1
#> 275      reproducible       1.1.0
#> 276      reproducible       1.1.0
#> 277           reshape       0.8.8
#> 278          reshape2       1.4.4
#> 279       revdepcheck  1.0.0.9001
#> 280               rex       1.2.0
#> 281             rgdal       1.4-8
#> 282             rgeos       0.5-3
#> 283         Rgraphviz      2.30.0
#> 284              rhub       1.1.1
#> 285               rio      0.5.16
#> 286             rJava      0.9-12
#> 287             rlang       0.4.6
#> 288        rmapshaper       0.4.4
#> 289         rmarkdown         2.1
#> 290               rms       5.1-4
#> 291          roxygen2       7.1.0
#> 292         RPostgres       1.2.0
#> 293         rprojroot       1.3-2
#> 294             RSAGA       1.3.0
#> 295            Rserve     1.7-3.1
#> 296           RSQLite       2.2.0
#> 297        rstudioapi        0.11
#> 298         Rsymphony      0.1-28
#> 299               rts      1.0-49
#> 300           rvcheck       0.1.8
#> 301         rversions       2.0.1
#> 302             rvest       0.3.5
#> 303          sandwich       2.5-1
#> 304            scales       1.1.1
#> 305               sdm      1.0-89
#> 306           selectr       0.4-2
#> 307       sessioninfo       1.1.1
#> 308                sf       0.9-3
#> 309        shapefiles         0.7
#> 310             shiny     1.4.0.2
#> 311           shinyBS        0.61
#> 312   shinycssloaders         0.3
#> 313    shinydashboard       0.7.1
#> 314           shinyjs         1.1
#> 315      shinyWidgets       0.5.2
#> 316          showtext         0.8
#> 317        showtextdb         2.0
#> 318            slackr       1.4.2
#> 319              slam      0.1-47
#> 320         snakecase      0.11.0
#> 321              snow       0.4-3
#> 322         SnowballC       0.7.0
#> 323       sourcetools       0.1.7
#> 324                sp       1.4-1
#> 325         spacetime       1.2-3
#> 326            SpaDES       2.0.4
#> 327     SpaDES.addins       0.1.2
#> 328       SpaDES.core       1.0.1
#> 329      SpaDES.shiny       0.2.0
#> 330      SpaDES.tools  0.3.5.9000
#> 331              spam       2.5-1
#> 332           SparseM        1.78
#> 333        spatialEco       1.3-1
#> 334       SpatialPack       0.3-8
#> 335          spatstat      1.64-1
#> 336     spatstat.data       1.4-3
#> 337    spatstat.utils      1.17-0
#> 338            spData       0.3.5
#> 339             spdep       1.1-3
#> 340          spelling         2.1
#> 341        stabledist       0.7-1
#> 342             stars       0.4-1
#> 343           statmod      1.4.34
#> 344         stopwords         2.0
#> 345        stringdist     0.9.5.5
#> 346           stringi       1.4.6
#> 347           stringr       1.4.0
#> 348            styler  1.3.2.9000
#> 349         SuppDists     1.1-9.5
#> 350             sylly       0.1-5
#> 351          sylly.en       0.1-3
#> 352               sys         3.3
#> 353          sysfonts       0.8.1
#> 354            tensor         1.5
#> 355            testit        0.11
#> 356          testthat       2.3.2
#> 357           TH.data      1.0-10
#> 358            tibble       3.0.1
#> 359            tictoc         1.0
#> 360             tidyr       1.0.3
#> 361        tidyselect       1.1.0
#> 362          tidytext       0.2.4
#> 363             tiler       0.2.4
#> 364          timeDate    3043.102
#> 365        timeSeries    3062.100
#> 366          TimeWarp      1.0.15
#> 367           tinytex        0.22
#> 368        tokenizers       0.2.1
#> 369         triebeard       0.3.0
#> 370         truncnorm       1.0-8
#> 371             trust       0.1-8
#> 372             udapi       0.1.3
#> 373             units       0.6-6
#> 374         unixtools       0.1-1
#> 375          urltools       1.7.3
#> 376        usefulFuns       0.1.0
#> 377           usethis       1.6.1
#> 378              utf8       1.1.4
#> 379              uuid       0.1-4
#> 380                V8       3.0.2
#> 381             vctrs       0.3.0
#> 382             vegan       2.5-6
#> 383             velox       0.2.0
#> 384          versions         0.3
#> 385              VGAM       1.1-3
#> 386           viridis       0.5.1
#> 387       viridisLite       0.3.0
#> 388        visNetwork       2.0.9
#> 389           webshot       0.5.2
#> 390           whisker         0.4
#> 391            whoami       1.3.0
#> 392             withr       2.2.0
#> 393    wordcountaddin  0.3.0.9000
#> 394              xfun        0.13
#> 395              xlsx       0.6.3
#> 396          xlsxjars       0.6.1
#> 397               XML    3.99-0.3
#> 398              xml2       1.3.2
#> 399      xmlparsedata       1.0.3
#> 400             xopen       1.0.0
#> 401            xtable       1.8-4
#> 402               xts      0.12-0
#> 403          yaImpute      1.0-32
#> 404              yaml       2.2.1
#> 405             yesno       0.1.1
#> 406               zip       2.0.4
#> 407               zoo       1.8-8
#> 408           littler       0.3.9
#> 409              base       3.6.3
#> 410              boot      1.3-25
#> 411             class      7.3-17
#> 412           cluster       2.1.0
#> 413         codetools      0.2-16
#> 414          compiler       3.6.3
#> 415          datasets       3.6.3
#> 416           foreign      0.8-76
#> 417          graphics       3.6.3
#> 418         grDevices       3.6.3
#> 419              grid       3.6.3
#> 420        KernSmooth     2.23-17
#> 421           lattice     0.20-41
#> 422              MASS    7.3-51.6
#> 423            Matrix      1.2-18
#> 424           methods       3.6.3
#> 425              mgcv      1.8-31
#> 426              nlme     3.1-147
#> 427              nnet      7.3-14
#> 428          parallel       3.6.3
#> 429             rpart      4.1-15
#> 430           spatial      7.3-12
#> 431           splines       3.6.3
#> 432             stats       3.6.3
#> 433            stats4       3.6.3
#> 434          survival      3.1-12
#> 435             tcltk       3.6.3
#> 436             tools       3.6.3
#> 437      translations       3.6.3
#> 438             utils       3.6.3

data.table::fread(pkgSnapFile)

#>          instPkgs   instVers
#>   1: reproducible      1.1.1
#>   2: reproducible      1.1.1
#>   3:        abind      1.4-5
#>   4:      acepack      1.4.1
#>   5:          amc 0.1.6.9000
#>  ---                        
#> 434:     survival     3.1-12
#> 435:        tcltk      3.6.3
#> 436:        tools      3.6.3
#> 437: translations      3.6.3
#> 438:        utils      3.6.3