Take a snapshot of all the packages and version numbers

This can be used later by installVersions to install or re-install the correct versions.

pkgSnapshot(packageVersionFile, libPath, standAlone = FALSE)

Arguments

packageVersionFile
A filename to save the packages and their currently installed version numbers. Defaults to ".packageVersions.txt".

libPath
The path to the local library where packages are installed. Defaults to the .libPaths()[1]

standAlone
Logical. If TRUE, all packages will be installed and loaded strictly from the libPaths only. If FALSE, all .libPaths will be used to find the correct versions. This can be create dramatically faster installs if the user has a substantial number of the packages already in their personal library. In the case of TRUE, there will be a hidden file place in the libPath directory that lists all the packages that were needed during the Require call. Default FALSE to minimize package installing.

Details

A file is written with the package names and versions of all packages within libPath. This can later be passed to Require.

Examples

pkgSnapFile <- tempfile()
pkgSnapshot(pkgSnapFile, .libPaths()[1])

#> There is no C:/Users/emcintir/AppData/Local/Temp/RtmpG8MQuU/reproducible/cache/.Require/C_Users_emcintir_AppData_Local_Temp_Rtmp4eGUga_temp_libpath29206a16559f._packageVersionsAuto.txt and standAlone is FALSE. This snapshot will not be accurate because it will include all packages in C:/Users/emcintir/AppData/Local/Temp/Rtmp4eGUga/temp_libpath29206a16559f, C:/Eliot/R/win-library/3.6, C:/Program Files/R/R-3.6.1/library

#>              instPkgs    instVers
#> 1        reproducible       1.0.0
#> 2        reproducible       1.0.0
#> 3                 amc  0.1.5.9000
#> 4           animation         2.6
#> 5           archivist       2.3.4
#> 6             askpass         1.1
#> 7          assertthat       0.2.1
#> 8           backports       1.1.5
#> 9           base64enc       0.1-3
#> 10                 BH    1.72.0-3
#> 11              bindr       0.1.1
#> 12           bindrcpp       0.2.2
#> 13        BiocManager     1.30.10
#> 14                bit    1.1-15.1
#> 15              bit64       0.9-7
#> 16             bitops       1.0-6
#> 17               blob       1.2.1
#> 18               boot      1.3-24
#> 19               boot      1.3-24
#> 20               brew       1.0-6
#> 21              broom       0.5.3
#> 22              callr       3.4.1
#> 23            caTools    1.17.1.4
#> 24         cellranger       1.1.0
#> 25          checkmate       1.9.4
#> 26          CircStats       0.2-6
#> 27           classInt       0.4-2
#> 28                cli       2.0.1
#> 29              clipr       0.7.0
#> 30         clisymbols       1.2.0
#> 31         colorspace       1.4-1
#> 32         commonmark         1.7
#> 33             config         0.3
#> 34               covr       3.4.0
#> 35             crayon       1.3.4
#> 36          crosstalk       1.0.0
#> 37               crul       0.9.0
#> 38               curl         4.3
#> 39         data.table      1.12.8
#> 40                DBI       1.1.0
#> 41             dbplyr       1.4.2
#> 42            DEoptim       2.2-4
#> 43               desc       1.2.0
#> 44           devtools       2.2.1
#> 45            diagram       1.6.4
#> 46         DiagrammeR       1.0.5
#> 47             digest      0.6.23
#> 48              dismo       1.1-4
#> 49         doParallel      1.0.15
#> 50         downloader         0.4
#> 51              dplyr       0.8.3
#> 52                 DT        0.11
#> 53              e1071       1.7-3
#> 54            effsize       0.7.6
#> 55           ellipsis       0.3.0
#> 56           evaluate        0.14
#> 57              fansi       0.4.1
#> 58             farver       2.0.3
#> 59         fastdigest       0.6-3
#> 60          fasterize       1.0.0
#> 61            fastmap       1.0.1
#> 62          fastmatch       1.1-0
#> 63            fastshp       0.1-2
#> 64                 ff      2.2-14
#> 65             ffbase      0.12.7
#> 66              flock         0.7
#> 67               fmsb       0.7.0
#> 68            forcats       0.4.0
#> 69            foreach       1.4.7
#> 70            foreign      0.8-74
#> 71            foreign      0.8-74
#> 72          fpCompare       0.2.3
#> 73                 fs       1.3.1
#> 74             future      1.16.0
#> 75       future.apply       1.4.0
#> 76       future.callr       0.5.0
#> 77             gamlss       5.1-5
#> 78        gamlss.data       5.1-4
#> 79        gamlss.dist       5.1-5
#> 80             gargle       0.4.0
#> 81          gdalUtils       2.0.3
#> 82           generics       0.0.2
#> 83            geojson       0.3.2
#> 84          geojsonio       0.8.0
#> 85        geojsonlint       0.3.0
#> 86            ggimage       0.2.7
#> 87            ggplot2       3.2.1
#> 88          ggplotify       0.0.4
#> 89                 gh       1.1.0
#> 90              git2r      0.26.1
#> 91               glmm       1.3.0
#> 92            globals      0.12.5
#> 93               glue       1.3.1
#> 94        googleAuthR       1.1.1
#> 95        googledrive       1.0.0
#> 96           googleID    0.0.9001
#> 97           gridBase       0.4-7
#> 98          gridExtra         2.3
#> 99       gridGraphics       0.4-1
#> 100            gtable       0.3.0
#> 101             haven       2.2.0
#> 102            hexbin      1.28.0
#> 103        hexSticker       0.4.6
#> 104         highlight       0.5.0
#> 105             highr         0.8
#> 106               hms       0.5.3
#> 107         htmltools       0.4.0
#> 108       htmlwidgets       1.5.1
#> 109          httpcode       0.2.0
#> 110            httpuv       1.5.2
#> 111              httr       1.4.1
#> 112          hunspell         3.0
#> 113            igraph     1.2.4.2
#> 114        influenceR       0.1.0
#> 115               ini       0.3.1
#> 116            inline      0.3.15
#> 117          installr      0.22.0
#> 118         iterators      1.0.12
#> 119         itertools       0.1-3
#> 120              jpeg     0.1-8.1
#> 121               jqr       1.1.0
#> 122          jsonlite         1.6
#> 123      jsonvalidate       1.1.0
#> 124        KernSmooth     2.23-16
#> 125        KernSmooth     2.23-16
#> 126             knitr        1.27
#> 127          kSamples       1.2-9
#> 128          labeling         0.3
#> 129             LandR  0.0.3.9001
#> 130          LandR.CS  0.0.0.9001
#> 131      LandWebUtils       0.0.2
#> 132             later       1.0.0
#> 133      latticeExtra      0.6-29
#> 134          lazyeval       0.2.2
#> 135             lazyR       0.0.9
#> 136           leaflet       2.0.3
#> 137    leaflet.extras       1.0.0
#> 138 leaflet.providers       1.9.0
#> 139         lifecycle       0.1.0
#> 140           listenv       0.8.0
#> 141              lme4      1.1-21
#> 142           logging    0.10-108
#> 143         lubridate       1.7.4
#> 144            magick         2.2
#> 145          magrittr         1.5
#> 146               map       0.0.1
#> 147          maptools       0.9-9
#> 148          markdown         1.1
#> 149              MASS    7.3-51.5
#> 150              MASS    7.3-51.5
#> 151            Matrix      1.2-18
#> 152            Matrix      1.2-18
#> 153           memoise       1.1.0
#> 154              meow       0.1.4
#> 155              mgcv      1.8-31
#> 156              mgcv      1.8-31
#> 157    microbenchmark       1.4-7
#> 158              mime         0.8
#> 159            miniUI     0.1.1.1
#> 160             minqa       1.2.4
#> 161            modelr       0.1.5
#> 162             MuMIn     1.43.15
#> 163           munsell       0.5.0
#> 164           mvtnorm      1.0-11
#> 165         networkD3         0.4
#> 166              nlme     3.1-143
#> 167              nlme     3.1-143
#> 168            nloptr       1.2.1
#> 169           openssl       1.4.1
#> 170           packrat       0.5.0
#> 171             pedev  0.0.1.9005
#> 172            pemisc  0.0.1.9004
#> 173            pillar       1.4.3
#> 174              pins       0.3.1
#> 175          pkgbuild       1.0.6
#> 176         pkgconfig       2.0.3
#> 177           pkgdown       1.4.1
#> 178           pkgload       1.0.2
#> 179             plogr       0.2.0
#> 180              plyr       1.8.5
#> 181               png       0.1-7
#> 182            praise       1.0.0
#> 183       prettyunits       1.1.1
#> 184          processx       3.4.1
#> 185           profvis       0.3.6
#> 186          progress       1.2.2
#> 187          promises       1.1.0
#> 188         protolite         2.1
#> 189              pryr       0.1.4
#> 190                ps       1.3.0
#> 191             purrr       0.3.3
#> 192                qs      0.20.2
#> 193         quickPlot  0.1.6.9000
#> 194         R.devices      2.16.1
#> 195       R.methodsS3       1.7.1
#> 196              R.oo      1.23.0
#> 197           R.utils       2.9.2
#> 198                R6       2.4.1
#> 199      RandomFields       3.3.7
#> 200 RandomFieldsUtils       0.5.3
#> 201     RApiSerialize       0.1.0
#> 202          rappdirs       0.3.1
#> 203            raster       3.0-7
#> 204         rasterVis        0.47
#> 205         rcmdcheck       1.3.3
#> 206      RColorBrewer       1.1-2
#> 207              Rcpp       1.0.3
#> 208     RcppArmadillo 0.9.800.4.0
#> 209         RcppEigen   0.3.3.7.0
#> 210             RCurl    1.98-1.1
#> 211       RDCOMClient      0.94-0
#> 212             readr       1.3.1
#> 213            readxl       1.3.1
#> 214           rematch       1.0.1
#> 215          rematch2       2.1.0
#> 216           remotes       2.1.0
#> 217            reprex       0.3.0
#> 218      reproducible       1.0.0
#> 219      reproducible       1.0.0
#> 220          reshape2       1.4.3
#> 221               rex       1.1.2
#> 222             rgdal       1.4-8
#> 223             rgeos       0.5-2
#> 224             rgexf      0.15.3
#> 225             rJava      0.9-11
#> 226             rlang       0.4.3
#> 227        rmapshaper       0.4.2
#> 228         rmarkdown         2.0
#> 229              Rook       1.1-1
#> 230          roxygen2       7.0.2
#> 231         RPostgres       1.2.0
#> 232         rprojroot       1.3-2
#> 233           RSQLite       2.2.0
#> 234        rstudioapi        0.10
#> 235           rvcheck       0.1.7
#> 236         rversions       2.0.1
#> 237             rvest       0.3.5
#> 238            scales       1.1.0
#> 239              scam       1.2-5
#> 240           selectr       0.4-2
#> 241       sessioninfo       1.1.1
#> 242                sf       0.8-0
#> 243             shape       1.4.4
#> 244             shiny       1.4.0
#> 245           shinyBS        0.61
#> 246   shinycssloaders         0.3
#> 247    shinydashboard       0.7.1
#> 248           shinyjs         1.1
#> 249      shinyWidgets       0.5.0
#> 250          showtext         0.7
#> 251        showtextdb         2.0
#> 252            slackr       1.4.2
#> 253              snow       0.4-3
#> 254       sourcetools       0.1.7
#> 255                sp       1.3-2
#> 256            SpaDES       2.0.4
#> 257     SpaDES.addins       0.1.2
#> 258       SpaDES.core  0.2.9.9000
#> 259 SpaDES.experiment  0.0.2.9000
#> 260      SpaDES.tools  0.3.4.9000
#> 261  SpaDES.Workshops       0.1.0
#> 262          spelling         2.1
#> 263           startup      0.14.0
#> 264           stringi       1.4.4
#> 265           stringr       1.4.0
#> 266         SuppDists     1.1-9.5
#> 267          survival       3.1-8
#> 268          survival       3.1-8
#> 269               sys         3.3
#> 270          sysfonts         0.8
#> 271          testthat       2.3.1
#> 272            tibble       2.1.3
#> 273             tidyr       1.0.2
#> 274        tidyselect       0.2.5
#> 275         tidyverse       1.3.0
#> 276             tiler       0.2.4
#> 277          TimeWarp      1.0.15
#> 278           tinytex        0.19
#> 279         triebeard       0.3.0
#> 280             trust       0.1-8
#> 281             units       0.6-5
#> 282          urltools       1.7.3
#> 283            usefun       0.1.0
#> 284           usethis       1.5.1
#> 285              utf8       1.1.4
#> 286              uuid       0.1-2
#> 287                V8         3.0
#> 288             vctrs       0.2.2
#> 289             velox       0.2.0
#> 290          versions         0.3
#> 291              VGAM       1.1-2
#> 292           viridis       0.5.1
#> 293       viridisLite       0.3.0
#> 294        visNetwork       2.0.9
#> 295           whisker         0.4
#> 296             withr       2.1.2
#> 297              xfun        0.12
#> 298              xlsx       0.6.1
#> 299          xlsxjars       0.6.1
#> 300               XML   3.98-1.20
#> 301              xml2       1.2.2
#> 302             xopen       1.0.0
#> 303            xtable       1.8-4
#> 304              yaml       2.2.0
#> 305           zeallot       0.1.0
#> 306               zip       2.0.4
#> 307               zoo       1.8-6
#> 308              base       3.6.1
#> 309              boot      1.3-22
#> 310              boot      1.3-22
#> 311             class      7.3-15
#> 312           cluster       2.1.0
#> 313         codetools      0.2-16
#> 314          compiler       3.6.1
#> 315          datasets       3.6.1
#> 316           foreign      0.8-71
#> 317           foreign      0.8-71
#> 318          graphics       3.6.1
#> 319         grDevices       3.6.1
#> 320              grid       3.6.1
#> 321        KernSmooth     2.23-15
#> 322        KernSmooth     2.23-15
#> 323           lattice     0.20-38
#> 324              MASS    7.3-51.4
#> 325              MASS    7.3-51.4
#> 326            Matrix      1.2-17
#> 327            Matrix      1.2-17
#> 328           methods       3.6.1
#> 329              mgcv      1.8-28
#> 330              mgcv      1.8-28
#> 331              nlme     3.1-140
#> 332              nlme     3.1-140
#> 333              nnet      7.3-12
#> 334          parallel       3.6.1
#> 335             rpart      4.1-15
#> 336           spatial      7.3-11
#> 337           splines       3.6.1
#> 338             stats       3.6.1
#> 339            stats4       3.6.1
#> 340          survival    2.44-1.1
#> 341          survival    2.44-1.1
#> 342             tcltk       3.6.1
#> 343             tools       3.6.1
#> 344      translations       3.6.1
#> 345             utils       3.6.1

data.table::fread(pkgSnapFile)

#>          instPkgs   instVers
#>   1: reproducible      1.0.0
#>   2: reproducible      1.0.0
#>   3:          amc 0.1.5.9000
#>   4:    animation        2.6
#>   5:    archivist      2.3.4
#>  ---                        
#> 341:     survival   2.44-1.1
#> 342:        tcltk      3.6.1
#> 343:        tools      3.6.1
#> 344: translations      3.6.1
#> 345:        utils      3.6.1