Take a snapshot of all the packages and version numbers

This can be used later by installVersions to install or re-install the correct versions.

pkgSnapshot(packageVersionFile, libPath, standAlone = FALSE)

Arguments

packageVersionFile
A filename to save the packages and their currently installed version numbers. Defaults to ".packageVersions.txt".

libPath
The path to the local library where packages are installed. Defaults to the .libPaths()[1]

standAlone
Logical. If TRUE, all packages will be installed and loaded strictly from the libPaths only. If FALSE, all .libPaths will be used to find the correct versions. This can be create dramatically faster installs if the user has a substantial number of the packages already in their personal library. In the case of TRUE, there will be a hidden file place in the libPath directory that lists all the packages that were needed during the Require call. Default FALSE to minimize package installing.

Details

A file is written with the package names and versions of all packages within libPath. This can later be passed to Require.

Examples

pkgSnapFile <- tempfile()
pkgSnapshot(pkgSnapFile, .libPaths()[1])

#> There is no /home/achubaty/R/x86_64-pc-linux-gnu-library/3.4/._packageVersionsAuto.txt and standAlone is FALSE. This snapshot will not be accurate because it will include all packages in /home/achubaty/R/x86_64-pc-linux-gnu-library/3.4, /usr/local/lib/R/site-library, /usr/lib/R/site-library, /usr/lib/R/library

#>                            instPkgs    instVers
#> blogdown                   blogdown         0.8
#> bookdown                   bookdown         0.7
#> callr                         callr       2.0.4
#> fs                               fs       1.2.3
#> highlight                 highlight     0.4.7.2
#> pkgdown                     pkgdown       1.1.0
#> PKI                             PKI     0.1-5.1
#> processx                   processx       3.1.0
#> rematch                     rematch       1.0.1
#> RJSONIO                     RJSONIO       1.3-0
#> rmarkdown                 rmarkdown        1.10
#> rsconnect                 rsconnect       0.8.8
#> servr                         servr        0.10
#> tinytex                     tinytex         0.6
#> xfun                           xfun         0.3
#> archivist                 archivist       2.3.1
#> assertthat               assertthat       0.2.0
#> backports                 backports       1.1.2
#> base64enc                 base64enc       0.1-3
#> BH                               BH    1.66.0-1
#> bindr                         bindr       0.1.1
#> bindrcpp                   bindrcpp       0.2.2
#> bit                             bit      1.1-14
#> bit64                         bit64       0.9-7
#> bitops                       bitops       1.0-6
#> blob                           blob       1.1.1
#> brew                           brew       1.0-6
#> Cairo                         Cairo       1.5-9
#> checkmate                 checkmate       1.8.5
#> CircStats                 CircStats       0.2-6
#> classInt                   classInt       0.2-3
#> cli                             cli       1.0.0
#> colorspace               colorspace       1.3-2
#> commonmark               commonmark         1.5
#> covr                           covr       3.1.0
#> crayon                       crayon       1.3.4
#> crosstalk                 crosstalk       1.0.0
#> curl                           curl         3.2
#> data.table               data.table      1.11.4
#> DBI                             DBI       1.0.0
#> DEoptim                     DEoptim       2.2-4
#> desc                           desc       1.2.0
#> devtools                   devtools      1.13.6
#> DiagrammeR               DiagrammeR       1.0.0
#> dichromat                 dichromat       2.0-0
#> digest                       digest      0.6.15
#> downloader               downloader         0.4
#> dplyr                         dplyr       0.7.6
#> DT                               DT         0.4
#> e1071                         e1071       1.6-8
#> evaluate                   evaluate        0.11
#> fansi                         fansi       0.2.3
#> fastdigest               fastdigest       0.6-3
#> fasterize                 fasterize       1.0.0
#> fastmatch                 fastmatch       1.1-0
#> ff                               ff      2.2-14
#> ffbase                       ffbase      0.12.3
#> flock                         flock         0.7
#> foreach                     foreach       1.4.4
#> fpCompare                 fpCompare       0.2.2
#> future                       future       1.9.0
#> gdalUtils                 gdalUtils    2.0.1.14
#> geojson                     geojson       0.2.0
#> geojsonio                 geojsonio       0.6.0
#> geojsonlint             geojsonlint       0.2.0
#> ggplot2                     ggplot2       3.0.0
#> git2r                         git2r      0.23.0
#> globals                     globals      0.12.1
#> glue                           glue       1.3.0
#> googleAuthR             googleAuthR       0.6.3
#> googledrive             googledrive       0.1.1
#> googleID                   googleID    0.0.9001
#> gridBase                   gridBase       0.4-7
#> gridExtra                 gridExtra         2.3
#> gtable                       gtable       0.2.0
#> hexbin                       hexbin      1.27.2
#> highr                         highr         0.7
#> hms                             hms       0.4.2
#> htmltools                 htmltools       0.3.6
#> htmlwidgets             htmlwidgets         1.2
#> httpuv                       httpuv       1.4.5
#> httr                           httr       1.3.1
#> igraph                       igraph       1.2.1
#> influenceR               influenceR       0.1.0
#> iterators                 iterators      1.0.10
#> jqr                             jqr       1.0.0
#> jsonlite                   jsonlite         1.5
#> jsonvalidate           jsonvalidate       1.0.0
#> knitr                         knitr        1.20
#> labeling                   labeling         0.3
#> later                         later       0.7.3
#> latticeExtra           latticeExtra      0.6-28
#> lazyeval                   lazyeval       0.2.1
#> leaflet                     leaflet       2.0.1
#> lintr                         lintr       1.0.2
#> listenv                     listenv       0.7.0
#> lubridate                 lubridate       1.7.4
#> magrittr                   magrittr         1.5
#> maptools                   maptools       0.9-2
#> markdown                   markdown         0.8
#> memoise                     memoise       1.1.0
#> mime                           mime         0.5
#> miniCRAN                   miniCRAN      0.2.11
#> miniUI                       miniUI     0.1.1.1
#> munsell                     munsell       0.5.0
#> openssl                     openssl       1.0.1
#> packrat                     packrat     0.4.9-3
#> pillar                       pillar       1.3.0
#> pkgconfig                 pkgconfig       2.0.1
#> plogr                         plogr       0.2.0
#> plyr                           plyr       1.8.4
#> png                             png       0.1-7
#> praise                       praise       1.0.0
#> prettyunits             prettyunits       1.0.2
#> promises                   promises       1.0.1
#> protolite                 protolite         1.8
#> purrr                         purrr       0.2.5
#> quickPlot                 quickPlot       0.1.4
#> R.devices                 R.devices      2.16.0
#> R.methodsS3             R.methodsS3       1.7.1
#> R.oo                           R.oo      1.22.0
#> R.utils                     R.utils       2.6.0
#> R6                               R6       2.2.2
#> RandomFields           RandomFields      3.1.50
#> RandomFieldsUtils RandomFieldsUtils      0.3.25
#> raster                       raster       2.6-7
#> rasterVis                 rasterVis        0.45
#> RColorBrewer           RColorBrewer       1.1-2
#> Rcpp                           Rcpp     0.12.18
#> RcppArmadillo         RcppArmadillo 0.8.600.0.0
#> RCurl                         RCurl   1.95-4.11
#> readr                         readr       1.1.1
#> reproducible           reproducible       0.2.2
#> reshape2                   reshape2       1.4.3
#> rex                             rex       1.1.2
#> rgdal                         rgdal       1.3-3
#> rgeos                         rgeos      0.3-28
#> rgexf                         rgexf      0.15.3
#> rlang                         rlang       0.2.1
#> rmapshaper               rmapshaper       0.4.0
#> Rook                           Rook       1.1-1
#> roxygen2                   roxygen2       6.0.1
#> rprojroot                 rprojroot       1.3-2
#> RSQLite                     RSQLite       2.1.1
#> rstudioapi               rstudioapi         0.7
#> scales                       scales       0.5.0
#> sf                               sf       0.6-3
#> shiny                         shiny       1.1.0
#> shinyBS                     shinyBS        0.61
#> shinycssloaders     shinycssloaders       0.2.0
#> shinydashboard       shinydashboard       0.7.0
#> shinyjs                     shinyjs         1.0
#> shinyWidgets           shinyWidgets       0.4.3
#> sourcetools             sourcetools       0.1.7
#> sp                               sp       1.3-1
#> SpaDES                       SpaDES       2.0.2
#> SpaDES.addins         SpaDES.addins       0.1.1
#> SpaDES.core             SpaDES.core       0.2.1
#> SpaDES.shiny           SpaDES.shiny  0.1.0.9006
#> SpaDES.tools           SpaDES.tools       0.3.0
#> spData                       spData     0.2.9.0
#> stringdist               stringdist     0.9.5.1
#> stringi                     stringi       1.2.4
#> stringr                     stringr       1.3.1
#> testthat                   testthat       2.0.0
#> tibble                       tibble       1.4.2
#> tidyr                         tidyr       0.8.1
#> tidyselect               tidyselect       0.2.4
#> units                         units       0.6-0
#> utf8                           utf8       1.1.4
#> uuid                           uuid       0.1-2
#> V8                               V8         1.5
#> velox                         velox       0.2.0
#> versions                   versions         0.3
#> VGAM                           VGAM       1.0-5
#> viridis                     viridis       0.5.1
#> viridisLite             viridisLite       0.3.0
#> visNetwork               visNetwork       2.0.4
#> whisker                     whisker       0.3-2
#> withr                         withr       2.1.2
#> xlsx                           xlsx       0.6.1
#> xlsxjars                   xlsxjars       0.6.1
#> XML                             XML   3.98-1.12
#> xml2                           xml2       1.2.0
#> xtable                       xtable       1.8-2
#> yaml                           yaml      2.1.19
#> zoo                             zoo       1.8-3
#> rJava                         rJava       0.9-9
#> base                           base       3.4.4
#> boot                           boot      1.3-20
#> class                         class      7.3-14
#> cluster                     cluster       2.0.6
#> codetools                 codetools      0.2-15
#> compiler                   compiler       3.4.4
#> datasets                   datasets       3.4.4
#> foreign                     foreign      0.8-69
#> graphics                   graphics       3.4.4
#> grDevices                 grDevices       3.4.4
#> grid                           grid       3.4.4
#> KernSmooth               KernSmooth     2.23-15
#> lattice                     lattice     0.20-35
#> MASS                           MASS      7.3-49
#> Matrix                       Matrix      1.2-12
#> methods                     methods       3.4.4
#> mgcv                           mgcv      1.8-23
#> nlme                           nlme     3.1-131
#> nnet                           nnet      7.3-12
#> parallel                   parallel       3.4.4
#> rpart                         rpart      4.1-13
#> spatial                     spatial      7.3-11
#> splines                     splines       3.4.4
#> stats                         stats       3.4.4
#> stats4                       stats4       3.4.4
#> survival                   survival      2.41-3
#> tcltk                         tcltk       3.4.4
#> tools                         tools       3.4.4
#> translations           translations       3.4.4
#> utils                         utils       3.4.4

data.table::fread(pkgSnapFile)

#>          instPkgs instVers
#>   1:     blogdown      0.8
#>   2:     bookdown      0.7
#>   3:        callr    2.0.4
#>   4:           fs    1.2.3
#>   5:    highlight  0.4.7.2
#>  ---                      
#> 214:     survival   2.41-3
#> 215:        tcltk    3.4.4
#> 216:        tools    3.4.4
#> 217: translations    3.4.4
#> 218:        utils    3.4.4