Take a snapshot of all the packages and version numbers

This can be used later by installVersions to install or re-install the correct versions.

pkgSnapshot(packageVersionFile, libPath, standAlone = FALSE)

Arguments

packageVersionFile
A filename to save the packages and their currently installed version numbers. Defaults to ".packageVersions.txt".

libPath
The path to the local library where packages are installed. Defaults to the .libPaths()[1]

standAlone
Logical. If TRUE, all packages will be installed and loaded strictly from the libPaths only. If FALSE, all .libPaths will be used to find the correct versions. This can be create dramatically faster installs if the user has a substantial number of the packages already in their personal library. In the case of TRUE, there will be a hidden file place in the libPath directory that lists all the packages that were needed during the Require call. Default FALSE to minimize package installing.

Details

A file is written with the package names and versions of all packages within libPath. This can later be passed to Require.

Examples

pkgSnapFile <- tempfile()
pkgSnapshot(pkgSnapFile, .libPaths()[1])

#> There is no /tmp/RtmpNbxQhX/reproducible/cache/.Require/_tmp_RtmpFchKkE_temp_libpath1950227b8ee6f._packageVersionsAuto.txt and standAlone is FALSE. This snapshot will not be accurate because it will include all packages in /tmp/RtmpFchKkE/temp_libpath1950227b8ee6f, /home/achubaty/R/x86_64-pc-linux-gnu-library/3.6, /usr/local/lib/R/site-library, /usr/lib/R/site-library, /usr/lib/R/library

#>              instPkgs    instVers
#> 1        reproducible       1.0.0
#> 2        reproducible       1.0.0
#> 3                 amc  0.1.5.9000
#> 4           animation         2.6
#> 5           archivist       2.3.4
#> 6             askpass         1.1
#> 7          assertthat       0.2.1
#> 8           available       1.0.4
#> 9           backports       1.1.5
#> 10          base64enc       0.1-3
#> 11                 BH    1.72.0-3
#> 12                bit    1.1-15.2
#> 13              bit64       0.9-7
#> 14             bitops       1.0-6
#> 15               blob       1.2.1
#> 16               brew       1.0-6
#> 17              callr       3.4.2
#> 18          checkmate       2.0.0
#> 19          CircStats       0.2-6
#> 20           classInt       0.4-2
#> 21                cli       2.0.1
#> 22              clipr       0.7.0
#> 23         clisymbols       1.2.0
#> 24         colorspace       1.4-1
#> 25         commonmark         1.7
#> 26             config         0.3
#> 27               covr       3.4.0
#> 28            cowplot       1.0.0
#> 29             crayon       1.3.4
#> 30          crosstalk       1.0.0
#> 31               crul       0.9.0
#> 32               curl         4.3
#> 33         data.table      1.12.8
#> 34                DBI       1.1.0
#> 35               desc       1.2.0
#> 36           devtools       2.2.2
#> 37         DiagrammeR       1.0.5
#> 38             digest      0.6.24
#> 39              dismo       1.1-4
#> 40         doParallel      1.0.15
#> 41         downloader         0.4
#> 42              dplyr       0.8.4
#> 43                 DT        0.12
#> 44              e1071       1.7-3
#> 45           ellipsis       0.3.0
#> 46           evaluate        0.14
#> 47              fansi       0.4.1
#> 48             farver       2.0.3
#> 49         fastdigest       0.6-3
#> 50          fasterize       1.0.0
#> 51            fastmap       1.0.1
#> 52          fastmatch       1.1-0
#> 53            fastshp       0.1-2
#> 54                 ff      2.2-14
#> 55             ffbase      0.12.7
#> 56              flock         0.7
#> 57            foreach       1.4.8
#> 58          fpCompare       0.2.3
#> 59                 fs       1.3.1
#> 60             future      1.16.0
#> 61       future.callr       0.5.0
#> 62             gamlss       5.1-6
#> 63        gamlss.data       5.1-4
#> 64        gamlss.dist       5.1-6
#> 65             gargle       0.4.0
#> 66          gdalUtils     2.0.3.2
#> 67           generics       0.0.2
#> 68            geojson       0.3.2
#> 69          geojsonio       0.9.0
#> 70        geojsonlint       0.4.0
#> 71            ggplot2       3.2.1
#> 72           ggthemes       4.2.0
#> 73                 gh       1.1.0
#> 74              git2r      0.26.1
#> 75               glmm       1.3.0
#> 76            globals      0.12.5
#> 77               glue       1.3.1
#> 78        googleAuthR       1.1.1
#> 79        googledrive       1.0.0
#> 80           googleID    0.0.9001
#> 81         grainscape       0.4.3
#> 82           gridBase       0.4-7
#> 83          gridExtra         2.3
#> 84             gtable       0.3.0
#> 85             hexbin      1.28.1
#> 86          highlight       0.5.0
#> 87              highr         0.8
#> 88                hms       0.5.3
#> 89          htmltools       0.4.0
#> 90        htmlwidgets       1.5.1
#> 91           httpcode       0.2.0
#> 92             httpuv       1.5.2
#> 93               httr       1.4.1
#> 94           hunspell         3.0
#> 95             igraph     1.2.4.2
#> 96         influenceR       0.1.0
#> 97                ini       0.3.1
#> 98           ISOcodes  2019.12.22
#> 99          iterators      1.0.12
#> 100         itertools       0.1-3
#> 101       janeaustenr       0.1.5
#> 102              jpeg     0.1-8.1
#> 103               jqr       1.1.0
#> 104          jsonlite       1.6.1
#> 105      jsonvalidate       1.1.0
#> 106        kableExtra       1.1.0
#> 107             knitr        1.28
#> 108          labeling         0.3
#> 109             LandR  0.0.3.9001
#> 110          LandR.CS  0.0.0.9001
#> 111      LandWebUtils       0.0.2
#> 112             later       1.0.0
#> 113      latticeExtra      0.6-29
#> 114          lazyeval       0.2.2
#> 115           leaflet       2.0.3
#> 116    leaflet.extras       1.0.0
#> 117 leaflet.providers       1.9.0
#> 118         lifecycle       0.1.0
#> 119           listenv       0.8.0
#> 120              lme4      1.1-21
#> 121           logging    0.10-108
#> 122         lubridate       1.7.4
#> 123            magick         2.3
#> 124          magrittr         1.5
#> 125               map       0.0.1
#> 126          maptools       0.9-9
#> 127          markdown         1.1
#> 128           memoise       1.1.0
#> 129    microbenchmark       1.4-7
#> 130              mime         0.9
#> 131             minqa       1.2.4
#> 132             MuMIn     1.43.15
#> 133           munsell       0.5.0
#> 134           mvtnorm      1.0-12
#> 135            nloptr       1.2.1
#> 136           openssl       1.4.1
#> 137           packrat       0.5.0
#> 138            pemisc  0.0.1.9005
#> 139            pillar       1.4.3
#> 140          pkgbuild       1.0.6
#> 141         pkgconfig       2.0.3
#> 142           pkgdown       1.4.1
#> 143           pkgload       1.0.2
#> 144             plogr       0.2.0
#> 145              plyr       1.8.5
#> 146               png       0.1-7
#> 147            praise       1.0.0
#> 148        prettycode       1.1.0
#> 149       prettyunits       1.1.1
#> 150          processx       3.4.2
#> 151          promises       1.1.0
#> 152         protolite         2.1
#> 153              pryr       0.1.4
#> 154                ps       1.3.2
#> 155             purrr       0.3.3
#> 156                qs      0.21.1
#> 157         quickPlot       0.1.6
#> 158         R.devices      2.16.1
#> 159       R.methodsS3       1.8.0
#> 160              R.oo      1.23.0
#> 161           R.utils       2.9.2
#> 162                R6       2.4.1
#> 163      RandomFields       3.3.8
#> 164 RandomFieldsUtils       0.5.3
#> 165     RApiSerialize       0.1.0
#> 166            raster      3.0-12
#> 167         rasterVis        0.47
#> 168         rcmdcheck       1.3.3
#> 169      RColorBrewer       1.1-2
#> 170              Rcpp       1.0.3
#> 171     RcppArmadillo 0.9.850.1.0
#> 172         RcppEigen   0.3.3.7.0
#> 173             RCurl    1.98-1.1
#> 174              rdoc       0.1.0
#> 175             readr       1.3.1
#> 176          rematch2       2.1.0
#> 177           remotes       2.1.1
#> 178      reproducible       1.0.0
#> 179      reproducible       1.0.0
#> 180          reshape2       1.4.3
#> 181               rex       1.1.2
#> 182             rgdal       1.4-8
#> 183             rgeos       0.5-2
#> 184             rJava      0.9-11
#> 185             rlang       0.4.4
#> 186        rmapshaper       0.4.3
#> 187         rmarkdown         2.1
#> 188          roxygen2       7.0.2
#> 189         RPostgres       1.2.0
#> 190         rprojroot       1.3-2
#> 191            Rserve     1.7-3.1
#> 192           RSQLite       2.2.0
#> 193        rstudioapi        0.11
#> 194         rversions       2.0.1
#> 195             rvest       0.3.5
#> 196            scales       1.1.0
#> 197           selectr       0.4-2
#> 198       sessioninfo       1.1.1
#> 199                sf       0.8-1
#> 200             shiny       1.4.0
#> 201           shinyBS        0.61
#> 202   shinycssloaders         0.3
#> 203    shinydashboard       0.7.1
#> 204           shinyjs         1.1
#> 205      shinyWidgets       0.5.0
#> 206            slackr       1.4.2
#> 207         SnowballC       0.6.0
#> 208       sourcetools       0.1.7
#> 209                sp       1.3-2
#> 210       SpaDES.core       0.2.9
#> 211      SpaDES.tools       0.3.4
#> 212          spelling         2.1
#> 213         stopwords         1.0
#> 214        stringdist     0.9.5.5
#> 215           stringi       1.4.6
#> 216           stringr       1.4.0
#> 217               sys         3.3
#> 218          testthat       2.3.1
#> 219            tibble       2.1.3
#> 220             tidyr       1.0.2
#> 221        tidyselect       1.0.0
#> 222          tidytext       0.2.2
#> 223             tiler       0.2.4
#> 224          TimeWarp      1.0.15
#> 225           tinytex        0.19
#> 226        tokenizers       0.2.1
#> 227         triebeard       0.3.0
#> 228             trust       0.1-8
#> 229             udapi       0.1.3
#> 230             units       0.6-5
#> 231          urltools       1.7.3
#> 232           usethis       1.5.1
#> 233              utf8       1.1.4
#> 234              uuid       0.1-2
#> 235                V8       3.0.1
#> 236             vctrs       0.2.2
#> 237          versions         0.3
#> 238              VGAM       1.1-2
#> 239           viridis       0.5.1
#> 240       viridisLite       0.3.0
#> 241        visNetwork       2.0.9
#> 242           webshot       0.5.2
#> 243           whisker         0.4
#> 244             withr       2.1.2
#> 245              xfun        0.12
#> 246              xlsx       0.6.2
#> 247          xlsxjars       0.6.1
#> 248               XML    3.99-0.3
#> 249              xml2       1.2.2
#> 250             xopen       1.0.0
#> 251            xtable       1.8-4
#> 252              yaml       2.2.1
#> 253             yesno       0.1.0
#> 254               zoo       1.8-7
#> 255           littler       0.3.9
#> 256              base       3.6.2
#> 257              boot      1.3-24
#> 258             class      7.3-15
#> 259           cluster       2.1.0
#> 260         codetools      0.2-16
#> 261          compiler       3.6.2
#> 262          datasets       3.6.2
#> 263           foreign      0.8-75
#> 264          graphics       3.6.2
#> 265         grDevices       3.6.2
#> 266              grid       3.6.2
#> 267        KernSmooth     2.23-16
#> 268           lattice     0.20-38
#> 269              MASS    7.3-51.5
#> 270            Matrix      1.2-18
#> 271           methods       3.6.2
#> 272              mgcv      1.8-31
#> 273              nlme     3.1-144
#> 274              nnet      7.3-12
#> 275          parallel       3.6.2
#> 276             rpart      4.1-15
#> 277           spatial      7.3-11
#> 278           splines       3.6.2
#> 279             stats       3.6.2
#> 280            stats4       3.6.2
#> 281          survival       3.1-8
#> 282             tcltk       3.6.2
#> 283             tools       3.6.2
#> 284      translations       3.6.2
#> 285             utils       3.6.2

data.table::fread(pkgSnapFile)

#>          instPkgs   instVers
#>   1: reproducible      1.0.0
#>   2: reproducible      1.0.0
#>   3:          amc 0.1.5.9000
#>   4:    animation        2.6
#>   5:    archivist      2.3.4
#>  ---                        
#> 281:     survival      3.1-8
#> 282:        tcltk      3.6.2
#> 283:        tools      3.6.2
#> 284: translations      3.6.2
#> 285:        utils      3.6.2