Take a snapshot of all the packages and version numbers

This can be used later by installVersions to install or re-install the correct versions.

pkgSnapshot(packageVersionFile, libPath, standAlone = FALSE)

Arguments

packageVersionFile
A filename to save the packages and their currently installed version numbers. Defaults to ".packageVersions.txt".

libPath
The path to the local library where packages are installed. Defaults to the .libPaths()[1]

standAlone
Logical. If TRUE, all packages will be installed and loaded strictly from the libPaths only. If FALSE, all .libPaths will be used to find the correct versions. This can be create dramatically faster installs if the user has a substantial number of the packages already in their personal library. In the case of TRUE, there will be a hidden file place in the libPath directory that lists all the packages that were needed during the Require call. Default FALSE to minimize package installing.

Details

A file is written with the package names and versions of all packages within libPath. This can later be passed to Require.

Examples

pkgSnapFile <- tempfile()
pkgSnapshot(pkgSnapFile, .libPaths()[1])

#> There is no C:/Users/emcintir.L-VIC-A023744/AppData/Local/Temp/RtmpkRPYKY/reproducible/cache/.Require/C_Users_emcintir.L-VIC-A023744_AppData_Local_Temp_RtmpGoESaj_temp_libpath387c2c5531a4._packageVersionsAuto.txt and standAlone is FALSE. This snapshot will not be accurate because it will include all packages in C:/Users/emcintir.L-VIC-A023744/AppData/Local/Temp/RtmpGoESaj/temp_libpath387c2c5531a4, C:/Eliot/R/win-library/3.6, C:/Program Files/R/R-3.6.2/library

#>              instPkgs    instVers
#> 1        reproducible       1.0.0
#> 2        reproducible       1.0.0
#> 3               abind       1.4-5
#> 4       alphavantager       0.1.1
#> 5                 amc  0.1.5.9000
#> 6           animation         2.6
#> 7           archivist       2.3.4
#> 8             askpass         1.1
#> 9          assertthat       0.2.1
#> 10          backports       1.1.5
#> 11          base64enc       0.1-3
#> 12                 BH    1.72.0-3
#> 13        BiocManager     1.30.10
#> 14                bit    1.1-15.1
#> 15              bit64       0.9-7
#> 16             bitops       1.0-6
#> 17               blob       1.2.1
#> 18               boot      1.3-24
#> 19               boot      1.3-24
#> 20               brew       1.0-6
#> 21              broom       0.5.3
#> 22              callr       3.4.0
#> 23         cellranger       1.1.0
#> 24          checkmate       1.9.4
#> 25          CircStats       0.2-6
#> 26           classInt       0.4-2
#> 27                cli       2.0.1
#> 28              clipr       0.7.0
#> 29         clisymbols       1.2.0
#> 30         colorspace       1.4-1
#> 31         commonmark         1.7
#> 32               covr       3.4.0
#> 33          crancache  0.0.0.9001
#> 34           cranlike       1.0.2
#> 35           cranlogs       2.1.1
#> 36             crayon       1.3.4
#> 37          crosstalk       1.0.0
#> 38               crul       0.9.0
#> 39               curl         4.3
#> 40         data.table      1.12.8
#> 41                DBI       1.1.0
#> 42             dbplyr       1.4.2
#> 43            debugme       1.1.0
#> 44            DEoptim       2.2-4
#> 45               desc       1.2.0
#> 46           devtools       2.2.1
#> 47            diagram       1.6.4
#> 48         DiagrammeR       1.0.5
#> 49             digest      0.6.23
#> 50              dismo       1.1-4
#> 51         doParallel      1.0.15
#> 52         downloader         0.4
#> 53              dplyr       0.8.3
#> 54                 DT        0.11
#> 55              e1071       1.7-3
#> 56            effsize       0.7.6
#> 57           ellipsis       0.3.0
#> 58       emptyPackage       0.1.0
#> 59           evaluate        0.14
#> 60              fansi       0.4.1
#> 61             farver       2.0.3
#> 62         fastdigest       0.6-3
#> 63          fasterize       1.0.0
#> 64            fastmap       1.0.1
#> 65          fastmatch       1.1-0
#> 66            fastshp       0.1-2
#> 67                 ff      2.2-14
#> 68             ffbase      0.12.7
#> 69            Finance         1.0
#> 70              flock         0.7
#> 71            forcats       0.4.0
#> 72            foreach       1.4.7
#> 73            foreign      0.8-72
#> 74            foreign      0.8-72
#> 75            formatR         1.7
#> 76          fpCompare       0.2.3
#> 77                 fs       1.3.1
#> 78                fst       0.9.0
#> 79      futile.logger       1.4.3
#> 80     futile.options       1.0.1
#> 81             future      1.16.0
#> 82       future.apply       1.3.0
#> 83       future.callr       0.5.0
#> 84             gamlss       5.1-5
#> 85        gamlss.data       5.1-4
#> 86        gamlss.dist       5.1-5
#> 87             gargle       0.4.0
#> 88          gdalUtils       2.0.3
#> 89           generics       0.0.2
#> 90            geojson       0.3.2
#> 91          geojsonio       0.8.0
#> 92        geojsonlint       0.3.0
#> 93            ggimage       0.2.5
#> 94              ggmap   3.0.0.901
#> 95            ggplot2       3.2.1
#> 96          ggplotify       0.0.4
#> 97                 gh       1.0.1
#> 98              git2r      0.26.1
#> 99               glmm       1.3.0
#> 100           globals      0.12.5
#> 101              glue       1.3.1
#> 102            gmailr       1.0.0
#> 103       googleAuthR       1.1.1
#> 104       googledrive       1.0.0
#> 105          googleID    0.0.9001
#> 106          gridBase       0.4-7
#> 107         gridExtra         2.3
#> 108      gridGraphics       0.4-1
#> 109            gtable       0.3.0
#> 110             haven       2.2.0
#> 111            hexbin      1.28.0
#> 112        hexSticker       0.4.6
#> 113         highlight       0.5.0
#> 114             highr         0.8
#> 115               hms       0.5.3
#> 116         htmltools       0.4.0
#> 117       htmlwidgets       1.5.1
#> 118          httpcode       0.2.0
#> 119            httpuv       1.5.2
#> 120              httr       1.4.1
#> 121          hunspell         3.0
#> 122            igraph     1.2.4.2
#> 123        influenceR       0.1.0
#> 124               ini       0.3.1
#> 125         iterators      1.0.12
#> 126         itertools       0.1-3
#> 127              jpeg     0.1-8.1
#> 128               jqr       1.1.0
#> 129          jsonlite         1.6
#> 130      jsonvalidate       1.1.0
#> 131        KernSmooth     2.23-16
#> 132        KernSmooth     2.23-16
#> 133             knitr        1.26
#> 134          labeling         0.3
#> 135          lambda.r       1.2.4
#> 136             LandR  0.0.3.9001
#> 137          LandR.CS  0.0.0.9000
#> 138             later       1.0.0
#> 139      latticeExtra      0.6-28
#> 140          lazyeval       0.2.2
#> 141           leaflet       2.0.3
#> 142    leaflet.extras       1.0.0
#> 143 leaflet.providers       1.9.0
#> 144         lifecycle       0.1.0
#> 145           listenv       0.8.0
#> 146              lme4      1.1-21
#> 147           logging    0.10-108
#> 148         lubridate       1.7.4
#> 149              lucr       0.2.0
#> 150            lwgeom       0.1-7
#> 151            magick         2.2
#> 152          magrittr         1.5
#> 153               map       0.0.1
#> 154          maptools       0.9-9
#> 155          markdown         1.1
#> 156              MASS    7.3-51.5
#> 157              MASS    7.3-51.5
#> 158            Matrix      1.2-18
#> 159            Matrix      1.2-18
#> 160           memoise       1.1.0
#> 161              mgcv      1.8-31
#> 162              mgcv      1.8-31
#> 163    microbenchmark       1.4-7
#> 164              mime         0.8
#> 165            miniUI     0.1.1.1
#> 166             minqa       1.2.4
#> 167            modelr       0.1.5
#> 168             MuMIn     1.43.14
#> 169           munsell       0.5.0
#> 170           mvtnorm      1.0-11
#> 171              nlme     3.1-143
#> 172              nlme     3.1-143
#> 173            nloptr       1.2.1
#> 174           numbers       0.7-5
#> 175           openssl       1.4.1
#> 176           packrat       0.5.0
#> 177              padr       0.5.0
#> 178         parsedate       1.2.0
#> 179             pedev  0.0.1.9005
#> 180            pemisc  0.0.1.9005
#> 181            pillar       1.4.3
#> 182          pkgbuild       1.0.6
#> 183         pkgconfig       2.0.3
#> 184           pkgdown       1.4.1
#> 185           pkgload       1.0.2
#> 186             plogr       0.2.0
#> 187              plyr       1.8.5
#> 188               png       0.1-7
#> 189            praise       1.0.0
#> 190       prettyunits       1.1.0
#> 191          processx       3.4.1
#> 192           profvis       0.3.6
#> 193          progress       1.2.2
#> 194          promises       1.1.0
#> 195         protolite         2.1
#> 196              pryr       0.1.4
#> 197                ps       1.3.0
#> 198             purrr       0.3.3
#> 199                qs      0.20.2
#> 200          quantmod      0.4-15
#> 201         quickPlot  0.1.6.9000
#> 202         R.devices      2.16.1
#> 203       R.methodsS3       1.7.1
#> 204              R.oo      1.23.0
#> 205           R.utils       2.9.2
#> 206                R6       2.4.1
#> 207      RandomFields       3.3.7
#> 208 RandomFieldsUtils       0.5.3
#> 209     RApiSerialize       0.1.0
#> 210          rappdirs       0.3.1
#> 211            raster       3.0-7
#> 212         rasterVis        0.47
#> 213         rcmdcheck       1.3.3
#> 214      RColorBrewer       1.1-2
#> 215              Rcpp       1.0.3
#> 216     RcppArmadillo 0.9.800.3.0
#> 217         RcppEigen   0.3.3.7.0
#> 218             RCurl    1.98-1.1
#> 219             readr       1.3.1
#> 220            readxl       1.3.1
#> 221           rematch       1.0.1
#> 222          rematch2       2.1.0
#> 223           remotes       2.1.0
#> 224            reprex       0.3.0
#> 225      reproducible       1.0.0
#> 226      reproducible       1.0.0
#> 227          reshape2       1.4.3
#> 228       revdepcheck  1.0.0.9001
#> 229               rex       1.1.2
#> 230             rgdal       1.4-8
#> 231             rgeos       0.5-2
#> 232             rgexf      0.15.3
#> 233       RgoogleMaps     1.4.5.2
#> 234              rhub       1.1.1
#> 235             rJava      0.9-11
#> 236             rjson      0.2.20
#> 237             rlang       0.4.2
#> 238        rmapshaper       0.4.3
#> 239         rmarkdown         2.0
#> 240              Rook       1.1-1
#> 241          roxygen2       7.0.2
#> 242         RPostgres       1.2.0
#> 243         rprojroot       1.3-2
#> 244           RSQLite       2.2.0
#> 245        rstudioapi        0.10
#> 246           rvcheck       0.1.7
#> 247         rversions       2.0.1
#> 248             rvest       0.3.5
#> 249            scales       1.1.0
#> 250              scam       1.2-5
#> 251           selectr       0.4-2
#> 252       sessioninfo       1.1.1
#> 253                sf       0.8-0
#> 254             shape       1.4.4
#> 255             shiny       1.4.0
#> 256           shinyBS        0.61
#> 257   shinycssloaders         0.3
#> 258    shinydashboard       0.7.1
#> 259           shinyjs         1.1
#> 260      shinyWidgets       0.5.0
#> 261          showtext         0.7
#> 262        showtextdb         2.0
#> 263              snow       0.4-3
#> 264       sourcetools       0.1.7
#> 265                sp       1.3-2
#> 266            SpaDES       2.0.4
#> 267     SpaDES.addins       0.1.2
#> 268       SpaDES.core  0.2.9.9000
#> 269   SpaDES.examples  0.0.0.9005
#> 270 SpaDES.experiment  0.0.2.9000
#> 271      SpaDES.shiny       0.2.0
#> 272      SpaDES.tools  0.3.4.9000
#> 273            spData       0.3.2
#> 274          spelling         2.1
#> 275             stars       0.4-0
#> 276           stringi       1.4.4
#> 277           stringr       1.4.0
#> 278          survival       3.1-8
#> 279          survival       3.1-8
#> 280               sys         3.3
#> 281          sysfonts         0.8
#> 282          testthat       2.3.1
#> 283            tibble       2.1.3
#> 284             tidyr       1.0.0
#> 285        tidyselect       0.2.5
#> 286         tidyverse       1.3.0
#> 287             tiler       0.2.4
#> 288            timetk       0.1.2
#> 289          TimeWarp      1.0.15
#> 290           tinytex        0.18
#> 291           tkrplot      0.0-24
#> 292         triebeard       0.3.0
#> 293             trust       0.1-7
#> 294       tryCatchLog       1.1.6
#> 295               TTR      0.23-6
#> 296             units       0.6-5
#> 297          urltools       1.7.3
#> 298            usefun       0.1.0
#> 299           usethis       1.5.1
#> 300              utf8       1.1.4
#> 301              uuid       0.1-2
#> 302                V8       3.0.1
#> 303             vctrs       0.2.1
#> 304          versions         0.3
#> 305           viridis       0.5.1
#> 306       viridisLite       0.3.0
#> 307        visNetwork       2.0.9
#> 308           whisker         0.4
#> 309            whoami       1.3.0
#> 310             withr       2.1.2
#> 311              xfun        0.11
#> 312              xlsx       0.6.1
#> 313          xlsxjars       0.6.1
#> 314               XML   3.98-1.20
#> 315              xml2       1.2.2
#> 316             xopen       1.0.0
#> 317            xtable       1.8-4
#> 318               xts      0.11-2
#> 319              yaml       2.2.0
#> 320           zeallot       0.1.0
#> 321               zoo       1.8-6
#> 322              base       3.6.2
#> 323              boot      1.3-23
#> 324              boot      1.3-23
#> 325             class      7.3-15
#> 326           cluster       2.1.0
#> 327         codetools      0.2-16
#> 328          compiler       3.6.2
#> 329          datasets       3.6.2
#> 330           foreign      0.8-72
#> 331           foreign      0.8-72
#> 332          graphics       3.6.2
#> 333         grDevices       3.6.2
#> 334              grid       3.6.2
#> 335        KernSmooth     2.23-16
#> 336        KernSmooth     2.23-16
#> 337           lattice     0.20-38
#> 338              MASS    7.3-51.4
#> 339              MASS    7.3-51.4
#> 340            Matrix      1.2-18
#> 341            Matrix      1.2-18
#> 342           methods       3.6.2
#> 343              mgcv      1.8-31
#> 344              mgcv      1.8-31
#> 345              nlme     3.1-142
#> 346              nlme     3.1-142
#> 347              nnet      7.3-12
#> 348          parallel       3.6.2
#> 349             rpart      4.1-15
#> 350           spatial      7.3-11
#> 351           splines       3.6.2
#> 352             stats       3.6.2
#> 353            stats4       3.6.2
#> 354          survival       3.1-8
#> 355          survival       3.1-8
#> 356             tcltk       3.6.2
#> 357             tools       3.6.2
#> 358      translations       3.6.2
#> 359             utils       3.6.2

data.table::fread(pkgSnapFile)

#>           instPkgs   instVers
#>   1:  reproducible      1.0.0
#>   2:  reproducible      1.0.0
#>   3:         abind      1.4-5
#>   4: alphavantager      0.1.1
#>   5:           amc 0.1.5.9000
#>  ---                         
#> 355:      survival      3.1-8
#> 356:         tcltk      3.6.2
#> 357:         tools      3.6.2
#> 358:  translations      3.6.2
#> 359:         utils      3.6.2