Take a snapshot of all the packages and version numbers

This can be used later by installVersions to install or re-install the correct versions.

pkgSnapshot(packageVersionFile, libPath, standAlone = FALSE)

Arguments

packageVersionFile
A filename to save the packages and their currently installed version numbers. Defaults to ".packageVersions.txt".

libPath
The path to the local library where packages are installed. Defaults to the .libPaths()[1]

standAlone
Logical. If TRUE, all packages will be installed and loaded strictly from the libPaths only. If FALSE, all .libPaths will be used to find the correct versions. This can be create dramatically faster installs if the user has a substantial number of the packages already in their personal library. In the case of TRUE, there will be a hidden file place in the libPath directory that lists all the packages that were needed during the Require call. Default FALSE to minimize package installing.

Details

A file is written with the package names and versions of all packages within libPath. This can later be passed to Require.

Examples

pkgSnapFile <- tempfile()
pkgSnapshot(pkgSnapFile, .libPaths()[1])

#> There is no /private/var/folders/l2/hy6b0sl977bcd8695nt6j7s80000gn/T/RtmpYVoW4N/reproducible/cache/.Require/_private_var_folders_l2_hy6b0sl977bcd8695nt6j7s80000gn_T_Rtmp4GaTxy_temp_libpath78e277991dbd._packageVersionsAuto.txt and standAlone is FALSE. This snapshot will not be accurate because it will include all packages in /private/var/folders/l2/hy6b0sl977bcd8695nt6j7s80000gn/T/Rtmp4GaTxy/temp_libpath78e277991dbd, /Library/Frameworks/R.framework/Versions/3.6/Resources/library

#>              instPkgs    instVers
#> 1        reproducible       1.0.0
#> 2        reproducible       1.0.0
#> 3               abind       1.4-5
#> 4             acepack       1.4.1
#> 5                 amc  0.1.5.9000
#> 6           animation         2.6
#> 7           archivist       2.3.4
#> 8             askpass         1.1
#> 9          assertthat       0.2.1
#> 10              audio       0.1-6
#> 11          available       1.0.4
#> 12          backports       1.1.5
#> 13               base       3.6.2
#> 14          base64enc       0.1-3
#> 15                 BH    1.72.0-3
#> 16              bindr       0.1.1
#> 17           bindrcpp       0.2.2
#> 18        BiocManager     1.30.10
#> 19        BiocVersion       3.9.0
#> 20                bit    1.1-15.2
#> 21              bit64       0.9-7
#> 22             bitops       1.0-6
#> 23               blob       1.2.1
#> 24           bookdown        0.17
#> 25               boot      1.3-24
#> 26               brew       1.0-6
#> 27              broom       0.5.4
#> 28              Cairo      1.5-10
#> 29              callr       3.4.2
#> 30                car       3.0-6
#> 31            carData       3.0-3
#> 32            caTools      1.18.0
#> 33         cellranger       1.1.0
#> 34          checkmate       2.0.0
#> 35          CircStats       0.2-6
#> 36              class      7.3-15
#> 37           classInt       0.4-2
#> 38                cli       2.0.1
#> 39              clipr       0.7.0
#> 40         clisymbols       1.2.0
#> 41            cluster       2.1.0
#> 42          codetools      0.2-16
#> 43         colorspace       1.4-1
#> 44         commonmark         1.7
#> 45           compiler       3.6.2
#> 46               covr       3.4.0
#> 47            cowplot       1.0.0
#> 48           cranlike       1.0.2
#> 49             crayon       1.3.4
#> 50          crosstalk       1.0.0
#> 51               crul       0.9.0
#> 52               curl         4.3
#> 53          cyclocomp       1.1.0
#> 54         data.table      1.12.8
#> 55           datasets       3.6.2
#> 56                DBI       1.1.0
#> 57            debugme       1.1.0
#> 58             deldir      0.1-25
#> 59            DEoptim       2.2-4
#> 60               desc       1.2.0
#> 61           devtools       2.2.2
#> 62         DiagrammeR       1.0.5
#> 63          dichromat       2.0-0
#> 64             digest      0.6.24
#> 65              dismo       1.1-4
#> 66         downloader         0.4
#> 67              dplyr       0.8.4
#> 68                 DT        0.12
#> 69              e1071       1.7-3
#> 70           ellipsis       0.3.0
#> 71                enc       0.2.2
#> 72           evaluate        0.14
#> 73              fansi       0.4.1
#> 74             farver       2.0.3
#> 75         fastdigest       0.6-3
#> 76          fasterize       1.0.0
#> 77            fastmap       1.0.1
#> 78          fastmatch       1.1-0
#> 79            fastshp       0.1-2
#> 80                 ff      2.2-14
#> 81             ffbase      0.12.7
#> 82              flock         0.7
#> 83            forcats       0.4.0
#> 84            foreach       1.4.8
#> 85            foreign      0.8-75
#> 86            Formula       1.2-3
#> 87          fpCompare       0.2.3
#> 88                 fs       1.3.1
#> 89             future      1.16.0
#> 90       future.apply       1.4.0
#> 91       future.callr       0.5.0
#> 92             gargle       0.4.0
#> 93          gdalUtils     2.0.3.2
#> 94           generics       0.0.2
#> 95            geojson       0.3.2
#> 96          geojsonio       0.9.0
#> 97        geojsonlint       0.4.0
#> 98            ggimage       0.2.7
#> 99            ggplot2       3.2.1
#> 100         ggplotify       0.0.4
#> 101          ggthemes       4.2.0
#> 102                gh  1.1.0.9000
#> 103             git2r      0.26.1
#> 104           globals      0.12.5
#> 105              glue       1.3.1
#> 106            gmailr       1.0.0
#> 107           goftest       1.2-2
#> 108       googleAuthR       1.1.1
#> 109       googledrive       1.0.0
#> 110          googleID    0.0.9001
#> 111        grainscape       0.4.2
#> 112          graphics       3.6.2
#> 113         grDevices       3.6.2
#> 114              grid       3.6.2
#> 115          gridBase       0.4-7
#> 116         gridExtra         2.3
#> 117      gridGraphics       0.4-1
#> 118            gtable       0.3.0
#> 119             haven       2.2.0
#> 120            hexbin      1.28.1
#> 121        hexSticker       0.4.6
#> 122         highlight       0.5.0
#> 123          highlite  0.0.0.9000
#> 124             highr         0.8
#> 125             Hmisc       4.3-1
#> 126               hms       0.5.3
#> 127         htmlTable      1.13.3
#> 128         htmltools       0.4.0
#> 129       htmlwidgets       1.5.1
#> 130          httpcode       0.2.0
#> 131            httpuv       1.5.2
#> 132              httr       1.4.1
#> 133          hunspell         3.0
#> 134            igraph     1.2.4.2
#> 135        influenceR       0.1.0
#> 136               ini       0.3.1
#> 137            inline      0.3.15
#> 138          ISOcodes  2019.12.22
#> 139         iterators      1.0.12
#> 140       janeaustenr       0.1.5
#> 141              jpeg     0.1-8.1
#> 142               jqr       1.1.0
#> 143          jsonlite       1.6.1
#> 144      jsonvalidate       1.1.0
#> 145        kableExtra       1.1.0
#> 146        KernSmooth     2.23-16
#> 147             knitr        1.28
#> 148            koRpus      0.11-5
#> 149    koRpus.lang.en       0.1-3
#> 150          labeling         0.3
#> 151             LandR  0.0.3.9001
#> 152    landscapetools       0.5.0
#> 153      LandWebUtils       0.0.2
#> 154             later       1.0.0
#> 155           lattice     0.20-38
#> 156      latticeExtra      0.6-29
#> 157          lazyeval       0.2.2
#> 158           leaflet       2.0.3
#> 159    leaflet.extras       1.0.0
#> 160 leaflet.providers       1.9.0
#> 161         lifecycle       0.1.0
#> 162             lintr  2.0.0.9000
#> 163           listenv       0.8.0
#> 164              lme4      1.1-21
#> 165           logging    0.10-108
#> 166            lookup  0.0.0.9000
#> 167         lubridate       1.7.4
#> 168            magick         2.3
#> 169          magrittr         1.5
#> 170               map       0.0.1
#> 171          maptools       0.9-9
#> 172          markdown         1.1
#> 173              MASS    7.3-51.5
#> 174            Matrix      1.2-18
#> 175      MatrixModels       0.4-1
#> 176       matrixStats      0.55.0
#> 177           memoise  1.1.0.9000
#> 178              meow       0.1.4
#> 179           methods       3.6.2
#> 180              mgcv      1.8-31
#> 181    microbenchmark       1.4-7
#> 182              mime         0.9
#> 183            miniUI     0.1.1.1
#> 184             minqa       1.2.4
#> 185            mnormt       1.5-6
#> 186             MuMIn     1.43.15
#> 187           munsell       0.5.0
#> 188          NetLogoR       0.3.6
#> 189              nlme     3.1-144
#> 190              NLMR         1.0
#> 191            nloptr       1.2.1
#> 192              nnet      7.3-12
#> 193           openssl       1.4.1
#> 194          openxlsx       4.1.4
#> 195           packrat       0.5.0
#> 196          parallel       3.6.2
#> 197         parsedate       1.2.0
#> 198           pbapply       1.4-2
#> 199          pbkrtest       0.4-7
#> 200            pemisc  0.0.1.9005
#> 201            pillar       1.4.3
#> 202          pkgbuild       1.0.6
#> 203         pkgconfig       2.0.3
#> 204           pkgdown       1.4.1
#> 205           pkgload       1.0.2
#> 206               PKI     0.1-5.1
#> 207             plogr       0.2.0
#> 208              plyr       1.8.5
#> 209               png       0.1-7
#> 210          polyclip      1.10-0
#> 211            praise       1.0.0
#> 212        prettycode       1.1.0
#> 213       prettyunits       1.1.1
#> 214          processx       3.4.2
#> 215           profvis       0.3.6
#> 216          progress       1.2.2
#> 217          promises       1.1.0
#> 218         protolite         2.1
#> 219              pryr       0.1.4
#> 220                ps       1.3.2
#> 221             psych   1.9.12.31
#> 222             purrr       0.3.3
#> 223                qs      0.21.1
#> 224          quantreg        5.54
#> 225         quickPlot  0.1.6.9001
#> 226           R.cache      0.14.0
#> 227         R.devices      2.16.1
#> 228       R.methodsS3       1.8.0
#> 229              R.oo      1.23.0
#> 230           R.utils       2.9.2
#> 231                R6       2.4.1
#> 232      RandomFields       3.3.8
#> 233 RandomFieldsUtils       0.5.3
#> 234     RApiSerialize       0.1.0
#> 235          rappdirs       0.3.1
#> 236            raster      3.0-12
#> 237         rasterVis        0.47
#> 238         rcmdcheck       1.3.3
#> 239      RColorBrewer       1.1-2
#> 240              Rcpp       1.0.3
#> 241     RcppArmadillo 0.9.850.1.0
#> 242         RcppEigen   0.3.3.7.0
#> 243             RCurl    1.98-1.1
#> 244              rdoc       0.1.0
#> 245          reactlog       1.0.0
#> 246             readr       1.3.1
#> 247            readxl       1.3.1
#> 248           rematch       1.0.1
#> 249          rematch2       2.1.0
#> 250            remoji       0.1.0
#> 251           remotes       2.1.1
#> 252    ReplaceInFiles       0.1.0
#> 253      reproducible  1.0.0.9006
#> 254      reproducible  1.0.0.9006
#> 255          reshape2       1.4.3
#> 256               rex       1.1.2
#> 257             rgdal       1.4-8
#> 258           rgenoud     5.8-3.0
#> 259             rgeos       0.5-2
#> 260             rgexf      0.16.0
#> 261              rhub       1.1.1
#> 262               rio      0.5.16
#> 263             rJava      0.9-11
#> 264           RJSONIO     1.3-1.4
#> 265             rlang       0.4.4
#> 266        rmapshaper       0.4.3
#> 267         rmarkdown       2.1.1
#> 268              Rook       1.1-1
#> 269          roxygen2       7.0.2
#> 270             rpart      4.1-15
#> 271       RPostgreSQL       0.6-2
#> 272         rprojroot       1.3-2
#> 273         rsconnect      0.8.16
#> 274            Rserve     1.7-3.1
#> 275           RSQLite       2.2.0
#> 276        rstudioapi        0.11
#> 277           rvcheck       0.1.7
#> 278         rversions       2.0.1
#> 279             rvest       0.3.5
#> 280            scales       1.1.0
#> 281              scam       1.2-5
#> 282           selectr       0.4-2
#> 283             servr        0.15
#> 284       sessioninfo       1.1.1
#> 285                sf       0.8-1
#> 286             shiny       1.4.0
#> 287           shinyBS        0.61
#> 288   shinycssloaders         0.3
#> 289    shinydashboard       0.7.1
#> 290           shinyjs         1.1
#> 291      shinyWidgets       0.5.0
#> 292          showtext       0.7-1
#> 293        showtextdb         2.0
#> 294            slackr       1.4.2
#> 295         SnowballC       0.6.0
#> 296       sourcetools       0.1.7
#> 297                sp       1.3-2
#> 298            SpaDES       2.0.4
#> 299     SpaDES.addins       0.1.2
#> 300       SpaDES.core  0.2.9.9000
#> 301      SpaDES.shiny       0.2.0
#> 302      SpaDES.tools  0.3.4.9000
#> 303           SparseM        1.78
#> 304           spatial      7.3-11
#> 305          spatstat      1.63-0
#> 306     spatstat.data       1.4-3
#> 307    spatstat.utils      1.17-0
#> 308            spData       0.3.3
#> 309          spelling         2.1
#> 310           splines       3.6.2
#> 311             stats       3.6.2
#> 312            stats4       3.6.2
#> 313         stopwords         1.0
#> 314        stringdist     0.9.5.5
#> 315           stringi       1.4.5
#> 316           stringr       1.4.0
#> 317            styler       1.3.1
#> 318          survival       3.1-8
#> 319             sylly       0.1-5
#> 320          sylly.en       0.1-3
#> 321               sys         3.3
#> 322          sysfonts         0.8
#> 323             tcltk       3.6.2
#> 324            tensor         1.5
#> 325          testthat       2.3.1
#> 326            tibble       2.1.3
#> 327             tidyr       1.0.2
#> 328        tidyselect       1.0.0
#> 329          tidytext       0.2.2
#> 330             tiler       0.2.4
#> 331          TimeWarp      1.0.15
#> 332           tinytex        0.19
#> 333           tkrplot      0.0-24
#> 334        tokenizers       0.2.1
#> 335             tools       3.6.2
#> 336      translations       3.6.2
#> 337         triebeard       0.3.0
#> 338             udapi       0.1.3
#> 339          udunits2        0.13
#> 340             units       0.6-5
#> 341          urltools       1.7.3
#> 342           usethis       1.5.1
#> 343              utf8       1.1.4
#> 344             utils       3.6.2
#> 345              uuid       0.1-2
#> 346                V8       3.0.1
#> 347             vctrs       0.2.2
#> 348             velox       0.2.0
#> 349          versions         0.3
#> 350              VGAM       1.1-2
#> 351           viridis       0.5.1
#> 352       viridisLite       0.3.0
#> 353        visNetwork       2.0.9
#> 354        visualTest       1.0.0
#> 355           webshot       0.5.2
#> 356           whisker         0.4
#> 357            whoami       1.3.0
#> 358             withr       2.1.2
#> 359    wordcountaddin  0.3.0.9000
#> 360              xfun        0.12
#> 361              xlsx       0.6.2
#> 362          xlsxjars       0.6.1
#> 363               XML    3.99-0.3
#> 364              xml2       1.2.2
#> 365      xmlparsedata       1.0.3
#> 366             xopen       1.0.0
#> 367            xtable       1.8-4
#> 368              yaml       2.2.1
#> 369             yesno       0.1.0
#> 370           zeallot       0.1.0
#> 371               zip       2.0.4
#> 372               zoo       1.8-7

data.table::fread(pkgSnapFile)

#>          instPkgs   instVers
#>   1: reproducible      1.0.0
#>   2: reproducible      1.0.0
#>   3:        abind      1.4-5
#>   4:      acepack      1.4.1
#>   5:          amc 0.1.5.9000
#>  ---                        
#> 368:         yaml      2.2.1
#> 369:        yesno      0.1.0
#> 370:      zeallot      0.1.0
#> 371:          zip      2.0.4
#> 372:          zoo      1.8-7