Take a snapshot of all the packages and version numbers

This can be used later by installVersions to install or re-install the correct versions.

pkgSnapshot(packageVersionFile, libPath, standAlone = FALSE)

Arguments

packageVersionFile
A filename to save the packages and their currently installed version numbers. Defaults to ".packageVersions.txt".

libPath
The path to the local library where packages are installed. Defaults to the .libPaths()[1]

standAlone
Logical. If TRUE, all packages will be installed and loaded strictly from the libPaths only. If FALSE, all .libPaths will be used to find the correct versions. This can be create dramatically faster installs if the user has a substantial number of the packages already in their personal library. In the case of TRUE, there will be a hidden file place in the libPath directory that lists all the packages that were needed during the Require call. Default FALSE to minimize package installing.

Details

A file is written with the package names and versions of all packages within libPath. This can later be passed to Require.

Examples

pkgSnapFile <- tempfile()
pkgSnapshot(pkgSnapFile, .libPaths()[1])

#> There is no /tmp/RtmpIl2XAx/reproducible/cache/.Require/_home_emcintir_R_x86_64-pc-linux-gnu-library_3.6._packageVersionsAuto.txt and standAlone is FALSE. This snapshot will not be accurate because it will include all packages in /home/emcintir/R/x86_64-pc-linux-gnu-library/3.6, /usr/local/lib/R/site-library, /usr/lib/R/site-library, /usr/lib/R/library

#>              instPkgs    instVers
#> 1             caTools    1.17.1.2
#> 2           highlight     0.4.7.2
#> 3               LandR  0.0.2.9001
#> 4               LandR  0.0.2.9001
#> 5                 map       0.0.1
#> 6                 map       0.0.1
#> 7             packrat       0.5.0
#> 8             pkgdown       1.3.0
#> 9            rematch2       2.0.1
#> 10       reproducible  0.2.8.9006
#> 11       reproducible  0.2.8.9006
#> 12          rsconnect      0.8.13
#> 13           TimeWarp      1.0.15
#> 14           TimeWarp      1.0.15
#> 15      alphavantager       0.1.1
#> 16                amc  0.1.5.9000
#> 17          animation         2.6
#> 18          archivist       2.3.2
#> 19          base64enc       0.1-4
#> 20          base64enc       0.1-4
#> 21                bit      1.1-14
#> 22              bit64       0.9-7
#> 23               blob       1.1.1
#> 24              caret      6.0-84
#> 25          checkmate       1.9.3
#> 26          CircStats       0.2-6
#> 27           classInt       0.3-3
#> 28               covr       3.2.1
#> 29         data.table      1.12.2
#> 30                DBI       1.0.0
#> 31            DEoptim       2.2-4
#> 32         DiagrammeR       1.0.1
#> 33              dismo       1.1-4
#> 34         doParallel      1.0.14
#> 35         downloader         0.4
#> 36              e1071       1.7-2
#> 37         fastdigest       0.6-3
#> 38          fasterize       1.0.0
#> 39          fastmatch       1.1-0
#> 40            fastshp       0.1-2
#> 41                 ff      2.2-14
#> 42             ffbase      0.12.7
#> 43              flock         0.7
#> 44            foreach       1.4.4
#> 45           forecast         8.7
#> 46          fpCompare       0.2.2
#> 47           fracdiff       1.4-2
#> 48             future      1.13.0
#> 49          gdalUtils    2.0.1.14
#> 50           generics       0.0.2
#> 51            ggplot2       3.2.0
#> 52            ggplot2       3.2.0
#> 53               glmm       1.3.0
#> 54            globals      0.12.4
#> 55        googledrive       0.1.3
#> 56              gower       0.2.1
#> 57           gridBase       0.4-7
#> 58              highr       0.8.1
#> 59              highr       0.8.1
#> 60                hms       0.4.2
#> 61           hunspell         3.0
#> 62         influenceR       0.1.0
#> 63              ipred       0.9-9
#> 64          iterators      1.0.10
#> 65          itertools       0.1-3
#> 66              LandR  0.0.2.9001
#> 67              LandR  0.0.2.9001
#> 68       LandWebUtils       0.0.1
#> 69               lava       1.6.5
#> 70            listenv       0.7.0
#> 71             lmtest      0.9-37
#> 72          lubridate       1.7.4
#> 73               lucr       0.2.0
#> 74             magick         2.0
#> 75                map       0.0.1
#> 76                map       0.0.1
#> 77           maptools       0.9-5
#> 78     microbenchmark       1.4-6
#> 79               mime         0.7
#> 80               mime         0.7
#> 81       ModelMetrics       1.2.2
#> 82              MuMIn      1.43.8
#> 83            mvtnorm      1.0-10
#> 84           numDeriv  2016.8-1.1
#> 85               padr       0.5.0
#> 86              pedev  0.0.1.9005
#> 87             pemisc  0.0.1.9003
#> 88             praise       1.0.0
#> 89            prodlim  2018.04.18
#> 90            profvis       0.3.6
#> 91               pryr       0.1.4
#> 92           quadprog       1.5-7
#> 93           quantmod      0.4-15
#> 94          quickPlot  0.1.6.9000
#> 95        R.methodsS3       1.7.1
#> 96               R.oo      1.22.0
#> 97            R.utils       2.9.0
#> 98      RcppArmadillo 0.9.500.2.0
#> 99           RcppRoll       0.3.0
#> 100             RCurl   1.95-4.12
#> 101             readr       1.3.1
#> 102           recipes       0.1.5
#> 103           remotes       2.1.0
#> 104           remotes       2.1.0
#> 105      reproducible  0.2.8.9005
#> 106      reproducible  0.2.8.9005
#> 107               rex       1.1.2
#> 108             rgdal       1.4-4
#> 109           rgenoud     5.8-3.0
#> 110             rgeos       0.4-3
#> 111             rgexf      0.15.3
#> 112         rmarkdown        1.13
#> 113              Rook       1.1-1
#> 114         rprojroot       1.3-2
#> 115         rprojroot       1.3-2
#> 116           RSQLite       2.1.1
#> 117              scam       1.2-3
#> 118                sf       0.7-4
#> 119            SpaDES       2.0.3
#> 120     SpaDES.addins       0.1.2
#> 121       SpaDES.core  0.2.5.9007
#> 122      SpaDES.tools  0.3.2.9001
#> 123           SQUAREM   2017.10-1
#> 124          testthat       2.1.1
#> 125            tibble       2.1.3
#> 126            tibble       2.1.3
#> 127             tiler       0.2.1
#> 128          timeDate    3043.102
#> 129            timetk     0.1.1.1
#> 130          TimeWarp      1.0.15
#> 131          TimeWarp      1.0.15
#> 132           tinytex        0.13
#> 133             trust       0.1-7
#> 134           tseries     0.10-47
#> 135               TTR      0.23-4
#> 136             units       0.6-3
#> 137              urca       1.3-0
#> 138              uuid       0.1-2
#> 139          versions         0.3
#> 140              VGAM       1.1-1
#> 141        visNetwork       2.0.7
#> 142           xgboost      0.82.1
#> 143               XML   3.98-1.20
#> 144            xtable       1.8-4
#> 145            xtable       1.8-4
#> 146           askpass         1.1
#> 147        assertthat       0.2.1
#> 148         backports       1.1.4
#> 149         base64enc       0.1-3
#> 150         base64enc       0.1-3
#> 151                BH    1.69.0-1
#> 152            bitops       1.0-6
#> 153              brew       1.0-6
#> 154             callr       3.2.0
#> 155               cli       1.1.0
#> 156             clipr       0.6.0
#> 157        clisymbols       1.2.0
#> 158              coda      0.19-2
#> 159        colorspace       1.4-1
#> 160        commonmark         1.7
#> 161            crayon       1.3.4
#> 162         crosstalk       1.0.0
#> 163              curl         3.3
#> 164              desc       1.2.0
#> 165          devtools       2.0.2
#> 166            digest      0.6.19
#> 167             dplyr       0.8.1
#> 168          evaluate        0.14
#> 169             fansi       0.4.0
#> 170               FNN       1.1.3
#> 171                fs       1.3.1
#> 172           gdtools       0.1.9
#> 173              geoR   1.7-5.2.1
#> 174             ggmap       3.0.0
#> 175           ggplot2       3.2.0
#> 176           ggplot2       3.2.0
#> 177                gh       1.0.1
#> 178             git2r      0.25.2
#> 179              glue       1.3.1
#> 180         gridExtra         2.3
#> 181             gstat       2.0-2
#> 182            gtable       0.3.0
#> 183             highr         0.8
#> 184             highr         0.8
#> 185         htmltools       0.3.6
#> 186       htmlwidgets         1.3
#> 187            httpuv       1.5.1
#> 188              httr       1.4.0
#> 189            igraph     1.2.4.1
#> 190               ini       0.3.1
#> 191         intervals      0.15.1
#> 192              jpeg       0.1-8
#> 193          jsonlite         1.6
#> 194             knitr        1.23
#> 195          labeling         0.3
#> 196             later       0.8.0
#> 197          lazyeval       0.2.2
#> 198           leaflet       2.0.2
#> 199           littler       0.3.8
#> 200              lme4      1.1-21
#> 201          magrittr         1.5
#> 202  manipulateWidget      0.10.0
#> 203           mapdata       2.3.0
#> 204              maps       3.3.0
#> 205          markdown         0.9
#> 206           memoise       1.1.0
#> 207              mime         0.7
#> 208              mime         0.7
#> 209            miniUI     0.1.1.1
#> 210             minqa       1.2.4
#> 211           munsell       0.5.0
#> 212             ncdf4      1.16.1
#> 213            nloptr       1.2.1
#> 214           openssl         1.4
#> 215            pillar       1.4.1
#> 216          pkgbuild       1.0.3
#> 217         pkgconfig       2.0.2
#> 218         pkgKitten       0.1.4
#> 219           pkgload       1.0.2
#> 220             plogr       0.2.0
#> 221              plyr       1.8.4
#> 222               png       0.1-7
#> 223       prettyunits       1.0.2
#> 224          processx       3.3.1
#> 225          promises       1.0.1
#> 226                ps       1.3.0
#> 227             purrr       0.3.2
#> 228                R6       2.4.0
#> 229      RandomFields       3.3.6
#> 230 RandomFieldsUtils       0.5.3
#> 231            raster       2.9-5
#> 232         rcmdcheck       1.3.3
#> 233      RColorBrewer       1.1-2
#> 234              Rcpp       1.0.1
#> 235         RcppEigen   0.3.3.5.0
#> 236           remotes       2.0.4
#> 237           remotes       2.0.4
#> 238          reshape2       1.4.3
#> 239               rgl    0.100.19
#> 240       RgoogleMaps       1.4.3
#> 241             rjags         4-8
#> 242             rJava      0.9-11
#> 243             rjson      0.2.20
#> 244             rlang       0.3.4
#> 245          roxygen2       6.1.1
#> 246         rprojroot         1.2
#> 247         rprojroot         1.2
#> 248        rstudioapi        0.10
#> 249            scales       1.0.0
#> 250       sessioninfo       1.1.1
#> 251             shiny       1.3.2
#> 252              snow       0.4-3
#> 253       sourcetools       0.1.7
#> 254                sp       1.3-1
#> 255         spacetime       1.2-2
#> 256           splancs     2.01-40
#> 257           stringi       1.4.3
#> 258           stringr       1.4.0
#> 259           svglite       1.2.2
#> 260               sys         3.2
#> 261            tibble       2.1.3
#> 262            tibble       2.1.3
#> 263             tidyr       0.8.3
#> 264        tidyselect       0.2.5
#> 265           tkrplot      0.0-24
#> 266           usethis       1.5.0
#> 267              utf8       1.1.4
#> 268             vctrs       0.1.0
#> 269           viridis       0.5.1
#> 270       viridisLite       0.3.0
#> 271           webshot       0.5.1
#> 272           whisker       0.3-2
#> 273             withr       2.1.2
#> 274              xfun         0.7
#> 275              xlsx       0.6.1
#> 276          xlsxjars       0.6.1
#> 277              xml2       1.2.0
#> 278             xopen       1.0.0
#> 279            xtable       1.8-3
#> 280            xtable       1.8-3
#> 281               xts      0.11-2
#> 282              yaml       2.2.0
#> 283           zeallot       0.1.0
#> 284               zoo       1.8-6
#> 285              base       3.6.0
#> 286              boot      1.3-22
#> 287             class      7.3-15
#> 288           cluster       2.1.0
#> 289         codetools      0.2-16
#> 290          compiler       3.6.0
#> 291          datasets       3.6.0
#> 292           foreign      0.8-71
#> 293          graphics       3.6.0
#> 294         grDevices       3.6.0
#> 295              grid       3.6.0
#> 296        KernSmooth     2.23-15
#> 297           lattice     0.20-38
#> 298              MASS    7.3-51.4
#> 299            Matrix      1.2-17
#> 300           methods       3.6.0
#> 301              mgcv      1.8-28
#> 302              nlme     3.1-140
#> 303              nnet      7.3-12
#> 304          parallel       3.6.0
#> 305             rpart      4.1-15
#> 306           spatial      7.3-11
#> 307           splines       3.6.0
#> 308             stats       3.6.0
#> 309            stats4       3.6.0
#> 310          survival    2.44-1.1
#> 311             tcltk       3.6.0
#> 312             tools       3.6.0
#> 313      translations       3.6.0
#> 314             utils       3.6.0

data.table::fread(pkgSnapFile)

#>          instPkgs   instVers
#>   1:      caTools   1.17.1.2
#>   2:    highlight    0.4.7.2
#>   3:        LandR 0.0.2.9001
#>   4:        LandR 0.0.2.9001
#>   5:          map      0.0.1
#>  ---                        
#> 310:     survival   2.44-1.1
#> 311:        tcltk      3.6.0
#> 312:        tools      3.6.0
#> 313: translations      3.6.0
#> 314:        utils      3.6.0