Take a snapshot of all the packages and version numbers

This can be used later by installVersions to install or re-install the correct versions.

pkgSnapshot(packageVersionFile, libPath, standAlone = FALSE)

Arguments

packageVersionFile
A filename to save the packages and their currently installed version numbers. Defaults to ".packageVersions.txt".

libPath
The path to the local library where packages are installed. Defaults to the .libPaths()[1]

standAlone
Logical. If TRUE, all packages will be installed and loaded strictly from the libPaths only. If FALSE, all .libPaths will be used to find the correct versions. This can be create dramatically faster installs if the user has a substantial number of the packages already in their personal library. In the case of TRUE, there will be a hidden file place in the libPath directory that lists all the packages that were needed during the Require call. Default FALSE to minimize package installing.

Details

A file is written with the package names and versions of all packages within libPath. This can later be passed to Require.

Examples

pkgSnapFile <- tempfile()
pkgSnapshot(pkgSnapFile, .libPaths()[1])

#> There is no /tmp/RtmpEvs1Yv/reproducible/cache/.Require/_tmp_RtmpuMyl9A_temp_libpath41b0757c6080._packageVersionsAuto.txt and standAlone is FALSE. This snapshot will not be accurate because it will include all packages in /tmp/RtmpuMyl9A/temp_libpath41b0757c6080, /home/emcintir/R/x86_64-pc-linux-gnu-library/3.6, /usr/local/lib/R/site-library, /usr/lib/R/site-library, /usr/lib/R/library

#>                                instPkgs    instVers
#> 1             reproducible.reproducible      0.2.11
#> 2             reproducible.reproducible      0.2.11
#> 3                           abind.abind       1.4-5
#> 4                       acepack.acepack       1.4.1
#> 5                               amc.amc  0.1.5.9000
#> 6                               amc.amc  0.1.5.9000
#> 7                           bench.bench       1.0.4
#> 8                               car.car       3.0-5
#> 9                       carData.carData       3.0-3
#> 10                      caTools.caTools    1.17.1.2
#> 11                cellranger.cellranger       1.1.0
#> 12                  crancache.crancache  0.0.0.9001
#> 13                    cranlike.cranlike       1.0.2
#> 14                data.table.data.table      1.12.6
#> 15                data.table.data.table      1.12.6
#> 16                      debugme.debugme       1.1.0
#> 17                    devtools.devtools       2.2.1
#> 18                    devtools.devtools       2.2.1
#> 19                        digest.digest      0.6.22
#> 20                        digest.digest      0.6.22
#> 21                    ellipsis.ellipsis       0.3.0
#> 22                    ellipsis.ellipsis       0.3.0
#> 23                      fastmap.fastmap       1.0.1
#> 24                      Finance.Finance         1.0
#> 25                      forcats.forcats       0.4.0
#> 26                      Formula.Formula       1.2-3
#> 27                        future.future      1.15.0
#> 28                        future.future      1.15.0
#> 29            future.apply.future.apply       1.3.0
#> 30            future.callr.future.callr       0.5.0
#> 31                        gargle.gargle       0.4.0
#> 32                        gmailr.gmailr       1.0.0
#> 33                          haven.haven       2.2.0
#> 34                  highlight.highlight     0.4.7.2
#> 35                          Hmisc.Hmisc       4.3-0
#> 36                  htmlTable.htmlTable      1.13.2
#> 37                        inline.inline      0.3.15
#> 38                          LandR.LandR  0.0.2.9007
#> 39                          LandR.LandR  0.0.2.9007
#> 40                    LandR.CS.LandR.CS  0.0.0.9000
#> 41            LandWebUtils.LandWebUtils       0.0.2
#> 42            LandWebUtils.LandWebUtils       0.0.2
#> 43            latticeExtra.latticeExtra      0.6-28
#> 44                  lifecycle.lifecycle       0.1.0
#> 45                      littler.littler       0.3.9
#> 46                      littler.littler       0.3.9
#> 47                              map.map       0.0.1
#> 48                              map.map       0.0.1
#> 49            MatrixModels.MatrixModels       0.4-1
#> 50              matrixStats.matrixStats      0.55.0
#> 51                    NetLogoR.NetLogoR       0.3.5
#> 52                    openxlsx.openxlsx       4.1.3
#> 53                      packrat.packrat       0.5.0
#> 54                  parsedate.parsedate       1.2.0
#> 55                    pbkrtest.pbkrtest       0.4-7
#> 56                          pedev.pedev  0.0.1.9005
#> 57                          pedev.pedev  0.0.1.9005
#> 58                        pemisc.pemisc  0.0.1.9004
#> 59                        pemisc.pemisc  0.0.1.9004
#> 60                      pkgdown.pkgdown       1.4.1
#> 61                      profmem.profmem       0.5.0
#> 62                    progress.progress       1.2.2
#> 63                    quantreg.quantreg        5.52
#> 64                  quickPlot.quickPlot  0.1.6.9000
#> 65                  quickPlot.quickPlot  0.1.6.9000
#> 66                    rappdirs.rappdirs       0.3.1
#> 67                        readxl.readxl       1.3.1
#> 68                      rematch.rematch       1.0.1
#> 69                    rematch2.rematch2       2.1.0
#> 70            reproducible.reproducible      0.2.11
#> 71            reproducible.reproducible      0.2.11
#> 72              revdepcheck.revdepcheck  1.0.0.9001
#> 73                          rgeos.rgeos       0.5-2
#> 74                          rgeos.rgeos       0.5-2
#> 75                              rio.rio      0.5.16
#> 76                    roxygen2.roxygen2       7.0.0
#> 77                    roxygen2.roxygen2       7.0.0
#> 78                  rsconnect.rsconnect      0.8.15
#> 79                        SpaDES.SpaDES  2.0.4.9000
#> 80                        SpaDES.SpaDES  2.0.4.9000
#> 81              SpaDES.core.SpaDES.core  0.2.6.9006
#> 82              SpaDES.core.SpaDES.core  0.2.6.9006
#> 83  SpaDES.experiment.SpaDES.experiment  0.0.1.9000
#> 84            SpaDES.tools.SpaDES.tools  0.3.2.9002
#> 85            SpaDES.tools.SpaDES.tools  0.3.2.9002
#> 86                      SparseM.SparseM        1.77
#> 87                    spelling.spelling         2.1
#> 88                    testthat.testthat       2.3.0
#> 89                    testthat.testthat       2.3.0
#> 90                    testthat.testthat       2.3.0
#> 91                    TimeWarp.TimeWarp      1.0.15
#> 92                    TimeWarp.TimeWarp      1.0.15
#> 93                        whoami.whoami       1.3.0
#> 94                              zip.zip       2.0.4
#> 95          alphavantager.alphavantager       0.1.1
#> 96                              amc.amc  0.1.5.9000
#> 97                              amc.amc  0.1.5.9000
#> 98                  animation.animation         2.6
#> 99                  archivist.archivist       2.3.4
#> 100                 base64enc.base64enc       0.1-4
#> 101                 base64enc.base64enc       0.1-4
#> 102                             bit.bit      1.1-14
#> 103                         bit64.bit64       0.9-7
#> 104                           blob.blob       1.2.0
#> 105                         caret.caret      6.0-84
#> 106                 checkmate.checkmate       1.9.4
#> 107                 CircStats.CircStats       0.2-6
#> 108                   classInt.classInt       0.4-2
#> 109                           covr.covr       3.3.2
#> 110               data.table.data.table      1.12.2
#> 111               data.table.data.table      1.12.2
#> 112                             DBI.DBI       1.0.0
#> 113                     DEoptim.DEoptim       2.2-4
#> 114               DiagrammeR.DiagrammeR       1.0.1
#> 115                         dismo.dismo       1.1-4
#> 116               doParallel.doParallel      1.0.15
#> 117               downloader.downloader         0.4
#> 118                         e1071.e1071       1.7-2
#> 119               fastdigest.fastdigest       0.6-3
#> 120                 fasterize.fasterize       1.0.0
#> 121                 fastmatch.fastmatch       1.1-0
#> 122                     fastshp.fastshp       0.1-2
#> 123                               ff.ff      2.2-14
#> 124                       ffbase.ffbase      0.12.7
#> 125                         flock.flock         0.7
#> 126                     foreach.foreach       1.4.7
#> 127                   forecast.forecast         8.9
#> 128                 fpCompare.fpCompare       0.2.3
#> 129                   fracdiff.fracdiff       1.4-2
#> 130                       future.future      1.14.0
#> 131                       future.future      1.14.0
#> 132                 gdalUtils.gdalUtils    2.0.1.14
#> 133                   generics.generics       0.0.2
#> 134                     ggplot2.ggplot2       3.2.1
#> 135                     ggplot2.ggplot2       3.2.1
#> 136                           glmm.glmm       1.3.0
#> 137                     globals.globals      0.12.4
#> 138             googledrive.googledrive       1.0.0
#> 139                         gower.gower       0.2.1
#> 140                   gridBase.gridBase       0.4-7
#> 141                         highr.highr       0.8.1
#> 142                         highr.highr       0.8.1
#> 143                             hms.hms       0.5.2
#> 144                   hunspell.hunspell         3.0
#> 145               influenceR.influenceR       0.1.0
#> 146                         ipred.ipred       0.9-9
#> 147                 iterators.iterators      1.0.12
#> 148                 itertools.itertools       0.1-3
#> 149                         LandR.LandR  0.0.2.9001
#> 150                         LandR.LandR  0.0.2.9001
#> 151           LandWebUtils.LandWebUtils       0.0.1
#> 152           LandWebUtils.LandWebUtils       0.0.1
#> 153                           lava.lava       1.6.6
#> 154                     listenv.listenv       0.7.0
#> 155                       lmtest.lmtest      0.9-37
#> 156                 lubridate.lubridate       1.7.4
#> 157                           lucr.lucr       0.2.0
#> 158                       magick.magick         2.2
#> 159                             map.map       0.0.1
#> 160                             map.map       0.0.1
#> 161                   maptools.maptools       0.9-8
#> 162       microbenchmark.microbenchmark       1.4-7
#> 163                           mime.mime         0.7
#> 164                           mime.mime         0.7
#> 165           ModelMetrics.ModelMetrics       1.2.2
#> 166                         MuMIn.MuMIn      1.43.8
#> 167                     mvtnorm.mvtnorm      1.0-11
#> 168                   numDeriv.numDeriv  2016.8-1.1
#> 169                           padr.padr       0.5.0
#> 170                         pedev.pedev  0.0.1.9005
#> 171                         pedev.pedev  0.0.1.9005
#> 172                       pemisc.pemisc  0.0.1.9003
#> 173                       pemisc.pemisc  0.0.1.9003
#> 174                       praise.praise       1.0.0
#> 175                       praise.praise       1.0.0
#> 176                     prodlim.prodlim  2019.10.13
#> 177                     profvis.profvis       0.3.6
#> 178                           pryr.pryr       0.1.4
#> 179                   quadprog.quadprog       1.5-7
#> 180                   quantmod.quantmod      0.4-15
#> 181                 quickPlot.quickPlot  0.1.6.9000
#> 182                 quickPlot.quickPlot  0.1.6.9000
#> 183             R.methodsS3.R.methodsS3       1.7.1
#> 184                           R.oo.R.oo      1.23.0
#> 185                     R.utils.R.utils       2.9.0
#> 186         RcppArmadillo.RcppArmadillo 0.9.800.1.0
#> 187                   RcppRoll.RcppRoll       0.3.0
#> 188                         RCurl.RCurl   1.95-4.12
#> 189                         readr.readr       1.3.1
#> 190                     recipes.recipes       0.1.7
#> 191                     remotes.remotes       2.1.0
#> 192                     remotes.remotes       2.1.0
#> 193                             rex.rex       1.1.2
#> 194                         rgdal.rgdal       1.4-7
#> 195                     rgenoud.rgenoud     5.8-3.0
#> 196                         rgeos.rgeos       0.4-3
#> 197                         rgeos.rgeos       0.4-3
#> 198                         rgexf.rgexf      0.15.3
#> 199                 rmarkdown.rmarkdown        1.17
#> 200                           Rook.Rook       1.1-1
#> 201                 rprojroot.rprojroot       1.3-2
#> 202                 rprojroot.rprojroot       1.3-2
#> 203                     RSQLite.RSQLite       2.1.2
#> 204                           scam.scam       1.2-5
#> 205                               sf.sf       0.8-0
#> 206                       SpaDES.SpaDES       2.0.3
#> 207                       SpaDES.SpaDES       2.0.3
#> 208         SpaDES.addins.SpaDES.addins       0.1.2
#> 209             SpaDES.core.SpaDES.core  0.2.5.9007
#> 210             SpaDES.core.SpaDES.core  0.2.5.9007
#> 211           SpaDES.tools.SpaDES.tools  0.3.2.9001
#> 212           SpaDES.tools.SpaDES.tools  0.3.2.9001
#> 213                     SQUAREM.SQUAREM   2017.10-1
#> 214                   testthat.testthat       2.1.1
#> 215                   testthat.testthat       2.1.1
#> 216                   testthat.testthat       2.1.1
#> 217                       tibble.tibble       2.1.3
#> 218                       tibble.tibble       2.1.3
#> 219                         tiler.tiler       0.2.1
#> 220                   timeDate.timeDate    3043.102
#> 221                       timetk.timetk       0.1.2
#> 222                   TimeWarp.TimeWarp      1.0.15
#> 223                   TimeWarp.TimeWarp      1.0.15
#> 224                     tinytex.tinytex        0.17
#> 225                         trust.trust       0.1-7
#> 226                     tseries.tseries     0.10-47
#> 227                             TTR.TTR      0.23-5
#> 228                         units.units       0.6-5
#> 229                           urca.urca       1.3-0
#> 230                           uuid.uuid       0.1-2
#> 231                   versions.versions         0.3
#> 232                           VGAM.VGAM       1.1-1
#> 233               visNetwork.visNetwork       2.0.8
#> 234                     xgboost.xgboost    0.90.0.2
#> 235                             XML.XML   3.98-1.20
#> 236                       xtable.xtable       1.8-4
#> 237                       xtable.xtable       1.8-4
#> 238                     askpass.askpass         1.1
#> 239               assertthat.assertthat       0.2.1
#> 240                 backports.backports       1.1.5
#> 241                 base64enc.base64enc       0.1-3
#> 242                 base64enc.base64enc       0.1-3
#> 243                               BH.BH    1.69.0-1
#> 244                       bitops.bitops       1.0-6
#> 245                           brew.brew       1.0-6
#> 246                         callr.callr       3.3.2
#> 247                             cli.cli       1.1.0
#> 248                         clipr.clipr       0.7.0
#> 249               clisymbols.clisymbols       1.2.0
#> 250                           coda.coda      0.19-3
#> 251               colorspace.colorspace       1.4-1
#> 252               commonmark.commonmark         1.7
#> 253                       crayon.crayon       1.3.4
#> 254                 crosstalk.crosstalk       1.0.0
#> 255                           curl.curl         4.2
#> 256                           desc.desc       1.2.0
#> 257                   devtools.devtools       2.1.0
#> 258                   devtools.devtools       2.1.0
#> 259                       digest.digest      0.6.21
#> 260                       digest.digest      0.6.21
#> 261                         dplyr.dplyr       0.8.3
#> 262                               DT.DT        0.10
#> 263                   ellipsis.ellipsis     0.2.0.1
#> 264                   ellipsis.ellipsis     0.2.0.1
#> 265                   evaluate.evaluate        0.14
#> 266                         fansi.fansi       0.4.0
#> 267                             FNN.FNN       1.1.3
#> 268                               fs.fs       1.3.1
#> 269                     gdtools.gdtools       0.2.1
#> 270                           geoR.geoR   1.7-5.2.1
#> 271                         ggmap.ggmap       3.0.0
#> 272                     ggplot2.ggplot2       3.2.1
#> 273                     ggplot2.ggplot2       3.2.1
#> 274                               gh.gh       1.0.1
#> 275                         git2r.git2r      0.26.1
#> 276                           glue.glue       1.3.1
#> 277                 gridExtra.gridExtra         2.3
#> 278                         gstat.gstat       2.0-3
#> 279                       gtable.gtable       0.3.0
#> 280                         highr.highr         0.8
#> 281                         highr.highr         0.8
#> 282                 htmltools.htmltools       0.4.0
#> 283             htmlwidgets.htmlwidgets       1.5.1
#> 284                       httpuv.httpuv       1.5.2
#> 285                           httr.httr       1.4.1
#> 286                       igraph.igraph     1.2.4.1
#> 287                             ini.ini       0.3.1
#> 288                 intervals.intervals      0.15.1
#> 289                           jpeg.jpeg     0.1-8.1
#> 290                   jsonlite.jsonlite         1.6
#> 291                         knitr.knitr        1.26
#> 292                   labeling.labeling         0.3
#> 293                         later.later       1.0.0
#> 294                   lazyeval.lazyeval       0.2.2
#> 295                     leaflet.leaflet       2.0.3
#> 296 leaflet.providers.leaflet.providers       1.9.0
#> 297                     littler.littler       0.3.8
#> 298                     littler.littler       0.3.8
#> 299                           lme4.lme4      1.1-21
#> 300                   magrittr.magrittr         1.5
#> 301   manipulateWidget.manipulateWidget      0.10.0
#> 302                     mapdata.mapdata       2.3.0
#> 303                           maps.maps       3.3.0
#> 304                   markdown.markdown         1.1
#> 305                     memoise.memoise       1.1.0
#> 306                           mime.mime         0.7
#> 307                           mime.mime         0.7
#> 308                       miniUI.miniUI     0.1.1.1
#> 309                         minqa.minqa       1.2.4
#> 310                     munsell.munsell       0.5.0
#> 311                         ncdf4.ncdf4        1.17
#> 312                       nloptr.nloptr       1.2.1
#> 313                     openssl.openssl       1.4.1
#> 314                       pillar.pillar       1.4.2
#> 315                   pkgbuild.pkgbuild       1.0.6
#> 316                 pkgconfig.pkgconfig       2.0.3
#> 317                 pkgKitten.pkgKitten       0.1.5
#> 318                     pkgload.pkgload       1.0.2
#> 319                         plogr.plogr       0.2.0
#> 320                           plyr.plyr       1.8.4
#> 321                             png.png       0.1-7
#> 322                       praise.praise       1.0.0
#> 323                       praise.praise       1.0.0
#> 324             prettyunits.prettyunits       1.0.2
#> 325                   processx.processx       3.4.1
#> 326                   promises.promises       1.1.0
#> 327                               ps.ps       1.3.0
#> 328                         purrr.purrr       0.3.3
#> 329                               R6.R6       2.4.1
#> 330           RandomFields.RandomFields       3.3.6
#> 331 RandomFieldsUtils.RandomFieldsUtils       0.5.3
#> 332                       raster.raster       3.0-7
#> 333                 rcmdcheck.rcmdcheck       1.3.3
#> 334           RColorBrewer.RColorBrewer       1.1-2
#> 335                           Rcpp.Rcpp       1.0.3
#> 336                 RcppEigen.RcppEigen   0.3.3.7.0
#> 337                     remotes.remotes       2.1.0
#> 338                     remotes.remotes       2.1.0
#> 339                   reshape2.reshape2       1.4.3
#> 340                             rgl.rgl    0.100.30
#> 341             RgoogleMaps.RgoogleMaps       1.4.4
#> 342                         rjags.rjags        4-10
#> 343                         rJava.rJava      0.9-11
#> 344                         rjson.rjson      0.2.20
#> 345                         rlang.rlang       0.4.1
#> 346                   roxygen2.roxygen2       7.0.0
#> 347                   roxygen2.roxygen2       7.0.0
#> 348                 rprojroot.rprojroot         1.2
#> 349                 rprojroot.rprojroot         1.2
#> 350               rstudioapi.rstudioapi        0.10
#> 351                 rversions.rversions       2.0.0
#> 352                       scales.scales       1.0.0
#> 353             sessioninfo.sessioninfo       1.1.1
#> 354                         shiny.shiny       1.4.0
#> 355                           snow.snow       0.4-3
#> 356             sourcetools.sourcetools       0.1.7
#> 357                               sp.sp       1.3-2
#> 358                 spacetime.spacetime       1.2-2
#> 359                     splancs.splancs     2.01-40
#> 360                     stringi.stringi       1.4.3
#> 361                     stringr.stringr       1.4.0
#> 362                     svglite.svglite       1.2.2
#> 363                             sys.sys         3.3
#> 364             systemfonts.systemfonts       0.1.1
#> 365                   testthat.testthat       2.2.1
#> 366                   testthat.testthat       2.2.1
#> 367                   testthat.testthat       2.2.1
#> 368                       tibble.tibble       2.1.3
#> 369                       tibble.tibble       2.1.3
#> 370                         tidyr.tidyr       1.0.0
#> 371               tidyselect.tidyselect       0.2.5
#> 372                     tkrplot.tkrplot      0.0-24
#> 373                     usethis.usethis       1.5.1
#> 374                           utf8.utf8       1.1.4
#> 375                         vctrs.vctrs       0.2.0
#> 376                     viridis.viridis       0.5.1
#> 377             viridisLite.viridisLite       0.3.0
#> 378                     webshot.webshot       0.5.1
#> 379                     whisker.whisker         0.4
#> 380                         withr.withr       2.1.2
#> 381                           xfun.xfun        0.11
#> 382                           xlsx.xlsx       0.6.1
#> 383                   xlsxjars.xlsxjars       0.6.1
#> 384                           xml2.xml2       1.2.2
#> 385                         xopen.xopen       1.0.0
#> 386                       xtable.xtable       1.8-3
#> 387                       xtable.xtable       1.8-3
#> 388                             xts.xts      0.11-2
#> 389                           yaml.yaml       2.2.0
#> 390                     zeallot.zeallot       0.1.0
#> 391                             zoo.zoo       1.8-6
#> 392                           base.base       3.6.1
#> 393                           boot.boot      1.3-23
#> 394                         class.class      7.3-15
#> 395                     cluster.cluster       2.1.0
#> 396                 codetools.codetools      0.2-16
#> 397                   compiler.compiler       3.6.1
#> 398                   datasets.datasets       3.6.1
#> 399                     foreign.foreign      0.8-72
#> 400                   graphics.graphics       3.6.1
#> 401                 grDevices.grDevices       3.6.1
#> 402                           grid.grid       3.6.1
#> 403               KernSmooth.KernSmooth     2.23-16
#> 404                     lattice.lattice     0.20-38
#> 405                           MASS.MASS    7.3-51.4
#> 406                       Matrix.Matrix      1.2-17
#> 407                     methods.methods       3.6.1
#> 408                           mgcv.mgcv      1.8-31
#> 409                           nlme.nlme     3.1-142
#> 410                           nnet.nnet      7.3-12
#> 411                   parallel.parallel       3.6.1
#> 412                         rpart.rpart      4.1-15
#> 413                     spatial.spatial      7.3-11
#> 414                     splines.splines       3.6.1
#> 415                         stats.stats       3.6.1
#> 416                       stats4.stats4       3.6.1
#> 417                   survival.survival       3.1-7
#> 418                         tcltk.tcltk       3.6.1
#> 419                         tools.tools       3.6.1
#> 420           translations.translations       3.6.1
#> 421                         utils.utils       3.6.1

data.table::fread(pkgSnapFile)

#>                       instPkgs   instVers
#>   1: reproducible.reproducible     0.2.11
#>   2: reproducible.reproducible     0.2.11
#>   3:               abind.abind      1.4-5
#>   4:           acepack.acepack      1.4.1
#>   5:                   amc.amc 0.1.5.9000
#>  ---                                     
#> 417:         survival.survival      3.1-7
#> 418:               tcltk.tcltk      3.6.1
#> 419:               tools.tools      3.6.1
#> 420: translations.translations      3.6.1
#> 421:               utils.utils      3.6.1