Take a snapshot of all the packages and version numbers

This can be used later by installVersions to install or re-install the correct versions.

pkgSnapshot(packageVersionFile, libPath, standAlone = FALSE)

Arguments

packageVersionFile
A filename to save the packages and their currently installed version numbers. Defaults to ".packageVersions.txt".

libPath
The path to the local library where packages are installed. Defaults to the .libPaths()[1]

standAlone
Logical. If TRUE, all packages will be installed and loaded strictly from the libPaths only. If FALSE, all .libPaths will be used to find the correct versions. This can be create dramatically faster installs if the user has a substantial number of the packages already in their personal library. In the case of TRUE, there will be a hidden file place in the libPath directory that lists all the packages that were needed during the Require call. Default FALSE to minimize package installing.

Details

A file is written with the package names and versions of all packages within libPath. This can later be passed to Require.

Examples

pkgSnapFile <- tempfile()
pkgSnapshot(pkgSnapFile, .libPaths()[1])

#> There is no /tmp/RtmpO4wS3z/reproducible/cache/.Require/_tmp_RtmpYJaHHM_temp_libpath2f447e59b81._packageVersionsAuto.txt and standAlone is FALSE. This snapshot will not be accurate because it will include all packages in /tmp/RtmpYJaHHM/temp_libpath2f447e59b81, /home/emcintir/R/x86_64-pc-linux-gnu-library/3.6, /usr/local/lib/R/site-library, /usr/lib/R/site-library, /usr/lib/R/library

#>                                instPkgs    instVers
#> 1             reproducible.reproducible      0.2.11
#> 2             reproducible.reproducible      0.2.11
#> 3                           abind.abind       1.4-5
#> 4                       acepack.acepack       1.4.1
#> 5                               amc.amc  0.1.5.9000
#> 6                               amc.amc  0.1.5.9000
#> 7                           bench.bench       1.0.4
#> 8                               car.car       3.0-5
#> 9                       carData.carData       3.0-3
#> 10                      caTools.caTools    1.17.1.2
#> 11                cellranger.cellranger       1.1.0
#> 12                data.table.data.table      1.12.6
#> 13                data.table.data.table      1.12.6
#> 14                    devtools.devtools       2.2.1
#> 15                    devtools.devtools       2.2.1
#> 16                        digest.digest      0.6.22
#> 17                        digest.digest      0.6.22
#> 18                    ellipsis.ellipsis       0.3.0
#> 19                    ellipsis.ellipsis       0.3.0
#> 20                      fastmap.fastmap       1.0.1
#> 21                      Finance.Finance         1.0
#> 22                      forcats.forcats       0.4.0
#> 23                      Formula.Formula       1.2-3
#> 24                        future.future      1.15.0
#> 25                        future.future      1.15.0
#> 26            future.apply.future.apply       1.3.0
#> 27            future.callr.future.callr       0.5.0
#> 28                        gargle.gargle       0.4.0
#> 29                          haven.haven       2.2.0
#> 30                  highlight.highlight     0.4.7.2
#> 31                          Hmisc.Hmisc       4.3-0
#> 32                  htmlTable.htmlTable      1.13.2
#> 33                        inline.inline      0.3.15
#> 34                          LandR.LandR  0.0.2.9007
#> 35                          LandR.LandR  0.0.2.9007
#> 36                    LandR.CS.LandR.CS  0.0.0.9000
#> 37            LandWebUtils.LandWebUtils       0.0.2
#> 38            LandWebUtils.LandWebUtils       0.0.2
#> 39            latticeExtra.latticeExtra      0.6-28
#> 40                  lifecycle.lifecycle       0.1.0
#> 41                      littler.littler       0.3.9
#> 42                      littler.littler       0.3.9
#> 43                              map.map       0.0.1
#> 44                              map.map       0.0.1
#> 45            MatrixModels.MatrixModels       0.4-1
#> 46              matrixStats.matrixStats      0.55.0
#> 47                    NetLogoR.NetLogoR       0.3.5
#> 48                    openxlsx.openxlsx       4.1.3
#> 49                      packrat.packrat       0.5.0
#> 50                    pbkrtest.pbkrtest       0.4-7
#> 51                          pedev.pedev  0.0.1.9005
#> 52                          pedev.pedev  0.0.1.9005
#> 53                        pemisc.pemisc  0.0.1.9004
#> 54                        pemisc.pemisc  0.0.1.9004
#> 55                      pkgdown.pkgdown       1.4.1
#> 56                      profmem.profmem       0.5.0
#> 57                    progress.progress       1.2.2
#> 58                    quantreg.quantreg        5.52
#> 59                  quickPlot.quickPlot  0.1.6.9000
#> 60                  quickPlot.quickPlot  0.1.6.9000
#> 61                        readxl.readxl       1.3.1
#> 62                      rematch.rematch       1.0.1
#> 63                    rematch2.rematch2       2.1.0
#> 64            reproducible.reproducible 0.2.10.9007
#> 65            reproducible.reproducible 0.2.10.9007
#> 66                          rgeos.rgeos       0.5-2
#> 67                          rgeos.rgeos       0.5-2
#> 68                              rio.rio      0.5.16
#> 69                    roxygen2.roxygen2       7.0.0
#> 70                    roxygen2.roxygen2       7.0.0
#> 71                  rsconnect.rsconnect      0.8.15
#> 72                        SpaDES.SpaDES  2.0.4.9000
#> 73                        SpaDES.SpaDES  2.0.4.9000
#> 74              SpaDES.core.SpaDES.core  0.2.6.9006
#> 75              SpaDES.core.SpaDES.core  0.2.6.9006
#> 76  SpaDES.experiment.SpaDES.experiment  0.0.1.9000
#> 77            SpaDES.tools.SpaDES.tools  0.3.2.9002
#> 78            SpaDES.tools.SpaDES.tools  0.3.2.9002
#> 79                      SparseM.SparseM        1.77
#> 80                    spelling.spelling         2.1
#> 81                    testthat.testthat       2.3.0
#> 82                    testthat.testthat       2.3.0
#> 83                    testthat.testthat       2.3.0
#> 84                    TimeWarp.TimeWarp      1.0.15
#> 85                    TimeWarp.TimeWarp      1.0.15
#> 86                              zip.zip       2.0.4
#> 87          alphavantager.alphavantager       0.1.1
#> 88                              amc.amc  0.1.5.9000
#> 89                              amc.amc  0.1.5.9000
#> 90                  animation.animation         2.6
#> 91                  archivist.archivist       2.3.4
#> 92                  base64enc.base64enc       0.1-4
#> 93                  base64enc.base64enc       0.1-4
#> 94                              bit.bit      1.1-14
#> 95                          bit64.bit64       0.9-7
#> 96                            blob.blob       1.2.0
#> 97                          caret.caret      6.0-84
#> 98                  checkmate.checkmate       1.9.4
#> 99                  CircStats.CircStats       0.2-6
#> 100                   classInt.classInt       0.4-2
#> 101                           covr.covr       3.3.2
#> 102               data.table.data.table      1.12.2
#> 103               data.table.data.table      1.12.2
#> 104                             DBI.DBI       1.0.0
#> 105                     DEoptim.DEoptim       2.2-4
#> 106               DiagrammeR.DiagrammeR       1.0.1
#> 107                         dismo.dismo       1.1-4
#> 108               doParallel.doParallel      1.0.15
#> 109               downloader.downloader         0.4
#> 110                         e1071.e1071       1.7-2
#> 111               fastdigest.fastdigest       0.6-3
#> 112                 fasterize.fasterize       1.0.0
#> 113                 fastmatch.fastmatch       1.1-0
#> 114                     fastshp.fastshp       0.1-2
#> 115                               ff.ff      2.2-14
#> 116                       ffbase.ffbase      0.12.7
#> 117                         flock.flock         0.7
#> 118                     foreach.foreach       1.4.7
#> 119                   forecast.forecast         8.9
#> 120                 fpCompare.fpCompare       0.2.3
#> 121                   fracdiff.fracdiff       1.4-2
#> 122                       future.future      1.14.0
#> 123                       future.future      1.14.0
#> 124                 gdalUtils.gdalUtils    2.0.1.14
#> 125                   generics.generics       0.0.2
#> 126                     ggplot2.ggplot2       3.2.1
#> 127                     ggplot2.ggplot2       3.2.1
#> 128                           glmm.glmm       1.3.0
#> 129                     globals.globals      0.12.4
#> 130             googledrive.googledrive       1.0.0
#> 131                         gower.gower       0.2.1
#> 132                   gridBase.gridBase       0.4-7
#> 133                         highr.highr       0.8.1
#> 134                         highr.highr       0.8.1
#> 135                             hms.hms       0.5.2
#> 136                   hunspell.hunspell         3.0
#> 137               influenceR.influenceR       0.1.0
#> 138                         ipred.ipred       0.9-9
#> 139                 iterators.iterators      1.0.12
#> 140                 itertools.itertools       0.1-3
#> 141                         LandR.LandR  0.0.2.9001
#> 142                         LandR.LandR  0.0.2.9001
#> 143           LandWebUtils.LandWebUtils       0.0.1
#> 144           LandWebUtils.LandWebUtils       0.0.1
#> 145                           lava.lava       1.6.6
#> 146                     listenv.listenv       0.7.0
#> 147                       lmtest.lmtest      0.9-37
#> 148                 lubridate.lubridate       1.7.4
#> 149                           lucr.lucr       0.2.0
#> 150                       magick.magick         2.2
#> 151                             map.map       0.0.1
#> 152                             map.map       0.0.1
#> 153                   maptools.maptools       0.9-8
#> 154       microbenchmark.microbenchmark       1.4-7
#> 155                           mime.mime         0.7
#> 156                           mime.mime         0.7
#> 157           ModelMetrics.ModelMetrics       1.2.2
#> 158                         MuMIn.MuMIn      1.43.8
#> 159                     mvtnorm.mvtnorm      1.0-11
#> 160                   numDeriv.numDeriv  2016.8-1.1
#> 161                           padr.padr       0.5.0
#> 162                         pedev.pedev  0.0.1.9005
#> 163                         pedev.pedev  0.0.1.9005
#> 164                       pemisc.pemisc  0.0.1.9003
#> 165                       pemisc.pemisc  0.0.1.9003
#> 166                       praise.praise       1.0.0
#> 167                       praise.praise       1.0.0
#> 168                     prodlim.prodlim  2019.10.13
#> 169                     profvis.profvis       0.3.6
#> 170                           pryr.pryr       0.1.4
#> 171                   quadprog.quadprog       1.5-7
#> 172                   quantmod.quantmod      0.4-15
#> 173                 quickPlot.quickPlot  0.1.6.9000
#> 174                 quickPlot.quickPlot  0.1.6.9000
#> 175             R.methodsS3.R.methodsS3       1.7.1
#> 176                           R.oo.R.oo      1.23.0
#> 177                     R.utils.R.utils       2.9.0
#> 178         RcppArmadillo.RcppArmadillo 0.9.800.1.0
#> 179                   RcppRoll.RcppRoll       0.3.0
#> 180                         RCurl.RCurl   1.95-4.12
#> 181                         readr.readr       1.3.1
#> 182                     recipes.recipes       0.1.7
#> 183                     remotes.remotes       2.1.0
#> 184                     remotes.remotes       2.1.0
#> 185                             rex.rex       1.1.2
#> 186                         rgdal.rgdal       1.4-7
#> 187                     rgenoud.rgenoud     5.8-3.0
#> 188                         rgeos.rgeos       0.4-3
#> 189                         rgeos.rgeos       0.4-3
#> 190                         rgexf.rgexf      0.15.3
#> 191                 rmarkdown.rmarkdown        1.17
#> 192                           Rook.Rook       1.1-1
#> 193                 rprojroot.rprojroot       1.3-2
#> 194                 rprojroot.rprojroot       1.3-2
#> 195                     RSQLite.RSQLite       2.1.2
#> 196                           scam.scam       1.2-5
#> 197                               sf.sf       0.8-0
#> 198                       SpaDES.SpaDES       2.0.3
#> 199                       SpaDES.SpaDES       2.0.3
#> 200         SpaDES.addins.SpaDES.addins       0.1.2
#> 201             SpaDES.core.SpaDES.core  0.2.5.9007
#> 202             SpaDES.core.SpaDES.core  0.2.5.9007
#> 203           SpaDES.tools.SpaDES.tools  0.3.2.9001
#> 204           SpaDES.tools.SpaDES.tools  0.3.2.9001
#> 205                     SQUAREM.SQUAREM   2017.10-1
#> 206                   testthat.testthat       2.1.1
#> 207                   testthat.testthat       2.1.1
#> 208                   testthat.testthat       2.1.1
#> 209                       tibble.tibble       2.1.3
#> 210                       tibble.tibble       2.1.3
#> 211                         tiler.tiler       0.2.1
#> 212                   timeDate.timeDate    3043.102
#> 213                       timetk.timetk       0.1.2
#> 214                   TimeWarp.TimeWarp      1.0.15
#> 215                   TimeWarp.TimeWarp      1.0.15
#> 216                     tinytex.tinytex        0.17
#> 217                         trust.trust       0.1-7
#> 218                     tseries.tseries     0.10-47
#> 219                             TTR.TTR      0.23-5
#> 220                         units.units       0.6-5
#> 221                           urca.urca       1.3-0
#> 222                           uuid.uuid       0.1-2
#> 223                   versions.versions         0.3
#> 224                           VGAM.VGAM       1.1-1
#> 225               visNetwork.visNetwork       2.0.8
#> 226                     xgboost.xgboost    0.90.0.2
#> 227                             XML.XML   3.98-1.20
#> 228                       xtable.xtable       1.8-4
#> 229                       xtable.xtable       1.8-4
#> 230                     askpass.askpass         1.1
#> 231               assertthat.assertthat       0.2.1
#> 232                 backports.backports       1.1.5
#> 233                 base64enc.base64enc       0.1-3
#> 234                 base64enc.base64enc       0.1-3
#> 235                               BH.BH    1.69.0-1
#> 236                       bitops.bitops       1.0-6
#> 237                           brew.brew       1.0-6
#> 238                         callr.callr       3.3.2
#> 239                             cli.cli       1.1.0
#> 240                         clipr.clipr       0.7.0
#> 241               clisymbols.clisymbols       1.2.0
#> 242                           coda.coda      0.19-3
#> 243               colorspace.colorspace       1.4-1
#> 244               commonmark.commonmark         1.7
#> 245                       crayon.crayon       1.3.4
#> 246                 crosstalk.crosstalk       1.0.0
#> 247                           curl.curl         4.2
#> 248                           desc.desc       1.2.0
#> 249                   devtools.devtools       2.1.0
#> 250                   devtools.devtools       2.1.0
#> 251                       digest.digest      0.6.21
#> 252                       digest.digest      0.6.21
#> 253                         dplyr.dplyr       0.8.3
#> 254                               DT.DT        0.10
#> 255                   ellipsis.ellipsis     0.2.0.1
#> 256                   ellipsis.ellipsis     0.2.0.1
#> 257                   evaluate.evaluate        0.14
#> 258                         fansi.fansi       0.4.0
#> 259                             FNN.FNN       1.1.3
#> 260                               fs.fs       1.3.1
#> 261                     gdtools.gdtools       0.2.1
#> 262                           geoR.geoR   1.7-5.2.1
#> 263                         ggmap.ggmap       3.0.0
#> 264                     ggplot2.ggplot2       3.2.1
#> 265                     ggplot2.ggplot2       3.2.1
#> 266                               gh.gh       1.0.1
#> 267                         git2r.git2r      0.26.1
#> 268                           glue.glue       1.3.1
#> 269                 gridExtra.gridExtra         2.3
#> 270                         gstat.gstat       2.0-3
#> 271                       gtable.gtable       0.3.0
#> 272                         highr.highr         0.8
#> 273                         highr.highr         0.8
#> 274                 htmltools.htmltools       0.4.0
#> 275             htmlwidgets.htmlwidgets       1.5.1
#> 276                       httpuv.httpuv       1.5.2
#> 277                           httr.httr       1.4.1
#> 278                       igraph.igraph     1.2.4.1
#> 279                             ini.ini       0.3.1
#> 280                 intervals.intervals      0.15.1
#> 281                           jpeg.jpeg     0.1-8.1
#> 282                   jsonlite.jsonlite         1.6
#> 283                         knitr.knitr        1.26
#> 284                   labeling.labeling         0.3
#> 285                         later.later       1.0.0
#> 286                   lazyeval.lazyeval       0.2.2
#> 287                     leaflet.leaflet       2.0.3
#> 288 leaflet.providers.leaflet.providers       1.9.0
#> 289                     littler.littler       0.3.8
#> 290                     littler.littler       0.3.8
#> 291                           lme4.lme4      1.1-21
#> 292                   magrittr.magrittr         1.5
#> 293   manipulateWidget.manipulateWidget      0.10.0
#> 294                     mapdata.mapdata       2.3.0
#> 295                           maps.maps       3.3.0
#> 296                   markdown.markdown         1.1
#> 297                     memoise.memoise       1.1.0
#> 298                           mime.mime         0.7
#> 299                           mime.mime         0.7
#> 300                       miniUI.miniUI     0.1.1.1
#> 301                         minqa.minqa       1.2.4
#> 302                     munsell.munsell       0.5.0
#> 303                         ncdf4.ncdf4        1.17
#> 304                       nloptr.nloptr       1.2.1
#> 305                     openssl.openssl       1.4.1
#> 306                       pillar.pillar       1.4.2
#> 307                   pkgbuild.pkgbuild       1.0.6
#> 308                 pkgconfig.pkgconfig       2.0.3
#> 309                 pkgKitten.pkgKitten       0.1.5
#> 310                     pkgload.pkgload       1.0.2
#> 311                         plogr.plogr       0.2.0
#> 312                           plyr.plyr       1.8.4
#> 313                             png.png       0.1-7
#> 314                       praise.praise       1.0.0
#> 315                       praise.praise       1.0.0
#> 316             prettyunits.prettyunits       1.0.2
#> 317                   processx.processx       3.4.1
#> 318                   promises.promises       1.1.0
#> 319                               ps.ps       1.3.0
#> 320                         purrr.purrr       0.3.3
#> 321                               R6.R6       2.4.1
#> 322           RandomFields.RandomFields       3.3.6
#> 323 RandomFieldsUtils.RandomFieldsUtils       0.5.3
#> 324                       raster.raster       3.0-7
#> 325                 rcmdcheck.rcmdcheck       1.3.3
#> 326           RColorBrewer.RColorBrewer       1.1-2
#> 327                           Rcpp.Rcpp       1.0.3
#> 328                 RcppEigen.RcppEigen   0.3.3.7.0
#> 329                     remotes.remotes       2.1.0
#> 330                     remotes.remotes       2.1.0
#> 331                   reshape2.reshape2       1.4.3
#> 332                             rgl.rgl    0.100.30
#> 333             RgoogleMaps.RgoogleMaps       1.4.4
#> 334                         rjags.rjags        4-10
#> 335                         rJava.rJava      0.9-11
#> 336                         rjson.rjson      0.2.20
#> 337                         rlang.rlang       0.4.1
#> 338                   roxygen2.roxygen2       7.0.0
#> 339                   roxygen2.roxygen2       7.0.0
#> 340                 rprojroot.rprojroot         1.2
#> 341                 rprojroot.rprojroot         1.2
#> 342               rstudioapi.rstudioapi        0.10
#> 343                 rversions.rversions       2.0.0
#> 344                       scales.scales       1.0.0
#> 345             sessioninfo.sessioninfo       1.1.1
#> 346                         shiny.shiny       1.4.0
#> 347                           snow.snow       0.4-3
#> 348             sourcetools.sourcetools       0.1.7
#> 349                               sp.sp       1.3-2
#> 350                 spacetime.spacetime       1.2-2
#> 351                     splancs.splancs     2.01-40
#> 352                     stringi.stringi       1.4.3
#> 353                     stringr.stringr       1.4.0
#> 354                     svglite.svglite       1.2.2
#> 355                             sys.sys         3.3
#> 356             systemfonts.systemfonts       0.1.1
#> 357                   testthat.testthat       2.2.1
#> 358                   testthat.testthat       2.2.1
#> 359                   testthat.testthat       2.2.1
#> 360                       tibble.tibble       2.1.3
#> 361                       tibble.tibble       2.1.3
#> 362                         tidyr.tidyr       1.0.0
#> 363               tidyselect.tidyselect       0.2.5
#> 364                     tkrplot.tkrplot      0.0-24
#> 365                     usethis.usethis       1.5.1
#> 366                           utf8.utf8       1.1.4
#> 367                         vctrs.vctrs       0.2.0
#> 368                     viridis.viridis       0.5.1
#> 369             viridisLite.viridisLite       0.3.0
#> 370                     webshot.webshot       0.5.1
#> 371                     whisker.whisker         0.4
#> 372                         withr.withr       2.1.2
#> 373                           xfun.xfun        0.11
#> 374                           xlsx.xlsx       0.6.1
#> 375                   xlsxjars.xlsxjars       0.6.1
#> 376                           xml2.xml2       1.2.2
#> 377                         xopen.xopen       1.0.0
#> 378                       xtable.xtable       1.8-3
#> 379                       xtable.xtable       1.8-3
#> 380                             xts.xts      0.11-2
#> 381                           yaml.yaml       2.2.0
#> 382                     zeallot.zeallot       0.1.0
#> 383                             zoo.zoo       1.8-6
#> 384                           base.base       3.6.1
#> 385                           boot.boot      1.3-23
#> 386                         class.class      7.3-15
#> 387                     cluster.cluster       2.1.0
#> 388                 codetools.codetools      0.2-16
#> 389                   compiler.compiler       3.6.1
#> 390                   datasets.datasets       3.6.1
#> 391                     foreign.foreign      0.8-72
#> 392                   graphics.graphics       3.6.1
#> 393                 grDevices.grDevices       3.6.1
#> 394                           grid.grid       3.6.1
#> 395               KernSmooth.KernSmooth     2.23-16
#> 396                     lattice.lattice     0.20-38
#> 397                           MASS.MASS    7.3-51.4
#> 398                       Matrix.Matrix      1.2-17
#> 399                     methods.methods       3.6.1
#> 400                           mgcv.mgcv      1.8-31
#> 401                           nlme.nlme     3.1-142
#> 402                           nnet.nnet      7.3-12
#> 403                   parallel.parallel       3.6.1
#> 404                         rpart.rpart      4.1-15
#> 405                     spatial.spatial      7.3-11
#> 406                     splines.splines       3.6.1
#> 407                         stats.stats       3.6.1
#> 408                       stats4.stats4       3.6.1
#> 409                   survival.survival       3.1-7
#> 410                         tcltk.tcltk       3.6.1
#> 411                         tools.tools       3.6.1
#> 412           translations.translations       3.6.1
#> 413                         utils.utils       3.6.1

data.table::fread(pkgSnapFile)

#>                       instPkgs   instVers
#>   1: reproducible.reproducible     0.2.11
#>   2: reproducible.reproducible     0.2.11
#>   3:               abind.abind      1.4-5
#>   4:           acepack.acepack      1.4.1
#>   5:                   amc.amc 0.1.5.9000
#>  ---                                     
#> 409:         survival.survival      3.1-7
#> 410:               tcltk.tcltk      3.6.1
#> 411:               tools.tools      3.6.1
#> 412: translations.translations      3.6.1
#> 413:               utils.utils      3.6.1