Alternative to readLines that is faster

This alternative is from https://gist.github.com/hadley/6353939

readLinesRcpp(path)

Arguments

path
Path to text file to read.

Value

Similar to readLines. It may not return identical results.

Examples

readLinesRcpp(system.file(package = "reproducible", "DESCRIPTION"))

#>  [1] "Package: reproducible"                                                                                               
#>  [2] "Type: Package"                                                                                                       
#>  [3] "Title: A Set of Tools that Enhance Reproducibility Beyond Package Management"                                        
#>  [4] "Description: Collection of high-level, robust, machine- and OS-independent tools"                                    
#>  [5] "    for making deeply reproducible and reusable content in R."                                                       
#>  [6] "    This includes light weight package management (similar to 'packrat' and"                                         
#>  [7] "    'checkpoint', but more flexible, lightweight and simpler than both), tools"                                      
#>  [8] "    for caching, downloading and verifying or writing checksums, post-processing "                                   
#>  [9] "    of common spatial datasets, and accessing GitHub repositories. "                                                 
#> [10] "    Some features are still under active development."                                                               
#> [11] "URL: "                                                                                                               
#> [12] "    http://reproducible.predictiveecology.org,"                                                                      
#> [13] "    https://github.com/PredictiveEcology/reproducible"                                                               
#> [14] "Date: 2018-11-02"                                                                                                    
#> [15] "Version: 0.2.5"                                                                                                      
#> [16] "Authors@R: c("                                                                                                       
#> [17] "    person(\"Eliot J B\", \"McIntire\", email = \"eliot.mcintire@canada.ca\","                                       
#> [18] "    role = c(\"aut\", \"cre\"), comment=c(ORCID = \"0000-0002-6914-8316\")),"                                        
#> [19] "    person(\"Alex M\", \"Chubaty\", email = \"alex.chubaty@gmail.com\","                                             
#> [20] "    role = c(\"aut\"), comment = c(ORCID = \"0000-0001-7146-8135\")),"                                               
#> [21] "    person(\"Tati\", \"Micheletti\", email = \"tati.micheletti@gmail.com\","                                         
#> [22] "    role = c(\"ctb\"), comment=c(ORCID = \"0000-0003-4838-8342\")),"                                                 
#> [23] "    person(\"Ian\", \"Eddy\", email = \"ian.eddy@canada.com\","                                                      
#> [24] "    role = c(\"ctb\"), comment=c(ORCID = \"0000-0001-7397-2116\")),"                                                 
#> [25] "    person(\"Her Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada\","
#> [26] "    role = \"cph\")"                                                                                                 
#> [27] "    )"                                                                                                               
#> [28] "Depends:"                                                                                                            
#> [29] "    R (>= 3.3.0)"                                                                                                    
#> [30] "Imports:"                                                                                                            
#> [31] "    archivist (>= 2.1.2),"                                                                                           
#> [32] "    backports,"                                                                                                      
#> [33] "    crayon,"                                                                                                         
#> [34] "    data.table (>= 1.10.4),"                                                                                         
#> [35] "    devtools,"                                                                                                       
#> [36] "    digest,"                                                                                                         
#> [37] "    dplyr,"                                                                                                          
#> [38] "    fastdigest,"                                                                                                     
#> [39] "    fasterize,"                                                                                                      
#> [40] "    fpCompare,"                                                                                                      
#> [41] "    gdalUtils,"                                                                                                      
#> [42] "    git2r (>= 0.18),"                                                                                                
#> [43] "    googledrive,"                                                                                                    
#> [44] "    httr,"                                                                                                           
#> [45] "    magrittr,"                                                                                                       
#> [46] "    memoise,"                                                                                                        
#> [47] "    methods,"                                                                                                        
#> [48] "    parallel,"                                                                                                       
#> [49] "    quickPlot,"                                                                                                      
#> [50] "    R.utils,"                                                                                                        
#> [51] "    raster,"                                                                                                         
#> [52] "    Rcpp (>= 0.12.13),"                                                                                              
#> [53] "    RCurl (>= 1.95-4.8),"                                                                                            
#> [54] "    remotes,"                                                                                                        
#> [55] "    rgdal,"                                                                                                          
#> [56] "    rgeos,"                                                                                                          
#> [57] "    sf,"                                                                                                             
#> [58] "    sp,"                                                                                                             
#> [59] "    testthat,"                                                                                                       
#> [60] "    tibble,"                                                                                                         
#> [61] "    tools,"                                                                                                          
#> [62] "    utils,"                                                                                                          
#> [63] "    versions"                                                                                                        
#> [64] "Suggests:"                                                                                                           
#> [65] "    covr,"                                                                                                           
#> [66] "    future,"                                                                                                         
#> [67] "    knitr,"                                                                                                          
#> [68] "    rmarkdown,"                                                                                                      
#> [69] "    TimeWarp"                                                                                                        
#> [70] "Encoding: UTF-8"                                                                                                     
#> [71] "Language: en-CA"                                                                                                     
#> [72] "LinkingTo: Rcpp"                                                                                                     
#> [73] "License: GPL-3"                                                                                                      
#> [74] "VignetteBuilder: knitr, rmarkdown"                                                                                   
#> [75] "BugReports: https://github.com/PredictiveEcology/reproducible/issues"                                                
#> [76] "ByteCompile: yes"                                                                                                    
#> [77] "RoxygenNote: 6.1.0.9000"                                                                                             
#> [78] "Collate: "                                                                                                           
#> [79] "    'RcppExports.R'"                                                                                                 
#> [80] "    'cache-helpers.R'"                                                                                               
#> [81] "    'cache-tools.R'"                                                                                                 
#> [82] "    'robustDigest.R'"                                                                                                
#> [83] "    'cache.R'"                                                                                                       
#> [84] "    'checksums.R'"                                                                                                   
#> [85] "    'consistentPaths.R'"                                                                                             
#> [86] "    'convertPaths.R'"                                                                                                
#> [87] "    'download.R'"                                                                                                    
#> [88] "    'gis.R'"                                                                                                         
#> [89] "    'git.R'"                                                                                                         
#> [90] "    'helpers.R'"                                                                                                     
#> [91] "    'objectSize.R'"                                                                                                  
#> [92] "    'packages.R'"                                                                                                    
#> [93] "    'pipe.R'"                                                                                                        
#> [94] "    'postProcess.R'"                                                                                                 
#> [95] "    'preProcess.R'"                                                                                                  
#> [96] "    'prepInputs.R'"                                                                                                  
#> [97] "    'reproducible-package.R'"                                                                                        
#> [98] "    'search.R'"                                                                                                      
#> [99] "    'zzz.R'"