Alternative to readLines that is faster

This alternative is from https://gist.github.com/hadley/6353939

readLinesRcpp(path)

Arguments

path
Path to text file to read.

Value

Similar to readLines. It may not return identical results.

Examples

readLinesRcpp(system.file(package = "reproducible", "DESCRIPTION"))

#>   [1] "Package: reproducible"                                                                                               
#>   [2] "Type: Package"                                                                                                       
#>   [3] "Title: A Set of Tools that Enhance Reproducibility Beyond Package Management"                                        
#>   [4] "Description: Collection of high-level, robust, machine- and OS-independent tools"                                    
#>   [5] "    for making deeply reproducible and reusable content in R."                                                       
#>   [6] "    This includes light weight package management (similar to 'packrat' and"                                         
#>   [7] "    'checkpoint', but more flexible, lightweight and simpler than both), tools"                                      
#>   [8] "    for caching, downloading and verifying or writing checksums, post-processing "                                   
#>   [9] "    of common spatial datasets, and accessing GitHub repositories. "                                                 
#>  [10] "    Some features are still under active development."                                                               
#>  [11] "SystemRequirements: 'unrar' (Linux/macOS) or '7-Zip' (Windows)  to work with '.rar' files."                          
#>  [12] "URL: "                                                                                                               
#>  [13] "    http://reproducible.predictiveecology.org,"                                                                      
#>  [14] "    https://github.com/PredictiveEcology/reproducible"                                                               
#>  [15] "Date: 2019-01-25"                                                                                                    
#>  [16] "Version: 0.2.6"                                                                                                      
#>  [17] "Authors@R: c("                                                                                                       
#>  [18] "    person(\"Eliot J B\", \"McIntire\", email = \"eliot.mcintire@canada.ca\","                                       
#>  [19] "    role = c(\"aut\", \"cre\"), comment = c(ORCID = \"0000-0002-6914-8316\")),"                                      
#>  [20] "    person(\"Alex M\", \"Chubaty\", email = \"alex.chubaty@gmail.com\","                                             
#>  [21] "    role = c(\"aut\"), comment = c(ORCID = \"0000-0001-7146-8135\")),"                                               
#>  [22] "    person(\"Tati\", \"Micheletti\", email = \"tati.micheletti@gmail.com\","                                         
#>  [23] "    role = c(\"ctb\"), comment = c(ORCID = \"0000-0003-4838-8342\")),"                                               
#>  [24] "    person(\"Ceres\", \"Barros\", email = \"cbarros@mail.ubc.ca\","                                                  
#>  [25] "    role = c(\"ctb\"), comment = c(ORCID = \"0000-0003-4036-977X\")),"                                               
#>  [26] "    person(\"Ian\", \"Eddy\", email = \"ian.eddy@canada.com\","                                                      
#>  [27] "    role = c(\"ctb\"), comment = c(ORCID = \"0000-0001-7397-2116\")),"                                               
#>  [28] "    person(\"Her Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada\","
#>  [29] "    role = \"cph\")"                                                                                                 
#>  [30] "    )"                                                                                                               
#>  [31] "Depends:"                                                                                                            
#>  [32] "    R (>= 3.3)"                                                                                                      
#>  [33] "Imports:"                                                                                                            
#>  [34] "    archivist (>= 2.1.2),"                                                                                           
#>  [35] "    backports,"                                                                                                      
#>  [36] "    crayon,"                                                                                                         
#>  [37] "    data.table (>= 1.10.4),"                                                                                         
#>  [38] "    digest,"                                                                                                         
#>  [39] "    dplyr,"                                                                                                          
#>  [40] "    fastdigest,"                                                                                                     
#>  [41] "    fasterize,"                                                                                                      
#>  [42] "    fpCompare,"                                                                                                      
#>  [43] "    gdalUtils,"                                                                                                      
#>  [44] "    git2r (>= 0.18),"                                                                                                
#>  [45] "    googledrive,"                                                                                                    
#>  [46] "    httr,"                                                                                                           
#>  [47] "    magrittr,"                                                                                                       
#>  [48] "    memoise,"                                                                                                        
#>  [49] "    methods,"                                                                                                        
#>  [50] "    parallel,"                                                                                                       
#>  [51] "    quickPlot,"                                                                                                      
#>  [52] "    R.utils,"                                                                                                        
#>  [53] "    raster,"                                                                                                         
#>  [54] "    Rcpp (>= 0.12.13),"                                                                                              
#>  [55] "    RCurl (>= 1.95-4.8),"                                                                                            
#>  [56] "    remotes,"                                                                                                        
#>  [57] "    rgdal,"                                                                                                          
#>  [58] "    rgeos,"                                                                                                          
#>  [59] "    sf,"                                                                                                             
#>  [60] "    sp,"                                                                                                             
#>  [61] "    testthat,"                                                                                                       
#>  [62] "    tools,"                                                                                                          
#>  [63] "    utils,"                                                                                                          
#>  [64] "    versions"                                                                                                        
#>  [65] "Suggests:"                                                                                                           
#>  [66] "    covr,"                                                                                                           
#>  [67] "    future,"                                                                                                         
#>  [68] "    knitr,"                                                                                                          
#>  [69] "    rmarkdown,"                                                                                                      
#>  [70] "    TimeWarp"                                                                                                        
#>  [71] "Encoding: UTF-8"                                                                                                     
#>  [72] "Language: en-CA"                                                                                                     
#>  [73] "LinkingTo: Rcpp"                                                                                                     
#>  [74] "License: GPL-3"                                                                                                      
#>  [75] "VignetteBuilder: knitr, rmarkdown"                                                                                   
#>  [76] "BugReports: https://github.com/PredictiveEcology/reproducible/issues"                                                
#>  [77] "ByteCompile: yes"                                                                                                    
#>  [78] "RoxygenNote: 6.1.1"                                                                                                  
#>  [79] "Collate: "                                                                                                           
#>  [80] "    'RcppExports.R'"                                                                                                 
#>  [81] "    'cache-helpers.R'"                                                                                               
#>  [82] "    'cache-internals.R'"                                                                                             
#>  [83] "    'cache-tools.R'"                                                                                                 
#>  [84] "    'robustDigest.R'"                                                                                                
#>  [85] "    'cache.R'"                                                                                                       
#>  [86] "    'checksums.R'"                                                                                                   
#>  [87] "    'cloud.R'"                                                                                                       
#>  [88] "    'consistentPaths.R'"                                                                                             
#>  [89] "    'convertPaths.R'"                                                                                                
#>  [90] "    'download.R'"                                                                                                    
#>  [91] "    'gis.R'"                                                                                                         
#>  [92] "    'git.R'"                                                                                                         
#>  [93] "    'helpers.R'"                                                                                                     
#>  [94] "    'objectSize.R'"                                                                                                  
#>  [95] "    'options.R'"                                                                                                     
#>  [96] "    'packages.R'"                                                                                                    
#>  [97] "    'pipe.R'"                                                                                                        
#>  [98] "    'postProcess.R'"                                                                                                 
#>  [99] "    'preProcess.R'"                                                                                                  
#> [100] "    'prepInputs.R'"                                                                                                  
#> [101] "    'reproducible-package.R'"                                                                                        
#> [102] "    'search.R'"                                                                                                      
#> [103] "    'zzz.R'"