Alternative to readLines that is faster

This alternative is from https://gist.github.com/hadley/6353939

readLinesRcpp(path)

Arguments

path
Path to text file to read.

Value

Similar to readLines. It may not return identical results.

Examples

readLinesRcpp(system.file(package = "reproducible", "DESCRIPTION"))

#>  [1] "Package: reproducible"                                                                                               
#>  [2] "Type: Package"                                                                                                       
#>  [3] "Title: A Set of Tools that Enhance Reproducibility Beyond Package Management"                                        
#>  [4] "Description: Built on top of 'git2r' and 'archivist', this package aims at"                                          
#>  [5] "    making high-level, robust, machine and OS independent tools for making"                                          
#>  [6] "    deeply reproducible and reusable content in R. This includes light weight"                                       
#>  [7] "    package management (similar to 'packrat' and 'checkpoint', but more flexible,"                                   
#>  [8] "    lightweight and simpler than both), plus it includes"                                                            
#>  [9] "    tools for caching, and accessing GitHub repositories. There are many features"                                   
#> [10] "    that are still under active development, so behaviour may change in future"                                      
#> [11] "    versions."                                                                                                       
#> [12] "URL: "                                                                                                               
#> [13] "    http://reproducible.predictiveecology.org,"                                                                      
#> [14] "    https://github.com/PredictiveEcology/reproducible"                                                               
#> [15] "Date: 2018-06-12"                                                                                                    
#> [16] "Version: 0.2.0"                                                                                                      
#> [17] "Authors@R: c("                                                                                                       
#> [18] "    person(\"Eliot J B\", \"McIntire\", email = \"eliot.mcintire@canada.ca\","                                       
#> [19] "    role = c(\"aut\", \"cre\")),"                                                                                    
#> [20] "    person(\"Alex M\", \"Chubaty\", email = \"alex.chubaty@gmail.com\","                                             
#> [21] "    role = c(\"aut\")),"                                                                                             
#> [22] "    person(\"Her Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada\","
#> [23] "    role = \"cph\")"                                                                                                 
#> [24] "    )"                                                                                                               
#> [25] "Depends:"                                                                                                            
#> [26] "    R (>= 3.3.0)"                                                                                                    
#> [27] "Imports:"                                                                                                            
#> [28] "    archivist (>= 2.1.2),"                                                                                           
#> [29] "    crayon,"                                                                                                         
#> [30] "    data.table (>= 1.10.4),"                                                                                         
#> [31] "    devtools,"                                                                                                       
#> [32] "    digest,"                                                                                                         
#> [33] "    dplyr,"                                                                                                          
#> [34] "    fastdigest,"                                                                                                     
#> [35] "    fasterize,"                                                                                                      
#> [36] "    gdalUtils,"                                                                                                      
#> [37] "    git2r (>= 0.18),"                                                                                                
#> [38] "    googledrive,"                                                                                                    
#> [39] "    httr,"                                                                                                           
#> [40] "    magrittr,"                                                                                                       
#> [41] "    memoise,"                                                                                                        
#> [42] "    methods,"                                                                                                        
#> [43] "    raster,"                                                                                                         
#> [44] "    Rcpp (>= 0.12.13),"                                                                                              
#> [45] "    RCurl (>= 1.95-4.8),"                                                                                            
#> [46] "    R.utils,"                                                                                                        
#> [47] "    rgdal,"                                                                                                          
#> [48] "    rgeos,"                                                                                                          
#> [49] "    quickPlot,"                                                                                                      
#> [50] "    sf,"                                                                                                             
#> [51] "    sp,"                                                                                                             
#> [52] "    tools,"                                                                                                          
#> [53] "    utils,"                                                                                                          
#> [54] "    versions"                                                                                                        
#> [55] "Suggests:"                                                                                                           
#> [56] "    covr,"                                                                                                           
#> [57] "    hunspell,"                                                                                                       
#> [58] "    knitr,"                                                                                                          
#> [59] "    rmarkdown,"                                                                                                      
#> [60] "    testthat,"                                                                                                       
#> [61] "    TimeWarp"                                                                                                        
#> [62] "LinkingTo: Rcpp"                                                                                                     
#> [63] "License: GPL-3"                                                                                                      
#> [64] "BugReports: https://github.com/PredictiveEcology/reproducible/issues"                                                
#> [65] "ByteCompile: yes"                                                                                                    
#> [66] "RoxygenNote: 6.0.1"                                                                                                  
#> [67] "Collate: "                                                                                                           
#> [68] "    'RcppExports.R'"                                                                                                 
#> [69] "    'cache-helpers.R'"                                                                                               
#> [70] "    'cache-tools.R'"                                                                                                 
#> [71] "    'robustDigest.R'"                                                                                                
#> [72] "    'cache.R'"                                                                                                       
#> [73] "    'checksums.R'"                                                                                                   
#> [74] "    'consistentPaths.R'"                                                                                             
#> [75] "    'download.R'"                                                                                                    
#> [76] "    'gis.R'"                                                                                                         
#> [77] "    'git.R'"                                                                                                         
#> [78] "    'helpers.R'"                                                                                                     
#> [79] "    'objectSize.R'"                                                                                                  
#> [80] "    'packages.R'"                                                                                                    
#> [81] "    'pipe.R'"                                                                                                        
#> [82] "    'postProcess.R'"                                                                                                 
#> [83] "    'prepInputs.R'"                                                                                                  
#> [84] "    'reproducible-package.R'"                                                                                        
#> [85] "    'search.R'"                                                                                                      
#> [86] "    'zzz.R'"