Examining and modifying the cache

These are convenience wrappers around archivist package functions. They allow the user a bit of control over what is being cached.

clearCache(x, userTags = character(), after, before, ...)

# S4 method for ANY
clearCache(x, userTags = character(), after, before, ...)

showCache(x, userTags = character(), after, before, ...)

# S4 method for ANY
showCache(x, userTags = character(), after, before, ...)

keepCache(x, userTags = character(), after, before, ...)

# S4 method for ANY
keepCache(x, userTags = character(), after, before, ...)

Arguments

x
A simList or a directory containing a valid archivist repository

userTags
Character vector. If used, this will be used in place of the after and before. Specifying one or more userTag here will clear all objects that match those tags. Matching is via regular expresssion, meaning partial matches will work unless strict beginning (^) and end ($) of string characters are used. Matching will be against any of the 3 columns returned by showCache(), i.e., artifact, tagValue or tagName. Also, length userTags > 1, then matching is by `and`. For `or` matching, use | in a single character string. See examples.

after
A time (POSIX, character understandable by data.table). Objects cached after this time will be shown or deleted.

before
A time (POSIX, character understandable by data.table). Objects cached before this time will be shown or deleted.

...
Other arguments. Currently unused.

If neither after or before are provided, nor userTags, then all objects will be removed. If both after and before are specified, then all objects between after and before will be deleted. If userTags is used, this will override after or before.

Value

Will clear all (or that match userTags, or between after or before) objects from the repository located at cachePath of the sim object, if sim is provided, or located in cacheRepo. Also returns a data.table invisibly of the removed items.

Details

clearCache

remove items from the cache based on their userTag or times values.

keepCache

remove all cached items except those based on certain userTags or times values.

showCache

display the contents of the cache.

See also

splitTagsLocal.

Examples

library(raster)

tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir), silent = TRUE) # just to make sure it is clear

# Basic use
ranNumsA <- Cache(rnorm, 10, 16, cacheRepo = tmpDir)

# All same
ranNumsB <- Cache(rnorm, 10, 16, cacheRepo = tmpDir) # recovers cached copy

#>   loading cached result from previous rnorm call.

ranNumsC <- rnorm(10, 16) %>% Cache(cacheRepo = tmpDir) # recovers cached copy

#>   loading cached result from previous 'rnorm' pipe sequence call.

ranNumsD <- Cache(quote(rnorm(n = 10, 16)), cacheRepo = tmpDir) # recovers cached copy

#>   loading cached result from previous rnorm call.

# Any minor change makes it different
ranNumsE <- rnorm(10, 6) %>% Cache(cacheRepo = tmpDir) # different

## Example 1: basic cache use with tags
ranNumsA <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(runif, 4, cacheRepo = tmpDir, userTags = "objectName:b")

showCache(tmpDir, userTags = c("objectName"))

#>                             artifact         tagKey
#>  1: 516bc24d0c7e9a8095ddbf582454957d         format
#>  2: 516bc24d0c7e9a8095ddbf582454957d           name
#>  3: 516bc24d0c7e9a8095ddbf582454957d          class
#>  4: 516bc24d0c7e9a8095ddbf582454957d           date
#>  5: 516bc24d0c7e9a8095ddbf582454957d     objectName
#>  6: 516bc24d0c7e9a8095ddbf582454957d       function
#>  7: 516bc24d0c7e9a8095ddbf582454957d    object.size
#>  8: 516bc24d0c7e9a8095ddbf582454957d       accessed
#>  9: 516bc24d0c7e9a8095ddbf582454957d otherFunctions
#> 10: 516bc24d0c7e9a8095ddbf582454957d otherFunctions
#> 11: 516bc24d0c7e9a8095ddbf582454957d otherFunctions
#> 12: 516bc24d0c7e9a8095ddbf582454957d        cacheId
#> 13: 8c818d5140eea46644d8f1ecd22d27f7         format
#> 14: 8c818d5140eea46644d8f1ecd22d27f7           name
#> 15: 8c818d5140eea46644d8f1ecd22d27f7          class
#> 16: 8c818d5140eea46644d8f1ecd22d27f7           date
#> 17: 8c818d5140eea46644d8f1ecd22d27f7     objectName
#> 18: 8c818d5140eea46644d8f1ecd22d27f7       function
#> 19: 8c818d5140eea46644d8f1ecd22d27f7    object.size
#> 20: 8c818d5140eea46644d8f1ecd22d27f7       accessed
#> 21: 8c818d5140eea46644d8f1ecd22d27f7 otherFunctions
#> 22: 8c818d5140eea46644d8f1ecd22d27f7 otherFunctions
#> 23: 8c818d5140eea46644d8f1ecd22d27f7 otherFunctions
#> 24: 8c818d5140eea46644d8f1ecd22d27f7        cacheId
#>                             artifact         tagKey
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:48:06
#>  2:                               Cache 2018-01-25 08:48:06
#>  3:                             numeric 2018-01-25 08:48:06
#>  4:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#>  5:                                   b 2018-01-25 08:48:06
#>  6:                               runif 2018-01-25 08:48:06
#>  7:                                 528 2018-01-25 08:48:06
#>  8:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#>  9:      otherFunctions:build_reference 2018-01-25 08:48:06
#> 10: otherFunctions:data_reference_topic 2018-01-25 08:48:06
#> 11:                  otherFunctions:try 2018-01-25 08:48:06
#> 12:    d89a70080574eedee9d56a2a09d7ca01 2018-01-25 08:48:06
#> 13:                                 rda 2018-01-25 08:48:06
#> 14:                               Cache 2018-01-25 08:48:06
#> 15:                             numeric 2018-01-25 08:48:06
#> 16:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#> 17:                                   a 2018-01-25 08:48:06
#> 18:                               rnorm 2018-01-25 08:48:06
#> 19:                                 528 2018-01-25 08:48:06
#> 20:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#> 21:      otherFunctions:build_reference 2018-01-25 08:48:06
#> 22: otherFunctions:data_reference_topic 2018-01-25 08:48:06
#> 23:                  otherFunctions:try 2018-01-25 08:48:06
#> 24:    f05172cef0faa451e18d7ddc500dbd0e 2018-01-25 08:48:06
#>                                tagValue         createdDate

showCache(tmpDir, userTags = c("^a$")) # regular expression ... "a" exactly

#>                             artifact         tagKey
#>  1: 8c818d5140eea46644d8f1ecd22d27f7         format
#>  2: 8c818d5140eea46644d8f1ecd22d27f7           name
#>  3: 8c818d5140eea46644d8f1ecd22d27f7          class
#>  4: 8c818d5140eea46644d8f1ecd22d27f7           date
#>  5: 8c818d5140eea46644d8f1ecd22d27f7     objectName
#>  6: 8c818d5140eea46644d8f1ecd22d27f7       function
#>  7: 8c818d5140eea46644d8f1ecd22d27f7    object.size
#>  8: 8c818d5140eea46644d8f1ecd22d27f7       accessed
#>  9: 8c818d5140eea46644d8f1ecd22d27f7 otherFunctions
#> 10: 8c818d5140eea46644d8f1ecd22d27f7 otherFunctions
#> 11: 8c818d5140eea46644d8f1ecd22d27f7 otherFunctions
#> 12: 8c818d5140eea46644d8f1ecd22d27f7        cacheId
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:48:06
#>  2:                               Cache 2018-01-25 08:48:06
#>  3:                             numeric 2018-01-25 08:48:06
#>  4:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#>  5:                                   a 2018-01-25 08:48:06
#>  6:                               rnorm 2018-01-25 08:48:06
#>  7:                                 528 2018-01-25 08:48:06
#>  8:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#>  9:      otherFunctions:build_reference 2018-01-25 08:48:06
#> 10: otherFunctions:data_reference_topic 2018-01-25 08:48:06
#> 11:                  otherFunctions:try 2018-01-25 08:48:06
#> 12:    f05172cef0faa451e18d7ddc500dbd0e 2018-01-25 08:48:06

showCache(tmpDir, userTags = c("runif")) # show only cached objects made during runif call

#>                             artifact         tagKey
#>  1: 516bc24d0c7e9a8095ddbf582454957d         format
#>  2: 516bc24d0c7e9a8095ddbf582454957d           name
#>  3: 516bc24d0c7e9a8095ddbf582454957d          class
#>  4: 516bc24d0c7e9a8095ddbf582454957d           date
#>  5: 516bc24d0c7e9a8095ddbf582454957d     objectName
#>  6: 516bc24d0c7e9a8095ddbf582454957d       function
#>  7: 516bc24d0c7e9a8095ddbf582454957d    object.size
#>  8: 516bc24d0c7e9a8095ddbf582454957d       accessed
#>  9: 516bc24d0c7e9a8095ddbf582454957d otherFunctions
#> 10: 516bc24d0c7e9a8095ddbf582454957d otherFunctions
#> 11: 516bc24d0c7e9a8095ddbf582454957d otherFunctions
#> 12: 516bc24d0c7e9a8095ddbf582454957d        cacheId
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:48:06
#>  2:                               Cache 2018-01-25 08:48:06
#>  3:                             numeric 2018-01-25 08:48:06
#>  4:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#>  5:                                   b 2018-01-25 08:48:06
#>  6:                               runif 2018-01-25 08:48:06
#>  7:                                 528 2018-01-25 08:48:06
#>  8:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#>  9:      otherFunctions:build_reference 2018-01-25 08:48:06
#> 10: otherFunctions:data_reference_topic 2018-01-25 08:48:06
#> 11:                  otherFunctions:try 2018-01-25 08:48:06
#> 12:    d89a70080574eedee9d56a2a09d7ca01 2018-01-25 08:48:06

clearCache(tmpDir, userTags = c("runif")) # remove only cached objects made during runif call
showCache(tmpDir) # only those made during rnorm call

#>                             artifact         tagKey
#>  1: 22560797dae6678c31b186f8b6759dd1         format
#>  2: 22560797dae6678c31b186f8b6759dd1           name
#>  3: 22560797dae6678c31b186f8b6759dd1          class
#>  4: 22560797dae6678c31b186f8b6759dd1           date
#>  5: 22560797dae6678c31b186f8b6759dd1       function
#>  6: 22560797dae6678c31b186f8b6759dd1    object.size
#>  7: 22560797dae6678c31b186f8b6759dd1       accessed
#>  8: 22560797dae6678c31b186f8b6759dd1 otherFunctions
#>  9: 22560797dae6678c31b186f8b6759dd1 otherFunctions
#> 10: 22560797dae6678c31b186f8b6759dd1 otherFunctions
#> 11: 22560797dae6678c31b186f8b6759dd1        cacheId
#> 12: 22560797dae6678c31b186f8b6759dd1       accessed
#> 13: 22560797dae6678c31b186f8b6759dd1       accessed
#> 14: 22560797dae6678c31b186f8b6759dd1       accessed
#> 15: 8c818d5140eea46644d8f1ecd22d27f7         format
#> 16: 8c818d5140eea46644d8f1ecd22d27f7           name
#> 17: 8c818d5140eea46644d8f1ecd22d27f7          class
#> 18: 8c818d5140eea46644d8f1ecd22d27f7           date
#> 19: 8c818d5140eea46644d8f1ecd22d27f7     objectName
#> 20: 8c818d5140eea46644d8f1ecd22d27f7       function
#> 21: 8c818d5140eea46644d8f1ecd22d27f7    object.size
#> 22: 8c818d5140eea46644d8f1ecd22d27f7       accessed
#> 23: 8c818d5140eea46644d8f1ecd22d27f7 otherFunctions
#> 24: 8c818d5140eea46644d8f1ecd22d27f7 otherFunctions
#> 25: 8c818d5140eea46644d8f1ecd22d27f7 otherFunctions
#> 26: 8c818d5140eea46644d8f1ecd22d27f7        cacheId
#> 27: 9b3d704a25e49d9d52bd2d2e8e85dd71         format
#> 28: 9b3d704a25e49d9d52bd2d2e8e85dd71           name
#> 29: 9b3d704a25e49d9d52bd2d2e8e85dd71          class
#> 30: 9b3d704a25e49d9d52bd2d2e8e85dd71           date
#> 31: 9b3d704a25e49d9d52bd2d2e8e85dd71       function
#> 32: 9b3d704a25e49d9d52bd2d2e8e85dd71    object.size
#> 33: 9b3d704a25e49d9d52bd2d2e8e85dd71       accessed
#> 34: 9b3d704a25e49d9d52bd2d2e8e85dd71 otherFunctions
#> 35: 9b3d704a25e49d9d52bd2d2e8e85dd71 otherFunctions
#> 36: 9b3d704a25e49d9d52bd2d2e8e85dd71 otherFunctions
#> 37: 9b3d704a25e49d9d52bd2d2e8e85dd71        cacheId
#>                             artifact         tagKey
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:48:06
#>  2:                               Cache 2018-01-25 08:48:06
#>  3:                             numeric 2018-01-25 08:48:06
#>  4:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#>  5:                               rnorm 2018-01-25 08:48:06
#>  6:                                 624 2018-01-25 08:48:06
#>  7:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#>  8:      otherFunctions:build_reference 2018-01-25 08:48:06
#>  9: otherFunctions:data_reference_topic 2018-01-25 08:48:06
#> 10:                  otherFunctions:try 2018-01-25 08:48:06
#> 11:    c9b9e3d462fdcd976aac998bbbc60c7a 2018-01-25 08:48:06
#> 12:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#> 13:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#> 14:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#> 15:                                 rda 2018-01-25 08:48:06
#> 16:                               Cache 2018-01-25 08:48:06
#> 17:                             numeric 2018-01-25 08:48:06
#> 18:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#> 19:                                   a 2018-01-25 08:48:06
#> 20:                               rnorm 2018-01-25 08:48:06
#> 21:                                 528 2018-01-25 08:48:06
#> 22:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#> 23:      otherFunctions:build_reference 2018-01-25 08:48:06
#> 24: otherFunctions:data_reference_topic 2018-01-25 08:48:06
#> 25:                  otherFunctions:try 2018-01-25 08:48:06
#> 26:    f05172cef0faa451e18d7ddc500dbd0e 2018-01-25 08:48:06
#> 27:                                 rda 2018-01-25 08:48:06
#> 28:                               Cache 2018-01-25 08:48:06
#> 29:                             numeric 2018-01-25 08:48:06
#> 30:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#> 31:                 rnorm pipe sequence 2018-01-25 08:48:06
#> 32:                                 624 2018-01-25 08:48:06
#> 33:                 2018-01-25 08:48:06 2018-01-25 08:48:06
#> 34:      otherFunctions:build_reference 2018-01-25 08:48:06
#> 35: otherFunctions:data_reference_topic 2018-01-25 08:48:06
#> 36:                  otherFunctions:try 2018-01-25 08:48:06
#> 37:    25729d525b445f9a94475253cd095f48 2018-01-25 08:48:06
#>                                tagValue         createdDate

clearCache(tmpDir)

## Example 2: using the "accessed" tag
ranNumsA <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(runif, 4, cacheRepo = tmpDir, userTags = "objectName:b")

# access it again, from Cache
ranNumsA <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:a")

#>   loading cached result from previous rnorm call.

wholeCache <- showCache(tmpDir)

# keep only items accessed "recently" (i.e., only objectName:a)
onlyRecentlyAccessed <- showCache(tmpDir, userTags = max(wholeCache[tagKey == "accessed"]$tagValue))

# inverse join with 2 data.tables ... using: a[!b]
# i.e., return all of wholeCache that was not recently accessed
toRemove <- unique(wholeCache[!onlyRecentlyAccessed], by = "artifact")$artifact
clearCache(tmpDir, toRemove) # remove ones not recently accessed
showCache(tmpDir) # still has more recently accessed

#> Empty data.table (0 rows) of 3 cols: md5hash,name,createdDate

clearCache(tmpDir)

## Example 3: using keepCache
ranNumsA <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(runif, 4, cacheRepo = tmpDir, userTags = "objectName:b")

# keep only those cached items from the last 24 hours
oneDay <- 60 * 60 * 24
keepCache(tmpDir, after = Sys.time() - oneDay)

#>                             artifact         tagKey
#>  1: 072057415efce1620f66d7e38f5234fc         format
#>  2: 072057415efce1620f66d7e38f5234fc           name
#>  3: 072057415efce1620f66d7e38f5234fc          class
#>  4: 072057415efce1620f66d7e38f5234fc           date
#>  5: 072057415efce1620f66d7e38f5234fc     objectName
#>  6: 072057415efce1620f66d7e38f5234fc       function
#>  7: 072057415efce1620f66d7e38f5234fc    object.size
#>  8: 072057415efce1620f66d7e38f5234fc       accessed
#>  9: 072057415efce1620f66d7e38f5234fc otherFunctions
#> 10: 072057415efce1620f66d7e38f5234fc otherFunctions
#> 11: 072057415efce1620f66d7e38f5234fc otherFunctions
#> 12: 072057415efce1620f66d7e38f5234fc        cacheId
#> 13: ea1ee2cfea9b49ba415adb4745d4a45e         format
#> 14: ea1ee2cfea9b49ba415adb4745d4a45e           name
#> 15: ea1ee2cfea9b49ba415adb4745d4a45e          class
#> 16: ea1ee2cfea9b49ba415adb4745d4a45e           date
#> 17: ea1ee2cfea9b49ba415adb4745d4a45e     objectName
#> 18: ea1ee2cfea9b49ba415adb4745d4a45e       function
#> 19: ea1ee2cfea9b49ba415adb4745d4a45e    object.size
#> 20: ea1ee2cfea9b49ba415adb4745d4a45e       accessed
#> 21: ea1ee2cfea9b49ba415adb4745d4a45e otherFunctions
#> 22: ea1ee2cfea9b49ba415adb4745d4a45e otherFunctions
#> 23: ea1ee2cfea9b49ba415adb4745d4a45e otherFunctions
#> 24: ea1ee2cfea9b49ba415adb4745d4a45e        cacheId
#>                             artifact         tagKey
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:48:07
#>  2:                               Cache 2018-01-25 08:48:07
#>  3:                             numeric 2018-01-25 08:48:07
#>  4:                 2018-01-25 08:48:07 2018-01-25 08:48:07
#>  5:                                   b 2018-01-25 08:48:07
#>  6:                               runif 2018-01-25 08:48:07
#>  7:                                 528 2018-01-25 08:48:07
#>  8:                 2018-01-25 08:48:07 2018-01-25 08:48:07
#>  9:      otherFunctions:build_reference 2018-01-25 08:48:07
#> 10: otherFunctions:data_reference_topic 2018-01-25 08:48:07
#> 11:                  otherFunctions:try 2018-01-25 08:48:07
#> 12:    d89a70080574eedee9d56a2a09d7ca01 2018-01-25 08:48:07
#> 13:                                 rda 2018-01-25 08:48:07
#> 14:                               Cache 2018-01-25 08:48:07
#> 15:                             numeric 2018-01-25 08:48:07
#> 16:                 2018-01-25 08:48:07 2018-01-25 08:48:07
#> 17:                                   a 2018-01-25 08:48:07
#> 18:                               rnorm 2018-01-25 08:48:07
#> 19:                                 528 2018-01-25 08:48:07
#> 20:                 2018-01-25 08:48:07 2018-01-25 08:48:07
#> 21:      otherFunctions:build_reference 2018-01-25 08:48:07
#> 22: otherFunctions:data_reference_topic 2018-01-25 08:48:07
#> 23:                  otherFunctions:try 2018-01-25 08:48:07
#> 24:    f05172cef0faa451e18d7ddc500dbd0e 2018-01-25 08:48:07
#>                                tagValue         createdDate

# Keep all Cache items created with an rnorm() call
keepCache(tmpDir, userTags = "rnorm")

#>                             artifact         tagKey
#>  1: ea1ee2cfea9b49ba415adb4745d4a45e         format
#>  2: ea1ee2cfea9b49ba415adb4745d4a45e           name
#>  3: ea1ee2cfea9b49ba415adb4745d4a45e          class
#>  4: ea1ee2cfea9b49ba415adb4745d4a45e           date
#>  5: ea1ee2cfea9b49ba415adb4745d4a45e     objectName
#>  6: ea1ee2cfea9b49ba415adb4745d4a45e       function
#>  7: ea1ee2cfea9b49ba415adb4745d4a45e    object.size
#>  8: ea1ee2cfea9b49ba415adb4745d4a45e       accessed
#>  9: ea1ee2cfea9b49ba415adb4745d4a45e otherFunctions
#> 10: ea1ee2cfea9b49ba415adb4745d4a45e otherFunctions
#> 11: ea1ee2cfea9b49ba415adb4745d4a45e otherFunctions
#> 12: ea1ee2cfea9b49ba415adb4745d4a45e        cacheId
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:48:07
#>  2:                               Cache 2018-01-25 08:48:07
#>  3:                             numeric 2018-01-25 08:48:07
#>  4:                 2018-01-25 08:48:07 2018-01-25 08:48:07
#>  5:                                   a 2018-01-25 08:48:07
#>  6:                               rnorm 2018-01-25 08:48:07
#>  7:                                 528 2018-01-25 08:48:07
#>  8:                 2018-01-25 08:48:07 2018-01-25 08:48:07
#>  9:      otherFunctions:build_reference 2018-01-25 08:48:07
#> 10: otherFunctions:data_reference_topic 2018-01-25 08:48:07
#> 11:                  otherFunctions:try 2018-01-25 08:48:07
#> 12:    f05172cef0faa451e18d7ddc500dbd0e 2018-01-25 08:48:07

# Remove all Cache items that happened within a rnorm() call
clearCache(tmpDir, userTags = "rnorm")

showCache(tmpDir) ## empty

#> Empty data.table (0 rows) of 3 cols: md5hash,name,createdDate

## Example 4: searching for multiple objects in the cache

# default userTags is "and" matching; for "or" matching use |
ranNumsA <- Cache(runif, 4, cacheRepo = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:b")

# show all objects (runif and rnorm in this case)
showCache(tmpDir)

#>                             artifact         tagKey
#>  1: 08ba34427b7e41a32b00dda5d9e86ef6         format
#>  2: 08ba34427b7e41a32b00dda5d9e86ef6           name
#>  3: 08ba34427b7e41a32b00dda5d9e86ef6          class
#>  4: 08ba34427b7e41a32b00dda5d9e86ef6           date
#>  5: 08ba34427b7e41a32b00dda5d9e86ef6     objectName
#>  6: 08ba34427b7e41a32b00dda5d9e86ef6       function
#>  7: 08ba34427b7e41a32b00dda5d9e86ef6    object.size
#>  8: 08ba34427b7e41a32b00dda5d9e86ef6       accessed
#>  9: 08ba34427b7e41a32b00dda5d9e86ef6 otherFunctions
#> 10: 08ba34427b7e41a32b00dda5d9e86ef6 otherFunctions
#> 11: 08ba34427b7e41a32b00dda5d9e86ef6 otherFunctions
#> 12: 08ba34427b7e41a32b00dda5d9e86ef6        cacheId
#> 13: 2368af2b773058278da3761451900945         format
#> 14: 2368af2b773058278da3761451900945           name
#> 15: 2368af2b773058278da3761451900945          class
#> 16: 2368af2b773058278da3761451900945           date
#> 17: 2368af2b773058278da3761451900945     objectName
#> 18: 2368af2b773058278da3761451900945       function
#> 19: 2368af2b773058278da3761451900945    object.size
#> 20: 2368af2b773058278da3761451900945       accessed
#> 21: 2368af2b773058278da3761451900945 otherFunctions
#> 22: 2368af2b773058278da3761451900945 otherFunctions
#> 23: 2368af2b773058278da3761451900945 otherFunctions
#> 24: 2368af2b773058278da3761451900945        cacheId
#>                             artifact         tagKey
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:48:07
#>  2:                               Cache 2018-01-25 08:48:07
#>  3:                             numeric 2018-01-25 08:48:07
#>  4:                 2018-01-25 08:48:07 2018-01-25 08:48:07
#>  5:                                   b 2018-01-25 08:48:08
#>  6:                               rnorm 2018-01-25 08:48:08
#>  7:                                 528 2018-01-25 08:48:08
#>  8:                 2018-01-25 08:48:07 2018-01-25 08:48:08
#>  9:      otherFunctions:build_reference 2018-01-25 08:48:08
#> 10: otherFunctions:data_reference_topic 2018-01-25 08:48:08
#> 11:                  otherFunctions:try 2018-01-25 08:48:08
#> 12:    f05172cef0faa451e18d7ddc500dbd0e 2018-01-25 08:48:08
#> 13:                                 rda 2018-01-25 08:48:07
#> 14:                               Cache 2018-01-25 08:48:07
#> 15:                             numeric 2018-01-25 08:48:07
#> 16:                 2018-01-25 08:48:07 2018-01-25 08:48:07
#> 17:                                   a 2018-01-25 08:48:07
#> 18:                               runif 2018-01-25 08:48:07
#> 19:                                 528 2018-01-25 08:48:07
#> 20:                 2018-01-25 08:48:07 2018-01-25 08:48:07
#> 21:      otherFunctions:build_reference 2018-01-25 08:48:07
#> 22: otherFunctions:data_reference_topic 2018-01-25 08:48:07
#> 23:                  otherFunctions:try 2018-01-25 08:48:07
#> 24:    d89a70080574eedee9d56a2a09d7ca01 2018-01-25 08:48:07
#>                                tagValue         createdDate

# show objects that are both runif and rnorm
# (i.e., none in this case, because objecs are either or, not both)
showCache(tmpDir, userTags = c("runif", "rnorm")) ## empty

#> Empty data.table (0 rows) of 4 cols: artifact,tagKey,tagValue,createdDate

# show objects that are either runif or rnorm ("or" search)
showCache(tmpDir, userTags = "runif|rnorm")

#>                             artifact         tagKey
#>  1: 08ba34427b7e41a32b00dda5d9e86ef6         format
#>  2: 08ba34427b7e41a32b00dda5d9e86ef6           name
#>  3: 08ba34427b7e41a32b00dda5d9e86ef6          class
#>  4: 08ba34427b7e41a32b00dda5d9e86ef6           date
#>  5: 08ba34427b7e41a32b00dda5d9e86ef6     objectName
#>  6: 08ba34427b7e41a32b00dda5d9e86ef6       function
#>  7: 08ba34427b7e41a32b00dda5d9e86ef6    object.size
#>  8: 08ba34427b7e41a32b00dda5d9e86ef6       accessed
#>  9: 08ba34427b7e41a32b00dda5d9e86ef6 otherFunctions
#> 10: 08ba34427b7e41a32b00dda5d9e86ef6 otherFunctions
#> 11: 08ba34427b7e41a32b00dda5d9e86ef6 otherFunctions
#> 12: 08ba34427b7e41a32b00dda5d9e86ef6        cacheId
#> 13: 2368af2b773058278da3761451900945         format
#> 14: 2368af2b773058278da3761451900945           name
#> 15: 2368af2b773058278da3761451900945          class
#> 16: 2368af2b773058278da3761451900945           date
#> 17: 2368af2b773058278da3761451900945     objectName
#> 18: 2368af2b773058278da3761451900945       function
#> 19: 2368af2b773058278da3761451900945    object.size
#> 20: 2368af2b773058278da3761451900945       accessed
#> 21: 2368af2b773058278da3761451900945 otherFunctions
#> 22: 2368af2b773058278da3761451900945 otherFunctions
#> 23: 2368af2b773058278da3761451900945 otherFunctions
#> 24: 2368af2b773058278da3761451900945        cacheId
#>                             artifact         tagKey
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:48:07
#>  2:                               Cache 2018-01-25 08:48:07
#>  3:                             numeric 2018-01-25 08:48:07
#>  4:                 2018-01-25 08:48:07 2018-01-25 08:48:07
#>  5:                                   b 2018-01-25 08:48:08
#>  6:                               rnorm 2018-01-25 08:48:08
#>  7:                                 528 2018-01-25 08:48:08
#>  8:                 2018-01-25 08:48:07 2018-01-25 08:48:08
#>  9:      otherFunctions:build_reference 2018-01-25 08:48:08
#> 10: otherFunctions:data_reference_topic 2018-01-25 08:48:08
#> 11:                  otherFunctions:try 2018-01-25 08:48:08
#> 12:    f05172cef0faa451e18d7ddc500dbd0e 2018-01-25 08:48:08
#> 13:                                 rda 2018-01-25 08:48:07
#> 14:                               Cache 2018-01-25 08:48:07
#> 15:                             numeric 2018-01-25 08:48:07
#> 16:                 2018-01-25 08:48:07 2018-01-25 08:48:07
#> 17:                                   a 2018-01-25 08:48:07
#> 18:                               runif 2018-01-25 08:48:07
#> 19:                                 528 2018-01-25 08:48:07
#> 20:                 2018-01-25 08:48:07 2018-01-25 08:48:07
#> 21:      otherFunctions:build_reference 2018-01-25 08:48:07
#> 22: otherFunctions:data_reference_topic 2018-01-25 08:48:07
#> 23:                  otherFunctions:try 2018-01-25 08:48:07
#> 24:    d89a70080574eedee9d56a2a09d7ca01 2018-01-25 08:48:07
#>                                tagValue         createdDate

# keep only objects that are either runif or rnorm ("or" search)
keepCache(tmpDir, userTags = "runif|rnorm")

#>                             artifact         tagKey
#>  1: 08ba34427b7e41a32b00dda5d9e86ef6         format
#>  2: 08ba34427b7e41a32b00dda5d9e86ef6           name
#>  3: 08ba34427b7e41a32b00dda5d9e86ef6          class
#>  4: 08ba34427b7e41a32b00dda5d9e86ef6           date
#>  5: 08ba34427b7e41a32b00dda5d9e86ef6     objectName
#>  6: 08ba34427b7e41a32b00dda5d9e86ef6       function
#>  7: 08ba34427b7e41a32b00dda5d9e86ef6    object.size
#>  8: 08ba34427b7e41a32b00dda5d9e86ef6       accessed
#>  9: 08ba34427b7e41a32b00dda5d9e86ef6 otherFunctions
#> 10: 08ba34427b7e41a32b00dda5d9e86ef6 otherFunctions
#> 11: 08ba34427b7e41a32b00dda5d9e86ef6 otherFunctions
#> 12: 08ba34427b7e41a32b00dda5d9e86ef6        cacheId
#> 13: 2368af2b773058278da3761451900945         format
#> 14: 2368af2b773058278da3761451900945           name
#> 15: 2368af2b773058278da3761451900945          class
#> 16: 2368af2b773058278da3761451900945           date
#> 17: 2368af2b773058278da3761451900945     objectName
#> 18: 2368af2b773058278da3761451900945       function
#> 19: 2368af2b773058278da3761451900945    object.size
#> 20: 2368af2b773058278da3761451900945       accessed
#> 21: 2368af2b773058278da3761451900945 otherFunctions
#> 22: 2368af2b773058278da3761451900945 otherFunctions
#> 23: 2368af2b773058278da3761451900945 otherFunctions
#> 24: 2368af2b773058278da3761451900945        cacheId
#>                             artifact         tagKey
#>                                tagValue         createdDate
#>  1:                                 rda 2018-01-25 08:48:07
#>  2:                               Cache 2018-01-25 08:48:07
#>  3:                             numeric 2018-01-25 08:48:07
#>  4:                 2018-01-25 08:48:07 2018-01-25 08:48:07
#>  5:                                   b 2018-01-25 08:48:08
#>  6:                               rnorm 2018-01-25 08:48:08
#>  7:                                 528 2018-01-25 08:48:08
#>  8:                 2018-01-25 08:48:07 2018-01-25 08:48:08
#>  9:      otherFunctions:build_reference 2018-01-25 08:48:08
#> 10: otherFunctions:data_reference_topic 2018-01-25 08:48:08
#> 11:                  otherFunctions:try 2018-01-25 08:48:08
#> 12:    f05172cef0faa451e18d7ddc500dbd0e 2018-01-25 08:48:08
#> 13:                                 rda 2018-01-25 08:48:07
#> 14:                               Cache 2018-01-25 08:48:07
#> 15:                             numeric 2018-01-25 08:48:07
#> 16:                 2018-01-25 08:48:07 2018-01-25 08:48:07
#> 17:                                   a 2018-01-25 08:48:07
#> 18:                               runif 2018-01-25 08:48:07
#> 19:                                 528 2018-01-25 08:48:07
#> 20:                 2018-01-25 08:48:07 2018-01-25 08:48:07
#> 21:      otherFunctions:build_reference 2018-01-25 08:48:07
#> 22: otherFunctions:data_reference_topic 2018-01-25 08:48:07
#> 23:                  otherFunctions:try 2018-01-25 08:48:07
#> 24:    d89a70080574eedee9d56a2a09d7ca01 2018-01-25 08:48:07
#>                                tagValue         createdDate

clearCache(tmpDir)

## Example 5: using caching to speed up rerunning expensive computations
ras <- raster(extent(0, 10, 0, 10), res = 1,
              vals = sample(1:5, replace = TRUE, size = 1e2),
              crs = "+proj=lcc +lat_1=48 +lat_2=33 +lon_0=-100 +ellps=WGS84")

# A slow operation, like GIS operation
notCached <- suppressWarnings(
  # project raster generates warnings when run non-interactively
  projectRaster(ras, crs = crs(ras), res = 5, cacheRepo = tmpDir)
)

cached <- suppressWarnings(
  # project raster generates warnings when run non-interactively
  Cache(projectRaster, ras, crs = crs(ras), res = 5, cacheRepo = tmpDir)
)

# second time is much faster
reRun <- suppressWarnings(
  # project raster generates warnings when run non-interactively
  Cache(projectRaster, ras, crs = crs(ras), res = 5, cacheRepo = tmpDir)
)

#>   loading cached result from previous projectRaster call.

# recovered cached version is same as non-cached version
all.equal(notCached, reRun) ## TRUE

#> [1] TRUE

## Example 6: working with file paths

# if passing a character string, it will take 2 complete passes to before
#  a cached copy is used when it is a save event (read or load is different)
obj <- 1:10
fname <- tempfile(fileext = ".RData")
Cache(saveRDS, obj, file = fname, cacheRepo = tmpDir)

#> NULL

Cache(saveRDS, obj, file = fname, cacheRepo = tmpDir)

#> NULL

Cache(saveRDS, obj, file = fname, cacheRepo = tmpDir) # cached copy is loaded

#>   loading cached result from previous saveRDS call.

#> [1] "NULL"
#> attr(,"tags")
#> [1] "cacheId:256e9b67000fa93742bffc0f38015cea"
#> attr(,"call")
#> [1] ""

# however, using asPath(), cached version retrieved after being run once
fname1 <- tempfile(fileext = ".RData")
Cache(saveRDS, obj, file = asPath(fname1), cacheRepo = tmpDir)

#> NULL

Cache(saveRDS, obj, file = asPath(fname1), cacheRepo = tmpDir) # cached copy is loaded

#> NULL

clearCache(tmpDir)

## cleanup
unlink(c("filename.rda", "filename1.rda"))
unlink(dirname(tmpDir), recursive = TRUE)