Examining and modifying the cache

These are convenience wrappers around archivist package functions. They allow the user a bit of control over what is being cached.

clearCache(x, userTags = character(), after, before, ...)

# S4 method for ANY
clearCache(x, userTags, after, before,
  ask = getOption("reproducible.ask"), ...)

showCache(x, userTags = character(), after, before, ...)

# S4 method for ANY
showCache(x, userTags = character(), after, before, ...)

keepCache(x, userTags = character(), after, before, ask, ...)

# S4 method for ANY
keepCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), ...)

Arguments

x
A simList or a directory containing a valid archivist repository

userTags
Character vector. If used, this will be used in place of the after and before. Specifying one or more userTag here will clear all objects that match those tags. Matching is via regular expression, meaning partial matches will work unless strict beginning (^) and end ($) of string characters are used. Matching will be against any of the 3 columns returned by showCache(), i.e., artifact, tagValue or tagName. Also, length userTags > 1, then matching is by `and`. For `or` matching, use | in a single character string. See examples.

If neither after or before are provided, nor userTags, then all objects will be removed. If both after and before are specified, then all objects between after and before will be deleted. If userTags is used, this will override after or before.

after
A time (POSIX, character understandable by data.table). Objects cached after this time will be shown or deleted.

before
A time (POSIX, character understandable by data.table). Objects cached before this time will be shown or deleted.

...
Other arguments. Currently, regexp, a logical, can be provided. This must be TRUE if the use is passing a regular expression. Otherwise, userTags will need to be exact matches. Default is missing, which is the same as TRUE. If there are errors due to regular expression problem, try FALSE.

ask
Logical. If FALSE, then it will not ask to confirm deletions using clearCache or keepCache. Default is TRUE

Value

Will clear all objects (or those that match userTags, or those between after or before) from the repository located at cachePath of the sim object, if sim is provided, or located in cacheRepo. Invisibly returns a data.table of the removed items.

Details

clearCache

remove items from the cache based on their userTag or times values.

keepCache

remove all cached items except those based on certain userTags or times values.

showCache

display the contents of the cache.

Note

If the cache is larger than 10MB, and clearCache is used, there will be a message and a pause, if interactive, to prevent accidentally deleting of a large cache repository.

See also

Examples

library(raster)
try(detach("package:magrittr", unload = TRUE), silent = TRUE) # magrittr,
                                    #if loaded, gives an error below

tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir), silent = TRUE) # just to make sure it is clear

# Basic use
ranNumsA <- Cache(rnorm, 10, 16, cacheRepo = tmpDir)

# All same
ranNumsB <- Cache(rnorm, 10, 16, cacheRepo = tmpDir) # recovers cached copy

#>   loading cached result from previous rnorm call, adding to memoised copy

ranNumsC <- rnorm(10, 16) %>% Cache(cacheRepo = tmpDir) # recovers cached copy

#>   loading memoised result from previous 'rnorm' pipe sequence call.

ranNumsD <- Cache(quote(rnorm(n = 10, 16)), cacheRepo = tmpDir) # recovers cached copy

#>   loading memoised result from previous rnorm call.

# Any minor change makes it different
ranNumsE <- rnorm(10, 6) %>% Cache(cacheRepo = tmpDir) # different

## Example 1: basic cache use with tags
ranNumsA <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(runif, 4, cacheRepo = tmpDir, userTags = "objectName:b")
ranNumsC <- Cache(runif, 40, cacheRepo = tmpDir, userTags = "objectName:b")

showCache(tmpDir, userTags = c("objectName"))

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 786 bytes

#>                             artifact         tagKey
#>  1: 87fe39b13a982710a11d32c99b27c283         format
#>  2: 87fe39b13a982710a11d32c99b27c283           name
#>  3: 87fe39b13a982710a11d32c99b27c283          class
#>  4: 87fe39b13a982710a11d32c99b27c283           date
#>  5: 87fe39b13a982710a11d32c99b27c283        cacheId
#>  6: 87fe39b13a982710a11d32c99b27c283     objectName
#>  7: 87fe39b13a982710a11d32c99b27c283       function
#>  8: 87fe39b13a982710a11d32c99b27c283    object.size
#>  9: 87fe39b13a982710a11d32c99b27c283       accessed
#> 10: 87fe39b13a982710a11d32c99b27c283 otherFunctions
#> 11: 87fe39b13a982710a11d32c99b27c283 otherFunctions
#> 12: 87fe39b13a982710a11d32c99b27c283 otherFunctions
#> 13: 87fe39b13a982710a11d32c99b27c283 otherFunctions
#> 14: 87fe39b13a982710a11d32c99b27c283 otherFunctions
#> 15: 87fe39b13a982710a11d32c99b27c283 otherFunctions
#> 16: 87fe39b13a982710a11d32c99b27c283 otherFunctions
#> 17: 87fe39b13a982710a11d32c99b27c283      preDigest
#> 18: 87fe39b13a982710a11d32c99b27c283      preDigest
#> 19: a9c5a995a523ed149261cc60b2519c28         format
#> 20: a9c5a995a523ed149261cc60b2519c28           name
#> 21: a9c5a995a523ed149261cc60b2519c28          class
#> 22: a9c5a995a523ed149261cc60b2519c28           date
#> 23: a9c5a995a523ed149261cc60b2519c28        cacheId
#> 24: a9c5a995a523ed149261cc60b2519c28     objectName
#> 25: a9c5a995a523ed149261cc60b2519c28       function
#> 26: a9c5a995a523ed149261cc60b2519c28    object.size
#> 27: a9c5a995a523ed149261cc60b2519c28       accessed
#> 28: a9c5a995a523ed149261cc60b2519c28 otherFunctions
#> 29: a9c5a995a523ed149261cc60b2519c28 otherFunctions
#> 30: a9c5a995a523ed149261cc60b2519c28 otherFunctions
#> 31: a9c5a995a523ed149261cc60b2519c28 otherFunctions
#> 32: a9c5a995a523ed149261cc60b2519c28 otherFunctions
#> 33: a9c5a995a523ed149261cc60b2519c28 otherFunctions
#> 34: a9c5a995a523ed149261cc60b2519c28 otherFunctions
#> 35: a9c5a995a523ed149261cc60b2519c28      preDigest
#> 36: a9c5a995a523ed149261cc60b2519c28      preDigest
#> 37: ab03a01b8f2c049d06b4426bb0627b0a         format
#> 38: ab03a01b8f2c049d06b4426bb0627b0a           name
#> 39: ab03a01b8f2c049d06b4426bb0627b0a          class
#> 40: ab03a01b8f2c049d06b4426bb0627b0a           date
#> 41: ab03a01b8f2c049d06b4426bb0627b0a        cacheId
#> 42: ab03a01b8f2c049d06b4426bb0627b0a     objectName
#> 43: ab03a01b8f2c049d06b4426bb0627b0a       function
#> 44: ab03a01b8f2c049d06b4426bb0627b0a    object.size
#> 45: ab03a01b8f2c049d06b4426bb0627b0a       accessed
#> 46: ab03a01b8f2c049d06b4426bb0627b0a otherFunctions
#> 47: ab03a01b8f2c049d06b4426bb0627b0a otherFunctions
#> 48: ab03a01b8f2c049d06b4426bb0627b0a otherFunctions
#> 49: ab03a01b8f2c049d06b4426bb0627b0a otherFunctions
#> 50: ab03a01b8f2c049d06b4426bb0627b0a otherFunctions
#> 51: ab03a01b8f2c049d06b4426bb0627b0a otherFunctions
#> 52: ab03a01b8f2c049d06b4426bb0627b0a otherFunctions
#> 53: ab03a01b8f2c049d06b4426bb0627b0a      preDigest
#> 54: ab03a01b8f2c049d06b4426bb0627b0a      preDigest
#>                             artifact         tagKey
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-08-05 21:53:16
#>  2:      87fe39b13a982710a11d32c99b27c283 2018-08-05 21:53:16
#>  3:                               numeric 2018-08-05 21:53:16
#>  4:                   2018-08-05 21:53:16 2018-08-05 21:53:16
#>  5:      e37bb635c97bc2eeecab63816b881bbc 2018-08-05 21:53:16
#>  6:                                     b 2018-08-05 21:53:16
#>  7:                                 runif 2018-08-05 21:53:16
#>  8:                                   952 2018-08-05 21:53:16
#>  9:                   2018-08-05 21:53:16 2018-08-05 21:53:16
#> 10:                   withCallingHandlers 2018-08-05 21:53:16
#> 11:                               saveRDS 2018-08-05 21:53:16
#> 12:                               do.call 2018-08-05 21:53:16
#> 13:                      build_site_local 2018-08-05 21:53:16
#> 14:                       build_reference 2018-08-05 21:53:16
#> 15:                  data_reference_topic 2018-08-05 21:53:16
#> 16:                                   try 2018-08-05 21:53:16
#> 17:    n:969a49ec15bcd4323ff31538af321264 2018-08-05 21:53:16
#> 18: .FUN:d2631d24c3b38b89c7bdd4ab7faaaac3 2018-08-05 21:53:16
#> 19:                                   rda 2018-08-05 21:53:16
#> 20:      a9c5a995a523ed149261cc60b2519c28 2018-08-05 21:53:16
#> 21:                               numeric 2018-08-05 21:53:16
#> 22:                   2018-08-05 21:53:16 2018-08-05 21:53:16
#> 23:      b419232b60eb6eedb59532cf0b592010 2018-08-05 21:53:16
#> 24:                                     b 2018-08-05 21:53:16
#> 25:                                 runif 2018-08-05 21:53:16
#> 26:                                  1240 2018-08-05 21:53:16
#> 27:                   2018-08-05 21:53:16 2018-08-05 21:53:16
#> 28:                   withCallingHandlers 2018-08-05 21:53:16
#> 29:                               saveRDS 2018-08-05 21:53:16
#> 30:                               do.call 2018-08-05 21:53:16
#> 31:                      build_site_local 2018-08-05 21:53:16
#> 32:                       build_reference 2018-08-05 21:53:16
#> 33:                  data_reference_topic 2018-08-05 21:53:16
#> 34:                                   try 2018-08-05 21:53:16
#> 35:    n:9d7bd914b9878238bccd5686a5a41903 2018-08-05 21:53:16
#> 36: .FUN:d2631d24c3b38b89c7bdd4ab7faaaac3 2018-08-05 21:53:16
#> 37:                                   rda 2018-08-05 21:53:16
#> 38:      ab03a01b8f2c049d06b4426bb0627b0a 2018-08-05 21:53:16
#> 39:                               numeric 2018-08-05 21:53:16
#> 40:                   2018-08-05 21:53:16 2018-08-05 21:53:16
#> 41:      85874f26b2e0c1ef689a7d379d275ebf 2018-08-05 21:53:16
#> 42:                                     a 2018-08-05 21:53:16
#> 43:                                 rnorm 2018-08-05 21:53:16
#> 44:                                   952 2018-08-05 21:53:16
#> 45:                   2018-08-05 21:53:16 2018-08-05 21:53:16
#> 46:                   withCallingHandlers 2018-08-05 21:53:16
#> 47:                               saveRDS 2018-08-05 21:53:16
#> 48:                               do.call 2018-08-05 21:53:16
#> 49:                      build_site_local 2018-08-05 21:53:16
#> 50:                       build_reference 2018-08-05 21:53:16
#> 51:                  data_reference_topic 2018-08-05 21:53:16
#> 52:                                   try 2018-08-05 21:53:16
#> 53:    n:969a49ec15bcd4323ff31538af321264 2018-08-05 21:53:16
#> 54: .FUN:7e9a928f110f80b3612e71883a6ec1f4 2018-08-05 21:53:16
#>                                  tagValue         createdDate

showCache(tmpDir, userTags = c("^a$")) # regular expression ... "a" exactly

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 238 bytes

#>                             artifact         tagKey
#>  1: ab03a01b8f2c049d06b4426bb0627b0a         format
#>  2: ab03a01b8f2c049d06b4426bb0627b0a           name
#>  3: ab03a01b8f2c049d06b4426bb0627b0a          class
#>  4: ab03a01b8f2c049d06b4426bb0627b0a           date
#>  5: ab03a01b8f2c049d06b4426bb0627b0a        cacheId
#>  6: ab03a01b8f2c049d06b4426bb0627b0a     objectName
#>  7: ab03a01b8f2c049d06b4426bb0627b0a       function
#>  8: ab03a01b8f2c049d06b4426bb0627b0a    object.size
#>  9: ab03a01b8f2c049d06b4426bb0627b0a       accessed
#> 10: ab03a01b8f2c049d06b4426bb0627b0a otherFunctions
#> 11: ab03a01b8f2c049d06b4426bb0627b0a otherFunctions
#> 12: ab03a01b8f2c049d06b4426bb0627b0a otherFunctions
#> 13: ab03a01b8f2c049d06b4426bb0627b0a otherFunctions
#> 14: ab03a01b8f2c049d06b4426bb0627b0a otherFunctions
#> 15: ab03a01b8f2c049d06b4426bb0627b0a otherFunctions
#> 16: ab03a01b8f2c049d06b4426bb0627b0a otherFunctions
#> 17: ab03a01b8f2c049d06b4426bb0627b0a      preDigest
#> 18: ab03a01b8f2c049d06b4426bb0627b0a      preDigest
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-08-05 21:53:16
#>  2:      ab03a01b8f2c049d06b4426bb0627b0a 2018-08-05 21:53:16
#>  3:                               numeric 2018-08-05 21:53:16
#>  4:                   2018-08-05 21:53:16 2018-08-05 21:53:16
#>  5:      85874f26b2e0c1ef689a7d379d275ebf 2018-08-05 21:53:16
#>  6:                                     a 2018-08-05 21:53:16
#>  7:                                 rnorm 2018-08-05 21:53:16
#>  8:                                   952 2018-08-05 21:53:16
#>  9:                   2018-08-05 21:53:16 2018-08-05 21:53:16
#> 10:                   withCallingHandlers 2018-08-05 21:53:16
#> 11:                               saveRDS 2018-08-05 21:53:16
#> 12:                               do.call 2018-08-05 21:53:16
#> 13:                      build_site_local 2018-08-05 21:53:16
#> 14:                       build_reference 2018-08-05 21:53:16
#> 15:                  data_reference_topic 2018-08-05 21:53:16
#> 16:                                   try 2018-08-05 21:53:16
#> 17:    n:969a49ec15bcd4323ff31538af321264 2018-08-05 21:53:16
#> 18: .FUN:7e9a928f110f80b3612e71883a6ec1f4 2018-08-05 21:53:16

# Fine control of cache elements -- pick out only the large runif object, and remove it
cache1 <- showCache(tmpDir, userTags = c("runif")) # show only cached objects made during runif

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 548 bytes

toRemove <- cache1[tagKey=="object.size"][as.numeric(tagValue) > 700]$artifact
clearCache(tmpDir, userTags = toRemove)

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 0 bytes

cacheAfter <- showCache(tmpDir, userTags = c("runif")) # Only the small one is left

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 548 bytes