Examining and modifying the cache

These are convenience wrappers around archivist package functions. They allow the user a bit of control over what is being cached.

clearCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), useCloud = FALSE,
  cloudFolderID = NULL, ...)

# S4 method for ANY
clearCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), useCloud = FALSE,
  cloudFolderID = getOption("reproducible.cloudFolderID", NULL), ...)

cc(secs, ...)

showCache(x, userTags = character(), after, before, ...)

# S4 method for ANY
showCache(x, userTags = character(), after, before, ...)

keepCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), ...)

# S4 method for ANY
keepCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), ...)

Arguments

x
A simList or a directory containing a valid archivist repository. Note: For compatibility with Cache argument, cacheRepo can also be used instead of x, though x will take precedence.

userTags
Character vector. If used, this will be used in place of the after and before. Specifying one or more userTag here will clear all objects that match those tags. Matching is via regular expression, meaning partial matches will work unless strict beginning (^) and end ($) of string characters are used. Matching will be against any of the 3 columns returned by showCache(), i.e., artifact, tagValue or tagName. Also, length userTags > 1, then matching is by `and`. For `or` matching, use | in a single character string. See examples.

after
A time (POSIX, character understandable by data.table). Objects cached after this time will be shown or deleted.

before
A time (POSIX, character understandable by data.table). Objects cached before this time will be shown or deleted.

ask
Logical. If FALSE, then it will not ask to confirm deletions using clearCache or keepCache. Default is TRUE

useCloud
Logical. If TRUE, then every object that is deleted locally will also be deleted in the cloudFolderID, if it is non-NULL

cloudFolderID
A googledrive id of a folder, e.g., using drive_mkdir(). If left as NULL, the function will create a cloud folder with a warning. The warning will have the cloudFolderID that should be used in subsequent calls. It will also be added to options("reproducible.cloudFolderID"), but this will not persist across sessions.

...
Other arguments. Currently, regexp, a logical, can be provided. This must be TRUE if the use is passing a regular expression. Otherwise, userTags will need to be exact matches. Default is missing, which is the same as TRUE. If there are errors due to regular expression problem, try FALSE. For cc, it is passed to clearCache, e.g., ask, userTags

secs
Currently 3 options: the number of seconds to pass to clearCache(after = secs), a POSIXct time e.g., from Sys.time(), or missing. If missing, the default, then it will delete the most recent entry in the Cache.

Value

Will clear all objects (or those that match userTags, or those between after or before) from the repository located at cachePath of the sim object, if sim is provided, or located in cacheRepo. Invisibly returns a data.table of the removed items.

Details

If neither after or before are provided, nor userTags, then all objects will be removed. If both after and before are specified, then all objects between after and before will be deleted. If userTags is used, this will override after or before.

cc(secs) is just a shortcut for clearCache(repo = Paths$cachePath, after = secs), i.e., to remove any cache entries touched in the last secs seconds.

clearCache

remove items from the cache based on their userTag or times values.

keepCache

remove all cached items except those based on certain userTags or times values.

showCache

display the contents of the cache.

Note

If the cache is larger than 10MB, and clearCache is used, there will be a message and a pause, if interactive, to prevent accidentally deleting of a large cache repository.

See also

Examples

library(raster)

tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir, ask = FALSE), silent = TRUE) # just to make sure it is clear

# Basic use
ranNumsA <- Cache(rnorm, 10, 16, cacheRepo = tmpDir)

# All same
ranNumsB <- Cache(rnorm, 10, 16, cacheRepo = tmpDir) # recovers cached copy

#>   loading cached result from previous rnorm call, adding to memoised copy

ranNumsC <- Cache(cacheRepo = tmpDir) %C% rnorm(10, 16)  # recovers cached copy

#> Error in match.call(definition, call, expand.dots, envir): unused argument (envir = <environment>)

ranNumsD <- Cache(quote(rnorm(n = 10, 16)), cacheRepo = tmpDir) # recovers cached copy

#>   loading memoised result from previous rnorm call.

# Any minor change makes it different
ranNumsE <- Cache(cacheRepo = tmpDir) %C% rnorm(10, 6)# different

#> Error in match.call(definition, call, expand.dots, envir): unused argument (envir = <environment>)

## Example 1: basic cache use with tags
ranNumsA <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(runif, 4, cacheRepo = tmpDir, userTags = "objectName:b")
ranNumsC <- Cache(runif, 40, cacheRepo = tmpDir, userTags = "objectName:b")

showCache(tmpDir, userTags = c("objectName"))

#> Cache size: 

#>   Total (including Rasters): 1.1 Kb

#>   Selected objects (not including Rasters): 828 bytes

#>                             artifact         tagKey
#>  1: 07cbd8db5b53b094b56b704cf2884dd9         format
#>  2: 07cbd8db5b53b094b56b704cf2884dd9           name
#>  3: 07cbd8db5b53b094b56b704cf2884dd9          class
#>  4: 07cbd8db5b53b094b56b704cf2884dd9           date
#>  5: 07cbd8db5b53b094b56b704cf2884dd9        cacheId
#>  6: 07cbd8db5b53b094b56b704cf2884dd9     objectName
#>  7: 07cbd8db5b53b094b56b704cf2884dd9       function
#>  8: 07cbd8db5b53b094b56b704cf2884dd9    object.size
#>  9: 07cbd8db5b53b094b56b704cf2884dd9       accessed
#> 10: 07cbd8db5b53b094b56b704cf2884dd9 otherFunctions
#> 11: 07cbd8db5b53b094b56b704cf2884dd9 otherFunctions
#> 12: 07cbd8db5b53b094b56b704cf2884dd9 otherFunctions
#> 13: 07cbd8db5b53b094b56b704cf2884dd9 otherFunctions
#> 14: 07cbd8db5b53b094b56b704cf2884dd9 otherFunctions
#> 15: 07cbd8db5b53b094b56b704cf2884dd9 otherFunctions
#> 16: 07cbd8db5b53b094b56b704cf2884dd9 otherFunctions
#> 17: 07cbd8db5b53b094b56b704cf2884dd9 otherFunctions
#> 18: 07cbd8db5b53b094b56b704cf2884dd9 otherFunctions
#> 19: 07cbd8db5b53b094b56b704cf2884dd9 otherFunctions
#> 20: 07cbd8db5b53b094b56b704cf2884dd9 otherFunctions
#> 21: 07cbd8db5b53b094b56b704cf2884dd9      preDigest
#> 22: 07cbd8db5b53b094b56b704cf2884dd9      preDigest
#> 23: 102c7c7f1e5478e954322fd9813d8d73         format
#> 24: 102c7c7f1e5478e954322fd9813d8d73           name
#> 25: 102c7c7f1e5478e954322fd9813d8d73          class
#> 26: 102c7c7f1e5478e954322fd9813d8d73           date
#> 27: 102c7c7f1e5478e954322fd9813d8d73        cacheId
#> 28: 102c7c7f1e5478e954322fd9813d8d73     objectName
#> 29: 102c7c7f1e5478e954322fd9813d8d73       function
#> 30: 102c7c7f1e5478e954322fd9813d8d73    object.size
#> 31: 102c7c7f1e5478e954322fd9813d8d73       accessed
#> 32: 102c7c7f1e5478e954322fd9813d8d73 otherFunctions
#> 33: 102c7c7f1e5478e954322fd9813d8d73 otherFunctions
#> 34: 102c7c7f1e5478e954322fd9813d8d73 otherFunctions
#> 35: 102c7c7f1e5478e954322fd9813d8d73 otherFunctions
#> 36: 102c7c7f1e5478e954322fd9813d8d73 otherFunctions
#> 37: 102c7c7f1e5478e954322fd9813d8d73 otherFunctions
#> 38: 102c7c7f1e5478e954322fd9813d8d73 otherFunctions
#> 39: 102c7c7f1e5478e954322fd9813d8d73 otherFunctions
#> 40: 102c7c7f1e5478e954322fd9813d8d73 otherFunctions
#> 41: 102c7c7f1e5478e954322fd9813d8d73 otherFunctions
#> 42: 102c7c7f1e5478e954322fd9813d8d73 otherFunctions
#> 43: 102c7c7f1e5478e954322fd9813d8d73      preDigest
#> 44: 102c7c7f1e5478e954322fd9813d8d73      preDigest
#> 45: 8a60a18f4c112d748c90c24578d668e2         format
#> 46: 8a60a18f4c112d748c90c24578d668e2           name
#> 47: 8a60a18f4c112d748c90c24578d668e2          class
#> 48: 8a60a18f4c112d748c90c24578d668e2           date
#> 49: 8a60a18f4c112d748c90c24578d668e2        cacheId
#> 50: 8a60a18f4c112d748c90c24578d668e2     objectName
#> 51: 8a60a18f4c112d748c90c24578d668e2       function
#> 52: 8a60a18f4c112d748c90c24578d668e2    object.size
#> 53: 8a60a18f4c112d748c90c24578d668e2       accessed
#> 54: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 55: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 56: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 57: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 58: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 59: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 60: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 61: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 62: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 63: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 64: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 65: 8a60a18f4c112d748c90c24578d668e2      preDigest
#> 66: 8a60a18f4c112d748c90c24578d668e2      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-09-05 12:26:40
#>  2: 07cbd8db5b53b094b56b704cf2884dd9 2019-09-05 12:26:40
#>  3:                          numeric 2019-09-05 12:26:40
#>  4:              2019-09-05 12:26:40 2019-09-05 12:26:40
#>  5:                 7225b47fe5dc51d0 2019-09-05 12:26:40
#>  6:                                b 2019-09-05 12:26:40
#>  7:                            runif 2019-09-05 12:26:40
#>  8:                             1296 2019-09-05 12:26:40
#>  9:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 10:                          saveRDS 2019-09-05 12:26:40
#> 11:                          do.call 2019-09-05 12:26:40
#> 12:                 build_site_local 2019-09-05 12:26:40
#> 13:                  build_reference 2019-09-05 12:26:40
#> 14:             data_reference_topic 2019-09-05 12:26:40
#> 15:                     run_examples 2019-09-05 12:26:40
#> 16:               highlight_examples 2019-09-05 12:26:40
#> 17:                        timing_fn 2019-09-05 12:26:40
#> 18:                           handle 2019-09-05 12:26:40
#> 19:                              try 2019-09-05 12:26:40
#> 20:                      withVisible 2019-09-05 12:26:40
#> 21:               n:1393aef18608c8be 2019-09-05 12:26:40
#> 22:            .FUN:881ec847b7161f3c 2019-09-05 12:26:40
#> 23:                              rda 2019-09-05 12:26:40
#> 24: 102c7c7f1e5478e954322fd9813d8d73 2019-09-05 12:26:40
#> 25:                          numeric 2019-09-05 12:26:40
#> 26:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 27:                 3aef38d1fc02aee5 2019-09-05 12:26:40
#> 28:                                b 2019-09-05 12:26:40
#> 29:                            runif 2019-09-05 12:26:40
#> 30:                             1008 2019-09-05 12:26:40
#> 31:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 32:                          saveRDS 2019-09-05 12:26:40
#> 33:                          do.call 2019-09-05 12:26:40
#> 34:                 build_site_local 2019-09-05 12:26:40
#> 35:                  build_reference 2019-09-05 12:26:40
#> 36:             data_reference_topic 2019-09-05 12:26:40
#> 37:                     run_examples 2019-09-05 12:26:40
#> 38:               highlight_examples 2019-09-05 12:26:40
#> 39:                        timing_fn 2019-09-05 12:26:40
#> 40:                           handle 2019-09-05 12:26:40
#> 41:                              try 2019-09-05 12:26:40
#> 42:                      withVisible 2019-09-05 12:26:40
#> 43:               n:7eef4eae85fd9229 2019-09-05 12:26:40
#> 44:            .FUN:881ec847b7161f3c 2019-09-05 12:26:40
#> 45:                              rda 2019-09-05 12:26:40
#> 46: 8a60a18f4c112d748c90c24578d668e2 2019-09-05 12:26:40
#> 47:                          numeric 2019-09-05 12:26:40
#> 48:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 49:                 f7bee22047b8d0c1 2019-09-05 12:26:40
#> 50:                                a 2019-09-05 12:26:40
#> 51:                            rnorm 2019-09-05 12:26:40
#> 52:                             1008 2019-09-05 12:26:40
#> 53:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 54:                          saveRDS 2019-09-05 12:26:40
#> 55:                          do.call 2019-09-05 12:26:40
#> 56:                 build_site_local 2019-09-05 12:26:40
#> 57:                  build_reference 2019-09-05 12:26:40
#> 58:             data_reference_topic 2019-09-05 12:26:40
#> 59:                     run_examples 2019-09-05 12:26:40
#> 60:               highlight_examples 2019-09-05 12:26:40
#> 61:                        timing_fn 2019-09-05 12:26:40
#> 62:                           handle 2019-09-05 12:26:40
#> 63:                              try 2019-09-05 12:26:40
#> 64:                      withVisible 2019-09-05 12:26:40
#> 65:               n:7eef4eae85fd9229 2019-09-05 12:26:40
#> 66:            .FUN:4f604aa46882b368 2019-09-05 12:26:40
#>                             tagValue         createdDate

showCache(tmpDir, userTags = c("^a$")) # regular expression ... "a" exactly

#> Cache size: 

#>   Total (including Rasters): 1.1 Kb

#>   Selected objects (not including Rasters): 252 bytes

#>                             artifact         tagKey
#>  1: 8a60a18f4c112d748c90c24578d668e2         format
#>  2: 8a60a18f4c112d748c90c24578d668e2           name
#>  3: 8a60a18f4c112d748c90c24578d668e2          class
#>  4: 8a60a18f4c112d748c90c24578d668e2           date
#>  5: 8a60a18f4c112d748c90c24578d668e2        cacheId
#>  6: 8a60a18f4c112d748c90c24578d668e2     objectName
#>  7: 8a60a18f4c112d748c90c24578d668e2       function
#>  8: 8a60a18f4c112d748c90c24578d668e2    object.size
#>  9: 8a60a18f4c112d748c90c24578d668e2       accessed
#> 10: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 11: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 12: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 13: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 14: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 15: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 16: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 17: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 18: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 19: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 20: 8a60a18f4c112d748c90c24578d668e2 otherFunctions
#> 21: 8a60a18f4c112d748c90c24578d668e2      preDigest
#> 22: 8a60a18f4c112d748c90c24578d668e2      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-09-05 12:26:40
#>  2: 8a60a18f4c112d748c90c24578d668e2 2019-09-05 12:26:40
#>  3:                          numeric 2019-09-05 12:26:40
#>  4:              2019-09-05 12:26:40 2019-09-05 12:26:40
#>  5:                 f7bee22047b8d0c1 2019-09-05 12:26:40
#>  6:                                a 2019-09-05 12:26:40
#>  7:                            rnorm 2019-09-05 12:26:40
#>  8:                             1008 2019-09-05 12:26:40
#>  9:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 10:                          saveRDS 2019-09-05 12:26:40
#> 11:                          do.call 2019-09-05 12:26:40
#> 12:                 build_site_local 2019-09-05 12:26:40
#> 13:                  build_reference 2019-09-05 12:26:40
#> 14:             data_reference_topic 2019-09-05 12:26:40
#> 15:                     run_examples 2019-09-05 12:26:40
#> 16:               highlight_examples 2019-09-05 12:26:40
#> 17:                        timing_fn 2019-09-05 12:26:40
#> 18:                           handle 2019-09-05 12:26:40
#> 19:                              try 2019-09-05 12:26:40
#> 20:                      withVisible 2019-09-05 12:26:40
#> 21:               n:7eef4eae85fd9229 2019-09-05 12:26:40
#> 22:            .FUN:4f604aa46882b368 2019-09-05 12:26:40
#>                             tagValue         createdDate

# Fine control of cache elements -- pick out only the large runif object, and remove it
cache1 <- showCache(tmpDir, userTags = c("runif")) # show only cached objects made during runif

#> Cache size: 

#>   Total (including Rasters): 1.1 Kb

#>   Selected objects (not including Rasters): 576 bytes

toRemove <- cache1[tagKey == "object.size"][as.numeric(tagValue) > 700]$artifact
clearCache(tmpDir, userTags = toRemove, ask = FALSE)

#> Cache size: 

#>   Total (including Rasters): 1.1 Kb

#>   Selected objects (not including Rasters): 0 bytes

cacheAfter <- showCache(tmpDir, userTags = c("runif")) # Only the small one is left

#> Cache size: 

#>   Total (including Rasters): 1.1 Kb

#>   Selected objects (not including Rasters): 576 bytes

tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir, ask = FALSE), silent = TRUE) # just to make sure it is clear

Cache(rnorm, 1, cacheRepo = tmpDir)

#> [1] -0.6109455
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:7072c305d8c69df0"
#> attr(,"call")
#> [1] ""

thisTime <- Sys.time()
Cache(rnorm, 2, cacheRepo = tmpDir)

#> [1] 1.2518842 0.2503195
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:3b347dee8f6305c7"
#> attr(,"call")
#> [1] ""

Cache(rnorm, 3, cacheRepo = tmpDir)

#> [1] -1.7055817 -0.8554131 -0.1449016
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:f0da91a44b839434"
#> attr(,"call")
#> [1] ""

Cache(rnorm, 4, cacheRepo = tmpDir)

#> [1] -0.3244470 -0.1725649 -1.2360629 -1.9023042
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:f7bee22047b8d0c1"
#> attr(,"call")
#> [1] ""

showCache(x = tmpDir) # shows all 4 entries

#> Cache size: 

#>   Total (including Rasters): 998 bytes

#>   Selected objects (not including Rasters): 998 bytes

#>                             artifact         tagKey
#>  1: 5895f04248d35050b3cd2fa031b9366c         format
#>  2: 5895f04248d35050b3cd2fa031b9366c           name
#>  3: 5895f04248d35050b3cd2fa031b9366c          class
#>  4: 5895f04248d35050b3cd2fa031b9366c           date
#>  5: 5895f04248d35050b3cd2fa031b9366c        cacheId
#>  6: 5895f04248d35050b3cd2fa031b9366c       function
#>  7: 5895f04248d35050b3cd2fa031b9366c    object.size
#>  8: 5895f04248d35050b3cd2fa031b9366c       accessed
#>  9: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 10: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 11: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 12: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 13: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 14: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 15: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 16: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 17: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 18: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 19: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 20: 5895f04248d35050b3cd2fa031b9366c      preDigest
#> 21: 5895f04248d35050b3cd2fa031b9366c      preDigest
#> 22: 5e52f410b4f7f1843a9a5e1280bc5576         format
#> 23: 5e52f410b4f7f1843a9a5e1280bc5576           name
#> 24: 5e52f410b4f7f1843a9a5e1280bc5576          class
#> 25: 5e52f410b4f7f1843a9a5e1280bc5576           date
#> 26: 5e52f410b4f7f1843a9a5e1280bc5576        cacheId
#> 27: 5e52f410b4f7f1843a9a5e1280bc5576       function
#> 28: 5e52f410b4f7f1843a9a5e1280bc5576    object.size
#> 29: 5e52f410b4f7f1843a9a5e1280bc5576       accessed
#> 30: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 31: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 32: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 33: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 34: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 35: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 36: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 37: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 38: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 39: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 40: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 41: 5e52f410b4f7f1843a9a5e1280bc5576      preDigest
#> 42: 5e52f410b4f7f1843a9a5e1280bc5576      preDigest
#> 43: 75d6cdb892950c5734bde85d803e0fcc         format
#> 44: 75d6cdb892950c5734bde85d803e0fcc           name
#> 45: 75d6cdb892950c5734bde85d803e0fcc          class
#> 46: 75d6cdb892950c5734bde85d803e0fcc           date
#> 47: 75d6cdb892950c5734bde85d803e0fcc        cacheId
#> 48: 75d6cdb892950c5734bde85d803e0fcc       function
#> 49: 75d6cdb892950c5734bde85d803e0fcc    object.size
#> 50: 75d6cdb892950c5734bde85d803e0fcc       accessed
#> 51: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 52: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 53: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 54: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 55: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 56: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 57: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 58: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 59: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 60: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 61: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 62: 75d6cdb892950c5734bde85d803e0fcc      preDigest
#> 63: 75d6cdb892950c5734bde85d803e0fcc      preDigest
#> 64: a00fc7ac4c955f5963d9e4ff89b6d323         format
#> 65: a00fc7ac4c955f5963d9e4ff89b6d323           name
#> 66: a00fc7ac4c955f5963d9e4ff89b6d323          class
#> 67: a00fc7ac4c955f5963d9e4ff89b6d323           date
#> 68: a00fc7ac4c955f5963d9e4ff89b6d323        cacheId
#> 69: a00fc7ac4c955f5963d9e4ff89b6d323       function
#> 70: a00fc7ac4c955f5963d9e4ff89b6d323    object.size
#> 71: a00fc7ac4c955f5963d9e4ff89b6d323       accessed
#> 72: a00fc7ac4c955f5963d9e4ff89b6d323 otherFunctions
#> 73: a00fc7ac4c955f5963d9e4ff89b6d323 otherFunctions
#> 74: a00fc7ac4c955f5963d9e4ff89b6d323 otherFunctions
#> 75: a00fc7ac4c955f5963d9e4ff89b6d323 otherFunctions
#> 76: a00fc7ac4c955f5963d9e4ff89b6d323 otherFunctions
#> 77: a00fc7ac4c955f5963d9e4ff89b6d323 otherFunctions
#> 78: a00fc7ac4c955f5963d9e4ff89b6d323 otherFunctions
#> 79: a00fc7ac4c955f5963d9e4ff89b6d323 otherFunctions
#> 80: a00fc7ac4c955f5963d9e4ff89b6d323 otherFunctions
#> 81: a00fc7ac4c955f5963d9e4ff89b6d323 otherFunctions
#> 82: a00fc7ac4c955f5963d9e4ff89b6d323 otherFunctions
#> 83: a00fc7ac4c955f5963d9e4ff89b6d323      preDigest
#> 84: a00fc7ac4c955f5963d9e4ff89b6d323      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-09-05 12:26:40
#>  2: 5895f04248d35050b3cd2fa031b9366c 2019-09-05 12:26:40
#>  3:                          numeric 2019-09-05 12:26:40
#>  4:              2019-09-05 12:26:40 2019-09-05 12:26:40
#>  5:                 f0da91a44b839434 2019-09-05 12:26:40
#>  6:                            rnorm 2019-09-05 12:26:40
#>  7:                             1008 2019-09-05 12:26:40
#>  8:              2019-09-05 12:26:40 2019-09-05 12:26:40
#>  9:                          saveRDS 2019-09-05 12:26:40
#> 10:                          do.call 2019-09-05 12:26:40
#> 11:                 build_site_local 2019-09-05 12:26:40
#> 12:                  build_reference 2019-09-05 12:26:40
#> 13:             data_reference_topic 2019-09-05 12:26:40
#> 14:                     run_examples 2019-09-05 12:26:40
#> 15:               highlight_examples 2019-09-05 12:26:40
#> 16:                        timing_fn 2019-09-05 12:26:40
#> 17:                           handle 2019-09-05 12:26:40
#> 18:                              try 2019-09-05 12:26:40
#> 19:                      withVisible 2019-09-05 12:26:40
#> 20:               n:7f12988bd88a0fb8 2019-09-05 12:26:40
#> 21:            .FUN:4f604aa46882b368 2019-09-05 12:26:40
#> 22:                              rda 2019-09-05 12:26:40
#> 23: 5e52f410b4f7f1843a9a5e1280bc5576 2019-09-05 12:26:40
#> 24:                          numeric 2019-09-05 12:26:40
#> 25:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 26:                 f7bee22047b8d0c1 2019-09-05 12:26:40
#> 27:                            rnorm 2019-09-05 12:26:40
#> 28:                             1008 2019-09-05 12:26:40
#> 29:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 30:                          saveRDS 2019-09-05 12:26:40
#> 31:                          do.call 2019-09-05 12:26:40
#> 32:                 build_site_local 2019-09-05 12:26:40
#> 33:                  build_reference 2019-09-05 12:26:40
#> 34:             data_reference_topic 2019-09-05 12:26:40
#> 35:                     run_examples 2019-09-05 12:26:40
#> 36:               highlight_examples 2019-09-05 12:26:40
#> 37:                        timing_fn 2019-09-05 12:26:40
#> 38:                           handle 2019-09-05 12:26:40
#> 39:                              try 2019-09-05 12:26:40
#> 40:                      withVisible 2019-09-05 12:26:40
#> 41:               n:7eef4eae85fd9229 2019-09-05 12:26:40
#> 42:            .FUN:4f604aa46882b368 2019-09-05 12:26:40
#> 43:                              rda 2019-09-05 12:26:40
#> 44: 75d6cdb892950c5734bde85d803e0fcc 2019-09-05 12:26:40
#> 45:                          numeric 2019-09-05 12:26:40
#> 46:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 47:                 7072c305d8c69df0 2019-09-05 12:26:40
#> 48:                            rnorm 2019-09-05 12:26:40
#> 49:                              984 2019-09-05 12:26:40
#> 50:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 51:                          saveRDS 2019-09-05 12:26:40
#> 52:                          do.call 2019-09-05 12:26:40
#> 53:                 build_site_local 2019-09-05 12:26:40
#> 54:                  build_reference 2019-09-05 12:26:40
#> 55:             data_reference_topic 2019-09-05 12:26:40
#> 56:                     run_examples 2019-09-05 12:26:40
#> 57:               highlight_examples 2019-09-05 12:26:40
#> 58:                        timing_fn 2019-09-05 12:26:40
#> 59:                           handle 2019-09-05 12:26:40
#> 60:                              try 2019-09-05 12:26:40
#> 61:                      withVisible 2019-09-05 12:26:40
#> 62:               n:853b1797f54b229c 2019-09-05 12:26:40
#> 63:            .FUN:4f604aa46882b368 2019-09-05 12:26:40
#> 64:                              rda 2019-09-05 12:26:40
#> 65: a00fc7ac4c955f5963d9e4ff89b6d323 2019-09-05 12:26:40
#> 66:                          numeric 2019-09-05 12:26:40
#> 67:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 68:                 3b347dee8f6305c7 2019-09-05 12:26:40
#> 69:                            rnorm 2019-09-05 12:26:40
#> 70:                              992 2019-09-05 12:26:40
#> 71:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 72:                          saveRDS 2019-09-05 12:26:40
#> 73:                          do.call 2019-09-05 12:26:40
#> 74:                 build_site_local 2019-09-05 12:26:40
#> 75:                  build_reference 2019-09-05 12:26:40
#> 76:             data_reference_topic 2019-09-05 12:26:40
#> 77:                     run_examples 2019-09-05 12:26:40
#> 78:               highlight_examples 2019-09-05 12:26:40
#> 79:                        timing_fn 2019-09-05 12:26:40
#> 80:                           handle 2019-09-05 12:26:40
#> 81:                              try 2019-09-05 12:26:40
#> 82:                      withVisible 2019-09-05 12:26:40
#> 83:               n:82dc709f2b91918a 2019-09-05 12:26:40
#> 84:            .FUN:4f604aa46882b368 2019-09-05 12:26:40
#>                             tagValue         createdDate

cc(ask = FALSE, x = tmpDir)

#> No time provided; removing the most recent entry to the Cache

#> Cache size: 

#>   Total (including Rasters): 998 bytes

#>   Selected objects (not including Rasters): 248 bytes

showCache(x = tmpDir) # most recent is gone

#> Cache size: 

#>   Total (including Rasters): 750 bytes

#>   Selected objects (not including Rasters): 750 bytes

#>                             artifact         tagKey
#>  1: 5895f04248d35050b3cd2fa031b9366c         format
#>  2: 5895f04248d35050b3cd2fa031b9366c           name
#>  3: 5895f04248d35050b3cd2fa031b9366c          class
#>  4: 5895f04248d35050b3cd2fa031b9366c           date
#>  5: 5895f04248d35050b3cd2fa031b9366c        cacheId
#>  6: 5895f04248d35050b3cd2fa031b9366c       function
#>  7: 5895f04248d35050b3cd2fa031b9366c    object.size
#>  8: 5895f04248d35050b3cd2fa031b9366c       accessed
#>  9: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 10: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 11: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 12: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 13: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 14: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 15: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 16: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 17: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 18: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 19: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 20: 5895f04248d35050b3cd2fa031b9366c      preDigest
#> 21: 5895f04248d35050b3cd2fa031b9366c      preDigest
#> 22: 5e52f410b4f7f1843a9a5e1280bc5576         format
#> 23: 5e52f410b4f7f1843a9a5e1280bc5576           name
#> 24: 5e52f410b4f7f1843a9a5e1280bc5576          class
#> 25: 5e52f410b4f7f1843a9a5e1280bc5576           date
#> 26: 5e52f410b4f7f1843a9a5e1280bc5576        cacheId
#> 27: 5e52f410b4f7f1843a9a5e1280bc5576       function
#> 28: 5e52f410b4f7f1843a9a5e1280bc5576    object.size
#> 29: 5e52f410b4f7f1843a9a5e1280bc5576       accessed
#> 30: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 31: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 32: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 33: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 34: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 35: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 36: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 37: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 38: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 39: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 40: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 41: 5e52f410b4f7f1843a9a5e1280bc5576      preDigest
#> 42: 5e52f410b4f7f1843a9a5e1280bc5576      preDigest
#> 43: 75d6cdb892950c5734bde85d803e0fcc         format
#> 44: 75d6cdb892950c5734bde85d803e0fcc           name
#> 45: 75d6cdb892950c5734bde85d803e0fcc          class
#> 46: 75d6cdb892950c5734bde85d803e0fcc           date
#> 47: 75d6cdb892950c5734bde85d803e0fcc        cacheId
#> 48: 75d6cdb892950c5734bde85d803e0fcc       function
#> 49: 75d6cdb892950c5734bde85d803e0fcc    object.size
#> 50: 75d6cdb892950c5734bde85d803e0fcc       accessed
#> 51: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 52: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 53: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 54: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 55: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 56: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 57: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 58: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 59: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 60: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 61: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 62: 75d6cdb892950c5734bde85d803e0fcc      preDigest
#> 63: 75d6cdb892950c5734bde85d803e0fcc      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-09-05 12:26:40
#>  2: 5895f04248d35050b3cd2fa031b9366c 2019-09-05 12:26:40
#>  3:                          numeric 2019-09-05 12:26:40
#>  4:              2019-09-05 12:26:40 2019-09-05 12:26:40
#>  5:                 f0da91a44b839434 2019-09-05 12:26:40
#>  6:                            rnorm 2019-09-05 12:26:40
#>  7:                             1008 2019-09-05 12:26:40
#>  8:              2019-09-05 12:26:40 2019-09-05 12:26:40
#>  9:                          saveRDS 2019-09-05 12:26:40
#> 10:                          do.call 2019-09-05 12:26:40
#> 11:                 build_site_local 2019-09-05 12:26:40
#> 12:                  build_reference 2019-09-05 12:26:40
#> 13:             data_reference_topic 2019-09-05 12:26:40
#> 14:                     run_examples 2019-09-05 12:26:40
#> 15:               highlight_examples 2019-09-05 12:26:40
#> 16:                        timing_fn 2019-09-05 12:26:40
#> 17:                           handle 2019-09-05 12:26:40
#> 18:                              try 2019-09-05 12:26:40
#> 19:                      withVisible 2019-09-05 12:26:40
#> 20:               n:7f12988bd88a0fb8 2019-09-05 12:26:40
#> 21:            .FUN:4f604aa46882b368 2019-09-05 12:26:40
#> 22:                              rda 2019-09-05 12:26:40
#> 23: 5e52f410b4f7f1843a9a5e1280bc5576 2019-09-05 12:26:40
#> 24:                          numeric 2019-09-05 12:26:40
#> 25:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 26:                 f7bee22047b8d0c1 2019-09-05 12:26:40
#> 27:                            rnorm 2019-09-05 12:26:40
#> 28:                             1008 2019-09-05 12:26:40
#> 29:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 30:                          saveRDS 2019-09-05 12:26:40
#> 31:                          do.call 2019-09-05 12:26:40
#> 32:                 build_site_local 2019-09-05 12:26:40
#> 33:                  build_reference 2019-09-05 12:26:40
#> 34:             data_reference_topic 2019-09-05 12:26:40
#> 35:                     run_examples 2019-09-05 12:26:40
#> 36:               highlight_examples 2019-09-05 12:26:40
#> 37:                        timing_fn 2019-09-05 12:26:40
#> 38:                           handle 2019-09-05 12:26:40
#> 39:                              try 2019-09-05 12:26:40
#> 40:                      withVisible 2019-09-05 12:26:40
#> 41:               n:7eef4eae85fd9229 2019-09-05 12:26:40
#> 42:            .FUN:4f604aa46882b368 2019-09-05 12:26:40
#> 43:                              rda 2019-09-05 12:26:40
#> 44: 75d6cdb892950c5734bde85d803e0fcc 2019-09-05 12:26:40
#> 45:                          numeric 2019-09-05 12:26:40
#> 46:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 47:                 7072c305d8c69df0 2019-09-05 12:26:40
#> 48:                            rnorm 2019-09-05 12:26:40
#> 49:                              984 2019-09-05 12:26:40
#> 50:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 51:                          saveRDS 2019-09-05 12:26:40
#> 52:                          do.call 2019-09-05 12:26:40
#> 53:                 build_site_local 2019-09-05 12:26:40
#> 54:                  build_reference 2019-09-05 12:26:40
#> 55:             data_reference_topic 2019-09-05 12:26:40
#> 56:                     run_examples 2019-09-05 12:26:40
#> 57:               highlight_examples 2019-09-05 12:26:40
#> 58:                        timing_fn 2019-09-05 12:26:40
#> 59:                           handle 2019-09-05 12:26:40
#> 60:                              try 2019-09-05 12:26:40
#> 61:                      withVisible 2019-09-05 12:26:40
#> 62:               n:853b1797f54b229c 2019-09-05 12:26:40
#> 63:            .FUN:4f604aa46882b368 2019-09-05 12:26:40
#>                             tagValue         createdDate

cc(thisTime, ask = FALSE, x = tmpDir)

#> Cache size: 

#>   Total (including Rasters): 750 bytes

#>   Selected objects (not including Rasters): 0 bytes

showCache(x = tmpDir) # all those after thisTime gone, i.e., only 1 left

#> Cache size: 

#>   Total (including Rasters): 750 bytes

#>   Selected objects (not including Rasters): 750 bytes

#>                             artifact         tagKey
#>  1: 5895f04248d35050b3cd2fa031b9366c         format
#>  2: 5895f04248d35050b3cd2fa031b9366c           name
#>  3: 5895f04248d35050b3cd2fa031b9366c          class
#>  4: 5895f04248d35050b3cd2fa031b9366c           date
#>  5: 5895f04248d35050b3cd2fa031b9366c        cacheId
#>  6: 5895f04248d35050b3cd2fa031b9366c       function
#>  7: 5895f04248d35050b3cd2fa031b9366c    object.size
#>  8: 5895f04248d35050b3cd2fa031b9366c       accessed
#>  9: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 10: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 11: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 12: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 13: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 14: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 15: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 16: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 17: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 18: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 19: 5895f04248d35050b3cd2fa031b9366c otherFunctions
#> 20: 5895f04248d35050b3cd2fa031b9366c      preDigest
#> 21: 5895f04248d35050b3cd2fa031b9366c      preDigest
#> 22: 5e52f410b4f7f1843a9a5e1280bc5576         format
#> 23: 5e52f410b4f7f1843a9a5e1280bc5576           name
#> 24: 5e52f410b4f7f1843a9a5e1280bc5576          class
#> 25: 5e52f410b4f7f1843a9a5e1280bc5576           date
#> 26: 5e52f410b4f7f1843a9a5e1280bc5576        cacheId
#> 27: 5e52f410b4f7f1843a9a5e1280bc5576       function
#> 28: 5e52f410b4f7f1843a9a5e1280bc5576    object.size
#> 29: 5e52f410b4f7f1843a9a5e1280bc5576       accessed
#> 30: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 31: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 32: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 33: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 34: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 35: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 36: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 37: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 38: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 39: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 40: 5e52f410b4f7f1843a9a5e1280bc5576 otherFunctions
#> 41: 5e52f410b4f7f1843a9a5e1280bc5576      preDigest
#> 42: 5e52f410b4f7f1843a9a5e1280bc5576      preDigest
#> 43: 75d6cdb892950c5734bde85d803e0fcc         format
#> 44: 75d6cdb892950c5734bde85d803e0fcc           name
#> 45: 75d6cdb892950c5734bde85d803e0fcc          class
#> 46: 75d6cdb892950c5734bde85d803e0fcc           date
#> 47: 75d6cdb892950c5734bde85d803e0fcc        cacheId
#> 48: 75d6cdb892950c5734bde85d803e0fcc       function
#> 49: 75d6cdb892950c5734bde85d803e0fcc    object.size
#> 50: 75d6cdb892950c5734bde85d803e0fcc       accessed
#> 51: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 52: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 53: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 54: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 55: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 56: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 57: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 58: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 59: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 60: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 61: 75d6cdb892950c5734bde85d803e0fcc otherFunctions
#> 62: 75d6cdb892950c5734bde85d803e0fcc      preDigest
#> 63: 75d6cdb892950c5734bde85d803e0fcc      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-09-05 12:26:40
#>  2: 5895f04248d35050b3cd2fa031b9366c 2019-09-05 12:26:40
#>  3:                          numeric 2019-09-05 12:26:40
#>  4:              2019-09-05 12:26:40 2019-09-05 12:26:40
#>  5:                 f0da91a44b839434 2019-09-05 12:26:40
#>  6:                            rnorm 2019-09-05 12:26:40
#>  7:                             1008 2019-09-05 12:26:40
#>  8:              2019-09-05 12:26:40 2019-09-05 12:26:40
#>  9:                          saveRDS 2019-09-05 12:26:40
#> 10:                          do.call 2019-09-05 12:26:40
#> 11:                 build_site_local 2019-09-05 12:26:40
#> 12:                  build_reference 2019-09-05 12:26:40
#> 13:             data_reference_topic 2019-09-05 12:26:40
#> 14:                     run_examples 2019-09-05 12:26:40
#> 15:               highlight_examples 2019-09-05 12:26:40
#> 16:                        timing_fn 2019-09-05 12:26:40
#> 17:                           handle 2019-09-05 12:26:40
#> 18:                              try 2019-09-05 12:26:40
#> 19:                      withVisible 2019-09-05 12:26:40
#> 20:               n:7f12988bd88a0fb8 2019-09-05 12:26:40
#> 21:            .FUN:4f604aa46882b368 2019-09-05 12:26:40
#> 22:                              rda 2019-09-05 12:26:40
#> 23: 5e52f410b4f7f1843a9a5e1280bc5576 2019-09-05 12:26:40
#> 24:                          numeric 2019-09-05 12:26:40
#> 25:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 26:                 f7bee22047b8d0c1 2019-09-05 12:26:40
#> 27:                            rnorm 2019-09-05 12:26:40
#> 28:                             1008 2019-09-05 12:26:40
#> 29:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 30:                          saveRDS 2019-09-05 12:26:40
#> 31:                          do.call 2019-09-05 12:26:40
#> 32:                 build_site_local 2019-09-05 12:26:40
#> 33:                  build_reference 2019-09-05 12:26:40
#> 34:             data_reference_topic 2019-09-05 12:26:40
#> 35:                     run_examples 2019-09-05 12:26:40
#> 36:               highlight_examples 2019-09-05 12:26:40
#> 37:                        timing_fn 2019-09-05 12:26:40
#> 38:                           handle 2019-09-05 12:26:40
#> 39:                              try 2019-09-05 12:26:40
#> 40:                      withVisible 2019-09-05 12:26:40
#> 41:               n:7eef4eae85fd9229 2019-09-05 12:26:40
#> 42:            .FUN:4f604aa46882b368 2019-09-05 12:26:40
#> 43:                              rda 2019-09-05 12:26:40
#> 44: 75d6cdb892950c5734bde85d803e0fcc 2019-09-05 12:26:40
#> 45:                          numeric 2019-09-05 12:26:40
#> 46:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 47:                 7072c305d8c69df0 2019-09-05 12:26:40
#> 48:                            rnorm 2019-09-05 12:26:40
#> 49:                              984 2019-09-05 12:26:40
#> 50:              2019-09-05 12:26:40 2019-09-05 12:26:40
#> 51:                          saveRDS 2019-09-05 12:26:40
#> 52:                          do.call 2019-09-05 12:26:40
#> 53:                 build_site_local 2019-09-05 12:26:40
#> 54:                  build_reference 2019-09-05 12:26:40
#> 55:             data_reference_topic 2019-09-05 12:26:40
#> 56:                     run_examples 2019-09-05 12:26:40
#> 57:               highlight_examples 2019-09-05 12:26:40
#> 58:                        timing_fn 2019-09-05 12:26:40
#> 59:                           handle 2019-09-05 12:26:40
#> 60:                              try 2019-09-05 12:26:40
#> 61:                      withVisible 2019-09-05 12:26:40
#> 62:               n:853b1797f54b229c 2019-09-05 12:26:40
#> 63:            .FUN:4f604aa46882b368 2019-09-05 12:26:40
#>                             tagValue         createdDate

cc(ask = FALSE, x = tmpDir) # Cache is

#> No time provided; removing the most recent entry to the Cache

#> Cache size: 

#>   Total (including Rasters): 750 bytes

#>   Selected objects (not including Rasters): 246 bytes

cc(ask = FALSE, x = tmpDir) # Cache is already empty

#> No time provided; removing the most recent entry to the Cache

#> Cache size: 

#>   Total (including Rasters): 504 bytes

#>   Selected objects (not including Rasters): 252 bytes