Examining and modifying the cache

These are convenience wrappers around archivist package functions. They allow the user a bit of control over what is being cached.

clearCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), ...)

# S4 method for ANY
clearCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), ...)

showCache(x, userTags = character(), after, before, ...)

# S4 method for ANY
showCache(x, userTags = character(), after, before, ...)

keepCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), ...)

# S4 method for ANY
keepCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), ...)

Arguments

x
A simList or a directory containing a valid archivist repository

userTags
Character vector. If used, this will be used in place of the after and before. Specifying one or more userTag here will clear all objects that match those tags. Matching is via regular expression, meaning partial matches will work unless strict beginning (^) and end ($) of string characters are used. Matching will be against any of the 3 columns returned by showCache(), i.e., artifact, tagValue or tagName. Also, length userTags > 1, then matching is by `and`. For `or` matching, use | in a single character string. See examples.

If neither after or before are provided, nor userTags, then all objects will be removed. If both after and before are specified, then all objects between after and before will be deleted. If userTags is used, this will override after or before.

after
A time (POSIX, character understandable by data.table). Objects cached after this time will be shown or deleted.

before
A time (POSIX, character understandable by data.table). Objects cached before this time will be shown or deleted.

ask
Logical. If FALSE, then it will not ask to confirm deletions using clearCache or keepCache. Default is TRUE

...
Other arguments. Currently, regexp, a logical, can be provided. This must be TRUE if the use is passing a regular expression. Otherwise, userTags will need to be exact matches. Default is missing, which is the same as TRUE. If there are errors due to regular expression problem, try FALSE.

Value

Will clear all objects (or those that match userTags, or those between after or before) from the repository located at cachePath of the sim object, if sim is provided, or located in cacheRepo. Invisibly returns a data.table of the removed items.

Details

clearCache

remove items from the cache based on their userTag or times values.

keepCache

remove all cached items except those based on certain userTags or times values.

showCache

display the contents of the cache.

Note

If the cache is larger than 10MB, and clearCache is used, there will be a message and a pause, if interactive, to prevent accidentally deleting of a large cache repository.

See also

Examples

library(raster)
try(detach("package:magrittr", unload = TRUE), silent = TRUE) # magrittr,

#> Warning: ‘magrittr’ namespace cannot be unloaded:
#>   namespace ‘magrittr’ is imported by ‘purrr’, ‘archivist’, ‘testthat’, ‘igraph’, ‘reproducible’, ‘dplyr’, ‘stringr’, ‘pkgdown’, ‘prettyunits’, ‘sf’ so cannot be unloaded

                                    #if loaded, gives an error below

tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir), silent = TRUE) # just to make sure it is clear

# Basic use
ranNumsA <- Cache(rnorm, 10, 16, cacheRepo = tmpDir)

# All same
ranNumsB <- Cache(rnorm, 10, 16, cacheRepo = tmpDir) # recovers cached copy

#>   loading cached result from previous rnorm call, adding to memoised copy

ranNumsC <- rnorm(10, 16) %>% Cache(cacheRepo = tmpDir) # recovers cached copy

#>   loading memoised result from previous 'rnorm' pipe sequence call.

ranNumsD <- Cache(quote(rnorm(n = 10, 16)), cacheRepo = tmpDir) # recovers cached copy

#>   loading memoised result from previous rnorm call.

# Any minor change makes it different
ranNumsE <- rnorm(10, 6) %>% Cache(cacheRepo = tmpDir) # different

## Example 1: basic cache use with tags
ranNumsA <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(runif, 4, cacheRepo = tmpDir, userTags = "objectName:b")
ranNumsC <- Cache(runif, 40, cacheRepo = tmpDir, userTags = "objectName:b")

showCache(tmpDir, userTags = c("objectName"))

#> Cache size: 

#>   Total (including Rasters): 1.4 Kb

#>   Selected objects (not including Rasters): 840 bytes

#>                             artifact         tagKey
#>  1: 058593e99ed9430b20f4c15cb5439bbd         format
#>  2: 058593e99ed9430b20f4c15cb5439bbd           name
#>  3: 058593e99ed9430b20f4c15cb5439bbd          class
#>  4: 058593e99ed9430b20f4c15cb5439bbd           date
#>  5: 058593e99ed9430b20f4c15cb5439bbd        cacheId
#>  6: 058593e99ed9430b20f4c15cb5439bbd     objectName
#>  7: 058593e99ed9430b20f4c15cb5439bbd       function
#>  8: 058593e99ed9430b20f4c15cb5439bbd    object.size
#>  9: 058593e99ed9430b20f4c15cb5439bbd       accessed
#> 10: 058593e99ed9430b20f4c15cb5439bbd otherFunctions
#> 11: 058593e99ed9430b20f4c15cb5439bbd otherFunctions
#> 12: 058593e99ed9430b20f4c15cb5439bbd otherFunctions
#> 13: 058593e99ed9430b20f4c15cb5439bbd otherFunctions
#> 14: 058593e99ed9430b20f4c15cb5439bbd otherFunctions
#> 15: 058593e99ed9430b20f4c15cb5439bbd otherFunctions
#> 16: 058593e99ed9430b20f4c15cb5439bbd otherFunctions
#> 17: 058593e99ed9430b20f4c15cb5439bbd otherFunctions
#> 18: 058593e99ed9430b20f4c15cb5439bbd otherFunctions
#> 19: 058593e99ed9430b20f4c15cb5439bbd otherFunctions
#> 20: 058593e99ed9430b20f4c15cb5439bbd otherFunctions
#> 21: 058593e99ed9430b20f4c15cb5439bbd otherFunctions
#> 22: 058593e99ed9430b20f4c15cb5439bbd      preDigest
#> 23: 058593e99ed9430b20f4c15cb5439bbd      preDigest
#> 24: 4150296eba68a37eedbf0fb403ba4003         format
#> 25: 4150296eba68a37eedbf0fb403ba4003           name
#> 26: 4150296eba68a37eedbf0fb403ba4003          class
#> 27: 4150296eba68a37eedbf0fb403ba4003           date
#> 28: 4150296eba68a37eedbf0fb403ba4003        cacheId
#> 29: 4150296eba68a37eedbf0fb403ba4003     objectName
#> 30: 4150296eba68a37eedbf0fb403ba4003       function
#> 31: 4150296eba68a37eedbf0fb403ba4003    object.size
#> 32: 4150296eba68a37eedbf0fb403ba4003       accessed
#> 33: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 34: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 35: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 36: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 37: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 38: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 39: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 40: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 41: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 42: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 43: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 44: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 45: 4150296eba68a37eedbf0fb403ba4003      preDigest
#> 46: 4150296eba68a37eedbf0fb403ba4003      preDigest
#> 47: c7ce0243bcdc9ea9172561458259e860         format
#> 48: c7ce0243bcdc9ea9172561458259e860           name
#> 49: c7ce0243bcdc9ea9172561458259e860          class
#> 50: c7ce0243bcdc9ea9172561458259e860           date
#> 51: c7ce0243bcdc9ea9172561458259e860        cacheId
#> 52: c7ce0243bcdc9ea9172561458259e860     objectName
#> 53: c7ce0243bcdc9ea9172561458259e860       function
#> 54: c7ce0243bcdc9ea9172561458259e860    object.size
#> 55: c7ce0243bcdc9ea9172561458259e860       accessed
#> 56: c7ce0243bcdc9ea9172561458259e860 otherFunctions
#> 57: c7ce0243bcdc9ea9172561458259e860 otherFunctions
#> 58: c7ce0243bcdc9ea9172561458259e860 otherFunctions
#> 59: c7ce0243bcdc9ea9172561458259e860 otherFunctions
#> 60: c7ce0243bcdc9ea9172561458259e860 otherFunctions
#> 61: c7ce0243bcdc9ea9172561458259e860 otherFunctions
#> 62: c7ce0243bcdc9ea9172561458259e860 otherFunctions
#> 63: c7ce0243bcdc9ea9172561458259e860 otherFunctions
#> 64: c7ce0243bcdc9ea9172561458259e860 otherFunctions
#> 65: c7ce0243bcdc9ea9172561458259e860 otherFunctions
#> 66: c7ce0243bcdc9ea9172561458259e860 otherFunctions
#> 67: c7ce0243bcdc9ea9172561458259e860 otherFunctions
#> 68: c7ce0243bcdc9ea9172561458259e860      preDigest
#> 69: c7ce0243bcdc9ea9172561458259e860      preDigest
#>                             artifact         tagKey
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-11-06 14:17:06
#>  2:      058593e99ed9430b20f4c15cb5439bbd 2018-11-06 14:17:06
#>  3:                               numeric 2018-11-06 14:17:06
#>  4:                   2018-11-06 14:17:06 2018-11-06 14:17:06
#>  5:      c540fe4e542ac025a1cb8dff2755fc7f 2018-11-06 14:17:06
#>  6:                                     b 2018-11-06 14:17:06
#>  7:                                 runif 2018-11-06 14:17:06
#>  8:                                  1312 2018-11-06 14:17:06
#>  9:                   2018-11-06 14:17:06 2018-11-06 14:17:06
#> 10:                   withCallingHandlers 2018-11-06 14:17:06
#> 11:                               saveRDS 2018-11-06 14:17:06
#> 12:                               do.call 2018-11-06 14:17:06
#> 13:                      build_site_local 2018-11-06 14:17:06
#> 14:                       build_reference 2018-11-06 14:17:06
#> 15:                  data_reference_topic 2018-11-06 14:17:06
#> 16:                               as_data 2018-11-06 14:17:06
#> 17:                  as_data.tag_examples 2018-11-06 14:17:06
#> 18:                             timing_fn 2018-11-06 14:17:06
#> 19:                                handle 2018-11-06 14:17:06
#> 20:                                   try 2018-11-06 14:17:06
#> 21:                           withVisible 2018-11-06 14:17:06
#> 22:    n:c42c56d63e6a88c428b18dee0f4489d7 2018-11-06 14:17:06
#> 23: .FUN:179092fbad3ec261e58bc89bc8a703c3 2018-11-06 14:17:06
#> 24:                                   rda 2018-11-06 14:17:05
#> 25:      4150296eba68a37eedbf0fb403ba4003 2018-11-06 14:17:05
#> 26:                               numeric 2018-11-06 14:17:05
#> 27:                   2018-11-06 14:17:05 2018-11-06 14:17:05
#> 28:      f55d44974d91fbf61b6cb524983b1ebc 2018-11-06 14:17:05
#> 29:                                     a 2018-11-06 14:17:05
#> 30:                                 rnorm 2018-11-06 14:17:05
#> 31:                                  1024 2018-11-06 14:17:05
#> 32:                   2018-11-06 14:17:05 2018-11-06 14:17:05
#> 33:                   withCallingHandlers 2018-11-06 14:17:05
#> 34:                               saveRDS 2018-11-06 14:17:05
#> 35:                               do.call 2018-11-06 14:17:05
#> 36:                      build_site_local 2018-11-06 14:17:05
#> 37:                       build_reference 2018-11-06 14:17:05
#> 38:                  data_reference_topic 2018-11-06 14:17:05
#> 39:                               as_data 2018-11-06 14:17:05
#> 40:                  as_data.tag_examples 2018-11-06 14:17:05
#> 41:                             timing_fn 2018-11-06 14:17:05
#> 42:                                handle 2018-11-06 14:17:05
#> 43:                                   try 2018-11-06 14:17:05
#> 44:                           withVisible 2018-11-06 14:17:05
#> 45:    n:c1de94a5a8d2a5813e07152fb7ea2038 2018-11-06 14:17:05
#> 46: .FUN:a979d0700a6ef8c7c18736096ff1d522 2018-11-06 14:17:05
#> 47:                                   rda 2018-11-06 14:17:05
#> 48:      c7ce0243bcdc9ea9172561458259e860 2018-11-06 14:17:05
#> 49:                               numeric 2018-11-06 14:17:05
#> 50:                   2018-11-06 14:17:05 2018-11-06 14:17:05
#> 51:      7cb09aa387ba235db013863648303b34 2018-11-06 14:17:05
#> 52:                                     b 2018-11-06 14:17:05
#> 53:                                 runif 2018-11-06 14:17:05
#> 54:                                  1024 2018-11-06 14:17:05
#> 55:                   2018-11-06 14:17:05 2018-11-06 14:17:05
#> 56:                   withCallingHandlers 2018-11-06 14:17:05
#> 57:                               saveRDS 2018-11-06 14:17:06
#> 58:                               do.call 2018-11-06 14:17:06
#> 59:                      build_site_local 2018-11-06 14:17:06
#> 60:                       build_reference 2018-11-06 14:17:06
#> 61:                  data_reference_topic 2018-11-06 14:17:06
#> 62:                               as_data 2018-11-06 14:17:06
#> 63:                  as_data.tag_examples 2018-11-06 14:17:06
#> 64:                             timing_fn 2018-11-06 14:17:06
#> 65:                                handle 2018-11-06 14:17:06
#> 66:                                   try 2018-11-06 14:17:06
#> 67:                           withVisible 2018-11-06 14:17:06
#> 68:    n:c1de94a5a8d2a5813e07152fb7ea2038 2018-11-06 14:17:06
#> 69: .FUN:179092fbad3ec261e58bc89bc8a703c3 2018-11-06 14:17:06
#>                                  tagValue         createdDate

showCache(tmpDir, userTags = c("^a$")) # regular expression ... "a" exactly

#> Cache size: 

#>   Total (including Rasters): 1.4 Kb

#>   Selected objects (not including Rasters): 256 bytes

#>                             artifact         tagKey
#>  1: 4150296eba68a37eedbf0fb403ba4003         format
#>  2: 4150296eba68a37eedbf0fb403ba4003           name
#>  3: 4150296eba68a37eedbf0fb403ba4003          class
#>  4: 4150296eba68a37eedbf0fb403ba4003           date
#>  5: 4150296eba68a37eedbf0fb403ba4003        cacheId
#>  6: 4150296eba68a37eedbf0fb403ba4003     objectName
#>  7: 4150296eba68a37eedbf0fb403ba4003       function
#>  8: 4150296eba68a37eedbf0fb403ba4003    object.size
#>  9: 4150296eba68a37eedbf0fb403ba4003       accessed
#> 10: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 11: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 12: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 13: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 14: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 15: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 16: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 17: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 18: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 19: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 20: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 21: 4150296eba68a37eedbf0fb403ba4003 otherFunctions
#> 22: 4150296eba68a37eedbf0fb403ba4003      preDigest
#> 23: 4150296eba68a37eedbf0fb403ba4003      preDigest
#>                             artifact         tagKey
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-11-06 14:17:05
#>  2:      4150296eba68a37eedbf0fb403ba4003 2018-11-06 14:17:05
#>  3:                               numeric 2018-11-06 14:17:05
#>  4:                   2018-11-06 14:17:05 2018-11-06 14:17:05
#>  5:      f55d44974d91fbf61b6cb524983b1ebc 2018-11-06 14:17:05
#>  6:                                     a 2018-11-06 14:17:05
#>  7:                                 rnorm 2018-11-06 14:17:05
#>  8:                                  1024 2018-11-06 14:17:05
#>  9:                   2018-11-06 14:17:05 2018-11-06 14:17:05
#> 10:                   withCallingHandlers 2018-11-06 14:17:05
#> 11:                               saveRDS 2018-11-06 14:17:05
#> 12:                               do.call 2018-11-06 14:17:05
#> 13:                      build_site_local 2018-11-06 14:17:05
#> 14:                       build_reference 2018-11-06 14:17:05
#> 15:                  data_reference_topic 2018-11-06 14:17:05
#> 16:                               as_data 2018-11-06 14:17:05
#> 17:                  as_data.tag_examples 2018-11-06 14:17:05
#> 18:                             timing_fn 2018-11-06 14:17:05
#> 19:                                handle 2018-11-06 14:17:05
#> 20:                                   try 2018-11-06 14:17:05
#> 21:                           withVisible 2018-11-06 14:17:05
#> 22:    n:c1de94a5a8d2a5813e07152fb7ea2038 2018-11-06 14:17:05
#> 23: .FUN:a979d0700a6ef8c7c18736096ff1d522 2018-11-06 14:17:05
#>                                  tagValue         createdDate

# Fine control of cache elements -- pick out only the large runif object, and remove it
cache1 <- showCache(tmpDir, userTags = c("runif")) # show only cached objects made during runif

#> Cache size: 

#>   Total (including Rasters): 1.4 Kb

#>   Selected objects (not including Rasters): 584 bytes

toRemove <- cache1[tagKey=="object.size"][as.numeric(tagValue) > 700]$artifact
clearCache(tmpDir, userTags = toRemove)

#> Cache size: 

#>   Total (including Rasters): 1.4 Kb

#>   Selected objects (not including Rasters): 0 bytes

cacheAfter <- showCache(tmpDir, userTags = c("runif")) # Only the small one is left

#> Cache size: 

#>   Total (including Rasters): 1.4 Kb

#>   Selected objects (not including Rasters): 584 bytes