Examining and modifying the cache

These are convenience wrappers around archivist package functions. They allow the user a bit of control over what is being cached.

clearCache(x, userTags = character(), after, before, ...)

# S4 method for ANY
clearCache(x, userTags, after, before,
  ask = getOption("reproducible.ask"), ...)

showCache(x, userTags = character(), after, before, ...)

# S4 method for ANY
showCache(x, userTags = character(), after, before, ...)

keepCache(x, userTags = character(), after, before, ask, ...)

# S4 method for ANY
keepCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), ...)

Arguments

x
A simList or a directory containing a valid archivist repository

userTags
Character vector. If used, this will be used in place of the after and before. Specifying one or more userTag here will clear all objects that match those tags. Matching is via regular expression, meaning partial matches will work unless strict beginning (^) and end ($) of string characters are used. Matching will be against any of the 3 columns returned by showCache(), i.e., artifact, tagValue or tagName. Also, length userTags > 1, then matching is by `and`. For `or` matching, use | in a single character string. See examples.

If neither after or before are provided, nor userTags, then all objects will be removed. If both after and before are specified, then all objects between after and before will be deleted. If userTags is used, this will override after or before.

after
A time (POSIX, character understandable by data.table). Objects cached after this time will be shown or deleted.

before
A time (POSIX, character understandable by data.table). Objects cached before this time will be shown or deleted.

...
Other arguments. Currently, regexp, a logical, can be provided. This must be TRUE if the use is passing a regular expression. Otherwise, userTags will need to be exact matches. Default is missing, which is the same as TRUE. If there are errors due to regular expression problem, try FALSE.

ask
Logical. If FALSE, then it will not ask to confirm deletions using clearCache or keepCache. Default is TRUE

Value

Will clear all objects (or those that match userTags, or those between after or before) from the repository located at cachePath of the sim object, if sim is provided, or located in cacheRepo. Invisibly returns a data.table of the removed items.

Details

clearCache

remove items from the cache based on their userTag or times values.

keepCache

remove all cached items except those based on certain userTags or times values.

showCache

display the contents of the cache.

Note

If the cache is larger than 10MB, and clearCache is used, there will be a message and a pause, if interactive, to prevent accidentally deleting of a large cache repository.

See also

Examples

library(raster)
try(detach("package:magrittr", unload = TRUE), silent = TRUE) # magrittr,
                                    #if loaded, gives an error below

tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir), silent = TRUE) # just to make sure it is clear

# Basic use
ranNumsA <- Cache(rnorm, 10, 16, cacheRepo = tmpDir)

# All same
ranNumsB <- Cache(rnorm, 10, 16, cacheRepo = tmpDir) # recovers cached copy

#>   loading cached result from previous rnorm call, adding to memoised copy

ranNumsC <- rnorm(10, 16) %>% Cache(cacheRepo = tmpDir) # recovers cached copy

#>   loading memoised result from previous 'rnorm' pipe sequence call.

ranNumsD <- Cache(quote(rnorm(n = 10, 16)), cacheRepo = tmpDir) # recovers cached copy

#>   loading memoised result from previous rnorm call.

# Any minor change makes it different
ranNumsE <- rnorm(10, 6) %>% Cache(cacheRepo = tmpDir) # different

## Example 1: basic cache use with tags
ranNumsA <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(runif, 4, cacheRepo = tmpDir, userTags = "objectName:b")
ranNumsC <- Cache(runif, 40, cacheRepo = tmpDir, userTags = "objectName:b")

showCache(tmpDir, userTags = c("objectName"))

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 786 bytes

#>                             artifact         tagKey
#>  1: 47caa8b524649ff8aa21f9663b053cb1         format
#>  2: 47caa8b524649ff8aa21f9663b053cb1           name
#>  3: 47caa8b524649ff8aa21f9663b053cb1          class
#>  4: 47caa8b524649ff8aa21f9663b053cb1           date
#>  5: 47caa8b524649ff8aa21f9663b053cb1        cacheId
#>  6: 47caa8b524649ff8aa21f9663b053cb1     objectName
#>  7: 47caa8b524649ff8aa21f9663b053cb1       function
#>  8: 47caa8b524649ff8aa21f9663b053cb1    object.size
#>  9: 47caa8b524649ff8aa21f9663b053cb1       accessed
#> 10: 47caa8b524649ff8aa21f9663b053cb1 otherFunctions
#> 11: 47caa8b524649ff8aa21f9663b053cb1 otherFunctions
#> 12: 47caa8b524649ff8aa21f9663b053cb1 otherFunctions
#> 13: 47caa8b524649ff8aa21f9663b053cb1 otherFunctions
#> 14: 47caa8b524649ff8aa21f9663b053cb1 otherFunctions
#> 15: 47caa8b524649ff8aa21f9663b053cb1 otherFunctions
#> 16: 47caa8b524649ff8aa21f9663b053cb1 otherFunctions
#> 17: 47caa8b524649ff8aa21f9663b053cb1      preDigest
#> 18: 47caa8b524649ff8aa21f9663b053cb1      preDigest
#> 19: 51dfd95e1b0b05d00ec0746ba25eff66         format
#> 20: 51dfd95e1b0b05d00ec0746ba25eff66           name
#> 21: 51dfd95e1b0b05d00ec0746ba25eff66          class
#> 22: 51dfd95e1b0b05d00ec0746ba25eff66           date
#> 23: 51dfd95e1b0b05d00ec0746ba25eff66        cacheId
#> 24: 51dfd95e1b0b05d00ec0746ba25eff66     objectName
#> 25: 51dfd95e1b0b05d00ec0746ba25eff66       function
#> 26: 51dfd95e1b0b05d00ec0746ba25eff66    object.size
#> 27: 51dfd95e1b0b05d00ec0746ba25eff66       accessed
#> 28: 51dfd95e1b0b05d00ec0746ba25eff66 otherFunctions
#> 29: 51dfd95e1b0b05d00ec0746ba25eff66 otherFunctions
#> 30: 51dfd95e1b0b05d00ec0746ba25eff66 otherFunctions
#> 31: 51dfd95e1b0b05d00ec0746ba25eff66 otherFunctions
#> 32: 51dfd95e1b0b05d00ec0746ba25eff66 otherFunctions
#> 33: 51dfd95e1b0b05d00ec0746ba25eff66 otherFunctions
#> 34: 51dfd95e1b0b05d00ec0746ba25eff66 otherFunctions
#> 35: 51dfd95e1b0b05d00ec0746ba25eff66      preDigest
#> 36: 51dfd95e1b0b05d00ec0746ba25eff66      preDigest
#> 37: ca9e504d7f3dabee384f7c3f02179a00         format
#> 38: ca9e504d7f3dabee384f7c3f02179a00           name
#> 39: ca9e504d7f3dabee384f7c3f02179a00          class
#> 40: ca9e504d7f3dabee384f7c3f02179a00           date
#> 41: ca9e504d7f3dabee384f7c3f02179a00        cacheId
#> 42: ca9e504d7f3dabee384f7c3f02179a00     objectName
#> 43: ca9e504d7f3dabee384f7c3f02179a00       function
#> 44: ca9e504d7f3dabee384f7c3f02179a00    object.size
#> 45: ca9e504d7f3dabee384f7c3f02179a00       accessed
#> 46: ca9e504d7f3dabee384f7c3f02179a00 otherFunctions
#> 47: ca9e504d7f3dabee384f7c3f02179a00 otherFunctions
#> 48: ca9e504d7f3dabee384f7c3f02179a00 otherFunctions
#> 49: ca9e504d7f3dabee384f7c3f02179a00 otherFunctions
#> 50: ca9e504d7f3dabee384f7c3f02179a00 otherFunctions
#> 51: ca9e504d7f3dabee384f7c3f02179a00 otherFunctions
#> 52: ca9e504d7f3dabee384f7c3f02179a00 otherFunctions
#> 53: ca9e504d7f3dabee384f7c3f02179a00      preDigest
#> 54: ca9e504d7f3dabee384f7c3f02179a00      preDigest
#>                             artifact         tagKey
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-07-23 11:24:20
#>  2:      47caa8b524649ff8aa21f9663b053cb1 2018-07-23 11:24:20
#>  3:                               numeric 2018-07-23 11:24:20
#>  4:                   2018-07-23 11:24:20 2018-07-23 11:24:20
#>  5:      e37bb635c97bc2eeecab63816b881bbc 2018-07-23 11:24:20
#>  6:                                     b 2018-07-23 11:24:20
#>  7:                                 runif 2018-07-23 11:24:20
#>  8:                                   952 2018-07-23 11:24:20
#>  9:                   2018-07-23 11:24:20 2018-07-23 11:24:20
#> 10:                   withCallingHandlers 2018-07-23 11:24:20
#> 11:                               saveRDS 2018-07-23 11:24:20
#> 12:                               do.call 2018-07-23 11:24:20
#> 13:                      build_site_local 2018-07-23 11:24:20
#> 14:                       build_reference 2018-07-23 11:24:20
#> 15:                  data_reference_topic 2018-07-23 11:24:20
#> 16:                                   try 2018-07-23 11:24:20
#> 17:    n:969a49ec15bcd4323ff31538af321264 2018-07-23 11:24:20
#> 18: .FUN:d2631d24c3b38b89c7bdd4ab7faaaac3 2018-07-23 11:24:20
#> 19:                                   rda 2018-07-23 11:24:20
#> 20:      51dfd95e1b0b05d00ec0746ba25eff66 2018-07-23 11:24:20
#> 21:                               numeric 2018-07-23 11:24:20
#> 22:                   2018-07-23 11:24:20 2018-07-23 11:24:20
#> 23:      85874f26b2e0c1ef689a7d379d275ebf 2018-07-23 11:24:20
#> 24:                                     a 2018-07-23 11:24:20
#> 25:                                 rnorm 2018-07-23 11:24:20
#> 26:                                   952 2018-07-23 11:24:20
#> 27:                   2018-07-23 11:24:20 2018-07-23 11:24:20
#> 28:                   withCallingHandlers 2018-07-23 11:24:20
#> 29:                               saveRDS 2018-07-23 11:24:20
#> 30:                               do.call 2018-07-23 11:24:20
#> 31:                      build_site_local 2018-07-23 11:24:20
#> 32:                       build_reference 2018-07-23 11:24:20
#> 33:                  data_reference_topic 2018-07-23 11:24:20
#> 34:                                   try 2018-07-23 11:24:20
#> 35:    n:969a49ec15bcd4323ff31538af321264 2018-07-23 11:24:20
#> 36: .FUN:7e9a928f110f80b3612e71883a6ec1f4 2018-07-23 11:24:20
#> 37:                                   rda 2018-07-23 11:24:20
#> 38:      ca9e504d7f3dabee384f7c3f02179a00 2018-07-23 11:24:20
#> 39:                               numeric 2018-07-23 11:24:20
#> 40:                   2018-07-23 11:24:20 2018-07-23 11:24:20
#> 41:      b419232b60eb6eedb59532cf0b592010 2018-07-23 11:24:20
#> 42:                                     b 2018-07-23 11:24:20
#> 43:                                 runif 2018-07-23 11:24:20
#> 44:                                  1240 2018-07-23 11:24:20
#> 45:                   2018-07-23 11:24:20 2018-07-23 11:24:20
#> 46:                   withCallingHandlers 2018-07-23 11:24:20
#> 47:                               saveRDS 2018-07-23 11:24:20
#> 48:                               do.call 2018-07-23 11:24:20
#> 49:                      build_site_local 2018-07-23 11:24:20
#> 50:                       build_reference 2018-07-23 11:24:20
#> 51:                  data_reference_topic 2018-07-23 11:24:20
#> 52:                                   try 2018-07-23 11:24:20
#> 53:    n:9d7bd914b9878238bccd5686a5a41903 2018-07-23 11:24:20
#> 54: .FUN:d2631d24c3b38b89c7bdd4ab7faaaac3 2018-07-23 11:24:20
#>                                  tagValue         createdDate

showCache(tmpDir, userTags = c("^a$")) # regular expression ... "a" exactly

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 238 bytes

#>                             artifact         tagKey
#>  1: 51dfd95e1b0b05d00ec0746ba25eff66         format
#>  2: 51dfd95e1b0b05d00ec0746ba25eff66           name
#>  3: 51dfd95e1b0b05d00ec0746ba25eff66          class
#>  4: 51dfd95e1b0b05d00ec0746ba25eff66           date
#>  5: 51dfd95e1b0b05d00ec0746ba25eff66        cacheId
#>  6: 51dfd95e1b0b05d00ec0746ba25eff66     objectName
#>  7: 51dfd95e1b0b05d00ec0746ba25eff66       function
#>  8: 51dfd95e1b0b05d00ec0746ba25eff66    object.size
#>  9: 51dfd95e1b0b05d00ec0746ba25eff66       accessed
#> 10: 51dfd95e1b0b05d00ec0746ba25eff66 otherFunctions
#> 11: 51dfd95e1b0b05d00ec0746ba25eff66 otherFunctions
#> 12: 51dfd95e1b0b05d00ec0746ba25eff66 otherFunctions
#> 13: 51dfd95e1b0b05d00ec0746ba25eff66 otherFunctions
#> 14: 51dfd95e1b0b05d00ec0746ba25eff66 otherFunctions
#> 15: 51dfd95e1b0b05d00ec0746ba25eff66 otherFunctions
#> 16: 51dfd95e1b0b05d00ec0746ba25eff66 otherFunctions
#> 17: 51dfd95e1b0b05d00ec0746ba25eff66      preDigest
#> 18: 51dfd95e1b0b05d00ec0746ba25eff66      preDigest
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-07-23 11:24:20
#>  2:      51dfd95e1b0b05d00ec0746ba25eff66 2018-07-23 11:24:20
#>  3:                               numeric 2018-07-23 11:24:20
#>  4:                   2018-07-23 11:24:20 2018-07-23 11:24:20
#>  5:      85874f26b2e0c1ef689a7d379d275ebf 2018-07-23 11:24:20
#>  6:                                     a 2018-07-23 11:24:20
#>  7:                                 rnorm 2018-07-23 11:24:20
#>  8:                                   952 2018-07-23 11:24:20
#>  9:                   2018-07-23 11:24:20 2018-07-23 11:24:20
#> 10:                   withCallingHandlers 2018-07-23 11:24:20
#> 11:                               saveRDS 2018-07-23 11:24:20
#> 12:                               do.call 2018-07-23 11:24:20
#> 13:                      build_site_local 2018-07-23 11:24:20
#> 14:                       build_reference 2018-07-23 11:24:20
#> 15:                  data_reference_topic 2018-07-23 11:24:20
#> 16:                                   try 2018-07-23 11:24:20
#> 17:    n:969a49ec15bcd4323ff31538af321264 2018-07-23 11:24:20
#> 18: .FUN:7e9a928f110f80b3612e71883a6ec1f4 2018-07-23 11:24:20

# Fine control of cache elements -- pick out only the large runif object, and remove it
cache1 <- showCache(tmpDir, userTags = c("runif")) # show only cached objects made during runif

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 548 bytes

toRemove <- cache1[tagKey=="object.size"][as.numeric(tagValue) > 700]$artifact
clearCache(tmpDir, userTags = toRemove)

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 0 bytes

cacheAfter <- showCache(tmpDir, userTags = c("runif")) # Only the small one is left

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 548 bytes