Examining and modifying the cache

These are convenience wrappers around archivist package functions. They allow the user a bit of control over what is being cached.

clearCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), ...)

# S4 method for ANY
clearCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), ...)

cc(secs, ...)

showCache(x, userTags = character(), after, before, ...)

# S4 method for ANY
showCache(x, userTags = character(), after, before, ...)

keepCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), ...)

# S4 method for ANY
keepCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), ...)

Arguments

x
A simList or a directory containing a valid archivist repository. Note: For compatibility with Cache argument, cacheRepo can also be used instead of x, though x will take precedence.

userTags
Character vector. If used, this will be used in place of the after and before. Specifying one or more userTag here will clear all objects that match those tags. Matching is via regular expression, meaning partial matches will work unless strict beginning (^) and end ($) of string characters are used. Matching will be against any of the 3 columns returned by showCache(), i.e., artifact, tagValue or tagName. Also, length userTags > 1, then matching is by `and`. For `or` matching, use | in a single character string. See examples.

If neither after or before are provided, nor userTags, then all objects will be removed. If both after and before are specified, then all objects between after and before will be deleted. If userTags is used, this will override after or before.

after
A time (POSIX, character understandable by data.table). Objects cached after this time will be shown or deleted.

before
A time (POSIX, character understandable by data.table). Objects cached before this time will be shown or deleted.

ask
Logical. If FALSE, then it will not ask to confirm deletions using clearCache or keepCache. Default is TRUE

...
Other arguments. Currently, regexp, a logical, can be provided. This must be TRUE if the use is passing a regular expression. Otherwise, userTags will need to be exact matches. Default is missing, which is the same as TRUE. If there are errors due to regular expression problem, try FALSE. For cc, it is passed to clearCache, e.g., ask, userTags

secs
Currently 3 options: the number of seconds to pass to clearCache(after = secs), a POSIXct time e.g., from Sys.time(), or missing. If missing, the default, then it will delete the most recent entry in the Cache.

Value

Will clear all objects (or those that match userTags, or those between after or before) from the repository located at cachePath of the sim object, if sim is provided, or located in cacheRepo. Invisibly returns a data.table of the removed items.

Details

cc(secs) is just a shortcut for clearCache(repo = Paths$cachePath, after = secs), i.e., to remove any cache entries touched in the last secs seconds.

clearCache

remove items from the cache based on their userTag or times values.

keepCache

remove all cached items except those based on certain userTags or times values.

showCache

display the contents of the cache.

Note

If the cache is larger than 10MB, and clearCache is used, there will be a message and a pause, if interactive, to prevent accidentally deleting of a large cache repository.

See also

Examples

library(raster)

tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir, ask = FALSE), silent = TRUE) # just to make sure it is clear

# Basic use
ranNumsA <- Cache(rnorm, 10, 16, cacheRepo = tmpDir)

# All same
ranNumsB <- Cache(rnorm, 10, 16, cacheRepo = tmpDir) # recovers cached copy

#>   loading cached result from previous rnorm call, adding to memoised copy

ranNumsC <- Cache(cacheRepo = tmpDir) %C% rnorm(10, 16)  # recovers cached copy

#>   loading memoised result from previous 'rnorm' pipe sequence call.

ranNumsD <- Cache(quote(rnorm(n = 10, 16)), cacheRepo = tmpDir) # recovers cached copy

#>   loading memoised result from previous rnorm call.

# Any minor change makes it different
ranNumsE <- Cache(cacheRepo = tmpDir) %C% rnorm(10, 6)# different

## Example 1: basic cache use with tags
ranNumsA <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(runif, 4, cacheRepo = tmpDir, userTags = "objectName:b")
ranNumsC <- Cache(runif, 40, cacheRepo = tmpDir, userTags = "objectName:b")

showCache(tmpDir, userTags = c("objectName"))

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 828 bytes

#>                             artifact         tagKey
#>  1: 1eab9b21af31caca043d76e38f265849         format
#>  2: 1eab9b21af31caca043d76e38f265849           name
#>  3: 1eab9b21af31caca043d76e38f265849          class
#>  4: 1eab9b21af31caca043d76e38f265849           date
#>  5: 1eab9b21af31caca043d76e38f265849        cacheId
#>  6: 1eab9b21af31caca043d76e38f265849     objectName
#>  7: 1eab9b21af31caca043d76e38f265849       function
#>  8: 1eab9b21af31caca043d76e38f265849    object.size
#>  9: 1eab9b21af31caca043d76e38f265849       accessed
#> 10: 1eab9b21af31caca043d76e38f265849 otherFunctions
#> 11: 1eab9b21af31caca043d76e38f265849 otherFunctions
#> 12: 1eab9b21af31caca043d76e38f265849 otherFunctions
#> 13: 1eab9b21af31caca043d76e38f265849 otherFunctions
#> 14: 1eab9b21af31caca043d76e38f265849 otherFunctions
#> 15: 1eab9b21af31caca043d76e38f265849 otherFunctions
#> 16: 1eab9b21af31caca043d76e38f265849 otherFunctions
#> 17: 1eab9b21af31caca043d76e38f265849 otherFunctions
#> 18: 1eab9b21af31caca043d76e38f265849 otherFunctions
#> 19: 1eab9b21af31caca043d76e38f265849 otherFunctions
#> 20: 1eab9b21af31caca043d76e38f265849 otherFunctions
#> 21: 1eab9b21af31caca043d76e38f265849      preDigest
#> 22: 1eab9b21af31caca043d76e38f265849      preDigest
#> 23: 8218ef6e0f955b99d494842a03c83ee0         format
#> 24: 8218ef6e0f955b99d494842a03c83ee0           name
#> 25: 8218ef6e0f955b99d494842a03c83ee0          class
#> 26: 8218ef6e0f955b99d494842a03c83ee0           date
#> 27: 8218ef6e0f955b99d494842a03c83ee0        cacheId
#> 28: 8218ef6e0f955b99d494842a03c83ee0     objectName
#> 29: 8218ef6e0f955b99d494842a03c83ee0       function
#> 30: 8218ef6e0f955b99d494842a03c83ee0    object.size
#> 31: 8218ef6e0f955b99d494842a03c83ee0       accessed
#> 32: 8218ef6e0f955b99d494842a03c83ee0 otherFunctions
#> 33: 8218ef6e0f955b99d494842a03c83ee0 otherFunctions
#> 34: 8218ef6e0f955b99d494842a03c83ee0 otherFunctions
#> 35: 8218ef6e0f955b99d494842a03c83ee0 otherFunctions
#> 36: 8218ef6e0f955b99d494842a03c83ee0 otherFunctions
#> 37: 8218ef6e0f955b99d494842a03c83ee0 otherFunctions
#> 38: 8218ef6e0f955b99d494842a03c83ee0 otherFunctions
#> 39: 8218ef6e0f955b99d494842a03c83ee0 otherFunctions
#> 40: 8218ef6e0f955b99d494842a03c83ee0 otherFunctions
#> 41: 8218ef6e0f955b99d494842a03c83ee0 otherFunctions
#> 42: 8218ef6e0f955b99d494842a03c83ee0 otherFunctions
#> 43: 8218ef6e0f955b99d494842a03c83ee0      preDigest
#> 44: 8218ef6e0f955b99d494842a03c83ee0      preDigest
#> 45: ce44cd3ecef9289d7e9632827644f309         format
#> 46: ce44cd3ecef9289d7e9632827644f309           name
#> 47: ce44cd3ecef9289d7e9632827644f309          class
#> 48: ce44cd3ecef9289d7e9632827644f309           date
#> 49: ce44cd3ecef9289d7e9632827644f309        cacheId
#> 50: ce44cd3ecef9289d7e9632827644f309     objectName
#> 51: ce44cd3ecef9289d7e9632827644f309       function
#> 52: ce44cd3ecef9289d7e9632827644f309    object.size
#> 53: ce44cd3ecef9289d7e9632827644f309       accessed
#> 54: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 55: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 56: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 57: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 58: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 59: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 60: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 61: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 62: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 63: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 64: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 65: ce44cd3ecef9289d7e9632827644f309      preDigest
#> 66: ce44cd3ecef9289d7e9632827644f309      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-03-18 10:06:17
#>  2: 1eab9b21af31caca043d76e38f265849 2019-03-18 10:06:17
#>  3:                          numeric 2019-03-18 10:06:17
#>  4:              2019-03-18 10:06:17 2019-03-18 10:06:17
#>  5:                 3aef38d1fc02aee5 2019-03-18 10:06:17
#>  6:                                b 2019-03-18 10:06:17
#>  7:                            runif 2019-03-18 10:06:17
#>  8:                             1008 2019-03-18 10:06:17
#>  9:              2019-03-18 10:06:17 2019-03-18 10:06:17
#> 10:                          saveRDS 2019-03-18 10:06:17
#> 11:                          do.call 2019-03-18 10:06:17
#> 12:                 build_site_local 2019-03-18 10:06:17
#> 13:                  build_reference 2019-03-18 10:06:17
#> 14:             data_reference_topic 2019-03-18 10:06:17
#> 15:                          as_data 2019-03-18 10:06:17
#> 16:             as_data.tag_examples 2019-03-18 10:06:17
#> 17:                        timing_fn 2019-03-18 10:06:17
#> 18:                           handle 2019-03-18 10:06:17
#> 19:                              try 2019-03-18 10:06:17
#> 20:                      withVisible 2019-03-18 10:06:17
#> 21:               n:7eef4eae85fd9229 2019-03-18 10:06:17
#> 22:            .FUN:881ec847b7161f3c 2019-03-18 10:06:17
#> 23:                              rda 2019-03-18 10:06:17
#> 24: 8218ef6e0f955b99d494842a03c83ee0 2019-03-18 10:06:17
#> 25:                          numeric 2019-03-18 10:06:17
#> 26:              2019-03-18 10:06:17 2019-03-18 10:06:17
#> 27:                 7225b47fe5dc51d0 2019-03-18 10:06:17
#> 28:                                b 2019-03-18 10:06:17
#> 29:                            runif 2019-03-18 10:06:17
#> 30:                             1296 2019-03-18 10:06:17
#> 31:              2019-03-18 10:06:17 2019-03-18 10:06:17
#> 32:                          saveRDS 2019-03-18 10:06:17
#> 33:                          do.call 2019-03-18 10:06:17
#> 34:                 build_site_local 2019-03-18 10:06:17
#> 35:                  build_reference 2019-03-18 10:06:17
#> 36:             data_reference_topic 2019-03-18 10:06:17
#> 37:                          as_data 2019-03-18 10:06:17
#> 38:             as_data.tag_examples 2019-03-18 10:06:17
#> 39:                        timing_fn 2019-03-18 10:06:17
#> 40:                           handle 2019-03-18 10:06:17
#> 41:                              try 2019-03-18 10:06:17
#> 42:                      withVisible 2019-03-18 10:06:17
#> 43:               n:1393aef18608c8be 2019-03-18 10:06:17
#> 44:            .FUN:881ec847b7161f3c 2019-03-18 10:06:17
#> 45:                              rda 2019-03-18 10:06:17
#> 46: ce44cd3ecef9289d7e9632827644f309 2019-03-18 10:06:17
#> 47:                          numeric 2019-03-18 10:06:17
#> 48:              2019-03-18 10:06:17 2019-03-18 10:06:17
#> 49:                 f7bee22047b8d0c1 2019-03-18 10:06:17
#> 50:                                a 2019-03-18 10:06:17
#> 51:                            rnorm 2019-03-18 10:06:17
#> 52:                             1008 2019-03-18 10:06:17
#> 53:              2019-03-18 10:06:17 2019-03-18 10:06:17
#> 54:                          saveRDS 2019-03-18 10:06:17
#> 55:                          do.call 2019-03-18 10:06:17
#> 56:                 build_site_local 2019-03-18 10:06:17
#> 57:                  build_reference 2019-03-18 10:06:17
#> 58:             data_reference_topic 2019-03-18 10:06:17
#> 59:                          as_data 2019-03-18 10:06:17
#> 60:             as_data.tag_examples 2019-03-18 10:06:17
#> 61:                        timing_fn 2019-03-18 10:06:17
#> 62:                           handle 2019-03-18 10:06:17
#> 63:                              try 2019-03-18 10:06:17
#> 64:                      withVisible 2019-03-18 10:06:17
#> 65:               n:7eef4eae85fd9229 2019-03-18 10:06:17
#> 66:            .FUN:4f604aa46882b368 2019-03-18 10:06:17
#>                             tagValue         createdDate

showCache(tmpDir, userTags = c("^a$")) # regular expression ... "a" exactly

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 252 bytes

#>                             artifact         tagKey
#>  1: ce44cd3ecef9289d7e9632827644f309         format
#>  2: ce44cd3ecef9289d7e9632827644f309           name
#>  3: ce44cd3ecef9289d7e9632827644f309          class
#>  4: ce44cd3ecef9289d7e9632827644f309           date
#>  5: ce44cd3ecef9289d7e9632827644f309        cacheId
#>  6: ce44cd3ecef9289d7e9632827644f309     objectName
#>  7: ce44cd3ecef9289d7e9632827644f309       function
#>  8: ce44cd3ecef9289d7e9632827644f309    object.size
#>  9: ce44cd3ecef9289d7e9632827644f309       accessed
#> 10: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 11: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 12: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 13: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 14: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 15: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 16: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 17: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 18: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 19: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 20: ce44cd3ecef9289d7e9632827644f309 otherFunctions
#> 21: ce44cd3ecef9289d7e9632827644f309      preDigest
#> 22: ce44cd3ecef9289d7e9632827644f309      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-03-18 10:06:17
#>  2: ce44cd3ecef9289d7e9632827644f309 2019-03-18 10:06:17
#>  3:                          numeric 2019-03-18 10:06:17
#>  4:              2019-03-18 10:06:17 2019-03-18 10:06:17
#>  5:                 f7bee22047b8d0c1 2019-03-18 10:06:17
#>  6:                                a 2019-03-18 10:06:17
#>  7:                            rnorm 2019-03-18 10:06:17
#>  8:                             1008 2019-03-18 10:06:17
#>  9:              2019-03-18 10:06:17 2019-03-18 10:06:17
#> 10:                          saveRDS 2019-03-18 10:06:17
#> 11:                          do.call 2019-03-18 10:06:17
#> 12:                 build_site_local 2019-03-18 10:06:17
#> 13:                  build_reference 2019-03-18 10:06:17
#> 14:             data_reference_topic 2019-03-18 10:06:17
#> 15:                          as_data 2019-03-18 10:06:17
#> 16:             as_data.tag_examples 2019-03-18 10:06:17
#> 17:                        timing_fn 2019-03-18 10:06:17
#> 18:                           handle 2019-03-18 10:06:17
#> 19:                              try 2019-03-18 10:06:17
#> 20:                      withVisible 2019-03-18 10:06:17
#> 21:               n:7eef4eae85fd9229 2019-03-18 10:06:17
#> 22:            .FUN:4f604aa46882b368 2019-03-18 10:06:17
#>                             tagValue         createdDate

# Fine control of cache elements -- pick out only the large runif object, and remove it
cache1 <- showCache(tmpDir, userTags = c("runif")) # show only cached objects made during runif

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 576 bytes

toRemove <- cache1[tagKey=="object.size"][as.numeric(tagValue) > 700]$artifact
clearCache(tmpDir, userTags = toRemove, ask = FALSE)

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 0 bytes

cacheAfter <- showCache(tmpDir, userTags = c("runif")) # Only the small one is left

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 576 bytes

tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir, ask = FALSE), silent = TRUE) # just to make sure it is clear

Cache(rnorm, 1, cacheRepo = tmpDir)

#> [1] 1.060524
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:7072c305d8c69df0"
#> attr(,"call")
#> [1] ""

thisTime <- Sys.time()
Cache(rnorm, 2, cacheRepo = tmpDir)

#> [1] -1.47870 -1.55156
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:3b347dee8f6305c7"
#> attr(,"call")
#> [1] ""

Cache(rnorm, 3, cacheRepo = tmpDir)

#> [1]  0.7775067  1.0684401 -0.1835877
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:f0da91a44b839434"
#> attr(,"call")
#> [1] ""

Cache(rnorm, 4, cacheRepo = tmpDir)

#> [1]  1.5582429 -0.2132424  0.9305353  0.4108118
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:f7bee22047b8d0c1"
#> attr(,"call")
#> [1] ""

showCache(x = tmpDir) # shows all 4 entries

#> Cache size: 

#>   Total (including Rasters): 998 bytes

#>   Selected objects (not including Rasters): 998 bytes

#>                             artifact         tagKey
#>  1: 67e008fb5897fce3a70e429e4bd54235         format
#>  2: 67e008fb5897fce3a70e429e4bd54235           name
#>  3: 67e008fb5897fce3a70e429e4bd54235          class
#>  4: 67e008fb5897fce3a70e429e4bd54235           date
#>  5: 67e008fb5897fce3a70e429e4bd54235        cacheId
#>  6: 67e008fb5897fce3a70e429e4bd54235       function
#>  7: 67e008fb5897fce3a70e429e4bd54235    object.size
#>  8: 67e008fb5897fce3a70e429e4bd54235       accessed
#>  9: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 10: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 11: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 12: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 13: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 14: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 15: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 16: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 17: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 18: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 19: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 20: 67e008fb5897fce3a70e429e4bd54235      preDigest
#> 21: 67e008fb5897fce3a70e429e4bd54235      preDigest
#> 22: 8622d8a9bc18f22433166e93253ca575         format
#> 23: 8622d8a9bc18f22433166e93253ca575           name
#> 24: 8622d8a9bc18f22433166e93253ca575          class
#> 25: 8622d8a9bc18f22433166e93253ca575           date
#> 26: 8622d8a9bc18f22433166e93253ca575        cacheId
#> 27: 8622d8a9bc18f22433166e93253ca575       function
#> 28: 8622d8a9bc18f22433166e93253ca575    object.size
#> 29: 8622d8a9bc18f22433166e93253ca575       accessed
#> 30: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 31: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 32: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 33: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 34: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 35: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 36: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 37: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 38: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 39: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 40: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 41: 8622d8a9bc18f22433166e93253ca575      preDigest
#> 42: 8622d8a9bc18f22433166e93253ca575      preDigest
#> 43: b5d57899bd723c7fbeeedada0c9760f9         format
#> 44: b5d57899bd723c7fbeeedada0c9760f9           name
#> 45: b5d57899bd723c7fbeeedada0c9760f9          class
#> 46: b5d57899bd723c7fbeeedada0c9760f9           date
#> 47: b5d57899bd723c7fbeeedada0c9760f9        cacheId
#> 48: b5d57899bd723c7fbeeedada0c9760f9       function
#> 49: b5d57899bd723c7fbeeedada0c9760f9    object.size
#> 50: b5d57899bd723c7fbeeedada0c9760f9       accessed
#> 51: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 52: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 53: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 54: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 55: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 56: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 57: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 58: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 59: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 60: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 61: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 62: b5d57899bd723c7fbeeedada0c9760f9      preDigest
#> 63: b5d57899bd723c7fbeeedada0c9760f9      preDigest
#> 64: e32f022ed31ebc781d02c719a95de823         format
#> 65: e32f022ed31ebc781d02c719a95de823           name
#> 66: e32f022ed31ebc781d02c719a95de823          class
#> 67: e32f022ed31ebc781d02c719a95de823           date
#> 68: e32f022ed31ebc781d02c719a95de823        cacheId
#> 69: e32f022ed31ebc781d02c719a95de823       function
#> 70: e32f022ed31ebc781d02c719a95de823    object.size
#> 71: e32f022ed31ebc781d02c719a95de823       accessed
#> 72: e32f022ed31ebc781d02c719a95de823 otherFunctions
#> 73: e32f022ed31ebc781d02c719a95de823 otherFunctions
#> 74: e32f022ed31ebc781d02c719a95de823 otherFunctions
#> 75: e32f022ed31ebc781d02c719a95de823 otherFunctions
#> 76: e32f022ed31ebc781d02c719a95de823 otherFunctions
#> 77: e32f022ed31ebc781d02c719a95de823 otherFunctions
#> 78: e32f022ed31ebc781d02c719a95de823 otherFunctions
#> 79: e32f022ed31ebc781d02c719a95de823 otherFunctions
#> 80: e32f022ed31ebc781d02c719a95de823 otherFunctions
#> 81: e32f022ed31ebc781d02c719a95de823 otherFunctions
#> 82: e32f022ed31ebc781d02c719a95de823 otherFunctions
#> 83: e32f022ed31ebc781d02c719a95de823      preDigest
#> 84: e32f022ed31ebc781d02c719a95de823      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-03-18 10:06:18
#>  2: 67e008fb5897fce3a70e429e4bd54235 2019-03-18 10:06:18
#>  3:                          numeric 2019-03-18 10:06:18
#>  4:              2019-03-18 10:06:18 2019-03-18 10:06:18
#>  5:                 f7bee22047b8d0c1 2019-03-18 10:06:18
#>  6:                            rnorm 2019-03-18 10:06:18
#>  7:                             1008 2019-03-18 10:06:18
#>  8:              2019-03-18 10:06:18 2019-03-18 10:06:18
#>  9:                          saveRDS 2019-03-18 10:06:18
#> 10:                          do.call 2019-03-18 10:06:18
#> 11:                 build_site_local 2019-03-18 10:06:18
#> 12:                  build_reference 2019-03-18 10:06:18
#> 13:             data_reference_topic 2019-03-18 10:06:18
#> 14:                          as_data 2019-03-18 10:06:18
#> 15:             as_data.tag_examples 2019-03-18 10:06:18
#> 16:                        timing_fn 2019-03-18 10:06:18
#> 17:                           handle 2019-03-18 10:06:18
#> 18:                              try 2019-03-18 10:06:18
#> 19:                      withVisible 2019-03-18 10:06:18
#> 20:               n:7eef4eae85fd9229 2019-03-18 10:06:18
#> 21:            .FUN:4f604aa46882b368 2019-03-18 10:06:18
#> 22:                              rda 2019-03-18 10:06:18
#> 23: 8622d8a9bc18f22433166e93253ca575 2019-03-18 10:06:18
#> 24:                          numeric 2019-03-18 10:06:18
#> 25:              2019-03-18 10:06:18 2019-03-18 10:06:18
#> 26:                 3b347dee8f6305c7 2019-03-18 10:06:18
#> 27:                            rnorm 2019-03-18 10:06:18
#> 28:                              992 2019-03-18 10:06:18
#> 29:              2019-03-18 10:06:18 2019-03-18 10:06:18
#> 30:                          saveRDS 2019-03-18 10:06:18
#> 31:                          do.call 2019-03-18 10:06:18
#> 32:                 build_site_local 2019-03-18 10:06:18
#> 33:                  build_reference 2019-03-18 10:06:18
#> 34:             data_reference_topic 2019-03-18 10:06:18
#> 35:                          as_data 2019-03-18 10:06:18
#> 36:             as_data.tag_examples 2019-03-18 10:06:18
#> 37:                        timing_fn 2019-03-18 10:06:18
#> 38:                           handle 2019-03-18 10:06:18
#> 39:                              try 2019-03-18 10:06:18
#> 40:                      withVisible 2019-03-18 10:06:18
#> 41:               n:82dc709f2b91918a 2019-03-18 10:06:18
#> 42:            .FUN:4f604aa46882b368 2019-03-18 10:06:18
#> 43:                              rda 2019-03-18 10:06:18
#> 44: b5d57899bd723c7fbeeedada0c9760f9 2019-03-18 10:06:18
#> 45:                          numeric 2019-03-18 10:06:18
#> 46:              2019-03-18 10:06:18 2019-03-18 10:06:18
#> 47:                 7072c305d8c69df0 2019-03-18 10:06:18
#> 48:                            rnorm 2019-03-18 10:06:18
#> 49:                              984 2019-03-18 10:06:18
#> 50:              2019-03-18 10:06:18 2019-03-18 10:06:18
#> 51:                          saveRDS 2019-03-18 10:06:18
#> 52:                          do.call 2019-03-18 10:06:18
#> 53:                 build_site_local 2019-03-18 10:06:18
#> 54:                  build_reference 2019-03-18 10:06:18
#> 55:             data_reference_topic 2019-03-18 10:06:18
#> 56:                          as_data 2019-03-18 10:06:18
#> 57:             as_data.tag_examples 2019-03-18 10:06:18
#> 58:                        timing_fn 2019-03-18 10:06:18
#> 59:                           handle 2019-03-18 10:06:18
#> 60:                              try 2019-03-18 10:06:18
#> 61:                      withVisible 2019-03-18 10:06:18
#> 62:               n:853b1797f54b229c 2019-03-18 10:06:18
#> 63:            .FUN:4f604aa46882b368 2019-03-18 10:06:18
#> 64:                              rda 2019-03-18 10:06:18
#> 65: e32f022ed31ebc781d02c719a95de823 2019-03-18 10:06:18
#> 66:                          numeric 2019-03-18 10:06:18
#> 67:              2019-03-18 10:06:18 2019-03-18 10:06:18
#> 68:                 f0da91a44b839434 2019-03-18 10:06:18
#> 69:                            rnorm 2019-03-18 10:06:18
#> 70:                             1008 2019-03-18 10:06:18
#> 71:              2019-03-18 10:06:18 2019-03-18 10:06:18
#> 72:                          saveRDS 2019-03-18 10:06:18
#> 73:                          do.call 2019-03-18 10:06:18
#> 74:                 build_site_local 2019-03-18 10:06:18
#> 75:                  build_reference 2019-03-18 10:06:18
#> 76:             data_reference_topic 2019-03-18 10:06:18
#> 77:                          as_data 2019-03-18 10:06:18
#> 78:             as_data.tag_examples 2019-03-18 10:06:18
#> 79:                        timing_fn 2019-03-18 10:06:18
#> 80:                           handle 2019-03-18 10:06:18
#> 81:                              try 2019-03-18 10:06:18
#> 82:                      withVisible 2019-03-18 10:06:18
#> 83:               n:7f12988bd88a0fb8 2019-03-18 10:06:18
#> 84:            .FUN:4f604aa46882b368 2019-03-18 10:06:18
#>                             tagValue         createdDate

cc(ask = FALSE, x = tmpDir)

#> No time provided; removing the most recent entry to the Cache

#> Cache size: 

#>   Total (including Rasters): 998 bytes

#>   Selected objects (not including Rasters): 252 bytes

showCache(x = tmpDir) # most recent is gone

#> Cache size: 

#>   Total (including Rasters): 746 bytes

#>   Selected objects (not including Rasters): 746 bytes

#>                             artifact         tagKey
#>  1: 67e008fb5897fce3a70e429e4bd54235         format
#>  2: 67e008fb5897fce3a70e429e4bd54235           name
#>  3: 67e008fb5897fce3a70e429e4bd54235          class
#>  4: 67e008fb5897fce3a70e429e4bd54235           date
#>  5: 67e008fb5897fce3a70e429e4bd54235        cacheId
#>  6: 67e008fb5897fce3a70e429e4bd54235       function
#>  7: 67e008fb5897fce3a70e429e4bd54235    object.size
#>  8: 67e008fb5897fce3a70e429e4bd54235       accessed
#>  9: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 10: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 11: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 12: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 13: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 14: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 15: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 16: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 17: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 18: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 19: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 20: 67e008fb5897fce3a70e429e4bd54235      preDigest
#> 21: 67e008fb5897fce3a70e429e4bd54235      preDigest
#> 22: 8622d8a9bc18f22433166e93253ca575         format
#> 23: 8622d8a9bc18f22433166e93253ca575           name
#> 24: 8622d8a9bc18f22433166e93253ca575          class
#> 25: 8622d8a9bc18f22433166e93253ca575           date
#> 26: 8622d8a9bc18f22433166e93253ca575        cacheId
#> 27: 8622d8a9bc18f22433166e93253ca575       function
#> 28: 8622d8a9bc18f22433166e93253ca575    object.size
#> 29: 8622d8a9bc18f22433166e93253ca575       accessed
#> 30: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 31: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 32: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 33: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 34: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 35: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 36: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 37: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 38: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 39: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 40: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 41: 8622d8a9bc18f22433166e93253ca575      preDigest
#> 42: 8622d8a9bc18f22433166e93253ca575      preDigest
#> 43: b5d57899bd723c7fbeeedada0c9760f9         format
#> 44: b5d57899bd723c7fbeeedada0c9760f9           name
#> 45: b5d57899bd723c7fbeeedada0c9760f9          class
#> 46: b5d57899bd723c7fbeeedada0c9760f9           date
#> 47: b5d57899bd723c7fbeeedada0c9760f9        cacheId
#> 48: b5d57899bd723c7fbeeedada0c9760f9       function
#> 49: b5d57899bd723c7fbeeedada0c9760f9    object.size
#> 50: b5d57899bd723c7fbeeedada0c9760f9       accessed
#> 51: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 52: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 53: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 54: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 55: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 56: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 57: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 58: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 59: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 60: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 61: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 62: b5d57899bd723c7fbeeedada0c9760f9      preDigest
#> 63: b5d57899bd723c7fbeeedada0c9760f9      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-03-18 10:06:18
#>  2: 67e008fb5897fce3a70e429e4bd54235 2019-03-18 10:06:18
#>  3:                          numeric 2019-03-18 10:06:18
#>  4:              2019-03-18 10:06:18 2019-03-18 10:06:18
#>  5:                 f7bee22047b8d0c1 2019-03-18 10:06:18
#>  6:                            rnorm 2019-03-18 10:06:18
#>  7:                             1008 2019-03-18 10:06:18
#>  8:              2019-03-18 10:06:18 2019-03-18 10:06:18
#>  9:                          saveRDS 2019-03-18 10:06:18
#> 10:                          do.call 2019-03-18 10:06:18
#> 11:                 build_site_local 2019-03-18 10:06:18
#> 12:                  build_reference 2019-03-18 10:06:18
#> 13:             data_reference_topic 2019-03-18 10:06:18
#> 14:                          as_data 2019-03-18 10:06:18
#> 15:             as_data.tag_examples 2019-03-18 10:06:18
#> 16:                        timing_fn 2019-03-18 10:06:18
#> 17:                           handle 2019-03-18 10:06:18
#> 18:                              try 2019-03-18 10:06:18
#> 19:                      withVisible 2019-03-18 10:06:18
#> 20:               n:7eef4eae85fd9229 2019-03-18 10:06:18
#> 21:            .FUN:4f604aa46882b368 2019-03-18 10:06:18
#> 22:                              rda 2019-03-18 10:06:18
#> 23: 8622d8a9bc18f22433166e93253ca575 2019-03-18 10:06:18
#> 24:                          numeric 2019-03-18 10:06:18
#> 25:              2019-03-18 10:06:18 2019-03-18 10:06:18
#> 26:                 3b347dee8f6305c7 2019-03-18 10:06:18
#> 27:                            rnorm 2019-03-18 10:06:18
#> 28:                              992 2019-03-18 10:06:18
#> 29:              2019-03-18 10:06:18 2019-03-18 10:06:18
#> 30:                          saveRDS 2019-03-18 10:06:18
#> 31:                          do.call 2019-03-18 10:06:18
#> 32:                 build_site_local 2019-03-18 10:06:18
#> 33:                  build_reference 2019-03-18 10:06:18
#> 34:             data_reference_topic 2019-03-18 10:06:18
#> 35:                          as_data 2019-03-18 10:06:18
#> 36:             as_data.tag_examples 2019-03-18 10:06:18
#> 37:                        timing_fn 2019-03-18 10:06:18
#> 38:                           handle 2019-03-18 10:06:18
#> 39:                              try 2019-03-18 10:06:18
#> 40:                      withVisible 2019-03-18 10:06:18
#> 41:               n:82dc709f2b91918a 2019-03-18 10:06:18
#> 42:            .FUN:4f604aa46882b368 2019-03-18 10:06:18
#> 43:                              rda 2019-03-18 10:06:18
#> 44: b5d57899bd723c7fbeeedada0c9760f9 2019-03-18 10:06:18
#> 45:                          numeric 2019-03-18 10:06:18
#> 46:              2019-03-18 10:06:18 2019-03-18 10:06:18
#> 47:                 7072c305d8c69df0 2019-03-18 10:06:18
#> 48:                            rnorm 2019-03-18 10:06:18
#> 49:                              984 2019-03-18 10:06:18
#> 50:              2019-03-18 10:06:18 2019-03-18 10:06:18
#> 51:                          saveRDS 2019-03-18 10:06:18
#> 52:                          do.call 2019-03-18 10:06:18
#> 53:                 build_site_local 2019-03-18 10:06:18
#> 54:                  build_reference 2019-03-18 10:06:18
#> 55:             data_reference_topic 2019-03-18 10:06:18
#> 56:                          as_data 2019-03-18 10:06:18
#> 57:             as_data.tag_examples 2019-03-18 10:06:18
#> 58:                        timing_fn 2019-03-18 10:06:18
#> 59:                           handle 2019-03-18 10:06:18
#> 60:                              try 2019-03-18 10:06:18
#> 61:                      withVisible 2019-03-18 10:06:18
#> 62:               n:853b1797f54b229c 2019-03-18 10:06:18
#> 63:            .FUN:4f604aa46882b368 2019-03-18 10:06:18
#>                             tagValue         createdDate

cc(thisTime, ask = FALSE, x = tmpDir)

#> Cache size: 

#>   Total (including Rasters): 746 bytes

#>   Selected objects (not including Rasters): 0 bytes

showCache(x = tmpDir) # all those after thisTime gone, i.e., only 1 left

#> Cache size: 

#>   Total (including Rasters): 746 bytes

#>   Selected objects (not including Rasters): 746 bytes

#>                             artifact         tagKey
#>  1: 67e008fb5897fce3a70e429e4bd54235         format
#>  2: 67e008fb5897fce3a70e429e4bd54235           name
#>  3: 67e008fb5897fce3a70e429e4bd54235          class
#>  4: 67e008fb5897fce3a70e429e4bd54235           date
#>  5: 67e008fb5897fce3a70e429e4bd54235        cacheId
#>  6: 67e008fb5897fce3a70e429e4bd54235       function
#>  7: 67e008fb5897fce3a70e429e4bd54235    object.size
#>  8: 67e008fb5897fce3a70e429e4bd54235       accessed
#>  9: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 10: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 11: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 12: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 13: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 14: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 15: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 16: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 17: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 18: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 19: 67e008fb5897fce3a70e429e4bd54235 otherFunctions
#> 20: 67e008fb5897fce3a70e429e4bd54235      preDigest
#> 21: 67e008fb5897fce3a70e429e4bd54235      preDigest
#> 22: 8622d8a9bc18f22433166e93253ca575         format
#> 23: 8622d8a9bc18f22433166e93253ca575           name
#> 24: 8622d8a9bc18f22433166e93253ca575          class
#> 25: 8622d8a9bc18f22433166e93253ca575           date
#> 26: 8622d8a9bc18f22433166e93253ca575        cacheId
#> 27: 8622d8a9bc18f22433166e93253ca575       function
#> 28: 8622d8a9bc18f22433166e93253ca575    object.size
#> 29: 8622d8a9bc18f22433166e93253ca575       accessed
#> 30: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 31: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 32: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 33: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 34: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 35: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 36: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 37: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 38: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 39: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 40: 8622d8a9bc18f22433166e93253ca575 otherFunctions
#> 41: 8622d8a9bc18f22433166e93253ca575      preDigest
#> 42: 8622d8a9bc18f22433166e93253ca575      preDigest
#> 43: b5d57899bd723c7fbeeedada0c9760f9         format
#> 44: b5d57899bd723c7fbeeedada0c9760f9           name
#> 45: b5d57899bd723c7fbeeedada0c9760f9          class
#> 46: b5d57899bd723c7fbeeedada0c9760f9           date
#> 47: b5d57899bd723c7fbeeedada0c9760f9        cacheId
#> 48: b5d57899bd723c7fbeeedada0c9760f9       function
#> 49: b5d57899bd723c7fbeeedada0c9760f9    object.size
#> 50: b5d57899bd723c7fbeeedada0c9760f9       accessed
#> 51: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 52: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 53: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 54: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 55: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 56: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 57: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 58: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 59: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 60: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 61: b5d57899bd723c7fbeeedada0c9760f9 otherFunctions
#> 62: b5d57899bd723c7fbeeedada0c9760f9      preDigest
#> 63: b5d57899bd723c7fbeeedada0c9760f9      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-03-18 10:06:18
#>  2: 67e008fb5897fce3a70e429e4bd54235 2019-03-18 10:06:18
#>  3:                          numeric 2019-03-18 10:06:18
#>  4:              2019-03-18 10:06:18 2019-03-18 10:06:18
#>  5:                 f7bee22047b8d0c1 2019-03-18 10:06:18
#>  6:                            rnorm 2019-03-18 10:06:18
#>  7:                             1008 2019-03-18 10:06:18
#>  8:              2019-03-18 10:06:18 2019-03-18 10:06:18
#>  9:                          saveRDS 2019-03-18 10:06:18
#> 10:                          do.call 2019-03-18 10:06:18
#> 11:                 build_site_local 2019-03-18 10:06:18
#> 12:                  build_reference 2019-03-18 10:06:18
#> 13:             data_reference_topic 2019-03-18 10:06:18
#> 14:                          as_data 2019-03-18 10:06:18
#> 15:             as_data.tag_examples 2019-03-18 10:06:18
#> 16:                        timing_fn 2019-03-18 10:06:18
#> 17:                           handle 2019-03-18 10:06:18
#> 18:                              try 2019-03-18 10:06:18
#> 19:                      withVisible 2019-03-18 10:06:18
#> 20:               n:7eef4eae85fd9229 2019-03-18 10:06:18
#> 21:            .FUN:4f604aa46882b368 2019-03-18 10:06:18
#> 22:                              rda 2019-03-18 10:06:18
#> 23: 8622d8a9bc18f22433166e93253ca575 2019-03-18 10:06:18
#> 24:                          numeric 2019-03-18 10:06:18
#> 25:              2019-03-18 10:06:18 2019-03-18 10:06:18
#> 26:                 3b347dee8f6305c7 2019-03-18 10:06:18
#> 27:                            rnorm 2019-03-18 10:06:18
#> 28:                              992 2019-03-18 10:06:18
#> 29:              2019-03-18 10:06:18 2019-03-18 10:06:18
#> 30:                          saveRDS 2019-03-18 10:06:18
#> 31:                          do.call 2019-03-18 10:06:18
#> 32:                 build_site_local 2019-03-18 10:06:18
#> 33:                  build_reference 2019-03-18 10:06:18
#> 34:             data_reference_topic 2019-03-18 10:06:18
#> 35:                          as_data 2019-03-18 10:06:18
#> 36:             as_data.tag_examples 2019-03-18 10:06:18
#> 37:                        timing_fn 2019-03-18 10:06:18
#> 38:                           handle 2019-03-18 10:06:18
#> 39:                              try 2019-03-18 10:06:18
#> 40:                      withVisible 2019-03-18 10:06:18
#> 41:               n:82dc709f2b91918a 2019-03-18 10:06:18
#> 42:            .FUN:4f604aa46882b368 2019-03-18 10:06:18
#> 43:                              rda 2019-03-18 10:06:18
#> 44: b5d57899bd723c7fbeeedada0c9760f9 2019-03-18 10:06:18
#> 45:                          numeric 2019-03-18 10:06:18
#> 46:              2019-03-18 10:06:18 2019-03-18 10:06:18
#> 47:                 7072c305d8c69df0 2019-03-18 10:06:18
#> 48:                            rnorm 2019-03-18 10:06:18
#> 49:                              984 2019-03-18 10:06:18
#> 50:              2019-03-18 10:06:18 2019-03-18 10:06:18
#> 51:                          saveRDS 2019-03-18 10:06:18
#> 52:                          do.call 2019-03-18 10:06:18
#> 53:                 build_site_local 2019-03-18 10:06:18
#> 54:                  build_reference 2019-03-18 10:06:18
#> 55:             data_reference_topic 2019-03-18 10:06:18
#> 56:                          as_data 2019-03-18 10:06:18
#> 57:             as_data.tag_examples 2019-03-18 10:06:18
#> 58:                        timing_fn 2019-03-18 10:06:18
#> 59:                           handle 2019-03-18 10:06:18
#> 60:                              try 2019-03-18 10:06:18
#> 61:                      withVisible 2019-03-18 10:06:18
#> 62:               n:853b1797f54b229c 2019-03-18 10:06:18
#> 63:            .FUN:4f604aa46882b368 2019-03-18 10:06:18
#>                             tagValue         createdDate

cc(ask = FALSE, x = tmpDir) # Cache is

#> No time provided; removing the most recent entry to the Cache

#> Cache size: 

#>   Total (including Rasters): 746 bytes

#>   Selected objects (not including Rasters): 246 bytes

cc(ask = FALSE, x = tmpDir) # Cache is already empty

#> No time provided; removing the most recent entry to the Cache

#> Cache size: 

#>   Total (including Rasters): 500 bytes

#>   Selected objects (not including Rasters): 248 bytes