Examining and modifying the cache

These are convenience wrappers around archivist package functions. They allow the user a bit of control over what is being cached.

clearCache(
  x,
  userTags = character(),
  after,
  before,
  ask = getOption("reproducible.ask"),
  useCloud = FALSE,
  cloudFolderID = NULL,
  ...
)

# S4 method for ANY
clearCache(
  x,
  userTags = character(),
  after,
  before,
  ask = getOption("reproducible.ask"),
  useCloud = FALSE,
  cloudFolderID = getOption("reproducible.cloudFolderID", NULL),
  ...
)

cc(secs, ...)

showCache(x, userTags = character(), after, before, ...)

# S4 method for ANY
showCache(x, userTags = character(), after, before, ...)

keepCache(
  x,
  userTags = character(),
  after,
  before,
  ask = getOption("reproducible.ask"),
  ...
)

# S4 method for ANY
keepCache(
  x,
  userTags = character(),
  after,
  before,
  ask = getOption("reproducible.ask"),
  ...
)

Arguments

x
A simList or a directory containing a valid archivist repository. Note: For compatibility with Cache argument, cacheRepo can also be used instead of x, though x will take precedence.

userTags
Character vector. If used, this will be used in place of the after and before. Specifying one or more userTag here will clear all objects that match those tags. Matching is via regular expression, meaning partial matches will work unless strict beginning (^) and end ($) of string characters are used. Matching will be against any of the 3 columns returned by showCache(), i.e., artifact, tagValue or tagName. Also, length userTags > 1, then matching is by `and`. For `or` matching, use | in a single character string. See examples.

after
A time (POSIX, character understandable by data.table). Objects cached after this time will be shown or deleted.

before
A time (POSIX, character understandable by data.table). Objects cached before this time will be shown or deleted.

ask
Logical. If FALSE, then it will not ask to confirm deletions using clearCache or keepCache. Default is TRUE

useCloud
Logical. If TRUE, then every object that is deleted locally will also be deleted in the cloudFolderID, if it is non-NULL

cloudFolderID
A googledrive id of a folder, e.g., using drive_mkdir(). If left as NULL, the function will create a cloud folder with a warning. The warning will have the cloudFolderID that should be used in subsequent calls. It will also be added to options("reproducible.cloudFolderID"), but this will not persist across sessions.

...
Other arguments. Currently, regexp, a logical, can be provided. This must be TRUE if the use is passing a regular expression. Otherwise, userTags will need to be exact matches. Default is missing, which is the same as TRUE. If there are errors due to regular expression problem, try FALSE. For cc, it is passed to clearCache, e.g., ask, userTags

secs
Currently 3 options: the number of seconds to pass to clearCache(after = secs), a POSIXct time e.g., from Sys.time(), or missing. If missing, the default, then it will delete the most recent entry in the Cache.

Value

Will clear all objects (or those that match userTags, or those between after or before) from the repository located at cachePath of the sim object, if sim is provided, or located in cacheRepo. Invisibly returns a data.table of the removed items.

Details

If neither after or before are provided, nor userTags, then all objects will be removed. If both after and before are specified, then all objects between after and before will be deleted. If userTags is used, this will override after or before.

cc(secs) is just a shortcut for clearCache(repo = Paths$cachePath, after = secs), i.e., to remove any cache entries touched in the last secs seconds.

clearCache

remove items from the cache based on their userTag or times values.

keepCache

remove all cached items except those based on certain userTags or times values.

showCache

display the contents of the cache.

Note

If the cache is larger than 10MB, and clearCache is used, there will be a message and a pause, if interactive, to prevent accidentally deleting of a large cache repository.

See also

Examples

library(raster)

tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir, ask = FALSE), silent = TRUE) # just to make sure it is clear

# Basic use
ranNumsA <- Cache(rnorm, 10, 16, cacheRepo = tmpDir)

# All same
ranNumsB <- Cache(rnorm, 10, 16, cacheRepo = tmpDir) # recovers cached copy

#>   loading cached result from previous rnorm call, adding to memoised copy

ranNumsC <- Cache(cacheRepo = tmpDir) %C% rnorm(10, 16)  # recovers cached copy

#> Error in match.call(definition, call, expand.dots, envir): unused argument (envir = <environment>)

ranNumsD <- Cache(quote(rnorm(n = 10, 16)), cacheRepo = tmpDir) # recovers cached copy

#>   loading memoised result from previous rnorm call.

# Any minor change makes it different
ranNumsE <- Cache(cacheRepo = tmpDir) %C% rnorm(10, 6)# different

#> Error in match.call(definition, call, expand.dots, envir): unused argument (envir = <environment>)

## Example 1: basic cache use with tags
ranNumsA <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(runif, 4, cacheRepo = tmpDir, userTags = "objectName:b")
ranNumsC <- Cache(runif, 40, cacheRepo = tmpDir, userTags = "objectName:b")

showCache(tmpDir, userTags = c("objectName"))

#> Cache size: 

#>   Total (including Rasters): 1.1 Kb

#>   Selected objects (not including Rasters): 828 bytes

#>                             artifact         tagKey
#>  1: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5         format
#>  2: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5           name
#>  3: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5          class
#>  4: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5           date
#>  5: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5        cacheId
#>  6: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5     objectName
#>  7: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5       function
#>  8: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5    object.size
#>  9: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5       accessed
#> 10: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 11: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 12: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 13: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 14: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 15: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 16: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 17: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 18: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 19: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 20: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 21: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5      preDigest
#> 22: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5      preDigest
#> 23: 506a2180b14e1d9efe08e8f36277ae23         format
#> 24: 506a2180b14e1d9efe08e8f36277ae23           name
#> 25: 506a2180b14e1d9efe08e8f36277ae23          class
#> 26: 506a2180b14e1d9efe08e8f36277ae23           date
#> 27: 506a2180b14e1d9efe08e8f36277ae23        cacheId
#> 28: 506a2180b14e1d9efe08e8f36277ae23     objectName
#> 29: 506a2180b14e1d9efe08e8f36277ae23       function
#> 30: 506a2180b14e1d9efe08e8f36277ae23    object.size
#> 31: 506a2180b14e1d9efe08e8f36277ae23       accessed
#> 32: 506a2180b14e1d9efe08e8f36277ae23 otherFunctions
#> 33: 506a2180b14e1d9efe08e8f36277ae23 otherFunctions
#> 34: 506a2180b14e1d9efe08e8f36277ae23 otherFunctions
#> 35: 506a2180b14e1d9efe08e8f36277ae23 otherFunctions
#> 36: 506a2180b14e1d9efe08e8f36277ae23 otherFunctions
#> 37: 506a2180b14e1d9efe08e8f36277ae23 otherFunctions
#> 38: 506a2180b14e1d9efe08e8f36277ae23 otherFunctions
#> 39: 506a2180b14e1d9efe08e8f36277ae23 otherFunctions
#> 40: 506a2180b14e1d9efe08e8f36277ae23 otherFunctions
#> 41: 506a2180b14e1d9efe08e8f36277ae23 otherFunctions
#> 42: 506a2180b14e1d9efe08e8f36277ae23 otherFunctions
#> 43: 506a2180b14e1d9efe08e8f36277ae23      preDigest
#> 44: 506a2180b14e1d9efe08e8f36277ae23      preDigest
#> 45: c673637dcba41c6053a1d46a7f09c8b2         format
#> 46: c673637dcba41c6053a1d46a7f09c8b2           name
#> 47: c673637dcba41c6053a1d46a7f09c8b2          class
#> 48: c673637dcba41c6053a1d46a7f09c8b2           date
#> 49: c673637dcba41c6053a1d46a7f09c8b2        cacheId
#> 50: c673637dcba41c6053a1d46a7f09c8b2     objectName
#> 51: c673637dcba41c6053a1d46a7f09c8b2       function
#> 52: c673637dcba41c6053a1d46a7f09c8b2    object.size
#> 53: c673637dcba41c6053a1d46a7f09c8b2       accessed
#> 54: c673637dcba41c6053a1d46a7f09c8b2 otherFunctions
#> 55: c673637dcba41c6053a1d46a7f09c8b2 otherFunctions
#> 56: c673637dcba41c6053a1d46a7f09c8b2 otherFunctions
#> 57: c673637dcba41c6053a1d46a7f09c8b2 otherFunctions
#> 58: c673637dcba41c6053a1d46a7f09c8b2 otherFunctions
#> 59: c673637dcba41c6053a1d46a7f09c8b2 otherFunctions
#> 60: c673637dcba41c6053a1d46a7f09c8b2 otherFunctions
#> 61: c673637dcba41c6053a1d46a7f09c8b2 otherFunctions
#> 62: c673637dcba41c6053a1d46a7f09c8b2 otherFunctions
#> 63: c673637dcba41c6053a1d46a7f09c8b2 otherFunctions
#> 64: c673637dcba41c6053a1d46a7f09c8b2 otherFunctions
#> 65: c673637dcba41c6053a1d46a7f09c8b2      preDigest
#> 66: c673637dcba41c6053a1d46a7f09c8b2      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-11-17 08:38:00
#>  2: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 2019-11-17 08:38:00
#>  3:                          numeric 2019-11-17 08:38:00
#>  4:              2019-11-17 08:38:00 2019-11-17 08:38:00
#>  5:                 f7bee22047b8d0c1 2019-11-17 08:38:00
#>  6:                                a 2019-11-17 08:38:00
#>  7:                            rnorm 2019-11-17 08:38:00
#>  8:                             1008 2019-11-17 08:38:00
#>  9:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 10:                          saveRDS 2019-11-17 08:38:00
#> 11:                          do.call 2019-11-17 08:38:00
#> 12:                 build_site_local 2019-11-17 08:38:00
#> 13:                  build_reference 2019-11-17 08:38:00
#> 14:             data_reference_topic 2019-11-17 08:38:00
#> 15:                     run_examples 2019-11-17 08:38:00
#> 16:               highlight_examples 2019-11-17 08:38:00
#> 17:                        timing_fn 2019-11-17 08:38:00
#> 18:                           handle 2019-11-17 08:38:00
#> 19:                              try 2019-11-17 08:38:00
#> 20:                      withVisible 2019-11-17 08:38:00
#> 21:               n:7eef4eae85fd9229 2019-11-17 08:38:00
#> 22:            .FUN:4f604aa46882b368 2019-11-17 08:38:00
#> 23:                              rda 2019-11-17 08:38:00
#> 24: 506a2180b14e1d9efe08e8f36277ae23 2019-11-17 08:38:00
#> 25:                          numeric 2019-11-17 08:38:00
#> 26:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 27:                 3aef38d1fc02aee5 2019-11-17 08:38:00
#> 28:                                b 2019-11-17 08:38:00
#> 29:                            runif 2019-11-17 08:38:00
#> 30:                             1008 2019-11-17 08:38:00
#> 31:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 32:                          saveRDS 2019-11-17 08:38:00
#> 33:                          do.call 2019-11-17 08:38:00
#> 34:                 build_site_local 2019-11-17 08:38:00
#> 35:                  build_reference 2019-11-17 08:38:00
#> 36:             data_reference_topic 2019-11-17 08:38:00
#> 37:                     run_examples 2019-11-17 08:38:00
#> 38:               highlight_examples 2019-11-17 08:38:00
#> 39:                        timing_fn 2019-11-17 08:38:00
#> 40:                           handle 2019-11-17 08:38:00
#> 41:                              try 2019-11-17 08:38:00
#> 42:                      withVisible 2019-11-17 08:38:00
#> 43:               n:7eef4eae85fd9229 2019-11-17 08:38:00
#> 44:            .FUN:881ec847b7161f3c 2019-11-17 08:38:00
#> 45:                              rda 2019-11-17 08:38:00
#> 46: c673637dcba41c6053a1d46a7f09c8b2 2019-11-17 08:38:00
#> 47:                          numeric 2019-11-17 08:38:00
#> 48:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 49:                 7225b47fe5dc51d0 2019-11-17 08:38:00
#> 50:                                b 2019-11-17 08:38:00
#> 51:                            runif 2019-11-17 08:38:00
#> 52:                             1296 2019-11-17 08:38:00
#> 53:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 54:                          saveRDS 2019-11-17 08:38:00
#> 55:                          do.call 2019-11-17 08:38:00
#> 56:                 build_site_local 2019-11-17 08:38:00
#> 57:                  build_reference 2019-11-17 08:38:00
#> 58:             data_reference_topic 2019-11-17 08:38:00
#> 59:                     run_examples 2019-11-17 08:38:00
#> 60:               highlight_examples 2019-11-17 08:38:00
#> 61:                        timing_fn 2019-11-17 08:38:00
#> 62:                           handle 2019-11-17 08:38:00
#> 63:                              try 2019-11-17 08:38:00
#> 64:                      withVisible 2019-11-17 08:38:00
#> 65:               n:1393aef18608c8be 2019-11-17 08:38:00
#> 66:            .FUN:881ec847b7161f3c 2019-11-17 08:38:00
#>                             tagValue         createdDate

showCache(tmpDir, userTags = c("^a$")) # regular expression ... "a" exactly

#> Cache size: 

#>   Total (including Rasters): 1.1 Kb

#>   Selected objects (not including Rasters): 252 bytes

#>                             artifact         tagKey
#>  1: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5         format
#>  2: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5           name
#>  3: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5          class
#>  4: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5           date
#>  5: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5        cacheId
#>  6: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5     objectName
#>  7: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5       function
#>  8: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5    object.size
#>  9: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5       accessed
#> 10: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 11: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 12: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 13: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 14: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 15: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 16: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 17: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 18: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 19: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 20: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 otherFunctions
#> 21: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5      preDigest
#> 22: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-11-17 08:38:00
#>  2: 2fc431aa1296a6c9d2c6ee5ef5c4a8a5 2019-11-17 08:38:00
#>  3:                          numeric 2019-11-17 08:38:00
#>  4:              2019-11-17 08:38:00 2019-11-17 08:38:00
#>  5:                 f7bee22047b8d0c1 2019-11-17 08:38:00
#>  6:                                a 2019-11-17 08:38:00
#>  7:                            rnorm 2019-11-17 08:38:00
#>  8:                             1008 2019-11-17 08:38:00
#>  9:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 10:                          saveRDS 2019-11-17 08:38:00
#> 11:                          do.call 2019-11-17 08:38:00
#> 12:                 build_site_local 2019-11-17 08:38:00
#> 13:                  build_reference 2019-11-17 08:38:00
#> 14:             data_reference_topic 2019-11-17 08:38:00
#> 15:                     run_examples 2019-11-17 08:38:00
#> 16:               highlight_examples 2019-11-17 08:38:00
#> 17:                        timing_fn 2019-11-17 08:38:00
#> 18:                           handle 2019-11-17 08:38:00
#> 19:                              try 2019-11-17 08:38:00
#> 20:                      withVisible 2019-11-17 08:38:00
#> 21:               n:7eef4eae85fd9229 2019-11-17 08:38:00
#> 22:            .FUN:4f604aa46882b368 2019-11-17 08:38:00
#>                             tagValue         createdDate

# Fine control of cache elements -- pick out only the large runif object, and remove it
cache1 <- showCache(tmpDir, userTags = c("runif")) # show only cached objects made during runif

#> Cache size: 

#>   Total (including Rasters): 1.1 Kb

#>   Selected objects (not including Rasters): 576 bytes

toRemove <- cache1[tagKey == "object.size"][as.numeric(tagValue) > 700]$artifact
clearCache(tmpDir, userTags = toRemove, ask = FALSE)

#> Cache size: 

#>   Total (including Rasters): 1.1 Kb

#>   Selected objects (not including Rasters): 0 bytes

cacheAfter <- showCache(tmpDir, userTags = c("runif")) # Only the small one is left

#> Cache size: 

#>   Total (including Rasters): 1.1 Kb

#>   Selected objects (not including Rasters): 576 bytes

tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir, ask = FALSE), silent = TRUE) # just to make sure it is clear

Cache(rnorm, 1, cacheRepo = tmpDir)

#> [1] -0.1449016
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:7072c305d8c69df0"
#> attr(,"call")
#> [1] ""

thisTime <- Sys.time()
Cache(rnorm, 2, cacheRepo = tmpDir)

#> [1] -0.3244470 -0.1725649
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:3b347dee8f6305c7"
#> attr(,"call")
#> [1] ""

Cache(rnorm, 3, cacheRepo = tmpDir)

#> [1] -1.23606292 -1.90230421 -0.09450402
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:f0da91a44b839434"
#> attr(,"call")
#> [1] ""

Cache(rnorm, 4, cacheRepo = tmpDir)

#> [1]  0.03255579  0.46129012  1.38140030 -0.41647627
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:f7bee22047b8d0c1"
#> attr(,"call")
#> [1] ""

showCache(x = tmpDir) # shows all 4 entries

#> Cache size: 

#>   Total (including Rasters): 998 bytes

#>   Selected objects (not including Rasters): 998 bytes

#>                             artifact         tagKey
#>  1: 5815c83e35ce20a509e5232d86e1904a         format
#>  2: 5815c83e35ce20a509e5232d86e1904a           name
#>  3: 5815c83e35ce20a509e5232d86e1904a          class
#>  4: 5815c83e35ce20a509e5232d86e1904a           date
#>  5: 5815c83e35ce20a509e5232d86e1904a        cacheId
#>  6: 5815c83e35ce20a509e5232d86e1904a       function
#>  7: 5815c83e35ce20a509e5232d86e1904a    object.size
#>  8: 5815c83e35ce20a509e5232d86e1904a       accessed
#>  9: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 10: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 11: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 12: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 13: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 14: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 15: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 16: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 17: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 18: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 19: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 20: 5815c83e35ce20a509e5232d86e1904a      preDigest
#> 21: 5815c83e35ce20a509e5232d86e1904a      preDigest
#> 22: 64375833af9c6c7b72c27fee7f204f0b         format
#> 23: 64375833af9c6c7b72c27fee7f204f0b           name
#> 24: 64375833af9c6c7b72c27fee7f204f0b          class
#> 25: 64375833af9c6c7b72c27fee7f204f0b           date
#> 26: 64375833af9c6c7b72c27fee7f204f0b        cacheId
#> 27: 64375833af9c6c7b72c27fee7f204f0b       function
#> 28: 64375833af9c6c7b72c27fee7f204f0b    object.size
#> 29: 64375833af9c6c7b72c27fee7f204f0b       accessed
#> 30: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 31: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 32: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 33: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 34: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 35: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 36: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 37: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 38: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 39: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 40: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 41: 64375833af9c6c7b72c27fee7f204f0b      preDigest
#> 42: 64375833af9c6c7b72c27fee7f204f0b      preDigest
#> 43: ae133608fe40487cf36af067e2c6aa59         format
#> 44: ae133608fe40487cf36af067e2c6aa59           name
#> 45: ae133608fe40487cf36af067e2c6aa59          class
#> 46: ae133608fe40487cf36af067e2c6aa59           date
#> 47: ae133608fe40487cf36af067e2c6aa59        cacheId
#> 48: ae133608fe40487cf36af067e2c6aa59       function
#> 49: ae133608fe40487cf36af067e2c6aa59    object.size
#> 50: ae133608fe40487cf36af067e2c6aa59       accessed
#> 51: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 52: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 53: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 54: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 55: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 56: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 57: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 58: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 59: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 60: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 61: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 62: ae133608fe40487cf36af067e2c6aa59      preDigest
#> 63: ae133608fe40487cf36af067e2c6aa59      preDigest
#> 64: c44832accfef2852aeebd490879bd01c         format
#> 65: c44832accfef2852aeebd490879bd01c           name
#> 66: c44832accfef2852aeebd490879bd01c          class
#> 67: c44832accfef2852aeebd490879bd01c           date
#> 68: c44832accfef2852aeebd490879bd01c        cacheId
#> 69: c44832accfef2852aeebd490879bd01c       function
#> 70: c44832accfef2852aeebd490879bd01c    object.size
#> 71: c44832accfef2852aeebd490879bd01c       accessed
#> 72: c44832accfef2852aeebd490879bd01c otherFunctions
#> 73: c44832accfef2852aeebd490879bd01c otherFunctions
#> 74: c44832accfef2852aeebd490879bd01c otherFunctions
#> 75: c44832accfef2852aeebd490879bd01c otherFunctions
#> 76: c44832accfef2852aeebd490879bd01c otherFunctions
#> 77: c44832accfef2852aeebd490879bd01c otherFunctions
#> 78: c44832accfef2852aeebd490879bd01c otherFunctions
#> 79: c44832accfef2852aeebd490879bd01c otherFunctions
#> 80: c44832accfef2852aeebd490879bd01c otherFunctions
#> 81: c44832accfef2852aeebd490879bd01c otherFunctions
#> 82: c44832accfef2852aeebd490879bd01c otherFunctions
#> 83: c44832accfef2852aeebd490879bd01c      preDigest
#> 84: c44832accfef2852aeebd490879bd01c      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-11-17 08:38:00
#>  2: 5815c83e35ce20a509e5232d86e1904a 2019-11-17 08:38:00
#>  3:                          numeric 2019-11-17 08:38:00
#>  4:              2019-11-17 08:38:00 2019-11-17 08:38:00
#>  5:                 f0da91a44b839434 2019-11-17 08:38:00
#>  6:                            rnorm 2019-11-17 08:38:00
#>  7:                             1008 2019-11-17 08:38:00
#>  8:              2019-11-17 08:38:00 2019-11-17 08:38:00
#>  9:                          saveRDS 2019-11-17 08:38:00
#> 10:                          do.call 2019-11-17 08:38:00
#> 11:                 build_site_local 2019-11-17 08:38:00
#> 12:                  build_reference 2019-11-17 08:38:00
#> 13:             data_reference_topic 2019-11-17 08:38:00
#> 14:                     run_examples 2019-11-17 08:38:00
#> 15:               highlight_examples 2019-11-17 08:38:00
#> 16:                        timing_fn 2019-11-17 08:38:00
#> 17:                           handle 2019-11-17 08:38:00
#> 18:                              try 2019-11-17 08:38:00
#> 19:                      withVisible 2019-11-17 08:38:00
#> 20:               n:7f12988bd88a0fb8 2019-11-17 08:38:00
#> 21:            .FUN:4f604aa46882b368 2019-11-17 08:38:00
#> 22:                              rda 2019-11-17 08:38:00
#> 23: 64375833af9c6c7b72c27fee7f204f0b 2019-11-17 08:38:00
#> 24:                          numeric 2019-11-17 08:38:00
#> 25:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 26:                 f7bee22047b8d0c1 2019-11-17 08:38:00
#> 27:                            rnorm 2019-11-17 08:38:00
#> 28:                             1008 2019-11-17 08:38:00
#> 29:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 30:                          saveRDS 2019-11-17 08:38:00
#> 31:                          do.call 2019-11-17 08:38:00
#> 32:                 build_site_local 2019-11-17 08:38:00
#> 33:                  build_reference 2019-11-17 08:38:00
#> 34:             data_reference_topic 2019-11-17 08:38:00
#> 35:                     run_examples 2019-11-17 08:38:00
#> 36:               highlight_examples 2019-11-17 08:38:00
#> 37:                        timing_fn 2019-11-17 08:38:00
#> 38:                           handle 2019-11-17 08:38:00
#> 39:                              try 2019-11-17 08:38:00
#> 40:                      withVisible 2019-11-17 08:38:00
#> 41:               n:7eef4eae85fd9229 2019-11-17 08:38:00
#> 42:            .FUN:4f604aa46882b368 2019-11-17 08:38:00
#> 43:                              rda 2019-11-17 08:38:00
#> 44: ae133608fe40487cf36af067e2c6aa59 2019-11-17 08:38:00
#> 45:                          numeric 2019-11-17 08:38:00
#> 46:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 47:                 3b347dee8f6305c7 2019-11-17 08:38:00
#> 48:                            rnorm 2019-11-17 08:38:00
#> 49:                              992 2019-11-17 08:38:00
#> 50:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 51:                          saveRDS 2019-11-17 08:38:00
#> 52:                          do.call 2019-11-17 08:38:00
#> 53:                 build_site_local 2019-11-17 08:38:00
#> 54:                  build_reference 2019-11-17 08:38:00
#> 55:             data_reference_topic 2019-11-17 08:38:00
#> 56:                     run_examples 2019-11-17 08:38:00
#> 57:               highlight_examples 2019-11-17 08:38:00
#> 58:                        timing_fn 2019-11-17 08:38:00
#> 59:                           handle 2019-11-17 08:38:00
#> 60:                              try 2019-11-17 08:38:00
#> 61:                      withVisible 2019-11-17 08:38:00
#> 62:               n:82dc709f2b91918a 2019-11-17 08:38:00
#> 63:            .FUN:4f604aa46882b368 2019-11-17 08:38:00
#> 64:                              rda 2019-11-17 08:38:00
#> 65: c44832accfef2852aeebd490879bd01c 2019-11-17 08:38:00
#> 66:                          numeric 2019-11-17 08:38:00
#> 67:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 68:                 7072c305d8c69df0 2019-11-17 08:38:00
#> 69:                            rnorm 2019-11-17 08:38:00
#> 70:                              984 2019-11-17 08:38:00
#> 71:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 72:                          saveRDS 2019-11-17 08:38:00
#> 73:                          do.call 2019-11-17 08:38:00
#> 74:                 build_site_local 2019-11-17 08:38:00
#> 75:                  build_reference 2019-11-17 08:38:00
#> 76:             data_reference_topic 2019-11-17 08:38:00
#> 77:                     run_examples 2019-11-17 08:38:00
#> 78:               highlight_examples 2019-11-17 08:38:00
#> 79:                        timing_fn 2019-11-17 08:38:00
#> 80:                           handle 2019-11-17 08:38:00
#> 81:                              try 2019-11-17 08:38:00
#> 82:                      withVisible 2019-11-17 08:38:00
#> 83:               n:853b1797f54b229c 2019-11-17 08:38:00
#> 84:            .FUN:4f604aa46882b368 2019-11-17 08:38:00
#>                             tagValue         createdDate

cc(ask = FALSE, x = tmpDir)

#> No time provided; removing the most recent entry to the Cache

#> Cache size: 

#>   Total (including Rasters): 998 bytes

#>   Selected objects (not including Rasters): 246 bytes

showCache(x = tmpDir) # most recent is gone

#> Cache size: 

#>   Total (including Rasters): 752 bytes

#>   Selected objects (not including Rasters): 752 bytes

#>                             artifact         tagKey
#>  1: 5815c83e35ce20a509e5232d86e1904a         format
#>  2: 5815c83e35ce20a509e5232d86e1904a           name
#>  3: 5815c83e35ce20a509e5232d86e1904a          class
#>  4: 5815c83e35ce20a509e5232d86e1904a           date
#>  5: 5815c83e35ce20a509e5232d86e1904a        cacheId
#>  6: 5815c83e35ce20a509e5232d86e1904a       function
#>  7: 5815c83e35ce20a509e5232d86e1904a    object.size
#>  8: 5815c83e35ce20a509e5232d86e1904a       accessed
#>  9: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 10: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 11: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 12: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 13: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 14: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 15: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 16: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 17: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 18: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 19: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 20: 5815c83e35ce20a509e5232d86e1904a      preDigest
#> 21: 5815c83e35ce20a509e5232d86e1904a      preDigest
#> 22: 64375833af9c6c7b72c27fee7f204f0b         format
#> 23: 64375833af9c6c7b72c27fee7f204f0b           name
#> 24: 64375833af9c6c7b72c27fee7f204f0b          class
#> 25: 64375833af9c6c7b72c27fee7f204f0b           date
#> 26: 64375833af9c6c7b72c27fee7f204f0b        cacheId
#> 27: 64375833af9c6c7b72c27fee7f204f0b       function
#> 28: 64375833af9c6c7b72c27fee7f204f0b    object.size
#> 29: 64375833af9c6c7b72c27fee7f204f0b       accessed
#> 30: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 31: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 32: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 33: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 34: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 35: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 36: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 37: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 38: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 39: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 40: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 41: 64375833af9c6c7b72c27fee7f204f0b      preDigest
#> 42: 64375833af9c6c7b72c27fee7f204f0b      preDigest
#> 43: ae133608fe40487cf36af067e2c6aa59         format
#> 44: ae133608fe40487cf36af067e2c6aa59           name
#> 45: ae133608fe40487cf36af067e2c6aa59          class
#> 46: ae133608fe40487cf36af067e2c6aa59           date
#> 47: ae133608fe40487cf36af067e2c6aa59        cacheId
#> 48: ae133608fe40487cf36af067e2c6aa59       function
#> 49: ae133608fe40487cf36af067e2c6aa59    object.size
#> 50: ae133608fe40487cf36af067e2c6aa59       accessed
#> 51: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 52: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 53: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 54: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 55: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 56: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 57: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 58: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 59: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 60: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 61: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 62: ae133608fe40487cf36af067e2c6aa59      preDigest
#> 63: ae133608fe40487cf36af067e2c6aa59      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-11-17 08:38:00
#>  2: 5815c83e35ce20a509e5232d86e1904a 2019-11-17 08:38:00
#>  3:                          numeric 2019-11-17 08:38:00
#>  4:              2019-11-17 08:38:00 2019-11-17 08:38:00
#>  5:                 f0da91a44b839434 2019-11-17 08:38:00
#>  6:                            rnorm 2019-11-17 08:38:00
#>  7:                             1008 2019-11-17 08:38:00
#>  8:              2019-11-17 08:38:00 2019-11-17 08:38:00
#>  9:                          saveRDS 2019-11-17 08:38:00
#> 10:                          do.call 2019-11-17 08:38:00
#> 11:                 build_site_local 2019-11-17 08:38:00
#> 12:                  build_reference 2019-11-17 08:38:00
#> 13:             data_reference_topic 2019-11-17 08:38:00
#> 14:                     run_examples 2019-11-17 08:38:00
#> 15:               highlight_examples 2019-11-17 08:38:00
#> 16:                        timing_fn 2019-11-17 08:38:00
#> 17:                           handle 2019-11-17 08:38:00
#> 18:                              try 2019-11-17 08:38:00
#> 19:                      withVisible 2019-11-17 08:38:00
#> 20:               n:7f12988bd88a0fb8 2019-11-17 08:38:00
#> 21:            .FUN:4f604aa46882b368 2019-11-17 08:38:00
#> 22:                              rda 2019-11-17 08:38:00
#> 23: 64375833af9c6c7b72c27fee7f204f0b 2019-11-17 08:38:00
#> 24:                          numeric 2019-11-17 08:38:00
#> 25:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 26:                 f7bee22047b8d0c1 2019-11-17 08:38:00
#> 27:                            rnorm 2019-11-17 08:38:00
#> 28:                             1008 2019-11-17 08:38:00
#> 29:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 30:                          saveRDS 2019-11-17 08:38:00
#> 31:                          do.call 2019-11-17 08:38:00
#> 32:                 build_site_local 2019-11-17 08:38:00
#> 33:                  build_reference 2019-11-17 08:38:00
#> 34:             data_reference_topic 2019-11-17 08:38:00
#> 35:                     run_examples 2019-11-17 08:38:00
#> 36:               highlight_examples 2019-11-17 08:38:00
#> 37:                        timing_fn 2019-11-17 08:38:00
#> 38:                           handle 2019-11-17 08:38:00
#> 39:                              try 2019-11-17 08:38:00
#> 40:                      withVisible 2019-11-17 08:38:00
#> 41:               n:7eef4eae85fd9229 2019-11-17 08:38:00
#> 42:            .FUN:4f604aa46882b368 2019-11-17 08:38:00
#> 43:                              rda 2019-11-17 08:38:00
#> 44: ae133608fe40487cf36af067e2c6aa59 2019-11-17 08:38:00
#> 45:                          numeric 2019-11-17 08:38:00
#> 46:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 47:                 3b347dee8f6305c7 2019-11-17 08:38:00
#> 48:                            rnorm 2019-11-17 08:38:00
#> 49:                              992 2019-11-17 08:38:00
#> 50:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 51:                          saveRDS 2019-11-17 08:38:00
#> 52:                          do.call 2019-11-17 08:38:00
#> 53:                 build_site_local 2019-11-17 08:38:00
#> 54:                  build_reference 2019-11-17 08:38:00
#> 55:             data_reference_topic 2019-11-17 08:38:00
#> 56:                     run_examples 2019-11-17 08:38:00
#> 57:               highlight_examples 2019-11-17 08:38:00
#> 58:                        timing_fn 2019-11-17 08:38:00
#> 59:                           handle 2019-11-17 08:38:00
#> 60:                              try 2019-11-17 08:38:00
#> 61:                      withVisible 2019-11-17 08:38:00
#> 62:               n:82dc709f2b91918a 2019-11-17 08:38:00
#> 63:            .FUN:4f604aa46882b368 2019-11-17 08:38:00
#>                             tagValue         createdDate

cc(thisTime, ask = FALSE, x = tmpDir)

#> Cache size: 

#>   Total (including Rasters): 752 bytes

#>   Selected objects (not including Rasters): 0 bytes

showCache(x = tmpDir) # all those after thisTime gone, i.e., only 1 left

#> Cache size: 

#>   Total (including Rasters): 752 bytes

#>   Selected objects (not including Rasters): 752 bytes

#>                             artifact         tagKey
#>  1: 5815c83e35ce20a509e5232d86e1904a         format
#>  2: 5815c83e35ce20a509e5232d86e1904a           name
#>  3: 5815c83e35ce20a509e5232d86e1904a          class
#>  4: 5815c83e35ce20a509e5232d86e1904a           date
#>  5: 5815c83e35ce20a509e5232d86e1904a        cacheId
#>  6: 5815c83e35ce20a509e5232d86e1904a       function
#>  7: 5815c83e35ce20a509e5232d86e1904a    object.size
#>  8: 5815c83e35ce20a509e5232d86e1904a       accessed
#>  9: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 10: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 11: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 12: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 13: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 14: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 15: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 16: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 17: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 18: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 19: 5815c83e35ce20a509e5232d86e1904a otherFunctions
#> 20: 5815c83e35ce20a509e5232d86e1904a      preDigest
#> 21: 5815c83e35ce20a509e5232d86e1904a      preDigest
#> 22: 64375833af9c6c7b72c27fee7f204f0b         format
#> 23: 64375833af9c6c7b72c27fee7f204f0b           name
#> 24: 64375833af9c6c7b72c27fee7f204f0b          class
#> 25: 64375833af9c6c7b72c27fee7f204f0b           date
#> 26: 64375833af9c6c7b72c27fee7f204f0b        cacheId
#> 27: 64375833af9c6c7b72c27fee7f204f0b       function
#> 28: 64375833af9c6c7b72c27fee7f204f0b    object.size
#> 29: 64375833af9c6c7b72c27fee7f204f0b       accessed
#> 30: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 31: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 32: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 33: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 34: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 35: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 36: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 37: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 38: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 39: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 40: 64375833af9c6c7b72c27fee7f204f0b otherFunctions
#> 41: 64375833af9c6c7b72c27fee7f204f0b      preDigest
#> 42: 64375833af9c6c7b72c27fee7f204f0b      preDigest
#> 43: ae133608fe40487cf36af067e2c6aa59         format
#> 44: ae133608fe40487cf36af067e2c6aa59           name
#> 45: ae133608fe40487cf36af067e2c6aa59          class
#> 46: ae133608fe40487cf36af067e2c6aa59           date
#> 47: ae133608fe40487cf36af067e2c6aa59        cacheId
#> 48: ae133608fe40487cf36af067e2c6aa59       function
#> 49: ae133608fe40487cf36af067e2c6aa59    object.size
#> 50: ae133608fe40487cf36af067e2c6aa59       accessed
#> 51: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 52: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 53: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 54: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 55: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 56: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 57: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 58: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 59: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 60: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 61: ae133608fe40487cf36af067e2c6aa59 otherFunctions
#> 62: ae133608fe40487cf36af067e2c6aa59      preDigest
#> 63: ae133608fe40487cf36af067e2c6aa59      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-11-17 08:38:00
#>  2: 5815c83e35ce20a509e5232d86e1904a 2019-11-17 08:38:00
#>  3:                          numeric 2019-11-17 08:38:00
#>  4:              2019-11-17 08:38:00 2019-11-17 08:38:00
#>  5:                 f0da91a44b839434 2019-11-17 08:38:00
#>  6:                            rnorm 2019-11-17 08:38:00
#>  7:                             1008 2019-11-17 08:38:00
#>  8:              2019-11-17 08:38:00 2019-11-17 08:38:00
#>  9:                          saveRDS 2019-11-17 08:38:00
#> 10:                          do.call 2019-11-17 08:38:00
#> 11:                 build_site_local 2019-11-17 08:38:00
#> 12:                  build_reference 2019-11-17 08:38:00
#> 13:             data_reference_topic 2019-11-17 08:38:00
#> 14:                     run_examples 2019-11-17 08:38:00
#> 15:               highlight_examples 2019-11-17 08:38:00
#> 16:                        timing_fn 2019-11-17 08:38:00
#> 17:                           handle 2019-11-17 08:38:00
#> 18:                              try 2019-11-17 08:38:00
#> 19:                      withVisible 2019-11-17 08:38:00
#> 20:               n:7f12988bd88a0fb8 2019-11-17 08:38:00
#> 21:            .FUN:4f604aa46882b368 2019-11-17 08:38:00
#> 22:                              rda 2019-11-17 08:38:00
#> 23: 64375833af9c6c7b72c27fee7f204f0b 2019-11-17 08:38:00
#> 24:                          numeric 2019-11-17 08:38:00
#> 25:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 26:                 f7bee22047b8d0c1 2019-11-17 08:38:00
#> 27:                            rnorm 2019-11-17 08:38:00
#> 28:                             1008 2019-11-17 08:38:00
#> 29:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 30:                          saveRDS 2019-11-17 08:38:00
#> 31:                          do.call 2019-11-17 08:38:00
#> 32:                 build_site_local 2019-11-17 08:38:00
#> 33:                  build_reference 2019-11-17 08:38:00
#> 34:             data_reference_topic 2019-11-17 08:38:00
#> 35:                     run_examples 2019-11-17 08:38:00
#> 36:               highlight_examples 2019-11-17 08:38:00
#> 37:                        timing_fn 2019-11-17 08:38:00
#> 38:                           handle 2019-11-17 08:38:00
#> 39:                              try 2019-11-17 08:38:00
#> 40:                      withVisible 2019-11-17 08:38:00
#> 41:               n:7eef4eae85fd9229 2019-11-17 08:38:00
#> 42:            .FUN:4f604aa46882b368 2019-11-17 08:38:00
#> 43:                              rda 2019-11-17 08:38:00
#> 44: ae133608fe40487cf36af067e2c6aa59 2019-11-17 08:38:00
#> 45:                          numeric 2019-11-17 08:38:00
#> 46:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 47:                 3b347dee8f6305c7 2019-11-17 08:38:00
#> 48:                            rnorm 2019-11-17 08:38:00
#> 49:                              992 2019-11-17 08:38:00
#> 50:              2019-11-17 08:38:00 2019-11-17 08:38:00
#> 51:                          saveRDS 2019-11-17 08:38:00
#> 52:                          do.call 2019-11-17 08:38:00
#> 53:                 build_site_local 2019-11-17 08:38:00
#> 54:                  build_reference 2019-11-17 08:38:00
#> 55:             data_reference_topic 2019-11-17 08:38:00
#> 56:                     run_examples 2019-11-17 08:38:00
#> 57:               highlight_examples 2019-11-17 08:38:00
#> 58:                        timing_fn 2019-11-17 08:38:00
#> 59:                           handle 2019-11-17 08:38:00
#> 60:                              try 2019-11-17 08:38:00
#> 61:                      withVisible 2019-11-17 08:38:00
#> 62:               n:82dc709f2b91918a 2019-11-17 08:38:00
#> 63:            .FUN:4f604aa46882b368 2019-11-17 08:38:00
#>                             tagValue         createdDate

cc(ask = FALSE, x = tmpDir) # Cache is

#> No time provided; removing the most recent entry to the Cache

#> Cache size: 

#>   Total (including Rasters): 752 bytes

#>   Selected objects (not including Rasters): 248 bytes

cc(ask = FALSE, x = tmpDir) # Cache is already empty

#> No time provided; removing the most recent entry to the Cache

#> Cache size: 

#>   Total (including Rasters): 504 bytes

#>   Selected objects (not including Rasters): 252 bytes