Examining and modifying the cache

These are convenience wrappers around archivist package functions. They allow the user a bit of control over what is being cached.

clearCache(x, userTags = character(), after, before, ...)

# S4 method for ANY
clearCache(x, userTags = character(), after, before, ...)

showCache(x, userTags = character(), after, before, ...)

# S4 method for ANY
showCache(x, userTags = character(), after, before, ...)

keepCache(x, userTags = character(), after, before, ...)

# S4 method for ANY
keepCache(x, userTags = character(), after, before, ...)

Arguments

x
A simList or a directory containing a valid archivist repository

userTags
Character vector. If used, this will be used in place of the after and before. Specifying one or more userTag here will clear all objects that match those tags. Matching is via regular expresssion, meaning partial matches will work unless strict beginning (^) and end ($) of string characters are used. Matching will be against any of the 3 columns returned by showCache(), i.e., artifact, tagValue or tagName. Also, length userTags > 1, then matching is by `and`. For `or` matching, use | in a single character string. See examples.

after
A time (POSIX, character understandable by data.table). Objects cached after this time will be shown or deleted.

before
A time (POSIX, character understandable by data.table). Objects cached before this time will be shown or deleted.

...
Other arguments. Currently unused.

If neither after or before are provided, nor userTags, then all objects will be removed. If both after and before are specified, then all objects between after and before will be deleted. If userTags is used, this will override after or before.

Value

Will clear all (or that match userTags, or between after or before) objects from the repository located at cachePath of the sim object, if sim is provided, or located in cacheRepo. Also returns a data.table invisibly of the removed items.

Details

clearCache

remove items from the cache based on their userTag or times values.

keepCache

remove all cached items except those based on certain userTags or times values.

showCache

display the contents of the cache.

Note

If the cache is larger than 10MB, and clearCache is used, there will be a message and a pause, if interactive, to prevent accidentally deleting of a large cache repository.

See also

Examples

library(raster)
try(detach("package:magrittr", unload = TRUE), silent = TRUE) # magrittr,
                                    #if loaded, gives an error below

tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir), silent = TRUE) # just to make sure it is clear

# Basic use
ranNumsA <- Cache(rnorm, 10, 16, cacheRepo = tmpDir)

# All same
ranNumsB <- Cache(rnorm, 10, 16, cacheRepo = tmpDir) # recovers cached copy

#>   loading cached result from previous rnorm call, adding to memoised copy

ranNumsC <- rnorm(10, 16) %>% Cache(cacheRepo = tmpDir) # recovers cached copy

#>   loading memoised result from previous 'rnorm' pipe sequence call.

ranNumsD <- Cache(quote(rnorm(n = 10, 16)), cacheRepo = tmpDir) # recovers cached copy

#>   loading memoised result from previous rnorm call.

# Any minor change makes it different
ranNumsE <- rnorm(10, 6) %>% Cache(cacheRepo = tmpDir) # different

## Example 1: basic cache use with tags
ranNumsA <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(runif, 4, cacheRepo = tmpDir, userTags = "objectName:b")
ranNumsC <- Cache(runif, 40, cacheRepo = tmpDir, userTags = "objectName:b")

showCache(tmpDir, userTags = c("objectName"))

#> Cache size: 

#>   Total (including Rasters): 21.1 Kb

#>   Selected objects (not including Rasters): 729 bytes

#>                             artifact         tagKey
#>  1: 1d6a5c43763d988ceae50569492c03f5         format
#>  2: 1d6a5c43763d988ceae50569492c03f5           name
#>  3: 1d6a5c43763d988ceae50569492c03f5          class
#>  4: 1d6a5c43763d988ceae50569492c03f5           date
#>  5: 1d6a5c43763d988ceae50569492c03f5        cacheId
#>  6: 1d6a5c43763d988ceae50569492c03f5     objectName
#>  7: 1d6a5c43763d988ceae50569492c03f5       function
#>  8: 1d6a5c43763d988ceae50569492c03f5    object.size
#>  9: 1d6a5c43763d988ceae50569492c03f5       accessed
#> 10: 1d6a5c43763d988ceae50569492c03f5 otherFunctions
#> 11: 1d6a5c43763d988ceae50569492c03f5 otherFunctions
#> 12: 1d6a5c43763d988ceae50569492c03f5 otherFunctions
#> 13: 1d6a5c43763d988ceae50569492c03f5 otherFunctions
#> 14: 1d6a5c43763d988ceae50569492c03f5 otherFunctions
#> 15: 1d6a5c43763d988ceae50569492c03f5 otherFunctions
#> 16: 1d6a5c43763d988ceae50569492c03f5 otherFunctions
#> 17: 1d6a5c43763d988ceae50569492c03f5      preDigest
#> 18: 1d6a5c43763d988ceae50569492c03f5      preDigest
#> 19: 501a49e75a0d174c82eebd7b3f804972         format
#> 20: 501a49e75a0d174c82eebd7b3f804972           name
#> 21: 501a49e75a0d174c82eebd7b3f804972          class
#> 22: 501a49e75a0d174c82eebd7b3f804972           date
#> 23: 501a49e75a0d174c82eebd7b3f804972        cacheId
#> 24: 501a49e75a0d174c82eebd7b3f804972     objectName
#> 25: 501a49e75a0d174c82eebd7b3f804972       function
#> 26: 501a49e75a0d174c82eebd7b3f804972    object.size
#> 27: 501a49e75a0d174c82eebd7b3f804972       accessed
#> 28: 501a49e75a0d174c82eebd7b3f804972 otherFunctions
#> 29: 501a49e75a0d174c82eebd7b3f804972 otherFunctions
#> 30: 501a49e75a0d174c82eebd7b3f804972 otherFunctions
#> 31: 501a49e75a0d174c82eebd7b3f804972 otherFunctions
#> 32: 501a49e75a0d174c82eebd7b3f804972 otherFunctions
#> 33: 501a49e75a0d174c82eebd7b3f804972 otherFunctions
#> 34: 501a49e75a0d174c82eebd7b3f804972 otherFunctions
#> 35: 501a49e75a0d174c82eebd7b3f804972      preDigest
#> 36: 501a49e75a0d174c82eebd7b3f804972      preDigest
#> 37: 8affdfbd0438a389e0509a3ce1e7db34         format
#> 38: 8affdfbd0438a389e0509a3ce1e7db34           name
#> 39: 8affdfbd0438a389e0509a3ce1e7db34          class
#> 40: 8affdfbd0438a389e0509a3ce1e7db34           date
#> 41: 8affdfbd0438a389e0509a3ce1e7db34        cacheId
#> 42: 8affdfbd0438a389e0509a3ce1e7db34     objectName
#> 43: 8affdfbd0438a389e0509a3ce1e7db34       function
#> 44: 8affdfbd0438a389e0509a3ce1e7db34    object.size
#> 45: 8affdfbd0438a389e0509a3ce1e7db34       accessed
#> 46: 8affdfbd0438a389e0509a3ce1e7db34 otherFunctions
#> 47: 8affdfbd0438a389e0509a3ce1e7db34 otherFunctions
#> 48: 8affdfbd0438a389e0509a3ce1e7db34 otherFunctions
#> 49: 8affdfbd0438a389e0509a3ce1e7db34 otherFunctions
#> 50: 8affdfbd0438a389e0509a3ce1e7db34 otherFunctions
#> 51: 8affdfbd0438a389e0509a3ce1e7db34 otherFunctions
#> 52: 8affdfbd0438a389e0509a3ce1e7db34 otherFunctions
#> 53: 8affdfbd0438a389e0509a3ce1e7db34      preDigest
#> 54: 8affdfbd0438a389e0509a3ce1e7db34      preDigest
#>                             artifact         tagKey
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-06-12 11:34:25
#>  2:                                 Cache 2018-06-12 11:34:25
#>  3:                               numeric 2018-06-12 11:34:25
#>  4:                   2018-06-12 11:34:25 2018-06-12 11:34:25
#>  5:      b419232b60eb6eedb59532cf0b592010 2018-06-12 11:34:25
#>  6:                                     b 2018-06-12 11:34:25
#>  7:                                 runif 2018-06-12 11:34:25
#>  8:                                   976 2018-06-12 11:34:25
#>  9:                   2018-06-12 11:34:25 2018-06-12 11:34:25
#> 10:                   withCallingHandlers 2018-06-12 11:34:25
#> 11:                               saveRDS 2018-06-12 11:34:25
#> 12:                               do.call 2018-06-12 11:34:25
#> 13:                      build_site_local 2018-06-12 11:34:25
#> 14:                       build_reference 2018-06-12 11:34:25
#> 15:                  data_reference_topic 2018-06-12 11:34:25
#> 16:                                   try 2018-06-12 11:34:25
#> 17:    n:9d7bd914b9878238bccd5686a5a41903 2018-06-12 11:34:25
#> 18: .FUN:d2631d24c3b38b89c7bdd4ab7faaaac3 2018-06-12 11:34:25
#> 19:                                   rda 2018-06-12 11:34:24
#> 20:                                 Cache 2018-06-12 11:34:24
#> 21:                               numeric 2018-06-12 11:34:24
#> 22:                   2018-06-12 11:34:24 2018-06-12 11:34:24
#> 23:      e37bb635c97bc2eeecab63816b881bbc 2018-06-12 11:34:24
#> 24:                                     b 2018-06-12 11:34:25
#> 25:                                 runif 2018-06-12 11:34:25
#> 26:                                   688 2018-06-12 11:34:25
#> 27:                   2018-06-12 11:34:24 2018-06-12 11:34:25
#> 28:                   withCallingHandlers 2018-06-12 11:34:25
#> 29:                               saveRDS 2018-06-12 11:34:25
#> 30:                               do.call 2018-06-12 11:34:25
#> 31:                      build_site_local 2018-06-12 11:34:25
#> 32:                       build_reference 2018-06-12 11:34:25
#> 33:                  data_reference_topic 2018-06-12 11:34:25
#> 34:                                   try 2018-06-12 11:34:25
#> 35:    n:969a49ec15bcd4323ff31538af321264 2018-06-12 11:34:25
#> 36: .FUN:d2631d24c3b38b89c7bdd4ab7faaaac3 2018-06-12 11:34:25
#> 37:                                   rda 2018-06-12 11:34:24
#> 38:                                 Cache 2018-06-12 11:34:24
#> 39:                               numeric 2018-06-12 11:34:24
#> 40:                   2018-06-12 11:34:24 2018-06-12 11:34:24
#> 41:      85874f26b2e0c1ef689a7d379d275ebf 2018-06-12 11:34:24
#> 42:                                     a 2018-06-12 11:34:24
#> 43:                                 rnorm 2018-06-12 11:34:24
#> 44:                                   688 2018-06-12 11:34:24
#> 45:                   2018-06-12 11:34:24 2018-06-12 11:34:24
#> 46:                   withCallingHandlers 2018-06-12 11:34:24
#> 47:                               saveRDS 2018-06-12 11:34:24
#> 48:                               do.call 2018-06-12 11:34:24
#> 49:                      build_site_local 2018-06-12 11:34:24
#> 50:                       build_reference 2018-06-12 11:34:24
#> 51:                  data_reference_topic 2018-06-12 11:34:24
#> 52:                                   try 2018-06-12 11:34:24
#> 53:    n:969a49ec15bcd4323ff31538af321264 2018-06-12 11:34:24
#> 54: .FUN:7e9a928f110f80b3612e71883a6ec1f4 2018-06-12 11:34:24
#>                                  tagValue         createdDate

showCache(tmpDir, userTags = c("^a$")) # regular expression ... "a" exactly

#> Cache size: 

#>   Total (including Rasters): 21.1 Kb

#>   Selected objects (not including Rasters): 172 bytes

#>                             artifact         tagKey
#>  1: 8affdfbd0438a389e0509a3ce1e7db34         format
#>  2: 8affdfbd0438a389e0509a3ce1e7db34           name
#>  3: 8affdfbd0438a389e0509a3ce1e7db34          class
#>  4: 8affdfbd0438a389e0509a3ce1e7db34           date
#>  5: 8affdfbd0438a389e0509a3ce1e7db34        cacheId
#>  6: 8affdfbd0438a389e0509a3ce1e7db34     objectName
#>  7: 8affdfbd0438a389e0509a3ce1e7db34       function
#>  8: 8affdfbd0438a389e0509a3ce1e7db34    object.size
#>  9: 8affdfbd0438a389e0509a3ce1e7db34       accessed
#> 10: 8affdfbd0438a389e0509a3ce1e7db34 otherFunctions
#> 11: 8affdfbd0438a389e0509a3ce1e7db34 otherFunctions
#> 12: 8affdfbd0438a389e0509a3ce1e7db34 otherFunctions
#> 13: 8affdfbd0438a389e0509a3ce1e7db34 otherFunctions
#> 14: 8affdfbd0438a389e0509a3ce1e7db34 otherFunctions
#> 15: 8affdfbd0438a389e0509a3ce1e7db34 otherFunctions
#> 16: 8affdfbd0438a389e0509a3ce1e7db34 otherFunctions
#> 17: 8affdfbd0438a389e0509a3ce1e7db34      preDigest
#> 18: 8affdfbd0438a389e0509a3ce1e7db34      preDigest
#>                                  tagValue         createdDate
#>  1:                                   rda 2018-06-12 11:34:24
#>  2:                                 Cache 2018-06-12 11:34:24
#>  3:                               numeric 2018-06-12 11:34:24
#>  4:                   2018-06-12 11:34:24 2018-06-12 11:34:24
#>  5:      85874f26b2e0c1ef689a7d379d275ebf 2018-06-12 11:34:24
#>  6:                                     a 2018-06-12 11:34:24
#>  7:                                 rnorm 2018-06-12 11:34:24
#>  8:                                   688 2018-06-12 11:34:24
#>  9:                   2018-06-12 11:34:24 2018-06-12 11:34:24
#> 10:                   withCallingHandlers 2018-06-12 11:34:24
#> 11:                               saveRDS 2018-06-12 11:34:24
#> 12:                               do.call 2018-06-12 11:34:24
#> 13:                      build_site_local 2018-06-12 11:34:24
#> 14:                       build_reference 2018-06-12 11:34:24
#> 15:                  data_reference_topic 2018-06-12 11:34:24
#> 16:                                   try 2018-06-12 11:34:24
#> 17:    n:969a49ec15bcd4323ff31538af321264 2018-06-12 11:34:24
#> 18: .FUN:7e9a928f110f80b3612e71883a6ec1f4 2018-06-12 11:34:24

# Fine control of cache elements -- pick out only the large runif object, and remove it
cache1 <- showCache(tmpDir, userTags = c("runif")) # show only cached objects made during runif

#> Cache size: 

#>   Total (including Rasters): 21.1 Kb

#>   Selected objects (not including Rasters): 557 bytes

toRemove <- cache1[tagKey=="object.size"][as.numeric(tagValue) > 700]$artifact
clearCache(tmpDir, userTags = toRemove)

#> Cache size: 

#>   Total (including Rasters): 21.1 Kb

#>   Selected objects (not including Rasters): 391 bytes

cacheAfter <- showCache(tmpDir, userTags = c("runif")) # Only the small one is left

#> Cache size: 

#>   Total (including Rasters): 20.8 Kb

#>   Selected objects (not including Rasters): 166 bytes