Examining and modifying the cache

These are convenience wrappers around archivist package functions. They allow the user a bit of control over what is being cached.

clearCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), ...)

# S4 method for ANY
clearCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), ...)

cc(secs, ...)

showCache(x, userTags = character(), after, before, ...)

# S4 method for ANY
showCache(x, userTags = character(), after, before, ...)

keepCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), ...)

# S4 method for ANY
keepCache(x, userTags = character(), after, before,
  ask = getOption("reproducible.ask"), ...)

Arguments

x
A simList or a directory containing a valid archivist repository. Note: For compatibility with Cache argument, cacheRepo can also be used instead of x, though x will take precedence.

userTags
Character vector. If used, this will be used in place of the after and before. Specifying one or more userTag here will clear all objects that match those tags. Matching is via regular expression, meaning partial matches will work unless strict beginning (^) and end ($) of string characters are used. Matching will be against any of the 3 columns returned by showCache(), i.e., artifact, tagValue or tagName. Also, length userTags > 1, then matching is by `and`. For `or` matching, use | in a single character string. See examples.

If neither after or before are provided, nor userTags, then all objects will be removed. If both after and before are specified, then all objects between after and before will be deleted. If userTags is used, this will override after or before.

after
A time (POSIX, character understandable by data.table). Objects cached after this time will be shown or deleted.

before
A time (POSIX, character understandable by data.table). Objects cached before this time will be shown or deleted.

ask
Logical. If FALSE, then it will not ask to confirm deletions using clearCache or keepCache. Default is TRUE

...
Other arguments. Currently, regexp, a logical, can be provided. This must be TRUE if the use is passing a regular expression. Otherwise, userTags will need to be exact matches. Default is missing, which is the same as TRUE. If there are errors due to regular expression problem, try FALSE. For cc, it is passed to clearCache, e.g., ask, userTags

secs
Currently 3 options: the number of seconds to pass to clearCache(after = secs), a POSIXct time e.g., from Sys.time(), or missing. If missing, the default, then it will delete the most recent entry in the Cache.

Value

Will clear all objects (or those that match userTags, or those between after or before) from the repository located at cachePath of the sim object, if sim is provided, or located in cacheRepo. Invisibly returns a data.table of the removed items.

Details

cc(secs) is just a shortcut for clearCache(repo = Paths$cachePath, after = secs), i.e., to remove any cache entries touched in the last secs seconds.

clearCache

remove items from the cache based on their userTag or times values.

keepCache

remove all cached items except those based on certain userTags or times values.

showCache

display the contents of the cache.

Note

If the cache is larger than 10MB, and clearCache is used, there will be a message and a pause, if interactive, to prevent accidentally deleting of a large cache repository.

See also

Examples

library(raster)

tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir, ask = FALSE), silent = TRUE) # just to make sure it is clear

# Basic use
ranNumsA <- Cache(rnorm, 10, 16, cacheRepo = tmpDir)

# All same
ranNumsB <- Cache(rnorm, 10, 16, cacheRepo = tmpDir) # recovers cached copy

#>   loading cached result from previous rnorm call, adding to memoised copy

ranNumsC <- Cache(cacheRepo = tmpDir) %C% rnorm(10, 16)  # recovers cached copy

#>   loading memoised result from previous 'rnorm' pipe sequence call.

ranNumsD <- Cache(quote(rnorm(n = 10, 16)), cacheRepo = tmpDir) # recovers cached copy

#>   loading memoised result from previous rnorm call.

# Any minor change makes it different
ranNumsE <- Cache(cacheRepo = tmpDir) %C% rnorm(10, 6)# different

## Example 1: basic cache use with tags
ranNumsA <- Cache(rnorm, 4, cacheRepo = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(runif, 4, cacheRepo = tmpDir, userTags = "objectName:b")
ranNumsC <- Cache(runif, 40, cacheRepo = tmpDir, userTags = "objectName:b")

showCache(tmpDir, userTags = c("objectName"))

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 828 bytes

#>                             artifact         tagKey
#>  1: 397b2f22818569361d663361be453f41         format
#>  2: 397b2f22818569361d663361be453f41           name
#>  3: 397b2f22818569361d663361be453f41          class
#>  4: 397b2f22818569361d663361be453f41           date
#>  5: 397b2f22818569361d663361be453f41        cacheId
#>  6: 397b2f22818569361d663361be453f41     objectName
#>  7: 397b2f22818569361d663361be453f41       function
#>  8: 397b2f22818569361d663361be453f41    object.size
#>  9: 397b2f22818569361d663361be453f41       accessed
#> 10: 397b2f22818569361d663361be453f41 otherFunctions
#> 11: 397b2f22818569361d663361be453f41 otherFunctions
#> 12: 397b2f22818569361d663361be453f41 otherFunctions
#> 13: 397b2f22818569361d663361be453f41 otherFunctions
#> 14: 397b2f22818569361d663361be453f41 otherFunctions
#> 15: 397b2f22818569361d663361be453f41 otherFunctions
#> 16: 397b2f22818569361d663361be453f41 otherFunctions
#> 17: 397b2f22818569361d663361be453f41 otherFunctions
#> 18: 397b2f22818569361d663361be453f41 otherFunctions
#> 19: 397b2f22818569361d663361be453f41 otherFunctions
#> 20: 397b2f22818569361d663361be453f41 otherFunctions
#> 21: 397b2f22818569361d663361be453f41      preDigest
#> 22: 397b2f22818569361d663361be453f41      preDigest
#> 23: 5d458c3e876c33326b67b72a83b78080         format
#> 24: 5d458c3e876c33326b67b72a83b78080           name
#> 25: 5d458c3e876c33326b67b72a83b78080          class
#> 26: 5d458c3e876c33326b67b72a83b78080           date
#> 27: 5d458c3e876c33326b67b72a83b78080        cacheId
#> 28: 5d458c3e876c33326b67b72a83b78080     objectName
#> 29: 5d458c3e876c33326b67b72a83b78080       function
#> 30: 5d458c3e876c33326b67b72a83b78080    object.size
#> 31: 5d458c3e876c33326b67b72a83b78080       accessed
#> 32: 5d458c3e876c33326b67b72a83b78080 otherFunctions
#> 33: 5d458c3e876c33326b67b72a83b78080 otherFunctions
#> 34: 5d458c3e876c33326b67b72a83b78080 otherFunctions
#> 35: 5d458c3e876c33326b67b72a83b78080 otherFunctions
#> 36: 5d458c3e876c33326b67b72a83b78080 otherFunctions
#> 37: 5d458c3e876c33326b67b72a83b78080 otherFunctions
#> 38: 5d458c3e876c33326b67b72a83b78080 otherFunctions
#> 39: 5d458c3e876c33326b67b72a83b78080 otherFunctions
#> 40: 5d458c3e876c33326b67b72a83b78080 otherFunctions
#> 41: 5d458c3e876c33326b67b72a83b78080 otherFunctions
#> 42: 5d458c3e876c33326b67b72a83b78080 otherFunctions
#> 43: 5d458c3e876c33326b67b72a83b78080      preDigest
#> 44: 5d458c3e876c33326b67b72a83b78080      preDigest
#> 45: e474c2fe676bb247d8f67ca15d705a59         format
#> 46: e474c2fe676bb247d8f67ca15d705a59           name
#> 47: e474c2fe676bb247d8f67ca15d705a59          class
#> 48: e474c2fe676bb247d8f67ca15d705a59           date
#> 49: e474c2fe676bb247d8f67ca15d705a59        cacheId
#> 50: e474c2fe676bb247d8f67ca15d705a59     objectName
#> 51: e474c2fe676bb247d8f67ca15d705a59       function
#> 52: e474c2fe676bb247d8f67ca15d705a59    object.size
#> 53: e474c2fe676bb247d8f67ca15d705a59       accessed
#> 54: e474c2fe676bb247d8f67ca15d705a59 otherFunctions
#> 55: e474c2fe676bb247d8f67ca15d705a59 otherFunctions
#> 56: e474c2fe676bb247d8f67ca15d705a59 otherFunctions
#> 57: e474c2fe676bb247d8f67ca15d705a59 otherFunctions
#> 58: e474c2fe676bb247d8f67ca15d705a59 otherFunctions
#> 59: e474c2fe676bb247d8f67ca15d705a59 otherFunctions
#> 60: e474c2fe676bb247d8f67ca15d705a59 otherFunctions
#> 61: e474c2fe676bb247d8f67ca15d705a59 otherFunctions
#> 62: e474c2fe676bb247d8f67ca15d705a59 otherFunctions
#> 63: e474c2fe676bb247d8f67ca15d705a59 otherFunctions
#> 64: e474c2fe676bb247d8f67ca15d705a59 otherFunctions
#> 65: e474c2fe676bb247d8f67ca15d705a59      preDigest
#> 66: e474c2fe676bb247d8f67ca15d705a59      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-03-06 07:42:51
#>  2: 397b2f22818569361d663361be453f41 2019-03-06 07:42:51
#>  3:                          numeric 2019-03-06 07:42:51
#>  4:              2019-03-06 07:42:51 2019-03-06 07:42:51
#>  5:                 f7bee22047b8d0c1 2019-03-06 07:42:51
#>  6:                                a 2019-03-06 07:42:51
#>  7:                            rnorm 2019-03-06 07:42:51
#>  8:                             1008 2019-03-06 07:42:51
#>  9:              2019-03-06 07:42:51 2019-03-06 07:42:51
#> 10:                          saveRDS 2019-03-06 07:42:51
#> 11:                          do.call 2019-03-06 07:42:51
#> 12:                 build_site_local 2019-03-06 07:42:51
#> 13:                  build_reference 2019-03-06 07:42:51
#> 14:             data_reference_topic 2019-03-06 07:42:51
#> 15:                          as_data 2019-03-06 07:42:51
#> 16:             as_data.tag_examples 2019-03-06 07:42:51
#> 17:                        timing_fn 2019-03-06 07:42:51
#> 18:                           handle 2019-03-06 07:42:51
#> 19:                              try 2019-03-06 07:42:51
#> 20:                      withVisible 2019-03-06 07:42:51
#> 21:               n:7eef4eae85fd9229 2019-03-06 07:42:51
#> 22:            .FUN:4f604aa46882b368 2019-03-06 07:42:51
#> 23:                              rda 2019-03-06 07:42:51
#> 24: 5d458c3e876c33326b67b72a83b78080 2019-03-06 07:42:51
#> 25:                          numeric 2019-03-06 07:42:51
#> 26:              2019-03-06 07:42:51 2019-03-06 07:42:51
#> 27:                 3aef38d1fc02aee5 2019-03-06 07:42:51
#> 28:                                b 2019-03-06 07:42:51
#> 29:                            runif 2019-03-06 07:42:51
#> 30:                             1008 2019-03-06 07:42:51
#> 31:              2019-03-06 07:42:51 2019-03-06 07:42:51
#> 32:                          saveRDS 2019-03-06 07:42:51
#> 33:                          do.call 2019-03-06 07:42:51
#> 34:                 build_site_local 2019-03-06 07:42:51
#> 35:                  build_reference 2019-03-06 07:42:51
#> 36:             data_reference_topic 2019-03-06 07:42:51
#> 37:                          as_data 2019-03-06 07:42:51
#> 38:             as_data.tag_examples 2019-03-06 07:42:51
#> 39:                        timing_fn 2019-03-06 07:42:51
#> 40:                           handle 2019-03-06 07:42:51
#> 41:                              try 2019-03-06 07:42:51
#> 42:                      withVisible 2019-03-06 07:42:51
#> 43:               n:7eef4eae85fd9229 2019-03-06 07:42:51
#> 44:            .FUN:881ec847b7161f3c 2019-03-06 07:42:51
#> 45:                              rda 2019-03-06 07:42:51
#> 46: e474c2fe676bb247d8f67ca15d705a59 2019-03-06 07:42:51
#> 47:                          numeric 2019-03-06 07:42:51
#> 48:              2019-03-06 07:42:51 2019-03-06 07:42:51
#> 49:                 7225b47fe5dc51d0 2019-03-06 07:42:51
#> 50:                                b 2019-03-06 07:42:51
#> 51:                            runif 2019-03-06 07:42:51
#> 52:                             1296 2019-03-06 07:42:51
#> 53:              2019-03-06 07:42:51 2019-03-06 07:42:51
#> 54:                          saveRDS 2019-03-06 07:42:51
#> 55:                          do.call 2019-03-06 07:42:51
#> 56:                 build_site_local 2019-03-06 07:42:51
#> 57:                  build_reference 2019-03-06 07:42:51
#> 58:             data_reference_topic 2019-03-06 07:42:51
#> 59:                          as_data 2019-03-06 07:42:51
#> 60:             as_data.tag_examples 2019-03-06 07:42:51
#> 61:                        timing_fn 2019-03-06 07:42:51
#> 62:                           handle 2019-03-06 07:42:51
#> 63:                              try 2019-03-06 07:42:51
#> 64:                      withVisible 2019-03-06 07:42:51
#> 65:               n:1393aef18608c8be 2019-03-06 07:42:51
#> 66:            .FUN:881ec847b7161f3c 2019-03-06 07:42:51
#>                             tagValue         createdDate

showCache(tmpDir, userTags = c("^a$")) # regular expression ... "a" exactly

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 252 bytes

#>                             artifact         tagKey
#>  1: 397b2f22818569361d663361be453f41         format
#>  2: 397b2f22818569361d663361be453f41           name
#>  3: 397b2f22818569361d663361be453f41          class
#>  4: 397b2f22818569361d663361be453f41           date
#>  5: 397b2f22818569361d663361be453f41        cacheId
#>  6: 397b2f22818569361d663361be453f41     objectName
#>  7: 397b2f22818569361d663361be453f41       function
#>  8: 397b2f22818569361d663361be453f41    object.size
#>  9: 397b2f22818569361d663361be453f41       accessed
#> 10: 397b2f22818569361d663361be453f41 otherFunctions
#> 11: 397b2f22818569361d663361be453f41 otherFunctions
#> 12: 397b2f22818569361d663361be453f41 otherFunctions
#> 13: 397b2f22818569361d663361be453f41 otherFunctions
#> 14: 397b2f22818569361d663361be453f41 otherFunctions
#> 15: 397b2f22818569361d663361be453f41 otherFunctions
#> 16: 397b2f22818569361d663361be453f41 otherFunctions
#> 17: 397b2f22818569361d663361be453f41 otherFunctions
#> 18: 397b2f22818569361d663361be453f41 otherFunctions
#> 19: 397b2f22818569361d663361be453f41 otherFunctions
#> 20: 397b2f22818569361d663361be453f41 otherFunctions
#> 21: 397b2f22818569361d663361be453f41      preDigest
#> 22: 397b2f22818569361d663361be453f41      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-03-06 07:42:51
#>  2: 397b2f22818569361d663361be453f41 2019-03-06 07:42:51
#>  3:                          numeric 2019-03-06 07:42:51
#>  4:              2019-03-06 07:42:51 2019-03-06 07:42:51
#>  5:                 f7bee22047b8d0c1 2019-03-06 07:42:51
#>  6:                                a 2019-03-06 07:42:51
#>  7:                            rnorm 2019-03-06 07:42:51
#>  8:                             1008 2019-03-06 07:42:51
#>  9:              2019-03-06 07:42:51 2019-03-06 07:42:51
#> 10:                          saveRDS 2019-03-06 07:42:51
#> 11:                          do.call 2019-03-06 07:42:51
#> 12:                 build_site_local 2019-03-06 07:42:51
#> 13:                  build_reference 2019-03-06 07:42:51
#> 14:             data_reference_topic 2019-03-06 07:42:51
#> 15:                          as_data 2019-03-06 07:42:51
#> 16:             as_data.tag_examples 2019-03-06 07:42:51
#> 17:                        timing_fn 2019-03-06 07:42:51
#> 18:                           handle 2019-03-06 07:42:51
#> 19:                              try 2019-03-06 07:42:51
#> 20:                      withVisible 2019-03-06 07:42:51
#> 21:               n:7eef4eae85fd9229 2019-03-06 07:42:51
#> 22:            .FUN:4f604aa46882b368 2019-03-06 07:42:51
#>                             tagValue         createdDate

# Fine control of cache elements -- pick out only the large runif object, and remove it
cache1 <- showCache(tmpDir, userTags = c("runif")) # show only cached objects made during runif

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 576 bytes

toRemove <- cache1[tagKey=="object.size"][as.numeric(tagValue) > 700]$artifact
clearCache(tmpDir, userTags = toRemove, ask = FALSE)

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 0 bytes

cacheAfter <- showCache(tmpDir, userTags = c("runif")) # Only the small one is left

#> Cache size: 

#>   Total (including Rasters): 1.3 Kb

#>   Selected objects (not including Rasters): 576 bytes

tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir, ask = FALSE), silent = TRUE) # just to make sure it is clear

Cache(rnorm, 1, cacheRepo = tmpDir)

#> [1] -1.4787
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:7072c305d8c69df0"
#> attr(,"call")
#> [1] ""

thisTime <- Sys.time()
Cache(rnorm, 2, cacheRepo = tmpDir)

#> [1] -1.5515602  0.7775067
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:3b347dee8f6305c7"
#> attr(,"call")
#> [1] ""

Cache(rnorm, 3, cacheRepo = tmpDir)

#> [1]  1.0684401 -0.1835877  1.5582429
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:f0da91a44b839434"
#> attr(,"call")
#> [1] ""

Cache(rnorm, 4, cacheRepo = tmpDir)

#> [1] -0.2132424  0.9305353  0.4108118 -1.2798443
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:f7bee22047b8d0c1"
#> attr(,"call")
#> [1] ""

showCache(x = tmpDir) # shows all 4 entries

#> Cache size: 

#>   Total (including Rasters): 998 bytes

#>   Selected objects (not including Rasters): 998 bytes

#>                             artifact         tagKey
#>  1: 06b0f8f5af2781d3f74786fb797dd985         format
#>  2: 06b0f8f5af2781d3f74786fb797dd985           name
#>  3: 06b0f8f5af2781d3f74786fb797dd985          class
#>  4: 06b0f8f5af2781d3f74786fb797dd985           date
#>  5: 06b0f8f5af2781d3f74786fb797dd985        cacheId
#>  6: 06b0f8f5af2781d3f74786fb797dd985       function
#>  7: 06b0f8f5af2781d3f74786fb797dd985    object.size
#>  8: 06b0f8f5af2781d3f74786fb797dd985       accessed
#>  9: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 10: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 11: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 12: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 13: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 14: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 15: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 16: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 17: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 18: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 19: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 20: 06b0f8f5af2781d3f74786fb797dd985      preDigest
#> 21: 06b0f8f5af2781d3f74786fb797dd985      preDigest
#> 22: 2c8492f352933627e962e697da788a62         format
#> 23: 2c8492f352933627e962e697da788a62           name
#> 24: 2c8492f352933627e962e697da788a62          class
#> 25: 2c8492f352933627e962e697da788a62           date
#> 26: 2c8492f352933627e962e697da788a62        cacheId
#> 27: 2c8492f352933627e962e697da788a62       function
#> 28: 2c8492f352933627e962e697da788a62    object.size
#> 29: 2c8492f352933627e962e697da788a62       accessed
#> 30: 2c8492f352933627e962e697da788a62 otherFunctions
#> 31: 2c8492f352933627e962e697da788a62 otherFunctions
#> 32: 2c8492f352933627e962e697da788a62 otherFunctions
#> 33: 2c8492f352933627e962e697da788a62 otherFunctions
#> 34: 2c8492f352933627e962e697da788a62 otherFunctions
#> 35: 2c8492f352933627e962e697da788a62 otherFunctions
#> 36: 2c8492f352933627e962e697da788a62 otherFunctions
#> 37: 2c8492f352933627e962e697da788a62 otherFunctions
#> 38: 2c8492f352933627e962e697da788a62 otherFunctions
#> 39: 2c8492f352933627e962e697da788a62 otherFunctions
#> 40: 2c8492f352933627e962e697da788a62 otherFunctions
#> 41: 2c8492f352933627e962e697da788a62      preDigest
#> 42: 2c8492f352933627e962e697da788a62      preDigest
#> 43: b0db9a12c38493c2c7c8622690b769a5         format
#> 44: b0db9a12c38493c2c7c8622690b769a5           name
#> 45: b0db9a12c38493c2c7c8622690b769a5          class
#> 46: b0db9a12c38493c2c7c8622690b769a5           date
#> 47: b0db9a12c38493c2c7c8622690b769a5        cacheId
#> 48: b0db9a12c38493c2c7c8622690b769a5       function
#> 49: b0db9a12c38493c2c7c8622690b769a5    object.size
#> 50: b0db9a12c38493c2c7c8622690b769a5       accessed
#> 51: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 52: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 53: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 54: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 55: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 56: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 57: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 58: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 59: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 60: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 61: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 62: b0db9a12c38493c2c7c8622690b769a5      preDigest
#> 63: b0db9a12c38493c2c7c8622690b769a5      preDigest
#> 64: f05a83fa21e025e7228c4b217589d260         format
#> 65: f05a83fa21e025e7228c4b217589d260           name
#> 66: f05a83fa21e025e7228c4b217589d260          class
#> 67: f05a83fa21e025e7228c4b217589d260           date
#> 68: f05a83fa21e025e7228c4b217589d260        cacheId
#> 69: f05a83fa21e025e7228c4b217589d260       function
#> 70: f05a83fa21e025e7228c4b217589d260    object.size
#> 71: f05a83fa21e025e7228c4b217589d260       accessed
#> 72: f05a83fa21e025e7228c4b217589d260 otherFunctions
#> 73: f05a83fa21e025e7228c4b217589d260 otherFunctions
#> 74: f05a83fa21e025e7228c4b217589d260 otherFunctions
#> 75: f05a83fa21e025e7228c4b217589d260 otherFunctions
#> 76: f05a83fa21e025e7228c4b217589d260 otherFunctions
#> 77: f05a83fa21e025e7228c4b217589d260 otherFunctions
#> 78: f05a83fa21e025e7228c4b217589d260 otherFunctions
#> 79: f05a83fa21e025e7228c4b217589d260 otherFunctions
#> 80: f05a83fa21e025e7228c4b217589d260 otherFunctions
#> 81: f05a83fa21e025e7228c4b217589d260 otherFunctions
#> 82: f05a83fa21e025e7228c4b217589d260 otherFunctions
#> 83: f05a83fa21e025e7228c4b217589d260      preDigest
#> 84: f05a83fa21e025e7228c4b217589d260      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-03-06 07:42:52
#>  2: 06b0f8f5af2781d3f74786fb797dd985 2019-03-06 07:42:52
#>  3:                          numeric 2019-03-06 07:42:52
#>  4:              2019-03-06 07:42:52 2019-03-06 07:42:52
#>  5:                 f7bee22047b8d0c1 2019-03-06 07:42:52
#>  6:                            rnorm 2019-03-06 07:42:52
#>  7:                             1008 2019-03-06 07:42:52
#>  8:              2019-03-06 07:42:52 2019-03-06 07:42:52
#>  9:                          saveRDS 2019-03-06 07:42:52
#> 10:                          do.call 2019-03-06 07:42:52
#> 11:                 build_site_local 2019-03-06 07:42:52
#> 12:                  build_reference 2019-03-06 07:42:52
#> 13:             data_reference_topic 2019-03-06 07:42:52
#> 14:                          as_data 2019-03-06 07:42:52
#> 15:             as_data.tag_examples 2019-03-06 07:42:52
#> 16:                        timing_fn 2019-03-06 07:42:52
#> 17:                           handle 2019-03-06 07:42:52
#> 18:                              try 2019-03-06 07:42:52
#> 19:                      withVisible 2019-03-06 07:42:52
#> 20:               n:7eef4eae85fd9229 2019-03-06 07:42:52
#> 21:            .FUN:4f604aa46882b368 2019-03-06 07:42:52
#> 22:                              rda 2019-03-06 07:42:52
#> 23: 2c8492f352933627e962e697da788a62 2019-03-06 07:42:52
#> 24:                          numeric 2019-03-06 07:42:52
#> 25:              2019-03-06 07:42:52 2019-03-06 07:42:52
#> 26:                 3b347dee8f6305c7 2019-03-06 07:42:52
#> 27:                            rnorm 2019-03-06 07:42:52
#> 28:                              992 2019-03-06 07:42:52
#> 29:              2019-03-06 07:42:52 2019-03-06 07:42:52
#> 30:                          saveRDS 2019-03-06 07:42:52
#> 31:                          do.call 2019-03-06 07:42:52
#> 32:                 build_site_local 2019-03-06 07:42:52
#> 33:                  build_reference 2019-03-06 07:42:52
#> 34:             data_reference_topic 2019-03-06 07:42:52
#> 35:                          as_data 2019-03-06 07:42:52
#> 36:             as_data.tag_examples 2019-03-06 07:42:52
#> 37:                        timing_fn 2019-03-06 07:42:52
#> 38:                           handle 2019-03-06 07:42:52
#> 39:                              try 2019-03-06 07:42:52
#> 40:                      withVisible 2019-03-06 07:42:52
#> 41:               n:82dc709f2b91918a 2019-03-06 07:42:52
#> 42:            .FUN:4f604aa46882b368 2019-03-06 07:42:52
#> 43:                              rda 2019-03-06 07:42:52
#> 44: b0db9a12c38493c2c7c8622690b769a5 2019-03-06 07:42:52
#> 45:                          numeric 2019-03-06 07:42:52
#> 46:              2019-03-06 07:42:52 2019-03-06 07:42:52
#> 47:                 f0da91a44b839434 2019-03-06 07:42:52
#> 48:                            rnorm 2019-03-06 07:42:52
#> 49:                             1008 2019-03-06 07:42:52
#> 50:              2019-03-06 07:42:52 2019-03-06 07:42:52
#> 51:                          saveRDS 2019-03-06 07:42:52
#> 52:                          do.call 2019-03-06 07:42:52
#> 53:                 build_site_local 2019-03-06 07:42:52
#> 54:                  build_reference 2019-03-06 07:42:52
#> 55:             data_reference_topic 2019-03-06 07:42:52
#> 56:                          as_data 2019-03-06 07:42:52
#> 57:             as_data.tag_examples 2019-03-06 07:42:52
#> 58:                        timing_fn 2019-03-06 07:42:52
#> 59:                           handle 2019-03-06 07:42:52
#> 60:                              try 2019-03-06 07:42:52
#> 61:                      withVisible 2019-03-06 07:42:52
#> 62:               n:7f12988bd88a0fb8 2019-03-06 07:42:52
#> 63:            .FUN:4f604aa46882b368 2019-03-06 07:42:52
#> 64:                              rda 2019-03-06 07:42:52
#> 65: f05a83fa21e025e7228c4b217589d260 2019-03-06 07:42:52
#> 66:                          numeric 2019-03-06 07:42:52
#> 67:              2019-03-06 07:42:52 2019-03-06 07:42:52
#> 68:                 7072c305d8c69df0 2019-03-06 07:42:52
#> 69:                            rnorm 2019-03-06 07:42:52
#> 70:                              984 2019-03-06 07:42:52
#> 71:              2019-03-06 07:42:52 2019-03-06 07:42:52
#> 72:                          saveRDS 2019-03-06 07:42:52
#> 73:                          do.call 2019-03-06 07:42:52
#> 74:                 build_site_local 2019-03-06 07:42:52
#> 75:                  build_reference 2019-03-06 07:42:52
#> 76:             data_reference_topic 2019-03-06 07:42:52
#> 77:                          as_data 2019-03-06 07:42:52
#> 78:             as_data.tag_examples 2019-03-06 07:42:52
#> 79:                        timing_fn 2019-03-06 07:42:52
#> 80:                           handle 2019-03-06 07:42:52
#> 81:                              try 2019-03-06 07:42:52
#> 82:                      withVisible 2019-03-06 07:42:52
#> 83:               n:853b1797f54b229c 2019-03-06 07:42:52
#> 84:            .FUN:4f604aa46882b368 2019-03-06 07:42:52
#>                             tagValue         createdDate

cc(ask = FALSE, x = tmpDir)

#> No time provided; removing the most recent entry to the Cache

#> Cache size: 

#>   Total (including Rasters): 998 bytes

#>   Selected objects (not including Rasters): 246 bytes

showCache(x = tmpDir) # most recent is gone

#> Cache size: 

#>   Total (including Rasters): 752 bytes

#>   Selected objects (not including Rasters): 752 bytes

#>                             artifact         tagKey
#>  1: 06b0f8f5af2781d3f74786fb797dd985         format
#>  2: 06b0f8f5af2781d3f74786fb797dd985           name
#>  3: 06b0f8f5af2781d3f74786fb797dd985          class
#>  4: 06b0f8f5af2781d3f74786fb797dd985           date
#>  5: 06b0f8f5af2781d3f74786fb797dd985        cacheId
#>  6: 06b0f8f5af2781d3f74786fb797dd985       function
#>  7: 06b0f8f5af2781d3f74786fb797dd985    object.size
#>  8: 06b0f8f5af2781d3f74786fb797dd985       accessed
#>  9: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 10: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 11: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 12: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 13: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 14: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 15: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 16: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 17: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 18: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 19: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 20: 06b0f8f5af2781d3f74786fb797dd985      preDigest
#> 21: 06b0f8f5af2781d3f74786fb797dd985      preDigest
#> 22: 2c8492f352933627e962e697da788a62         format
#> 23: 2c8492f352933627e962e697da788a62           name
#> 24: 2c8492f352933627e962e697da788a62          class
#> 25: 2c8492f352933627e962e697da788a62           date
#> 26: 2c8492f352933627e962e697da788a62        cacheId
#> 27: 2c8492f352933627e962e697da788a62       function
#> 28: 2c8492f352933627e962e697da788a62    object.size
#> 29: 2c8492f352933627e962e697da788a62       accessed
#> 30: 2c8492f352933627e962e697da788a62 otherFunctions
#> 31: 2c8492f352933627e962e697da788a62 otherFunctions
#> 32: 2c8492f352933627e962e697da788a62 otherFunctions
#> 33: 2c8492f352933627e962e697da788a62 otherFunctions
#> 34: 2c8492f352933627e962e697da788a62 otherFunctions
#> 35: 2c8492f352933627e962e697da788a62 otherFunctions
#> 36: 2c8492f352933627e962e697da788a62 otherFunctions
#> 37: 2c8492f352933627e962e697da788a62 otherFunctions
#> 38: 2c8492f352933627e962e697da788a62 otherFunctions
#> 39: 2c8492f352933627e962e697da788a62 otherFunctions
#> 40: 2c8492f352933627e962e697da788a62 otherFunctions
#> 41: 2c8492f352933627e962e697da788a62      preDigest
#> 42: 2c8492f352933627e962e697da788a62      preDigest
#> 43: b0db9a12c38493c2c7c8622690b769a5         format
#> 44: b0db9a12c38493c2c7c8622690b769a5           name
#> 45: b0db9a12c38493c2c7c8622690b769a5          class
#> 46: b0db9a12c38493c2c7c8622690b769a5           date
#> 47: b0db9a12c38493c2c7c8622690b769a5        cacheId
#> 48: b0db9a12c38493c2c7c8622690b769a5       function
#> 49: b0db9a12c38493c2c7c8622690b769a5    object.size
#> 50: b0db9a12c38493c2c7c8622690b769a5       accessed
#> 51: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 52: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 53: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 54: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 55: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 56: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 57: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 58: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 59: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 60: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 61: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 62: b0db9a12c38493c2c7c8622690b769a5      preDigest
#> 63: b0db9a12c38493c2c7c8622690b769a5      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-03-06 07:42:52
#>  2: 06b0f8f5af2781d3f74786fb797dd985 2019-03-06 07:42:52
#>  3:                          numeric 2019-03-06 07:42:52
#>  4:              2019-03-06 07:42:52 2019-03-06 07:42:52
#>  5:                 f7bee22047b8d0c1 2019-03-06 07:42:52
#>  6:                            rnorm 2019-03-06 07:42:52
#>  7:                             1008 2019-03-06 07:42:52
#>  8:              2019-03-06 07:42:52 2019-03-06 07:42:52
#>  9:                          saveRDS 2019-03-06 07:42:52
#> 10:                          do.call 2019-03-06 07:42:52
#> 11:                 build_site_local 2019-03-06 07:42:52
#> 12:                  build_reference 2019-03-06 07:42:52
#> 13:             data_reference_topic 2019-03-06 07:42:52
#> 14:                          as_data 2019-03-06 07:42:52
#> 15:             as_data.tag_examples 2019-03-06 07:42:52
#> 16:                        timing_fn 2019-03-06 07:42:52
#> 17:                           handle 2019-03-06 07:42:52
#> 18:                              try 2019-03-06 07:42:52
#> 19:                      withVisible 2019-03-06 07:42:52
#> 20:               n:7eef4eae85fd9229 2019-03-06 07:42:52
#> 21:            .FUN:4f604aa46882b368 2019-03-06 07:42:52
#> 22:                              rda 2019-03-06 07:42:52
#> 23: 2c8492f352933627e962e697da788a62 2019-03-06 07:42:52
#> 24:                          numeric 2019-03-06 07:42:52
#> 25:              2019-03-06 07:42:52 2019-03-06 07:42:52
#> 26:                 3b347dee8f6305c7 2019-03-06 07:42:52
#> 27:                            rnorm 2019-03-06 07:42:52
#> 28:                              992 2019-03-06 07:42:52
#> 29:              2019-03-06 07:42:52 2019-03-06 07:42:52
#> 30:                          saveRDS 2019-03-06 07:42:52
#> 31:                          do.call 2019-03-06 07:42:52
#> 32:                 build_site_local 2019-03-06 07:42:52
#> 33:                  build_reference 2019-03-06 07:42:52
#> 34:             data_reference_topic 2019-03-06 07:42:52
#> 35:                          as_data 2019-03-06 07:42:52
#> 36:             as_data.tag_examples 2019-03-06 07:42:52
#> 37:                        timing_fn 2019-03-06 07:42:52
#> 38:                           handle 2019-03-06 07:42:52
#> 39:                              try 2019-03-06 07:42:52
#> 40:                      withVisible 2019-03-06 07:42:52
#> 41:               n:82dc709f2b91918a 2019-03-06 07:42:52
#> 42:            .FUN:4f604aa46882b368 2019-03-06 07:42:52
#> 43:                              rda 2019-03-06 07:42:52
#> 44: b0db9a12c38493c2c7c8622690b769a5 2019-03-06 07:42:52
#> 45:                          numeric 2019-03-06 07:42:52
#> 46:              2019-03-06 07:42:52 2019-03-06 07:42:52
#> 47:                 f0da91a44b839434 2019-03-06 07:42:52
#> 48:                            rnorm 2019-03-06 07:42:52
#> 49:                             1008 2019-03-06 07:42:52
#> 50:              2019-03-06 07:42:52 2019-03-06 07:42:52
#> 51:                          saveRDS 2019-03-06 07:42:52
#> 52:                          do.call 2019-03-06 07:42:52
#> 53:                 build_site_local 2019-03-06 07:42:52
#> 54:                  build_reference 2019-03-06 07:42:52
#> 55:             data_reference_topic 2019-03-06 07:42:52
#> 56:                          as_data 2019-03-06 07:42:52
#> 57:             as_data.tag_examples 2019-03-06 07:42:52
#> 58:                        timing_fn 2019-03-06 07:42:52
#> 59:                           handle 2019-03-06 07:42:52
#> 60:                              try 2019-03-06 07:42:52
#> 61:                      withVisible 2019-03-06 07:42:52
#> 62:               n:7f12988bd88a0fb8 2019-03-06 07:42:52
#> 63:            .FUN:4f604aa46882b368 2019-03-06 07:42:52
#>                             tagValue         createdDate

cc(thisTime, ask = FALSE, x = tmpDir)

#> Cache size: 

#>   Total (including Rasters): 752 bytes

#>   Selected objects (not including Rasters): 0 bytes

showCache(x = tmpDir) # all those after thisTime gone, i.e., only 1 left

#> Cache size: 

#>   Total (including Rasters): 752 bytes

#>   Selected objects (not including Rasters): 752 bytes

#>                             artifact         tagKey
#>  1: 06b0f8f5af2781d3f74786fb797dd985         format
#>  2: 06b0f8f5af2781d3f74786fb797dd985           name
#>  3: 06b0f8f5af2781d3f74786fb797dd985          class
#>  4: 06b0f8f5af2781d3f74786fb797dd985           date
#>  5: 06b0f8f5af2781d3f74786fb797dd985        cacheId
#>  6: 06b0f8f5af2781d3f74786fb797dd985       function
#>  7: 06b0f8f5af2781d3f74786fb797dd985    object.size
#>  8: 06b0f8f5af2781d3f74786fb797dd985       accessed
#>  9: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 10: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 11: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 12: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 13: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 14: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 15: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 16: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 17: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 18: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 19: 06b0f8f5af2781d3f74786fb797dd985 otherFunctions
#> 20: 06b0f8f5af2781d3f74786fb797dd985      preDigest
#> 21: 06b0f8f5af2781d3f74786fb797dd985      preDigest
#> 22: 2c8492f352933627e962e697da788a62         format
#> 23: 2c8492f352933627e962e697da788a62           name
#> 24: 2c8492f352933627e962e697da788a62          class
#> 25: 2c8492f352933627e962e697da788a62           date
#> 26: 2c8492f352933627e962e697da788a62        cacheId
#> 27: 2c8492f352933627e962e697da788a62       function
#> 28: 2c8492f352933627e962e697da788a62    object.size
#> 29: 2c8492f352933627e962e697da788a62       accessed
#> 30: 2c8492f352933627e962e697da788a62 otherFunctions
#> 31: 2c8492f352933627e962e697da788a62 otherFunctions
#> 32: 2c8492f352933627e962e697da788a62 otherFunctions
#> 33: 2c8492f352933627e962e697da788a62 otherFunctions
#> 34: 2c8492f352933627e962e697da788a62 otherFunctions
#> 35: 2c8492f352933627e962e697da788a62 otherFunctions
#> 36: 2c8492f352933627e962e697da788a62 otherFunctions
#> 37: 2c8492f352933627e962e697da788a62 otherFunctions
#> 38: 2c8492f352933627e962e697da788a62 otherFunctions
#> 39: 2c8492f352933627e962e697da788a62 otherFunctions
#> 40: 2c8492f352933627e962e697da788a62 otherFunctions
#> 41: 2c8492f352933627e962e697da788a62      preDigest
#> 42: 2c8492f352933627e962e697da788a62      preDigest
#> 43: b0db9a12c38493c2c7c8622690b769a5         format
#> 44: b0db9a12c38493c2c7c8622690b769a5           name
#> 45: b0db9a12c38493c2c7c8622690b769a5          class
#> 46: b0db9a12c38493c2c7c8622690b769a5           date
#> 47: b0db9a12c38493c2c7c8622690b769a5        cacheId
#> 48: b0db9a12c38493c2c7c8622690b769a5       function
#> 49: b0db9a12c38493c2c7c8622690b769a5    object.size
#> 50: b0db9a12c38493c2c7c8622690b769a5       accessed
#> 51: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 52: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 53: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 54: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 55: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 56: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 57: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 58: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 59: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 60: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 61: b0db9a12c38493c2c7c8622690b769a5 otherFunctions
#> 62: b0db9a12c38493c2c7c8622690b769a5      preDigest
#> 63: b0db9a12c38493c2c7c8622690b769a5      preDigest
#>                             artifact         tagKey
#>                             tagValue         createdDate
#>  1:                              rda 2019-03-06 07:42:52
#>  2: 06b0f8f5af2781d3f74786fb797dd985 2019-03-06 07:42:52
#>  3:                          numeric 2019-03-06 07:42:52
#>  4:              2019-03-06 07:42:52 2019-03-06 07:42:52
#>  5:                 f7bee22047b8d0c1 2019-03-06 07:42:52
#>  6:                            rnorm 2019-03-06 07:42:52
#>  7:                             1008 2019-03-06 07:42:52
#>  8:              2019-03-06 07:42:52 2019-03-06 07:42:52
#>  9:                          saveRDS 2019-03-06 07:42:52
#> 10:                          do.call 2019-03-06 07:42:52
#> 11:                 build_site_local 2019-03-06 07:42:52
#> 12:                  build_reference 2019-03-06 07:42:52
#> 13:             data_reference_topic 2019-03-06 07:42:52
#> 14:                          as_data 2019-03-06 07:42:52
#> 15:             as_data.tag_examples 2019-03-06 07:42:52
#> 16:                        timing_fn 2019-03-06 07:42:52
#> 17:                           handle 2019-03-06 07:42:52
#> 18:                              try 2019-03-06 07:42:52
#> 19:                      withVisible 2019-03-06 07:42:52
#> 20:               n:7eef4eae85fd9229 2019-03-06 07:42:52
#> 21:            .FUN:4f604aa46882b368 2019-03-06 07:42:52
#> 22:                              rda 2019-03-06 07:42:52
#> 23: 2c8492f352933627e962e697da788a62 2019-03-06 07:42:52
#> 24:                          numeric 2019-03-06 07:42:52
#> 25:              2019-03-06 07:42:52 2019-03-06 07:42:52
#> 26:                 3b347dee8f6305c7 2019-03-06 07:42:52
#> 27:                            rnorm 2019-03-06 07:42:52
#> 28:                              992 2019-03-06 07:42:52
#> 29:              2019-03-06 07:42:52 2019-03-06 07:42:52
#> 30:                          saveRDS 2019-03-06 07:42:52
#> 31:                          do.call 2019-03-06 07:42:52
#> 32:                 build_site_local 2019-03-06 07:42:52
#> 33:                  build_reference 2019-03-06 07:42:52
#> 34:             data_reference_topic 2019-03-06 07:42:52
#> 35:                          as_data 2019-03-06 07:42:52
#> 36:             as_data.tag_examples 2019-03-06 07:42:52
#> 37:                        timing_fn 2019-03-06 07:42:52
#> 38:                           handle 2019-03-06 07:42:52
#> 39:                              try 2019-03-06 07:42:52
#> 40:                      withVisible 2019-03-06 07:42:52
#> 41:               n:82dc709f2b91918a 2019-03-06 07:42:52
#> 42:            .FUN:4f604aa46882b368 2019-03-06 07:42:52
#> 43:                              rda 2019-03-06 07:42:52
#> 44: b0db9a12c38493c2c7c8622690b769a5 2019-03-06 07:42:52
#> 45:                          numeric 2019-03-06 07:42:52
#> 46:              2019-03-06 07:42:52 2019-03-06 07:42:52
#> 47:                 f0da91a44b839434 2019-03-06 07:42:52
#> 48:                            rnorm 2019-03-06 07:42:52
#> 49:                             1008 2019-03-06 07:42:52
#> 50:              2019-03-06 07:42:52 2019-03-06 07:42:52
#> 51:                          saveRDS 2019-03-06 07:42:52
#> 52:                          do.call 2019-03-06 07:42:52
#> 53:                 build_site_local 2019-03-06 07:42:52
#> 54:                  build_reference 2019-03-06 07:42:52
#> 55:             data_reference_topic 2019-03-06 07:42:52
#> 56:                          as_data 2019-03-06 07:42:52
#> 57:             as_data.tag_examples 2019-03-06 07:42:52
#> 58:                        timing_fn 2019-03-06 07:42:52
#> 59:                           handle 2019-03-06 07:42:52
#> 60:                              try 2019-03-06 07:42:52
#> 61:                      withVisible 2019-03-06 07:42:52
#> 62:               n:7f12988bd88a0fb8 2019-03-06 07:42:52
#> 63:            .FUN:4f604aa46882b368 2019-03-06 07:42:52
#>                             tagValue         createdDate

cc(ask = FALSE, x = tmpDir) # Cache is

#> No time provided; removing the most recent entry to the Cache

#> Cache size: 

#>   Total (including Rasters): 752 bytes

#>   Selected objects (not including Rasters): 252 bytes

cc(ask = FALSE, x = tmpDir) # Cache is already empty

#> No time provided; removing the most recent entry to the Cache

#> Cache size: 

#>   Total (including Rasters): 500 bytes

#>   Selected objects (not including Rasters): 248 bytes