Examining and modifying the cache

These are convenience wrappers around DBI package functions. They allow the user a bit of control over what is being cached.

clearCache(
  x,
  userTags = character(),
  after = NULL,
  before = NULL,
  fun = NULL,
  cacheId = NULL,
  ask = getOption("reproducible.ask"),
  useCloud = FALSE,
  cloudFolderID = getOption("reproducible.cloudFolderID", NULL),
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  ...
)

# S4 method for class 'ANY'
clearCache(
  x,
  userTags = character(),
  after = NULL,
  before = NULL,
  fun = NULL,
  cacheId = NULL,
  ask = getOption("reproducible.ask"),
  useCloud = FALSE,
  cloudFolderID = getOption("reproducible.cloudFolderID", NULL),
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  ...
)

cc(secs, ..., verbose = getOption("reproducible.verbose"))

showCache(
  x,
  userTags = character(),
  after = NULL,
  before = NULL,
  fun = NULL,
  cacheId = NULL,
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  ...
)

# S4 method for class 'ANY'
showCache(
  x,
  userTags = character(),
  after = NULL,
  before = NULL,
  fun = NULL,
  cacheId = NULL,
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  ...
)

keepCache(
  x,
  userTags = character(),
  after = NULL,
  before = NULL,
  ask = getOption("reproducible.ask"),
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  ...
)

# S4 method for class 'ANY'
keepCache(
  x,
  userTags = character(),
  after = NULL,
  before = NULL,
  ask = getOption("reproducible.ask"),
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  ...
)

Arguments

x

A simList or a directory containing a valid Cache repository. Note: For compatibility with Cache argument, cachePath can also be used instead of x, though x will take precedence.

userTags

Character vector. If used, this will be used in place of the after and before. Specifying one or more userTag here will clear all objects that match those tags. Matching is via regular expression, meaning partial matches will work unless strict beginning (^) and end ($) of string characters are used. Matching will be against any of the 3 columns returned by showCache(), i.e., artifact, tagValue or tagName. Also, if length(userTags) > 1, then matching is by and. For or matching, use | in a single character string. See examples.

after

A time (POSIX, character understandable by data.table). Objects cached after this time will be shown or deleted.

before

A time (POSIX, character understandable by data.table). Objects cached before this time will be shown or deleted.

fun

An optional character vector describing the function name to extract. Only functions with this/these functions will be returned.

cacheId

An optional character vector describing the cacheIds to extract. Only entries with this/these cacheIds will be returned. If useDBI(FALSE), this will also be dramatically faster than using userTags, for a large cache.

ask

Logical. If FALSE, then it will not ask to confirm deletions using clearCache or keepCache. Default is TRUE

useCloud

Logical. If TRUE, then every object that is deleted locally will also be deleted in the cloudFolderID, if it is non-NULL

cloudFolderID

A googledrive dribble of a folder, e.g., using drive_mkdir(). If left as NULL, the function will create a cloud folder with name from last two folder levels of the cachePath path, : paste0(basename(dirname(cachePath)), "_", basename(cachePath)). This cloudFolderID will be added to options("reproducible.cloudFolderID"), but this will not persist across sessions. If this is a character string, it will treat this as a folder name to create or use on GoogleDrive.

drv

if using a database backend, drv must be an object that inherits from DBIDriver e.g., from package RSQLite, e.g., SQLite

conn

an optional DBIConnection object, as returned by dbConnect().

verbose

Numeric, -1 silent (where possible), 0 being very quiet, 1 showing more messaging, 2 being more messaging, etc. Default is 1. Above 3 will output much more information about the internals of Caching, which may help diagnose Caching challenges. Can set globally with an option, e.g., options('reproducible.verbose' = 0) to reduce to minimal

...

Other arguments. Can be in the form of tagKey = tagValue, such as, class = "numeric" to find all entries that are numerics in the cache. Note: the special cases of cacheId and fun have their own named arguments in these functions. Also can be regexp = xx, where xx is TRUE if the user is passing a regular expression. Otherwise, userTags will need to be exact matches. Default is missing, which is the same as TRUE. If there are errors due to regular expression problem, try FALSE. For cc, it is passed to clearCache, e.g., ask, userTags. For showCache, it can also be sorted = FALSE to return the object unsorted.

secs

Currently 3 options: the number of seconds to pass to clearCache(after = secs), a POSIXct time e.g., from Sys.time(), or missing. If missing, the default, then it will delete the most recent entry in the Cache.

Value

Will clear all objects (or those that match userTags, or those between after or before) from the repository located in cachePath. Invisibly returns a data.table of the removed items.

Details

If neither after or before are provided, nor userTags, then all objects will be removed. If both after and before are specified, then all objects between after and before will be deleted. If userTags is used, this will override after or before.

cc(secs) is just a shortcut for clearCache(repo = currentRepo, after = secs), i.e., to remove any cache entries touched in the last secs seconds. Since, secs can be missing, this is also be a shorthand for "remove most recent entry from the cache".

clearCache

remove items from the cache based on their userTag or times values.

keepCache

remove all cached items except those based on certain userTags or times values.

showCache

display the contents of the cache.

By default the return of showCache is sorted by cacheId. For convenience, a user can optionally have it unsorted (passing sorted = FALSE), which may be noticeably faster when the cache is large (> 1e4 entries).

Note

If the cache is larger than 10MB, and clearCache is used, there will be a message and a pause, if interactive, to prevent accidentally deleting of a large cache repository.

See also

Examples

data.table::setDTthreads(2)

tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir, ask = FALSE), silent = TRUE) # just to make sure it is clear

# Basic use
ranNumsA <- Cache(rnorm, 10, 16, cachePath = tmpDir)
#> Saved! Cache file: 4fca280cda001fc9.rds; fn: rnorm

# All same
ranNumsB <- Cache(rnorm, 10, 16, cachePath = tmpDir) # recovers cached copy
#> Object to retrieve (fn: rnorm, 4fca280cda001fc9.rds) ...
#> Loaded! Cached result from previous rnorm call
ranNumsD <- Cache(quote(rnorm(n = 10, 16)), cachePath = tmpDir) # recovers cached copy
#> Object to retrieve (fn: quote, 4fca280cda001fc9.rds) ...
#> Loaded! Cached result from previous quote call

# Any minor change makes it different
ranNumsE <- Cache(rnorm, 10, 6, cachePath = tmpDir) # different
#> Saved! Cache file: 5efb386c43c29ce1.rds; fn: rnorm

## Example 1: basic cache use with tags
ranNumsA <- Cache(rnorm, 4, cachePath = tmpDir, userTags = "objectName:a")
#> Saved! Cache file: ad0ea27476c50b66.rds; fn: rnorm
ranNumsB <- Cache(runif, 4, cachePath = tmpDir, userTags = "objectName:b")
#> Saved! Cache file: deaa37372f85861b.rds; fn: runif
ranNumsC <- Cache(runif, 40, cachePath = tmpDir, userTags = "objectName:b")
#> Saved! Cache file: a0603ff09bb4ce83.rds; fn: runif

showCache(tmpDir, userTags = c("objectName"))
#> Cache size:
#> Total (including Rasters): 828 bytes
#> Selected objects (not including Rasters): 828 bytes
#>              cacheId              tagKey                   tagValue
#>               <char>              <char>                     <char>
#>  1: a0603ff09bb4ce83          objectName                          b
#>  2: a0603ff09bb4ce83            function                      runif
#>  3: a0603ff09bb4ce83               class                    numeric
#>  4: a0603ff09bb4ce83         object.size                       1296
#>  5: a0603ff09bb4ce83            accessed 2024-12-12 06:16:02.232031
#>  6: a0603ff09bb4ce83             inCloud                      FALSE
#>  7: a0603ff09bb4ce83            fromDisk                      FALSE
#>  8: a0603ff09bb4ce83          resultHash                           
#>  9: a0603ff09bb4ce83   elapsedTimeDigest           0.001623392 secs
#> 10: a0603ff09bb4ce83 elapsedTimeFirstRun          6.151199e-05 secs
#> 11: a0603ff09bb4ce83      otherFunctions                 build_site
#> 12: a0603ff09bb4ce83      otherFunctions           build_site_local
#> 13: a0603ff09bb4ce83      otherFunctions            build_reference
#> 14: a0603ff09bb4ce83      otherFunctions         unwrap_purrr_error
#> 15: a0603ff09bb4ce83      otherFunctions                       map_
#> 16: a0603ff09bb4ce83      otherFunctions        with_indexed_errors
#> 17: a0603ff09bb4ce83      otherFunctions          call_with_cleanup
#> 18: a0603ff09bb4ce83      otherFunctions       data_reference_topic
#> 19: a0603ff09bb4ce83      otherFunctions               run_examples
#> 20: a0603ff09bb4ce83      otherFunctions         highlight_examples
#> 21: a0603ff09bb4ce83      otherFunctions              with_handlers
#> 22: a0603ff09bb4ce83           preDigest         n:1393aef18608c8be
#> 23: a0603ff09bb4ce83           preDigest       min:c40c00762a0dac94
#> 24: a0603ff09bb4ce83           preDigest       max:853b1797f54b229c
#> 25: a0603ff09bb4ce83           preDigest      .FUN:881ec847b7161f3c
#> 26: ad0ea27476c50b66          objectName                          a
#> 27: ad0ea27476c50b66            function                      rnorm
#> 28: ad0ea27476c50b66               class                    numeric
#> 29: ad0ea27476c50b66         object.size                       1008
#> 30: ad0ea27476c50b66            accessed 2024-12-12 06:16:02.195396
#> 31: ad0ea27476c50b66             inCloud                      FALSE
#> 32: ad0ea27476c50b66            fromDisk                      FALSE
#> 33: ad0ea27476c50b66          resultHash                           
#> 34: ad0ea27476c50b66   elapsedTimeDigest           0.001600981 secs
#> 35: ad0ea27476c50b66 elapsedTimeFirstRun          6.341934e-05 secs
#> 36: ad0ea27476c50b66      otherFunctions                 build_site
#> 37: ad0ea27476c50b66      otherFunctions           build_site_local
#> 38: ad0ea27476c50b66      otherFunctions            build_reference
#> 39: ad0ea27476c50b66      otherFunctions         unwrap_purrr_error
#> 40: ad0ea27476c50b66      otherFunctions                       map_
#> 41: ad0ea27476c50b66      otherFunctions        with_indexed_errors
#> 42: ad0ea27476c50b66      otherFunctions          call_with_cleanup
#> 43: ad0ea27476c50b66      otherFunctions       data_reference_topic
#> 44: ad0ea27476c50b66      otherFunctions               run_examples
#> 45: ad0ea27476c50b66      otherFunctions         highlight_examples
#> 46: ad0ea27476c50b66      otherFunctions              with_handlers
#> 47: ad0ea27476c50b66           preDigest         n:7eef4eae85fd9229
#> 48: ad0ea27476c50b66           preDigest      mean:c40c00762a0dac94
#> 49: ad0ea27476c50b66           preDigest        sd:853b1797f54b229c
#> 50: ad0ea27476c50b66           preDigest      .FUN:4f604aa46882b368
#> 51: deaa37372f85861b          objectName                          b
#> 52: deaa37372f85861b            function                      runif
#> 53: deaa37372f85861b               class                    numeric
#> 54: deaa37372f85861b         object.size                       1008
#> 55: deaa37372f85861b            accessed 2024-12-12 06:16:02.214557
#> 56: deaa37372f85861b             inCloud                      FALSE
#> 57: deaa37372f85861b            fromDisk                      FALSE
#> 58: deaa37372f85861b          resultHash                           
#> 59: deaa37372f85861b   elapsedTimeDigest           0.001597404 secs
#> 60: deaa37372f85861b elapsedTimeFirstRun          6.175041e-05 secs
#> 61: deaa37372f85861b      otherFunctions                 build_site
#> 62: deaa37372f85861b      otherFunctions           build_site_local
#> 63: deaa37372f85861b      otherFunctions            build_reference
#> 64: deaa37372f85861b      otherFunctions         unwrap_purrr_error
#> 65: deaa37372f85861b      otherFunctions                       map_
#> 66: deaa37372f85861b      otherFunctions        with_indexed_errors
#> 67: deaa37372f85861b      otherFunctions          call_with_cleanup
#> 68: deaa37372f85861b      otherFunctions       data_reference_topic
#> 69: deaa37372f85861b      otherFunctions               run_examples
#> 70: deaa37372f85861b      otherFunctions         highlight_examples
#> 71: deaa37372f85861b      otherFunctions              with_handlers
#> 72: deaa37372f85861b           preDigest         n:7eef4eae85fd9229
#> 73: deaa37372f85861b           preDigest       min:c40c00762a0dac94
#> 74: deaa37372f85861b           preDigest       max:853b1797f54b229c
#> 75: deaa37372f85861b           preDigest      .FUN:881ec847b7161f3c
#>              cacheId              tagKey                   tagValue
#>                    createdDate
#>                         <char>
#>  1: 2024-12-12 06:16:02.232647
#>  2: 2024-12-12 06:16:02.232647
#>  3: 2024-12-12 06:16:02.232647
#>  4: 2024-12-12 06:16:02.232647
#>  5: 2024-12-12 06:16:02.232647
#>  6: 2024-12-12 06:16:02.232647
#>  7: 2024-12-12 06:16:02.232647
#>  8: 2024-12-12 06:16:02.232647
#>  9: 2024-12-12 06:16:02.232647
#> 10: 2024-12-12 06:16:02.232647
#> 11: 2024-12-12 06:16:02.232647
#> 12: 2024-12-12 06:16:02.232647
#> 13: 2024-12-12 06:16:02.232647
#> 14: 2024-12-12 06:16:02.232647
#> 15: 2024-12-12 06:16:02.232647
#> 16: 2024-12-12 06:16:02.232647
#> 17: 2024-12-12 06:16:02.232647
#> 18: 2024-12-12 06:16:02.232647
#> 19: 2024-12-12 06:16:02.232647
#> 20: 2024-12-12 06:16:02.232647
#> 21: 2024-12-12 06:16:02.232647
#> 22: 2024-12-12 06:16:02.232647
#> 23: 2024-12-12 06:16:02.232647
#> 24: 2024-12-12 06:16:02.232647
#> 25: 2024-12-12 06:16:02.232647
#> 26: 2024-12-12 06:16:02.195987
#> 27: 2024-12-12 06:16:02.195987
#> 28: 2024-12-12 06:16:02.195987
#> 29: 2024-12-12 06:16:02.195987
#> 30: 2024-12-12 06:16:02.195987
#> 31: 2024-12-12 06:16:02.195987
#> 32: 2024-12-12 06:16:02.195987
#> 33: 2024-12-12 06:16:02.195987
#> 34: 2024-12-12 06:16:02.195987
#> 35: 2024-12-12 06:16:02.195987
#> 36: 2024-12-12 06:16:02.195987
#> 37: 2024-12-12 06:16:02.195987
#> 38: 2024-12-12 06:16:02.195987
#> 39: 2024-12-12 06:16:02.195987
#> 40: 2024-12-12 06:16:02.195987
#> 41: 2024-12-12 06:16:02.195987
#> 42: 2024-12-12 06:16:02.195987
#> 43: 2024-12-12 06:16:02.195987
#> 44: 2024-12-12 06:16:02.195987
#> 45: 2024-12-12 06:16:02.195987
#> 46: 2024-12-12 06:16:02.195987
#> 47: 2024-12-12 06:16:02.195987
#> 48: 2024-12-12 06:16:02.195987
#> 49: 2024-12-12 06:16:02.195987
#> 50: 2024-12-12 06:16:02.195987
#> 51: 2024-12-12 06:16:02.215147
#> 52: 2024-12-12 06:16:02.215147
#> 53: 2024-12-12 06:16:02.215147
#> 54: 2024-12-12 06:16:02.215147
#> 55: 2024-12-12 06:16:02.215147
#> 56: 2024-12-12 06:16:02.215147
#> 57: 2024-12-12 06:16:02.215147
#> 58: 2024-12-12 06:16:02.215147
#> 59: 2024-12-12 06:16:02.215147
#> 60: 2024-12-12 06:16:02.215147
#> 61: 2024-12-12 06:16:02.215147
#> 62: 2024-12-12 06:16:02.215147
#> 63: 2024-12-12 06:16:02.215147
#> 64: 2024-12-12 06:16:02.215147
#> 65: 2024-12-12 06:16:02.215147
#> 66: 2024-12-12 06:16:02.215147
#> 67: 2024-12-12 06:16:02.215147
#> 68: 2024-12-12 06:16:02.215147
#> 69: 2024-12-12 06:16:02.215147
#> 70: 2024-12-12 06:16:02.215147
#> 71: 2024-12-12 06:16:02.215147
#> 72: 2024-12-12 06:16:02.215147
#> 73: 2024-12-12 06:16:02.215147
#> 74: 2024-12-12 06:16:02.215147
#> 75: 2024-12-12 06:16:02.215147
#>                    createdDate
showCache(tmpDir, userTags = c("^a$")) # regular expression ... "a" exactly
#> Cache size:
#> Total (including Rasters): 252 bytes
#> Selected objects (not including Rasters): 252 bytes
#>              cacheId              tagKey                   tagValue
#>               <char>              <char>                     <char>
#>  1: ad0ea27476c50b66          objectName                          a
#>  2: ad0ea27476c50b66            function                      rnorm
#>  3: ad0ea27476c50b66               class                    numeric
#>  4: ad0ea27476c50b66         object.size                       1008
#>  5: ad0ea27476c50b66            accessed 2024-12-12 06:16:02.195396
#>  6: ad0ea27476c50b66             inCloud                      FALSE
#>  7: ad0ea27476c50b66            fromDisk                      FALSE
#>  8: ad0ea27476c50b66          resultHash                           
#>  9: ad0ea27476c50b66   elapsedTimeDigest           0.001600981 secs
#> 10: ad0ea27476c50b66 elapsedTimeFirstRun          6.341934e-05 secs
#> 11: ad0ea27476c50b66      otherFunctions                 build_site
#> 12: ad0ea27476c50b66      otherFunctions           build_site_local
#> 13: ad0ea27476c50b66      otherFunctions            build_reference
#> 14: ad0ea27476c50b66      otherFunctions         unwrap_purrr_error
#> 15: ad0ea27476c50b66      otherFunctions                       map_
#> 16: ad0ea27476c50b66      otherFunctions        with_indexed_errors
#> 17: ad0ea27476c50b66      otherFunctions          call_with_cleanup
#> 18: ad0ea27476c50b66      otherFunctions       data_reference_topic
#> 19: ad0ea27476c50b66      otherFunctions               run_examples
#> 20: ad0ea27476c50b66      otherFunctions         highlight_examples
#> 21: ad0ea27476c50b66      otherFunctions              with_handlers
#> 22: ad0ea27476c50b66           preDigest         n:7eef4eae85fd9229
#> 23: ad0ea27476c50b66           preDigest      mean:c40c00762a0dac94
#> 24: ad0ea27476c50b66           preDigest        sd:853b1797f54b229c
#> 25: ad0ea27476c50b66           preDigest      .FUN:4f604aa46882b368
#>              cacheId              tagKey                   tagValue
#>                    createdDate
#>                         <char>
#>  1: 2024-12-12 06:16:02.195987
#>  2: 2024-12-12 06:16:02.195987
#>  3: 2024-12-12 06:16:02.195987
#>  4: 2024-12-12 06:16:02.195987
#>  5: 2024-12-12 06:16:02.195987
#>  6: 2024-12-12 06:16:02.195987
#>  7: 2024-12-12 06:16:02.195987
#>  8: 2024-12-12 06:16:02.195987
#>  9: 2024-12-12 06:16:02.195987
#> 10: 2024-12-12 06:16:02.195987
#> 11: 2024-12-12 06:16:02.195987
#> 12: 2024-12-12 06:16:02.195987
#> 13: 2024-12-12 06:16:02.195987
#> 14: 2024-12-12 06:16:02.195987
#> 15: 2024-12-12 06:16:02.195987
#> 16: 2024-12-12 06:16:02.195987
#> 17: 2024-12-12 06:16:02.195987
#> 18: 2024-12-12 06:16:02.195987
#> 19: 2024-12-12 06:16:02.195987
#> 20: 2024-12-12 06:16:02.195987
#> 21: 2024-12-12 06:16:02.195987
#> 22: 2024-12-12 06:16:02.195987
#> 23: 2024-12-12 06:16:02.195987
#> 24: 2024-12-12 06:16:02.195987
#> 25: 2024-12-12 06:16:02.195987
#>                    createdDate

# Fine control of cache elements -- pick out only the large runif object, and remove it
cache1 <- showCache(tmpDir, userTags = c("runif")) # show only cached objects made during runif
#> Cache size:
#> Total (including Rasters): 576 bytes
#> Selected objects (not including Rasters): 576 bytes
toRemove <- cache1[tagKey == "object.size"][as.numeric(tagValue) > 700]$cacheId
clearCache(tmpDir, userTags = toRemove, ask = FALSE)
#> Cache size:
#> Total (including Rasters): 0 bytes
#> Selected objects (not including Rasters): 0 bytes
cacheAfter <- showCache(tmpDir, userTags = c("runif")) # Only the small one is left
#> Cache size:
#> Total (including Rasters): 576 bytes
#> Selected objects (not including Rasters): 576 bytes

data.table::setDTthreads(2)
tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir, ask = FALSE), silent = TRUE) # just to make sure it is clear

Cache(rnorm, 1, cachePath = tmpDir)
#> Saved! Cache file: 422bae4ed2f770cc.rds; fn: rnorm
#> [1] 0.1612428
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:422bae4ed2f770cc"
#> attr(,"call")
#> [1] ""
thisTime <- Sys.time()
Cache(rnorm, 2, cachePath = tmpDir)
#> Saved! Cache file: 66299c88b3924bf4.rds; fn: rnorm
#> [1] -0.3902589 -1.3527104
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:66299c88b3924bf4"
#> attr(,"call")
#> [1] ""
Cache(rnorm, 3, cachePath = tmpDir)
#> Saved! Cache file: 6724ef06570739aa.rds; fn: rnorm
#> [1]  1.7181970  0.6892219 -0.8328587
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:6724ef06570739aa"
#> attr(,"call")
#> [1] ""
Cache(rnorm, 4, cachePath = tmpDir)
#> Saved! Cache file: ad0ea27476c50b66.rds; fn: rnorm
#> [1] -0.3492281 -1.4799191 -0.1363727  1.6025934
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:ad0ea27476c50b66"
#> attr(,"call")
#> [1] ""
showCache(x = tmpDir) # shows all 4 entries
#> Cache size:
#> Total (including Rasters): 998 bytes
#> Selected objects (not including Rasters): 998 bytes
#>              cacheId              tagKey                   tagValue
#>               <char>              <char>                     <char>
#>  1: 422bae4ed2f770cc            function                      rnorm
#>  2: 422bae4ed2f770cc               class                    numeric
#>  3: 422bae4ed2f770cc         object.size                        984
#>  4: 422bae4ed2f770cc            accessed 2024-12-12 06:16:02.328365
#>  5: 422bae4ed2f770cc             inCloud                      FALSE
#>  6: 422bae4ed2f770cc            fromDisk                      FALSE
#>  7: 422bae4ed2f770cc          resultHash                           
#>  8: 422bae4ed2f770cc   elapsedTimeDigest           0.001799345 secs
#>  9: 422bae4ed2f770cc elapsedTimeFirstRun           7.43866e-05 secs
#> 10: 422bae4ed2f770cc      otherFunctions                 build_site
#> 11: 422bae4ed2f770cc      otherFunctions           build_site_local
#> 12: 422bae4ed2f770cc      otherFunctions            build_reference
#> 13: 422bae4ed2f770cc      otherFunctions         unwrap_purrr_error
#> 14: 422bae4ed2f770cc      otherFunctions                       map_
#> 15: 422bae4ed2f770cc      otherFunctions        with_indexed_errors
#> 16: 422bae4ed2f770cc      otherFunctions          call_with_cleanup
#> 17: 422bae4ed2f770cc      otherFunctions       data_reference_topic
#> 18: 422bae4ed2f770cc      otherFunctions               run_examples
#> 19: 422bae4ed2f770cc      otherFunctions         highlight_examples
#> 20: 422bae4ed2f770cc      otherFunctions              with_handlers
#> 21: 422bae4ed2f770cc           preDigest         n:853b1797f54b229c
#> 22: 422bae4ed2f770cc           preDigest      mean:c40c00762a0dac94
#> 23: 422bae4ed2f770cc           preDigest        sd:853b1797f54b229c
#> 24: 422bae4ed2f770cc           preDigest      .FUN:4f604aa46882b368
#> 25: 66299c88b3924bf4            function                      rnorm
#> 26: 66299c88b3924bf4               class                    numeric
#> 27: 66299c88b3924bf4         object.size                        992
#> 28: 66299c88b3924bf4            accessed 2024-12-12 06:16:02.346638
#> 29: 66299c88b3924bf4             inCloud                      FALSE
#> 30: 66299c88b3924bf4            fromDisk                      FALSE
#> 31: 66299c88b3924bf4          resultHash                           
#> 32: 66299c88b3924bf4   elapsedTimeDigest           0.001669884 secs
#> 33: 66299c88b3924bf4 elapsedTimeFirstRun          6.389618e-05 secs
#> 34: 66299c88b3924bf4      otherFunctions                 build_site
#> 35: 66299c88b3924bf4      otherFunctions           build_site_local
#> 36: 66299c88b3924bf4      otherFunctions            build_reference
#> 37: 66299c88b3924bf4      otherFunctions         unwrap_purrr_error
#> 38: 66299c88b3924bf4      otherFunctions                       map_
#> 39: 66299c88b3924bf4      otherFunctions        with_indexed_errors
#> 40: 66299c88b3924bf4      otherFunctions          call_with_cleanup
#> 41: 66299c88b3924bf4      otherFunctions       data_reference_topic
#> 42: 66299c88b3924bf4      otherFunctions               run_examples
#> 43: 66299c88b3924bf4      otherFunctions         highlight_examples
#> 44: 66299c88b3924bf4      otherFunctions              with_handlers
#> 45: 66299c88b3924bf4           preDigest         n:82dc709f2b91918a
#> 46: 66299c88b3924bf4           preDigest      mean:c40c00762a0dac94
#> 47: 66299c88b3924bf4           preDigest        sd:853b1797f54b229c
#> 48: 66299c88b3924bf4           preDigest      .FUN:4f604aa46882b368
#> 49: 6724ef06570739aa            function                      rnorm
#> 50: 6724ef06570739aa               class                    numeric
#> 51: 6724ef06570739aa         object.size                       1008
#> 52: 6724ef06570739aa            accessed 2024-12-12 06:16:02.364266
#> 53: 6724ef06570739aa             inCloud                      FALSE
#> 54: 6724ef06570739aa            fromDisk                      FALSE
#> 55: 6724ef06570739aa          resultHash                           
#> 56: 6724ef06570739aa   elapsedTimeDigest            0.00168252 secs
#> 57: 6724ef06570739aa elapsedTimeFirstRun          6.103516e-05 secs
#> 58: 6724ef06570739aa      otherFunctions                 build_site
#> 59: 6724ef06570739aa      otherFunctions           build_site_local
#> 60: 6724ef06570739aa      otherFunctions            build_reference
#> 61: 6724ef06570739aa      otherFunctions         unwrap_purrr_error
#> 62: 6724ef06570739aa      otherFunctions                       map_
#> 63: 6724ef06570739aa      otherFunctions        with_indexed_errors
#> 64: 6724ef06570739aa      otherFunctions          call_with_cleanup
#> 65: 6724ef06570739aa      otherFunctions       data_reference_topic
#> 66: 6724ef06570739aa      otherFunctions               run_examples
#> 67: 6724ef06570739aa      otherFunctions         highlight_examples
#> 68: 6724ef06570739aa      otherFunctions              with_handlers
#> 69: 6724ef06570739aa           preDigest         n:7f12988bd88a0fb8
#> 70: 6724ef06570739aa           preDigest      mean:c40c00762a0dac94
#> 71: 6724ef06570739aa           preDigest        sd:853b1797f54b229c
#> 72: 6724ef06570739aa           preDigest      .FUN:4f604aa46882b368
#> 73: ad0ea27476c50b66            function                      rnorm
#> 74: ad0ea27476c50b66               class                    numeric
#> 75: ad0ea27476c50b66         object.size                       1008
#> 76: ad0ea27476c50b66            accessed 2024-12-12 06:16:02.382416
#> 77: ad0ea27476c50b66             inCloud                      FALSE
#> 78: ad0ea27476c50b66            fromDisk                      FALSE
#> 79: ad0ea27476c50b66          resultHash                           
#> 80: ad0ea27476c50b66   elapsedTimeDigest           0.001615763 secs
#> 81: ad0ea27476c50b66 elapsedTimeFirstRun          6.079674e-05 secs
#> 82: ad0ea27476c50b66      otherFunctions                 build_site
#> 83: ad0ea27476c50b66      otherFunctions           build_site_local
#> 84: ad0ea27476c50b66      otherFunctions            build_reference
#> 85: ad0ea27476c50b66      otherFunctions         unwrap_purrr_error
#> 86: ad0ea27476c50b66      otherFunctions                       map_
#> 87: ad0ea27476c50b66      otherFunctions        with_indexed_errors
#> 88: ad0ea27476c50b66      otherFunctions          call_with_cleanup
#> 89: ad0ea27476c50b66      otherFunctions       data_reference_topic
#> 90: ad0ea27476c50b66      otherFunctions               run_examples
#> 91: ad0ea27476c50b66      otherFunctions         highlight_examples
#> 92: ad0ea27476c50b66      otherFunctions              with_handlers
#> 93: ad0ea27476c50b66           preDigest         n:7eef4eae85fd9229
#> 94: ad0ea27476c50b66           preDigest      mean:c40c00762a0dac94
#> 95: ad0ea27476c50b66           preDigest        sd:853b1797f54b229c
#> 96: ad0ea27476c50b66           preDigest      .FUN:4f604aa46882b368
#>              cacheId              tagKey                   tagValue
#>                    createdDate
#>                         <char>
#>  1: 2024-12-12 06:16:02.329004
#>  2: 2024-12-12 06:16:02.329004
#>  3: 2024-12-12 06:16:02.329004
#>  4: 2024-12-12 06:16:02.329004
#>  5: 2024-12-12 06:16:02.329004
#>  6: 2024-12-12 06:16:02.329004
#>  7: 2024-12-12 06:16:02.329004
#>  8: 2024-12-12 06:16:02.329004
#>  9: 2024-12-12 06:16:02.329004
#> 10: 2024-12-12 06:16:02.329004
#> 11: 2024-12-12 06:16:02.329004
#> 12: 2024-12-12 06:16:02.329004
#> 13: 2024-12-12 06:16:02.329004
#> 14: 2024-12-12 06:16:02.329004
#> 15: 2024-12-12 06:16:02.329004
#> 16: 2024-12-12 06:16:02.329004
#> 17: 2024-12-12 06:16:02.329004
#> 18: 2024-12-12 06:16:02.329004
#> 19: 2024-12-12 06:16:02.329004
#> 20: 2024-12-12 06:16:02.329004
#> 21: 2024-12-12 06:16:02.329004
#> 22: 2024-12-12 06:16:02.329004
#> 23: 2024-12-12 06:16:02.329004
#> 24: 2024-12-12 06:16:02.329004
#> 25: 2024-12-12 06:16:02.347223
#> 26: 2024-12-12 06:16:02.347223
#> 27: 2024-12-12 06:16:02.347223
#> 28: 2024-12-12 06:16:02.347223
#> 29: 2024-12-12 06:16:02.347223
#> 30: 2024-12-12 06:16:02.347223
#> 31: 2024-12-12 06:16:02.347223
#> 32: 2024-12-12 06:16:02.347223
#> 33: 2024-12-12 06:16:02.347223
#> 34: 2024-12-12 06:16:02.347223
#> 35: 2024-12-12 06:16:02.347223
#> 36: 2024-12-12 06:16:02.347223
#> 37: 2024-12-12 06:16:02.347223
#> 38: 2024-12-12 06:16:02.347223
#> 39: 2024-12-12 06:16:02.347223
#> 40: 2024-12-12 06:16:02.347223
#> 41: 2024-12-12 06:16:02.347223
#> 42: 2024-12-12 06:16:02.347223
#> 43: 2024-12-12 06:16:02.347223
#> 44: 2024-12-12 06:16:02.347223
#> 45: 2024-12-12 06:16:02.347223
#> 46: 2024-12-12 06:16:02.347223
#> 47: 2024-12-12 06:16:02.347223
#> 48: 2024-12-12 06:16:02.347223
#> 49: 2024-12-12 06:16:02.364922
#> 50: 2024-12-12 06:16:02.364922
#> 51: 2024-12-12 06:16:02.364922
#> 52: 2024-12-12 06:16:02.364922
#> 53: 2024-12-12 06:16:02.364922
#> 54: 2024-12-12 06:16:02.364922
#> 55: 2024-12-12 06:16:02.364922
#> 56: 2024-12-12 06:16:02.364922
#> 57: 2024-12-12 06:16:02.364922
#> 58: 2024-12-12 06:16:02.364922
#> 59: 2024-12-12 06:16:02.364922
#> 60: 2024-12-12 06:16:02.364922
#> 61: 2024-12-12 06:16:02.364922
#> 62: 2024-12-12 06:16:02.364922
#> 63: 2024-12-12 06:16:02.364922
#> 64: 2024-12-12 06:16:02.364922
#> 65: 2024-12-12 06:16:02.364922
#> 66: 2024-12-12 06:16:02.364922
#> 67: 2024-12-12 06:16:02.364922
#> 68: 2024-12-12 06:16:02.364922
#> 69: 2024-12-12 06:16:02.364922
#> 70: 2024-12-12 06:16:02.364922
#> 71: 2024-12-12 06:16:02.364922
#> 72: 2024-12-12 06:16:02.364922
#> 73: 2024-12-12 06:16:02.382993
#> 74: 2024-12-12 06:16:02.382993
#> 75: 2024-12-12 06:16:02.382993
#> 76: 2024-12-12 06:16:02.382993
#> 77: 2024-12-12 06:16:02.382993
#> 78: 2024-12-12 06:16:02.382993
#> 79: 2024-12-12 06:16:02.382993
#> 80: 2024-12-12 06:16:02.382993
#> 81: 2024-12-12 06:16:02.382993
#> 82: 2024-12-12 06:16:02.382993
#> 83: 2024-12-12 06:16:02.382993
#> 84: 2024-12-12 06:16:02.382993
#> 85: 2024-12-12 06:16:02.382993
#> 86: 2024-12-12 06:16:02.382993
#> 87: 2024-12-12 06:16:02.382993
#> 88: 2024-12-12 06:16:02.382993
#> 89: 2024-12-12 06:16:02.382993
#> 90: 2024-12-12 06:16:02.382993
#> 91: 2024-12-12 06:16:02.382993
#> 92: 2024-12-12 06:16:02.382993
#> 93: 2024-12-12 06:16:02.382993
#> 94: 2024-12-12 06:16:02.382993
#> 95: 2024-12-12 06:16:02.382993
#> 96: 2024-12-12 06:16:02.382993
#>                    createdDate
cc(ask = FALSE, x = tmpDir)
#> No time provided; removing the most recent entry to the Cache
#> Cache size:
#> Total (including Rasters): 252 bytes
#> Selected objects (not including Rasters): 252 bytes
showCache(x = tmpDir) # most recent is gone
#> Cache size:
#> Total (including Rasters): 746 bytes
#> Selected objects (not including Rasters): 746 bytes
#>              cacheId              tagKey                   tagValue
#>               <char>              <char>                     <char>
#>  1: 422bae4ed2f770cc            function                      rnorm
#>  2: 422bae4ed2f770cc               class                    numeric
#>  3: 422bae4ed2f770cc         object.size                        984
#>  4: 422bae4ed2f770cc            accessed 2024-12-12 06:16:02.328365
#>  5: 422bae4ed2f770cc             inCloud                      FALSE
#>  6: 422bae4ed2f770cc            fromDisk                      FALSE
#>  7: 422bae4ed2f770cc          resultHash                           
#>  8: 422bae4ed2f770cc   elapsedTimeDigest           0.001799345 secs
#>  9: 422bae4ed2f770cc elapsedTimeFirstRun           7.43866e-05 secs
#> 10: 422bae4ed2f770cc      otherFunctions                 build_site
#> 11: 422bae4ed2f770cc      otherFunctions           build_site_local
#> 12: 422bae4ed2f770cc      otherFunctions            build_reference
#> 13: 422bae4ed2f770cc      otherFunctions         unwrap_purrr_error
#> 14: 422bae4ed2f770cc      otherFunctions                       map_
#> 15: 422bae4ed2f770cc      otherFunctions        with_indexed_errors
#> 16: 422bae4ed2f770cc      otherFunctions          call_with_cleanup
#> 17: 422bae4ed2f770cc      otherFunctions       data_reference_topic
#> 18: 422bae4ed2f770cc      otherFunctions               run_examples
#> 19: 422bae4ed2f770cc      otherFunctions         highlight_examples
#> 20: 422bae4ed2f770cc      otherFunctions              with_handlers
#> 21: 422bae4ed2f770cc           preDigest         n:853b1797f54b229c
#> 22: 422bae4ed2f770cc           preDigest      mean:c40c00762a0dac94
#> 23: 422bae4ed2f770cc           preDigest        sd:853b1797f54b229c
#> 24: 422bae4ed2f770cc           preDigest      .FUN:4f604aa46882b368
#> 25: 66299c88b3924bf4            function                      rnorm
#> 26: 66299c88b3924bf4               class                    numeric
#> 27: 66299c88b3924bf4         object.size                        992
#> 28: 66299c88b3924bf4            accessed 2024-12-12 06:16:02.346638
#> 29: 66299c88b3924bf4             inCloud                      FALSE
#> 30: 66299c88b3924bf4            fromDisk                      FALSE
#> 31: 66299c88b3924bf4          resultHash                           
#> 32: 66299c88b3924bf4   elapsedTimeDigest           0.001669884 secs
#> 33: 66299c88b3924bf4 elapsedTimeFirstRun          6.389618e-05 secs
#> 34: 66299c88b3924bf4      otherFunctions                 build_site
#> 35: 66299c88b3924bf4      otherFunctions           build_site_local
#> 36: 66299c88b3924bf4      otherFunctions            build_reference
#> 37: 66299c88b3924bf4      otherFunctions         unwrap_purrr_error
#> 38: 66299c88b3924bf4      otherFunctions                       map_
#> 39: 66299c88b3924bf4      otherFunctions        with_indexed_errors
#> 40: 66299c88b3924bf4      otherFunctions          call_with_cleanup
#> 41: 66299c88b3924bf4      otherFunctions       data_reference_topic
#> 42: 66299c88b3924bf4      otherFunctions               run_examples
#> 43: 66299c88b3924bf4      otherFunctions         highlight_examples
#> 44: 66299c88b3924bf4      otherFunctions              with_handlers
#> 45: 66299c88b3924bf4           preDigest         n:82dc709f2b91918a
#> 46: 66299c88b3924bf4           preDigest      mean:c40c00762a0dac94
#> 47: 66299c88b3924bf4           preDigest        sd:853b1797f54b229c
#> 48: 66299c88b3924bf4           preDigest      .FUN:4f604aa46882b368
#> 49: 6724ef06570739aa            function                      rnorm
#> 50: 6724ef06570739aa               class                    numeric
#> 51: 6724ef06570739aa         object.size                       1008
#> 52: 6724ef06570739aa            accessed 2024-12-12 06:16:02.364266
#> 53: 6724ef06570739aa             inCloud                      FALSE
#> 54: 6724ef06570739aa            fromDisk                      FALSE
#> 55: 6724ef06570739aa          resultHash                           
#> 56: 6724ef06570739aa   elapsedTimeDigest            0.00168252 secs
#> 57: 6724ef06570739aa elapsedTimeFirstRun          6.103516e-05 secs
#> 58: 6724ef06570739aa      otherFunctions                 build_site
#> 59: 6724ef06570739aa      otherFunctions           build_site_local
#> 60: 6724ef06570739aa      otherFunctions            build_reference
#> 61: 6724ef06570739aa      otherFunctions         unwrap_purrr_error
#> 62: 6724ef06570739aa      otherFunctions                       map_
#> 63: 6724ef06570739aa      otherFunctions        with_indexed_errors
#> 64: 6724ef06570739aa      otherFunctions          call_with_cleanup
#> 65: 6724ef06570739aa      otherFunctions       data_reference_topic
#> 66: 6724ef06570739aa      otherFunctions               run_examples
#> 67: 6724ef06570739aa      otherFunctions         highlight_examples
#> 68: 6724ef06570739aa      otherFunctions              with_handlers
#> 69: 6724ef06570739aa           preDigest         n:7f12988bd88a0fb8
#> 70: 6724ef06570739aa           preDigest      mean:c40c00762a0dac94
#> 71: 6724ef06570739aa           preDigest        sd:853b1797f54b229c
#> 72: 6724ef06570739aa           preDigest      .FUN:4f604aa46882b368
#>              cacheId              tagKey                   tagValue
#>                    createdDate
#>                         <char>
#>  1: 2024-12-12 06:16:02.329004
#>  2: 2024-12-12 06:16:02.329004
#>  3: 2024-12-12 06:16:02.329004
#>  4: 2024-12-12 06:16:02.329004
#>  5: 2024-12-12 06:16:02.329004
#>  6: 2024-12-12 06:16:02.329004
#>  7: 2024-12-12 06:16:02.329004
#>  8: 2024-12-12 06:16:02.329004
#>  9: 2024-12-12 06:16:02.329004
#> 10: 2024-12-12 06:16:02.329004
#> 11: 2024-12-12 06:16:02.329004
#> 12: 2024-12-12 06:16:02.329004
#> 13: 2024-12-12 06:16:02.329004
#> 14: 2024-12-12 06:16:02.329004
#> 15: 2024-12-12 06:16:02.329004
#> 16: 2024-12-12 06:16:02.329004
#> 17: 2024-12-12 06:16:02.329004
#> 18: 2024-12-12 06:16:02.329004
#> 19: 2024-12-12 06:16:02.329004
#> 20: 2024-12-12 06:16:02.329004
#> 21: 2024-12-12 06:16:02.329004
#> 22: 2024-12-12 06:16:02.329004
#> 23: 2024-12-12 06:16:02.329004
#> 24: 2024-12-12 06:16:02.329004
#> 25: 2024-12-12 06:16:02.347223
#> 26: 2024-12-12 06:16:02.347223
#> 27: 2024-12-12 06:16:02.347223
#> 28: 2024-12-12 06:16:02.347223
#> 29: 2024-12-12 06:16:02.347223
#> 30: 2024-12-12 06:16:02.347223
#> 31: 2024-12-12 06:16:02.347223
#> 32: 2024-12-12 06:16:02.347223
#> 33: 2024-12-12 06:16:02.347223
#> 34: 2024-12-12 06:16:02.347223
#> 35: 2024-12-12 06:16:02.347223
#> 36: 2024-12-12 06:16:02.347223
#> 37: 2024-12-12 06:16:02.347223
#> 38: 2024-12-12 06:16:02.347223
#> 39: 2024-12-12 06:16:02.347223
#> 40: 2024-12-12 06:16:02.347223
#> 41: 2024-12-12 06:16:02.347223
#> 42: 2024-12-12 06:16:02.347223
#> 43: 2024-12-12 06:16:02.347223
#> 44: 2024-12-12 06:16:02.347223
#> 45: 2024-12-12 06:16:02.347223
#> 46: 2024-12-12 06:16:02.347223
#> 47: 2024-12-12 06:16:02.347223
#> 48: 2024-12-12 06:16:02.347223
#> 49: 2024-12-12 06:16:02.364922
#> 50: 2024-12-12 06:16:02.364922
#> 51: 2024-12-12 06:16:02.364922
#> 52: 2024-12-12 06:16:02.364922
#> 53: 2024-12-12 06:16:02.364922
#> 54: 2024-12-12 06:16:02.364922
#> 55: 2024-12-12 06:16:02.364922
#> 56: 2024-12-12 06:16:02.364922
#> 57: 2024-12-12 06:16:02.364922
#> 58: 2024-12-12 06:16:02.364922
#> 59: 2024-12-12 06:16:02.364922
#> 60: 2024-12-12 06:16:02.364922
#> 61: 2024-12-12 06:16:02.364922
#> 62: 2024-12-12 06:16:02.364922
#> 63: 2024-12-12 06:16:02.364922
#> 64: 2024-12-12 06:16:02.364922
#> 65: 2024-12-12 06:16:02.364922
#> 66: 2024-12-12 06:16:02.364922
#> 67: 2024-12-12 06:16:02.364922
#> 68: 2024-12-12 06:16:02.364922
#> 69: 2024-12-12 06:16:02.364922
#> 70: 2024-12-12 06:16:02.364922
#> 71: 2024-12-12 06:16:02.364922
#> 72: 2024-12-12 06:16:02.364922
#>                    createdDate
cc(thisTime, ask = FALSE, x = tmpDir)
#> Cache size:
#> Total (including Rasters): 500 bytes
#> Selected objects (not including Rasters): 500 bytes
showCache(x = tmpDir) # all those after thisTime gone, i.e., only 1 left
#> Cache size:
#> Total (including Rasters): 246 bytes
#> Selected objects (not including Rasters): 246 bytes
#>              cacheId              tagKey                   tagValue
#>               <char>              <char>                     <char>
#>  1: 422bae4ed2f770cc            function                      rnorm
#>  2: 422bae4ed2f770cc               class                    numeric
#>  3: 422bae4ed2f770cc         object.size                        984
#>  4: 422bae4ed2f770cc            accessed 2024-12-12 06:16:02.328365
#>  5: 422bae4ed2f770cc             inCloud                      FALSE
#>  6: 422bae4ed2f770cc            fromDisk                      FALSE
#>  7: 422bae4ed2f770cc          resultHash                           
#>  8: 422bae4ed2f770cc   elapsedTimeDigest           0.001799345 secs
#>  9: 422bae4ed2f770cc elapsedTimeFirstRun           7.43866e-05 secs
#> 10: 422bae4ed2f770cc      otherFunctions                 build_site
#> 11: 422bae4ed2f770cc      otherFunctions           build_site_local
#> 12: 422bae4ed2f770cc      otherFunctions            build_reference
#> 13: 422bae4ed2f770cc      otherFunctions         unwrap_purrr_error
#> 14: 422bae4ed2f770cc      otherFunctions                       map_
#> 15: 422bae4ed2f770cc      otherFunctions        with_indexed_errors
#> 16: 422bae4ed2f770cc      otherFunctions          call_with_cleanup
#> 17: 422bae4ed2f770cc      otherFunctions       data_reference_topic
#> 18: 422bae4ed2f770cc      otherFunctions               run_examples
#> 19: 422bae4ed2f770cc      otherFunctions         highlight_examples
#> 20: 422bae4ed2f770cc      otherFunctions              with_handlers
#> 21: 422bae4ed2f770cc           preDigest         n:853b1797f54b229c
#> 22: 422bae4ed2f770cc           preDigest      mean:c40c00762a0dac94
#> 23: 422bae4ed2f770cc           preDigest        sd:853b1797f54b229c
#> 24: 422bae4ed2f770cc           preDigest      .FUN:4f604aa46882b368
#>              cacheId              tagKey                   tagValue
#>                    createdDate
#>                         <char>
#>  1: 2024-12-12 06:16:02.329004
#>  2: 2024-12-12 06:16:02.329004
#>  3: 2024-12-12 06:16:02.329004
#>  4: 2024-12-12 06:16:02.329004
#>  5: 2024-12-12 06:16:02.329004
#>  6: 2024-12-12 06:16:02.329004
#>  7: 2024-12-12 06:16:02.329004
#>  8: 2024-12-12 06:16:02.329004
#>  9: 2024-12-12 06:16:02.329004
#> 10: 2024-12-12 06:16:02.329004
#> 11: 2024-12-12 06:16:02.329004
#> 12: 2024-12-12 06:16:02.329004
#> 13: 2024-12-12 06:16:02.329004
#> 14: 2024-12-12 06:16:02.329004
#> 15: 2024-12-12 06:16:02.329004
#> 16: 2024-12-12 06:16:02.329004
#> 17: 2024-12-12 06:16:02.329004
#> 18: 2024-12-12 06:16:02.329004
#> 19: 2024-12-12 06:16:02.329004
#> 20: 2024-12-12 06:16:02.329004
#> 21: 2024-12-12 06:16:02.329004
#> 22: 2024-12-12 06:16:02.329004
#> 23: 2024-12-12 06:16:02.329004
#> 24: 2024-12-12 06:16:02.329004
#>                    createdDate
cc(ask = FALSE, x = tmpDir) # Cache is
#> No time provided; removing the most recent entry to the Cache
#> Cache size:
#> Total (including Rasters): 246 bytes
#> Selected objects (not including Rasters): 246 bytes
cc(ask = FALSE, x = tmpDir) # Cache is already empty
#> No time provided; removing the most recent entry to the Cache
#> Cache already empty