Examining and modifying the cache

These are convenience wrappers around DBI package functions. They allow the user a bit of control over what is being cached.

clearCache(
  x,
  userTags = character(),
  after = NULL,
  before = NULL,
  ask = getOption("reproducible.ask"),
  useCloud = FALSE,
  cloudFolderID = getOption("reproducible.cloudFolderID", NULL),
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  ...
)

# S4 method for ANY
clearCache(
  x,
  userTags = character(),
  after = NULL,
  before = NULL,
  ask = getOption("reproducible.ask"),
  useCloud = FALSE,
  cloudFolderID = getOption("reproducible.cloudFolderID", NULL),
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  ...
)

cc(secs, ..., verbose = getOption("reproducible.verbose"))

showCache(
  x,
  userTags = character(),
  after = NULL,
  before = NULL,
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  ...
)

# S4 method for ANY
showCache(
  x,
  userTags = character(),
  after = NULL,
  before = NULL,
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  ...
)

keepCache(
  x,
  userTags = character(),
  after = NULL,
  before = NULL,
  ask = getOption("reproducible.ask"),
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  ...
)

# S4 method for ANY
keepCache(
  x,
  userTags = character(),
  after = NULL,
  before = NULL,
  ask = getOption("reproducible.ask"),
  drv = getDrv(getOption("reproducible.drv", NULL)),
  conn = getOption("reproducible.conn", NULL),
  verbose = getOption("reproducible.verbose"),
  ...
)

Arguments

x

A simList or a directory containing a valid Cache repository. Note: For compatibility with Cache argument, cachePath can also be used instead of x, though x will take precedence.

userTags

Character vector. If used, this will be used in place of the after and before. Specifying one or more userTag here will clear all objects that match those tags. Matching is via regular expression, meaning partial matches will work unless strict beginning (^) and end ($) of string characters are used. Matching will be against any of the 3 columns returned by showCache(), i.e., artifact, tagValue or tagName. Also, if length(userTags) > 1, then matching is by and. For or matching, use | in a single character string. See examples.

after

A time (POSIX, character understandable by data.table). Objects cached after this time will be shown or deleted.

before

A time (POSIX, character understandable by data.table). Objects cached before this time will be shown or deleted.

ask

Logical. If FALSE, then it will not ask to confirm deletions using clearCache or keepCache. Default is TRUE

useCloud

Logical. If TRUE, then every object that is deleted locally will also be deleted in the cloudFolderID, if it is non-NULL

cloudFolderID

A googledrive dribble of a folder, e.g., using drive_mkdir(). If left as NULL, the function will create a cloud folder with name from last two folder levels of the cachePath path, : paste0(basename(dirname(cachePath)), "_", basename(cachePath)). This cloudFolderID will be added to options("reproducible.cloudFolderID"), but this will not persist across sessions. If this is a character string, it will treat this as a folder name to create or use on GoogleDrive.

drv

if using a database backend, drv must be an object that inherits from DBIDriver e.g., from package RSQLite, e.g., SQLite

conn

an optional DBIConnection object, as returned by dbConnect().

verbose

Numeric, -1 silent (where possible), 0 being very quiet, 1 showing more messaging, 2 being more messaging, etc. Default is 1. Above 3 will output much more information about the internals of Caching, which may help diagnose Caching challenges. Can set globally with an option, e.g., options('reproducible.verbose' = 0) to reduce to minimal

...

Other arguments. Currently, regexp, a logical, can be provided. This must be TRUE if the use is passing a regular expression. Otherwise, userTags will need to be exact matches. Default is missing, which is the same as TRUE. If there are errors due to regular expression problem, try FALSE. For cc, it is passed to clearCache, e.g., ask, userTags. For showCache, it can also be sorted = FALSE to return the object unsorted.

secs

Currently 3 options: the number of seconds to pass to clearCache(after = secs), a POSIXct time e.g., from Sys.time(), or missing. If missing, the default, then it will delete the most recent entry in the Cache.

Value

Will clear all objects (or those that match userTags, or those between after or before) from the repository located at cachePath of the sim object, if sim is provided, or located in cachePath. Invisibly returns a data.table of the removed items.

Details

If neither after or before are provided, nor userTags, then all objects will be removed. If both after and before are specified, then all objects between after and before will be deleted. If userTags is used, this will override after or before.

cc(secs) is just a shortcut for clearCache(repo = currentRepo, after = secs), i.e., to remove any cache entries touched in the last secs seconds. Since, secs can be missing, this is also be a shorthand for "remove most recent entry from the cache".

clearCache

remove items from the cache based on their userTag or times values.

keepCache

remove all cached items except those based on certain userTags or times values.

showCache

display the contents of the cache.

By default the return of showCache is sorted by cacheId. For convenience, a user can optionally have it unsorted (passing sorted = FALSE), which may be noticeably faster when the cache is large (> 1e4 entries).

Note

If the cache is larger than 10MB, and clearCache is used, there will be a message and a pause, if interactive, to prevent accidentally deleting of a large cache repository.

See also

Examples

data.table::setDTthreads(2)

tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir, ask = FALSE), silent = TRUE) # just to make sure it is clear

# Basic use
ranNumsA <- Cache(rnorm, 10, 16, cachePath = tmpDir)

# All same
ranNumsB <- Cache(rnorm, 10, 16, cachePath = tmpDir) # recovers cached copy
#>   ...(Object to retrieve (fn: rnorm, 4fca280cda001fc9.rds))
#>      loaded cached result from previous rnorm call
ranNumsD <- Cache(quote(rnorm(n = 10, 16)), cachePath = tmpDir) # recovers cached copy
#>   ...(Object to retrieve (fn: quote, 4fca280cda001fc9.rds))
#>      loaded cached result from previous quote call

# Any minor change makes it different
ranNumsE <- Cache(rnorm, 10, 6, cachePath = tmpDir) # different

## Example 1: basic cache use with tags
ranNumsA <- Cache(rnorm, 4, cachePath = tmpDir, userTags = "objectName:a")
ranNumsB <- Cache(runif, 4, cachePath = tmpDir, userTags = "objectName:b")
ranNumsC <- Cache(runif, 40, cachePath = tmpDir, userTags = "objectName:b")

showCache(tmpDir, userTags = c("objectName"))
#> Cache size: 
#>   Total (including Rasters): 828 bytes
#>   Selected objects (not including Rasters): 828 bytes
#>              cacheId              tagKey                   tagValue
#>  1: a0603ff09bb4ce83          objectName                          b
#>  2: a0603ff09bb4ce83            function                      runif
#>  3: a0603ff09bb4ce83               class                    numeric
#>  4: a0603ff09bb4ce83         object.size                       1296
#>  5: a0603ff09bb4ce83            accessed 2023-11-22 17:00:31.727051
#>  6: a0603ff09bb4ce83             inCloud                      FALSE
#>  7: a0603ff09bb4ce83            fromDisk                      FALSE
#>  8: a0603ff09bb4ce83          resultHash                           
#>  9: a0603ff09bb4ce83   elapsedTimeDigest           0.001859903 secs
#> 10: a0603ff09bb4ce83 elapsedTimeFirstRun           2.69413e-05 secs
#> 11: a0603ff09bb4ce83      otherFunctions                 build_site
#> 12: a0603ff09bb4ce83      otherFunctions           build_site_local
#> 13: a0603ff09bb4ce83      otherFunctions            build_reference
#> 14: a0603ff09bb4ce83      otherFunctions                       map_
#> 15: a0603ff09bb4ce83      otherFunctions        with_indexed_errors
#> 16: a0603ff09bb4ce83      otherFunctions          call_with_cleanup
#> 17: a0603ff09bb4ce83      otherFunctions       data_reference_topic
#> 18: a0603ff09bb4ce83      otherFunctions               run_examples
#> 19: a0603ff09bb4ce83      otherFunctions         highlight_examples
#> 20: a0603ff09bb4ce83      otherFunctions                  timing_fn
#> 21: a0603ff09bb4ce83      otherFunctions                     handle
#> 22: a0603ff09bb4ce83      otherFunctions                        try
#> 23: a0603ff09bb4ce83           preDigest         n:1393aef18608c8be
#> 24: a0603ff09bb4ce83           preDigest       min:c40c00762a0dac94
#> 25: a0603ff09bb4ce83           preDigest       max:853b1797f54b229c
#> 26: a0603ff09bb4ce83           preDigest      .FUN:881ec847b7161f3c
#> 27: ad0ea27476c50b66          objectName                          a
#> 28: ad0ea27476c50b66            function                      rnorm
#> 29: ad0ea27476c50b66               class                    numeric
#> 30: ad0ea27476c50b66         object.size                       1008
#> 31: ad0ea27476c50b66            accessed 2023-11-22 17:00:31.693343
#> 32: ad0ea27476c50b66             inCloud                      FALSE
#> 33: ad0ea27476c50b66            fromDisk                      FALSE
#> 34: ad0ea27476c50b66          resultHash                           
#> 35: ad0ea27476c50b66   elapsedTimeDigest           0.001888037 secs
#> 36: ad0ea27476c50b66 elapsedTimeFirstRun          2.813339e-05 secs
#> 37: ad0ea27476c50b66      otherFunctions                 build_site
#> 38: ad0ea27476c50b66      otherFunctions           build_site_local
#> 39: ad0ea27476c50b66      otherFunctions            build_reference
#> 40: ad0ea27476c50b66      otherFunctions                       map_
#> 41: ad0ea27476c50b66      otherFunctions        with_indexed_errors
#> 42: ad0ea27476c50b66      otherFunctions          call_with_cleanup
#> 43: ad0ea27476c50b66      otherFunctions       data_reference_topic
#> 44: ad0ea27476c50b66      otherFunctions               run_examples
#> 45: ad0ea27476c50b66      otherFunctions         highlight_examples
#> 46: ad0ea27476c50b66      otherFunctions                  timing_fn
#> 47: ad0ea27476c50b66      otherFunctions                     handle
#> 48: ad0ea27476c50b66      otherFunctions                        try
#> 49: ad0ea27476c50b66           preDigest         n:7eef4eae85fd9229
#> 50: ad0ea27476c50b66           preDigest      mean:c40c00762a0dac94
#> 51: ad0ea27476c50b66           preDigest        sd:853b1797f54b229c
#> 52: ad0ea27476c50b66           preDigest      .FUN:4f604aa46882b368
#> 53: deaa37372f85861b          objectName                          b
#> 54: deaa37372f85861b            function                      runif
#> 55: deaa37372f85861b               class                    numeric
#> 56: deaa37372f85861b         object.size                       1008
#> 57: deaa37372f85861b            accessed 2023-11-22 17:00:31.710653
#> 58: deaa37372f85861b             inCloud                      FALSE
#> 59: deaa37372f85861b            fromDisk                      FALSE
#> 60: deaa37372f85861b          resultHash                           
#> 61: deaa37372f85861b   elapsedTimeDigest           0.001847267 secs
#> 62: deaa37372f85861b elapsedTimeFirstRun          2.670288e-05 secs
#> 63: deaa37372f85861b      otherFunctions                 build_site
#> 64: deaa37372f85861b      otherFunctions           build_site_local
#> 65: deaa37372f85861b      otherFunctions            build_reference
#> 66: deaa37372f85861b      otherFunctions                       map_
#> 67: deaa37372f85861b      otherFunctions        with_indexed_errors
#> 68: deaa37372f85861b      otherFunctions          call_with_cleanup
#> 69: deaa37372f85861b      otherFunctions       data_reference_topic
#> 70: deaa37372f85861b      otherFunctions               run_examples
#> 71: deaa37372f85861b      otherFunctions         highlight_examples
#> 72: deaa37372f85861b      otherFunctions                  timing_fn
#> 73: deaa37372f85861b      otherFunctions                     handle
#> 74: deaa37372f85861b      otherFunctions                        try
#> 75: deaa37372f85861b           preDigest         n:7eef4eae85fd9229
#> 76: deaa37372f85861b           preDigest       min:c40c00762a0dac94
#> 77: deaa37372f85861b           preDigest       max:853b1797f54b229c
#> 78: deaa37372f85861b           preDigest      .FUN:881ec847b7161f3c
#>              cacheId              tagKey                   tagValue
#>                    createdDate
#>  1: 2023-11-22 17:00:31.727535
#>  2: 2023-11-22 17:00:31.727535
#>  3: 2023-11-22 17:00:31.727535
#>  4: 2023-11-22 17:00:31.727535
#>  5: 2023-11-22 17:00:31.727535
#>  6: 2023-11-22 17:00:31.727535
#>  7: 2023-11-22 17:00:31.727535
#>  8: 2023-11-22 17:00:31.727535
#>  9: 2023-11-22 17:00:31.727535
#> 10: 2023-11-22 17:00:31.727535
#> 11: 2023-11-22 17:00:31.727535
#> 12: 2023-11-22 17:00:31.727535
#> 13: 2023-11-22 17:00:31.727535
#> 14: 2023-11-22 17:00:31.727535
#> 15: 2023-11-22 17:00:31.727535
#> 16: 2023-11-22 17:00:31.727535
#> 17: 2023-11-22 17:00:31.727535
#> 18: 2023-11-22 17:00:31.727535
#> 19: 2023-11-22 17:00:31.727535
#> 20: 2023-11-22 17:00:31.727535
#> 21: 2023-11-22 17:00:31.727535
#> 22: 2023-11-22 17:00:31.727535
#> 23: 2023-11-22 17:00:31.727535
#> 24: 2023-11-22 17:00:31.727535
#> 25: 2023-11-22 17:00:31.727535
#> 26: 2023-11-22 17:00:31.727535
#> 27: 2023-11-22 17:00:31.693831
#> 28: 2023-11-22 17:00:31.693831
#> 29: 2023-11-22 17:00:31.693831
#> 30: 2023-11-22 17:00:31.693831
#> 31: 2023-11-22 17:00:31.693831
#> 32: 2023-11-22 17:00:31.693831
#> 33: 2023-11-22 17:00:31.693831
#> 34: 2023-11-22 17:00:31.693831
#> 35: 2023-11-22 17:00:31.693831
#> 36: 2023-11-22 17:00:31.693831
#> 37: 2023-11-22 17:00:31.693831
#> 38: 2023-11-22 17:00:31.693831
#> 39: 2023-11-22 17:00:31.693831
#> 40: 2023-11-22 17:00:31.693831
#> 41: 2023-11-22 17:00:31.693831
#> 42: 2023-11-22 17:00:31.693831
#> 43: 2023-11-22 17:00:31.693831
#> 44: 2023-11-22 17:00:31.693831
#> 45: 2023-11-22 17:00:31.693831
#> 46: 2023-11-22 17:00:31.693831
#> 47: 2023-11-22 17:00:31.693831
#> 48: 2023-11-22 17:00:31.693831
#> 49: 2023-11-22 17:00:31.693831
#> 50: 2023-11-22 17:00:31.693831
#> 51: 2023-11-22 17:00:31.693831
#> 52: 2023-11-22 17:00:31.693831
#> 53: 2023-11-22 17:00:31.711149
#> 54: 2023-11-22 17:00:31.711149
#> 55: 2023-11-22 17:00:31.711149
#> 56: 2023-11-22 17:00:31.711149
#> 57: 2023-11-22 17:00:31.711149
#> 58: 2023-11-22 17:00:31.711149
#> 59: 2023-11-22 17:00:31.711149
#> 60: 2023-11-22 17:00:31.711149
#> 61: 2023-11-22 17:00:31.711149
#> 62: 2023-11-22 17:00:31.711149
#> 63: 2023-11-22 17:00:31.711149
#> 64: 2023-11-22 17:00:31.711149
#> 65: 2023-11-22 17:00:31.711149
#> 66: 2023-11-22 17:00:31.711149
#> 67: 2023-11-22 17:00:31.711149
#> 68: 2023-11-22 17:00:31.711149
#> 69: 2023-11-22 17:00:31.711149
#> 70: 2023-11-22 17:00:31.711149
#> 71: 2023-11-22 17:00:31.711149
#> 72: 2023-11-22 17:00:31.711149
#> 73: 2023-11-22 17:00:31.711149
#> 74: 2023-11-22 17:00:31.711149
#> 75: 2023-11-22 17:00:31.711149
#> 76: 2023-11-22 17:00:31.711149
#> 77: 2023-11-22 17:00:31.711149
#> 78: 2023-11-22 17:00:31.711149
#>                    createdDate
showCache(tmpDir, userTags = c("^a$")) # regular expression ... "a" exactly
#> Cache size: 
#>   Total (including Rasters): 252 bytes
#>   Selected objects (not including Rasters): 252 bytes
#>              cacheId              tagKey                   tagValue
#>  1: ad0ea27476c50b66          objectName                          a
#>  2: ad0ea27476c50b66            function                      rnorm
#>  3: ad0ea27476c50b66               class                    numeric
#>  4: ad0ea27476c50b66         object.size                       1008
#>  5: ad0ea27476c50b66            accessed 2023-11-22 17:00:31.693343
#>  6: ad0ea27476c50b66             inCloud                      FALSE
#>  7: ad0ea27476c50b66            fromDisk                      FALSE
#>  8: ad0ea27476c50b66          resultHash                           
#>  9: ad0ea27476c50b66   elapsedTimeDigest           0.001888037 secs
#> 10: ad0ea27476c50b66 elapsedTimeFirstRun          2.813339e-05 secs
#> 11: ad0ea27476c50b66      otherFunctions                 build_site
#> 12: ad0ea27476c50b66      otherFunctions           build_site_local
#> 13: ad0ea27476c50b66      otherFunctions            build_reference
#> 14: ad0ea27476c50b66      otherFunctions                       map_
#> 15: ad0ea27476c50b66      otherFunctions        with_indexed_errors
#> 16: ad0ea27476c50b66      otherFunctions          call_with_cleanup
#> 17: ad0ea27476c50b66      otherFunctions       data_reference_topic
#> 18: ad0ea27476c50b66      otherFunctions               run_examples
#> 19: ad0ea27476c50b66      otherFunctions         highlight_examples
#> 20: ad0ea27476c50b66      otherFunctions                  timing_fn
#> 21: ad0ea27476c50b66      otherFunctions                     handle
#> 22: ad0ea27476c50b66      otherFunctions                        try
#> 23: ad0ea27476c50b66           preDigest         n:7eef4eae85fd9229
#> 24: ad0ea27476c50b66           preDigest      mean:c40c00762a0dac94
#> 25: ad0ea27476c50b66           preDigest        sd:853b1797f54b229c
#> 26: ad0ea27476c50b66           preDigest      .FUN:4f604aa46882b368
#>              cacheId              tagKey                   tagValue
#>                    createdDate
#>  1: 2023-11-22 17:00:31.693831
#>  2: 2023-11-22 17:00:31.693831
#>  3: 2023-11-22 17:00:31.693831
#>  4: 2023-11-22 17:00:31.693831
#>  5: 2023-11-22 17:00:31.693831
#>  6: 2023-11-22 17:00:31.693831
#>  7: 2023-11-22 17:00:31.693831
#>  8: 2023-11-22 17:00:31.693831
#>  9: 2023-11-22 17:00:31.693831
#> 10: 2023-11-22 17:00:31.693831
#> 11: 2023-11-22 17:00:31.693831
#> 12: 2023-11-22 17:00:31.693831
#> 13: 2023-11-22 17:00:31.693831
#> 14: 2023-11-22 17:00:31.693831
#> 15: 2023-11-22 17:00:31.693831
#> 16: 2023-11-22 17:00:31.693831
#> 17: 2023-11-22 17:00:31.693831
#> 18: 2023-11-22 17:00:31.693831
#> 19: 2023-11-22 17:00:31.693831
#> 20: 2023-11-22 17:00:31.693831
#> 21: 2023-11-22 17:00:31.693831
#> 22: 2023-11-22 17:00:31.693831
#> 23: 2023-11-22 17:00:31.693831
#> 24: 2023-11-22 17:00:31.693831
#> 25: 2023-11-22 17:00:31.693831
#> 26: 2023-11-22 17:00:31.693831
#>                    createdDate

# Fine control of cache elements -- pick out only the large runif object, and remove it
cache1 <- showCache(tmpDir, userTags = c("runif")) # show only cached objects made during runif
#> Cache size: 
#>   Total (including Rasters): 576 bytes
#>   Selected objects (not including Rasters): 576 bytes
toRemove <- cache1[tagKey == "object.size"][as.numeric(tagValue) > 700]$cacheId
clearCache(tmpDir, userTags = toRemove, ask = FALSE)
#> Cache size: 
#>   Total (including Rasters): 0 bytes
#>   Selected objects (not including Rasters): 0 bytes
cacheAfter <- showCache(tmpDir, userTags = c("runif")) # Only the small one is left
#> Cache size: 
#>   Total (including Rasters): 576 bytes
#>   Selected objects (not including Rasters): 576 bytes

data.table::setDTthreads(2)
tmpDir <- file.path(tempdir(), "reproducible_examples", "Cache")
try(clearCache(tmpDir, ask = FALSE), silent = TRUE) # just to make sure it is clear

Cache(rnorm, 1, cachePath = tmpDir)
#> [1] 1.238532
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:422bae4ed2f770cc"
#> attr(,"call")
#> [1] ""
thisTime <- Sys.time()
Cache(rnorm, 2, cachePath = tmpDir)
#> [1] -0.3096662 -1.1562335
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:66299c88b3924bf4"
#> attr(,"call")
#> [1] ""
Cache(rnorm, 3, cachePath = tmpDir)
#> [1] -0.5636132  0.6491986 -1.3421304
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:6724ef06570739aa"
#> attr(,"call")
#> [1] ""
Cache(rnorm, 4, cachePath = tmpDir)
#> [1] -0.9056071 -0.8161012  0.6995249 -1.4031589
#> attr(,".Cache")
#> attr(,".Cache")$newCache
#> [1] TRUE
#> 
#> attr(,"tags")
#> [1] "cacheId:ad0ea27476c50b66"
#> attr(,"call")
#> [1] ""
showCache(x = tmpDir) # shows all 4 entries
#> Cache size: 
#>   Total (including Rasters): 998 bytes
#>   Selected objects (not including Rasters): 998 bytes
#>               cacheId              tagKey                   tagValue
#>   1: 422bae4ed2f770cc            function                      rnorm
#>   2: 422bae4ed2f770cc               class                    numeric
#>   3: 422bae4ed2f770cc         object.size                        984
#>   4: 422bae4ed2f770cc            accessed 2023-11-22 17:00:31.811389
#>   5: 422bae4ed2f770cc             inCloud                      FALSE
#>   6: 422bae4ed2f770cc            fromDisk                      FALSE
#>   7: 422bae4ed2f770cc          resultHash                           
#>   8: 422bae4ed2f770cc   elapsedTimeDigest           0.001922607 secs
#>   9: 422bae4ed2f770cc elapsedTimeFirstRun          2.908707e-05 secs
#>  10: 422bae4ed2f770cc      otherFunctions                 build_site
#>  11: 422bae4ed2f770cc      otherFunctions           build_site_local
#>  12: 422bae4ed2f770cc      otherFunctions            build_reference
#>  13: 422bae4ed2f770cc      otherFunctions                       map_
#>  14: 422bae4ed2f770cc      otherFunctions        with_indexed_errors
#>  15: 422bae4ed2f770cc      otherFunctions          call_with_cleanup
#>  16: 422bae4ed2f770cc      otherFunctions       data_reference_topic
#>  17: 422bae4ed2f770cc      otherFunctions               run_examples
#>  18: 422bae4ed2f770cc      otherFunctions         highlight_examples
#>  19: 422bae4ed2f770cc      otherFunctions                  timing_fn
#>  20: 422bae4ed2f770cc      otherFunctions                     handle
#>  21: 422bae4ed2f770cc      otherFunctions                        try
#>  22: 422bae4ed2f770cc           preDigest         n:853b1797f54b229c
#>  23: 422bae4ed2f770cc           preDigest      mean:c40c00762a0dac94
#>  24: 422bae4ed2f770cc           preDigest        sd:853b1797f54b229c
#>  25: 422bae4ed2f770cc           preDigest      .FUN:4f604aa46882b368
#>  26: 66299c88b3924bf4            function                      rnorm
#>  27: 66299c88b3924bf4               class                    numeric
#>  28: 66299c88b3924bf4         object.size                        992
#>  29: 66299c88b3924bf4            accessed 2023-11-22 17:00:31.829272
#>  30: 66299c88b3924bf4             inCloud                      FALSE
#>  31: 66299c88b3924bf4            fromDisk                      FALSE
#>  32: 66299c88b3924bf4          resultHash                           
#>  33: 66299c88b3924bf4   elapsedTimeDigest           0.001885653 secs
#>  34: 66299c88b3924bf4 elapsedTimeFirstRun          2.932549e-05 secs
#>  35: 66299c88b3924bf4      otherFunctions                 build_site
#>  36: 66299c88b3924bf4      otherFunctions           build_site_local
#>  37: 66299c88b3924bf4      otherFunctions            build_reference
#>  38: 66299c88b3924bf4      otherFunctions                       map_
#>  39: 66299c88b3924bf4      otherFunctions        with_indexed_errors
#>  40: 66299c88b3924bf4      otherFunctions          call_with_cleanup
#>  41: 66299c88b3924bf4      otherFunctions       data_reference_topic
#>  42: 66299c88b3924bf4      otherFunctions               run_examples
#>  43: 66299c88b3924bf4      otherFunctions         highlight_examples
#>  44: 66299c88b3924bf4      otherFunctions                  timing_fn
#>  45: 66299c88b3924bf4      otherFunctions                     handle
#>  46: 66299c88b3924bf4      otherFunctions                        try
#>  47: 66299c88b3924bf4           preDigest         n:82dc709f2b91918a
#>  48: 66299c88b3924bf4           preDigest      mean:c40c00762a0dac94
#>  49: 66299c88b3924bf4           preDigest        sd:853b1797f54b229c
#>  50: 66299c88b3924bf4           preDigest      .FUN:4f604aa46882b368
#>  51: 6724ef06570739aa            function                      rnorm
#>  52: 6724ef06570739aa               class                    numeric
#>  53: 6724ef06570739aa         object.size                       1008
#>  54: 6724ef06570739aa            accessed 2023-11-22 17:00:31.845715
#>  55: 6724ef06570739aa             inCloud                      FALSE
#>  56: 6724ef06570739aa            fromDisk                      FALSE
#>  57: 6724ef06570739aa          resultHash                           
#>  58: 6724ef06570739aa   elapsedTimeDigest           0.001798868 secs
#>  59: 6724ef06570739aa elapsedTimeFirstRun          2.717972e-05 secs
#>  60: 6724ef06570739aa      otherFunctions                 build_site
#>  61: 6724ef06570739aa      otherFunctions           build_site_local
#>  62: 6724ef06570739aa      otherFunctions            build_reference
#>  63: 6724ef06570739aa      otherFunctions                       map_
#>  64: 6724ef06570739aa      otherFunctions        with_indexed_errors
#>  65: 6724ef06570739aa      otherFunctions          call_with_cleanup
#>  66: 6724ef06570739aa      otherFunctions       data_reference_topic
#>  67: 6724ef06570739aa      otherFunctions               run_examples
#>  68: 6724ef06570739aa      otherFunctions         highlight_examples
#>  69: 6724ef06570739aa      otherFunctions                  timing_fn
#>  70: 6724ef06570739aa      otherFunctions                     handle
#>  71: 6724ef06570739aa      otherFunctions                        try
#>  72: 6724ef06570739aa           preDigest         n:7f12988bd88a0fb8
#>  73: 6724ef06570739aa           preDigest      mean:c40c00762a0dac94
#>  74: 6724ef06570739aa           preDigest        sd:853b1797f54b229c
#>  75: 6724ef06570739aa           preDigest      .FUN:4f604aa46882b368
#>  76: ad0ea27476c50b66            function                      rnorm
#>  77: ad0ea27476c50b66               class                    numeric
#>  78: ad0ea27476c50b66         object.size                       1008
#>  79: ad0ea27476c50b66            accessed 2023-11-22 17:00:31.881109
#>  80: ad0ea27476c50b66             inCloud                      FALSE
#>  81: ad0ea27476c50b66            fromDisk                      FALSE
#>  82: ad0ea27476c50b66          resultHash                           
#>  83: ad0ea27476c50b66   elapsedTimeDigest           0.001792192 secs
#>  84: ad0ea27476c50b66 elapsedTimeFirstRun          3.099442e-05 secs
#>  85: ad0ea27476c50b66      otherFunctions                 build_site
#>  86: ad0ea27476c50b66      otherFunctions           build_site_local
#>  87: ad0ea27476c50b66      otherFunctions            build_reference
#>  88: ad0ea27476c50b66      otherFunctions                       map_
#>  89: ad0ea27476c50b66      otherFunctions        with_indexed_errors
#>  90: ad0ea27476c50b66      otherFunctions          call_with_cleanup
#>  91: ad0ea27476c50b66      otherFunctions       data_reference_topic
#>  92: ad0ea27476c50b66      otherFunctions               run_examples
#>  93: ad0ea27476c50b66      otherFunctions         highlight_examples
#>  94: ad0ea27476c50b66      otherFunctions                  timing_fn
#>  95: ad0ea27476c50b66      otherFunctions                     handle
#>  96: ad0ea27476c50b66      otherFunctions                        try
#>  97: ad0ea27476c50b66           preDigest         n:7eef4eae85fd9229
#>  98: ad0ea27476c50b66           preDigest      mean:c40c00762a0dac94
#>  99: ad0ea27476c50b66           preDigest        sd:853b1797f54b229c
#> 100: ad0ea27476c50b66           preDigest      .FUN:4f604aa46882b368
#>               cacheId              tagKey                   tagValue
#>                     createdDate
#>   1: 2023-11-22 17:00:31.811894
#>   2: 2023-11-22 17:00:31.811894
#>   3: 2023-11-22 17:00:31.811894
#>   4: 2023-11-22 17:00:31.811894
#>   5: 2023-11-22 17:00:31.811894
#>   6: 2023-11-22 17:00:31.811894
#>   7: 2023-11-22 17:00:31.811894
#>   8: 2023-11-22 17:00:31.811894
#>   9: 2023-11-22 17:00:31.811894
#>  10: 2023-11-22 17:00:31.811894
#>  11: 2023-11-22 17:00:31.811894
#>  12: 2023-11-22 17:00:31.811894
#>  13: 2023-11-22 17:00:31.811894
#>  14: 2023-11-22 17:00:31.811894
#>  15: 2023-11-22 17:00:31.811894
#>  16: 2023-11-22 17:00:31.811894
#>  17: 2023-11-22 17:00:31.811894
#>  18: 2023-11-22 17:00:31.811894
#>  19: 2023-11-22 17:00:31.811894
#>  20: 2023-11-22 17:00:31.811894
#>  21: 2023-11-22 17:00:31.811894
#>  22: 2023-11-22 17:00:31.811894
#>  23: 2023-11-22 17:00:31.811894
#>  24: 2023-11-22 17:00:31.811894
#>  25: 2023-11-22 17:00:31.811894
#>  26: 2023-11-22 17:00:31.829778
#>  27: 2023-11-22 17:00:31.829778
#>  28: 2023-11-22 17:00:31.829778
#>  29: 2023-11-22 17:00:31.829778
#>  30: 2023-11-22 17:00:31.829778
#>  31: 2023-11-22 17:00:31.829778
#>  32: 2023-11-22 17:00:31.829778
#>  33: 2023-11-22 17:00:31.829778
#>  34: 2023-11-22 17:00:31.829778
#>  35: 2023-11-22 17:00:31.829778
#>  36: 2023-11-22 17:00:31.829778
#>  37: 2023-11-22 17:00:31.829778
#>  38: 2023-11-22 17:00:31.829778
#>  39: 2023-11-22 17:00:31.829778
#>  40: 2023-11-22 17:00:31.829778
#>  41: 2023-11-22 17:00:31.829778
#>  42: 2023-11-22 17:00:31.829778
#>  43: 2023-11-22 17:00:31.829778
#>  44: 2023-11-22 17:00:31.829778
#>  45: 2023-11-22 17:00:31.829778
#>  46: 2023-11-22 17:00:31.829778
#>  47: 2023-11-22 17:00:31.829778
#>  48: 2023-11-22 17:00:31.829778
#>  49: 2023-11-22 17:00:31.829778
#>  50: 2023-11-22 17:00:31.829778
#>  51: 2023-11-22 17:00:31.846189
#>  52: 2023-11-22 17:00:31.846189
#>  53: 2023-11-22 17:00:31.846189
#>  54: 2023-11-22 17:00:31.846189
#>  55: 2023-11-22 17:00:31.846189
#>  56: 2023-11-22 17:00:31.846189
#>  57: 2023-11-22 17:00:31.846189
#>  58: 2023-11-22 17:00:31.846189
#>  59: 2023-11-22 17:00:31.846189
#>  60: 2023-11-22 17:00:31.846189
#>  61: 2023-11-22 17:00:31.846189
#>  62: 2023-11-22 17:00:31.846189
#>  63: 2023-11-22 17:00:31.846189
#>  64: 2023-11-22 17:00:31.846189
#>  65: 2023-11-22 17:00:31.846189
#>  66: 2023-11-22 17:00:31.846189
#>  67: 2023-11-22 17:00:31.846189
#>  68: 2023-11-22 17:00:31.846189
#>  69: 2023-11-22 17:00:31.846189
#>  70: 2023-11-22 17:00:31.846189
#>  71: 2023-11-22 17:00:31.846189
#>  72: 2023-11-22 17:00:31.846189
#>  73: 2023-11-22 17:00:31.846189
#>  74: 2023-11-22 17:00:31.846189
#>  75: 2023-11-22 17:00:31.846189
#>  76: 2023-11-22 17:00:31.881618
#>  77: 2023-11-22 17:00:31.881618
#>  78: 2023-11-22 17:00:31.881618
#>  79: 2023-11-22 17:00:31.881618
#>  80: 2023-11-22 17:00:31.881618
#>  81: 2023-11-22 17:00:31.881618
#>  82: 2023-11-22 17:00:31.881618
#>  83: 2023-11-22 17:00:31.881618
#>  84: 2023-11-22 17:00:31.881618
#>  85: 2023-11-22 17:00:31.881618
#>  86: 2023-11-22 17:00:31.881618
#>  87: 2023-11-22 17:00:31.881618
#>  88: 2023-11-22 17:00:31.881618
#>  89: 2023-11-22 17:00:31.881618
#>  90: 2023-11-22 17:00:31.881618
#>  91: 2023-11-22 17:00:31.881618
#>  92: 2023-11-22 17:00:31.881618
#>  93: 2023-11-22 17:00:31.881618
#>  94: 2023-11-22 17:00:31.881618
#>  95: 2023-11-22 17:00:31.881618
#>  96: 2023-11-22 17:00:31.881618
#>  97: 2023-11-22 17:00:31.881618
#>  98: 2023-11-22 17:00:31.881618
#>  99: 2023-11-22 17:00:31.881618
#> 100: 2023-11-22 17:00:31.881618
#>                     createdDate
cc(ask = FALSE, x = tmpDir)
#> No time provided; removing the most recent entry to the Cache
#> Cache size: 
#>   Total (including Rasters): 252 bytes
#>   Selected objects (not including Rasters): 252 bytes
showCache(x = tmpDir) # most recent is gone
#> Cache size: 
#>   Total (including Rasters): 746 bytes
#>   Selected objects (not including Rasters): 746 bytes
#>              cacheId              tagKey                   tagValue
#>  1: 422bae4ed2f770cc            function                      rnorm
#>  2: 422bae4ed2f770cc               class                    numeric
#>  3: 422bae4ed2f770cc         object.size                        984
#>  4: 422bae4ed2f770cc            accessed 2023-11-22 17:00:31.811389
#>  5: 422bae4ed2f770cc             inCloud                      FALSE
#>  6: 422bae4ed2f770cc            fromDisk                      FALSE
#>  7: 422bae4ed2f770cc          resultHash                           
#>  8: 422bae4ed2f770cc   elapsedTimeDigest           0.001922607 secs
#>  9: 422bae4ed2f770cc elapsedTimeFirstRun          2.908707e-05 secs
#> 10: 422bae4ed2f770cc      otherFunctions                 build_site
#> 11: 422bae4ed2f770cc      otherFunctions           build_site_local
#> 12: 422bae4ed2f770cc      otherFunctions            build_reference
#> 13: 422bae4ed2f770cc      otherFunctions                       map_
#> 14: 422bae4ed2f770cc      otherFunctions        with_indexed_errors
#> 15: 422bae4ed2f770cc      otherFunctions          call_with_cleanup
#> 16: 422bae4ed2f770cc      otherFunctions       data_reference_topic
#> 17: 422bae4ed2f770cc      otherFunctions               run_examples
#> 18: 422bae4ed2f770cc      otherFunctions         highlight_examples
#> 19: 422bae4ed2f770cc      otherFunctions                  timing_fn
#> 20: 422bae4ed2f770cc      otherFunctions                     handle
#> 21: 422bae4ed2f770cc      otherFunctions                        try
#> 22: 422bae4ed2f770cc           preDigest         n:853b1797f54b229c
#> 23: 422bae4ed2f770cc           preDigest      mean:c40c00762a0dac94
#> 24: 422bae4ed2f770cc           preDigest        sd:853b1797f54b229c
#> 25: 422bae4ed2f770cc           preDigest      .FUN:4f604aa46882b368
#> 26: 66299c88b3924bf4            function                      rnorm
#> 27: 66299c88b3924bf4               class                    numeric
#> 28: 66299c88b3924bf4         object.size                        992
#> 29: 66299c88b3924bf4            accessed 2023-11-22 17:00:31.829272
#> 30: 66299c88b3924bf4             inCloud                      FALSE
#> 31: 66299c88b3924bf4            fromDisk                      FALSE
#> 32: 66299c88b3924bf4          resultHash                           
#> 33: 66299c88b3924bf4   elapsedTimeDigest           0.001885653 secs
#> 34: 66299c88b3924bf4 elapsedTimeFirstRun          2.932549e-05 secs
#> 35: 66299c88b3924bf4      otherFunctions                 build_site
#> 36: 66299c88b3924bf4      otherFunctions           build_site_local
#> 37: 66299c88b3924bf4      otherFunctions            build_reference
#> 38: 66299c88b3924bf4      otherFunctions                       map_
#> 39: 66299c88b3924bf4      otherFunctions        with_indexed_errors
#> 40: 66299c88b3924bf4      otherFunctions          call_with_cleanup
#> 41: 66299c88b3924bf4      otherFunctions       data_reference_topic
#> 42: 66299c88b3924bf4      otherFunctions               run_examples
#> 43: 66299c88b3924bf4      otherFunctions         highlight_examples
#> 44: 66299c88b3924bf4      otherFunctions                  timing_fn
#> 45: 66299c88b3924bf4      otherFunctions                     handle
#> 46: 66299c88b3924bf4      otherFunctions                        try
#> 47: 66299c88b3924bf4           preDigest         n:82dc709f2b91918a
#> 48: 66299c88b3924bf4           preDigest      mean:c40c00762a0dac94
#> 49: 66299c88b3924bf4           preDigest        sd:853b1797f54b229c
#> 50: 66299c88b3924bf4           preDigest      .FUN:4f604aa46882b368
#> 51: 6724ef06570739aa            function                      rnorm
#> 52: 6724ef06570739aa               class                    numeric
#> 53: 6724ef06570739aa         object.size                       1008
#> 54: 6724ef06570739aa            accessed 2023-11-22 17:00:31.845715
#> 55: 6724ef06570739aa             inCloud                      FALSE
#> 56: 6724ef06570739aa            fromDisk                      FALSE
#> 57: 6724ef06570739aa          resultHash                           
#> 58: 6724ef06570739aa   elapsedTimeDigest           0.001798868 secs
#> 59: 6724ef06570739aa elapsedTimeFirstRun          2.717972e-05 secs
#> 60: 6724ef06570739aa      otherFunctions                 build_site
#> 61: 6724ef06570739aa      otherFunctions           build_site_local
#> 62: 6724ef06570739aa      otherFunctions            build_reference
#> 63: 6724ef06570739aa      otherFunctions                       map_
#> 64: 6724ef06570739aa      otherFunctions        with_indexed_errors
#> 65: 6724ef06570739aa      otherFunctions          call_with_cleanup
#> 66: 6724ef06570739aa      otherFunctions       data_reference_topic
#> 67: 6724ef06570739aa      otherFunctions               run_examples
#> 68: 6724ef06570739aa      otherFunctions         highlight_examples
#> 69: 6724ef06570739aa      otherFunctions                  timing_fn
#> 70: 6724ef06570739aa      otherFunctions                     handle
#> 71: 6724ef06570739aa      otherFunctions                        try
#> 72: 6724ef06570739aa           preDigest         n:7f12988bd88a0fb8
#> 73: 6724ef06570739aa           preDigest      mean:c40c00762a0dac94
#> 74: 6724ef06570739aa           preDigest        sd:853b1797f54b229c
#> 75: 6724ef06570739aa           preDigest      .FUN:4f604aa46882b368
#>              cacheId              tagKey                   tagValue
#>                    createdDate
#>  1: 2023-11-22 17:00:31.811894
#>  2: 2023-11-22 17:00:31.811894
#>  3: 2023-11-22 17:00:31.811894
#>  4: 2023-11-22 17:00:31.811894
#>  5: 2023-11-22 17:00:31.811894
#>  6: 2023-11-22 17:00:31.811894
#>  7: 2023-11-22 17:00:31.811894
#>  8: 2023-11-22 17:00:31.811894
#>  9: 2023-11-22 17:00:31.811894
#> 10: 2023-11-22 17:00:31.811894
#> 11: 2023-11-22 17:00:31.811894
#> 12: 2023-11-22 17:00:31.811894
#> 13: 2023-11-22 17:00:31.811894
#> 14: 2023-11-22 17:00:31.811894
#> 15: 2023-11-22 17:00:31.811894
#> 16: 2023-11-22 17:00:31.811894
#> 17: 2023-11-22 17:00:31.811894
#> 18: 2023-11-22 17:00:31.811894
#> 19: 2023-11-22 17:00:31.811894
#> 20: 2023-11-22 17:00:31.811894
#> 21: 2023-11-22 17:00:31.811894
#> 22: 2023-11-22 17:00:31.811894
#> 23: 2023-11-22 17:00:31.811894
#> 24: 2023-11-22 17:00:31.811894
#> 25: 2023-11-22 17:00:31.811894
#> 26: 2023-11-22 17:00:31.829778
#> 27: 2023-11-22 17:00:31.829778
#> 28: 2023-11-22 17:00:31.829778
#> 29: 2023-11-22 17:00:31.829778
#> 30: 2023-11-22 17:00:31.829778
#> 31: 2023-11-22 17:00:31.829778
#> 32: 2023-11-22 17:00:31.829778
#> 33: 2023-11-22 17:00:31.829778
#> 34: 2023-11-22 17:00:31.829778
#> 35: 2023-11-22 17:00:31.829778
#> 36: 2023-11-22 17:00:31.829778
#> 37: 2023-11-22 17:00:31.829778
#> 38: 2023-11-22 17:00:31.829778
#> 39: 2023-11-22 17:00:31.829778
#> 40: 2023-11-22 17:00:31.829778
#> 41: 2023-11-22 17:00:31.829778
#> 42: 2023-11-22 17:00:31.829778
#> 43: 2023-11-22 17:00:31.829778
#> 44: 2023-11-22 17:00:31.829778
#> 45: 2023-11-22 17:00:31.829778
#> 46: 2023-11-22 17:00:31.829778
#> 47: 2023-11-22 17:00:31.829778
#> 48: 2023-11-22 17:00:31.829778
#> 49: 2023-11-22 17:00:31.829778
#> 50: 2023-11-22 17:00:31.829778
#> 51: 2023-11-22 17:00:31.846189
#> 52: 2023-11-22 17:00:31.846189
#> 53: 2023-11-22 17:00:31.846189
#> 54: 2023-11-22 17:00:31.846189
#> 55: 2023-11-22 17:00:31.846189
#> 56: 2023-11-22 17:00:31.846189
#> 57: 2023-11-22 17:00:31.846189
#> 58: 2023-11-22 17:00:31.846189
#> 59: 2023-11-22 17:00:31.846189
#> 60: 2023-11-22 17:00:31.846189
#> 61: 2023-11-22 17:00:31.846189
#> 62: 2023-11-22 17:00:31.846189
#> 63: 2023-11-22 17:00:31.846189
#> 64: 2023-11-22 17:00:31.846189
#> 65: 2023-11-22 17:00:31.846189
#> 66: 2023-11-22 17:00:31.846189
#> 67: 2023-11-22 17:00:31.846189
#> 68: 2023-11-22 17:00:31.846189
#> 69: 2023-11-22 17:00:31.846189
#> 70: 2023-11-22 17:00:31.846189
#> 71: 2023-11-22 17:00:31.846189
#> 72: 2023-11-22 17:00:31.846189
#> 73: 2023-11-22 17:00:31.846189
#> 74: 2023-11-22 17:00:31.846189
#> 75: 2023-11-22 17:00:31.846189
#>                    createdDate
cc(thisTime, ask = FALSE, x = tmpDir)
#> Cache size: 
#>   Total (including Rasters): 500 bytes
#>   Selected objects (not including Rasters): 500 bytes
showCache(x = tmpDir) # all those after thisTime gone, i.e., only 1 left
#> Cache size: 
#>   Total (including Rasters): 246 bytes
#>   Selected objects (not including Rasters): 246 bytes
#>              cacheId              tagKey                   tagValue
#>  1: 422bae4ed2f770cc            function                      rnorm
#>  2: 422bae4ed2f770cc               class                    numeric
#>  3: 422bae4ed2f770cc         object.size                        984
#>  4: 422bae4ed2f770cc            accessed 2023-11-22 17:00:31.811389
#>  5: 422bae4ed2f770cc             inCloud                      FALSE
#>  6: 422bae4ed2f770cc            fromDisk                      FALSE
#>  7: 422bae4ed2f770cc          resultHash                           
#>  8: 422bae4ed2f770cc   elapsedTimeDigest           0.001922607 secs
#>  9: 422bae4ed2f770cc elapsedTimeFirstRun          2.908707e-05 secs
#> 10: 422bae4ed2f770cc      otherFunctions                 build_site
#> 11: 422bae4ed2f770cc      otherFunctions           build_site_local
#> 12: 422bae4ed2f770cc      otherFunctions            build_reference
#> 13: 422bae4ed2f770cc      otherFunctions                       map_
#> 14: 422bae4ed2f770cc      otherFunctions        with_indexed_errors
#> 15: 422bae4ed2f770cc      otherFunctions          call_with_cleanup
#> 16: 422bae4ed2f770cc      otherFunctions       data_reference_topic
#> 17: 422bae4ed2f770cc      otherFunctions               run_examples
#> 18: 422bae4ed2f770cc      otherFunctions         highlight_examples
#> 19: 422bae4ed2f770cc      otherFunctions                  timing_fn
#> 20: 422bae4ed2f770cc      otherFunctions                     handle
#> 21: 422bae4ed2f770cc      otherFunctions                        try
#> 22: 422bae4ed2f770cc           preDigest         n:853b1797f54b229c
#> 23: 422bae4ed2f770cc           preDigest      mean:c40c00762a0dac94
#> 24: 422bae4ed2f770cc           preDigest        sd:853b1797f54b229c
#> 25: 422bae4ed2f770cc           preDigest      .FUN:4f604aa46882b368
#>              cacheId              tagKey                   tagValue
#>                    createdDate
#>  1: 2023-11-22 17:00:31.811894
#>  2: 2023-11-22 17:00:31.811894
#>  3: 2023-11-22 17:00:31.811894
#>  4: 2023-11-22 17:00:31.811894
#>  5: 2023-11-22 17:00:31.811894
#>  6: 2023-11-22 17:00:31.811894
#>  7: 2023-11-22 17:00:31.811894
#>  8: 2023-11-22 17:00:31.811894
#>  9: 2023-11-22 17:00:31.811894
#> 10: 2023-11-22 17:00:31.811894
#> 11: 2023-11-22 17:00:31.811894
#> 12: 2023-11-22 17:00:31.811894
#> 13: 2023-11-22 17:00:31.811894
#> 14: 2023-11-22 17:00:31.811894
#> 15: 2023-11-22 17:00:31.811894
#> 16: 2023-11-22 17:00:31.811894
#> 17: 2023-11-22 17:00:31.811894
#> 18: 2023-11-22 17:00:31.811894
#> 19: 2023-11-22 17:00:31.811894
#> 20: 2023-11-22 17:00:31.811894
#> 21: 2023-11-22 17:00:31.811894
#> 22: 2023-11-22 17:00:31.811894
#> 23: 2023-11-22 17:00:31.811894
#> 24: 2023-11-22 17:00:31.811894
#> 25: 2023-11-22 17:00:31.811894
#>                    createdDate
cc(ask = FALSE, x = tmpDir) # Cache is
#> No time provided; removing the most recent entry to the Cache
#> Cache size: 
#>   Total (including Rasters): 246 bytes
#>   Selected objects (not including Rasters): 246 bytes
cc(ask = FALSE, x = tmpDir) # Cache is already empty
#> No time provided; removing the most recent entry to the Cache
#> Cache already empty